Data tables with species parameter values for Mediterranean and US

A data sets of species parameter values, resulting from bibliographic search, fit to empirical data or expert-based guesses.

data("SpParamsMED")
data("SpParamsUS")

Format

A data frame with species as rows and the following columns (variables), grouped by categories.

Plant identity: Name: A factor with taxon names (mostly species names). IFNcodes: A factor with levels corresponding to species codes in the Third Spanish forest inventory (DGCN 2005). SpIndex: A numeric vector of the species index. Group: Either "Gymnosperm" or "Angiosperm". Order: Taxonomical order. Family: Taxonomical family. GrowthForm: Either Shrub, Tree and Tree/Shrub. Plant size: Hmax: Maximum plant height (in cm). Hmed: Median plant height (in cm). Z50: Rooting depth (in mm) corresponding to 50% of fine roots. Z95: Rooting depth (in mm) corresponding to 95% of fine roots. Shrub allometric coefficients (calibrated for Catalonia; see De CáceresCaceres et al. 2019): a_ash, b_ash: Allometric coefficients relating shrub height with the projected area of shrub individuals. a_bsh, b_bsh: Allometric coefficients relating phytovolume with dry weight of shrub individuals. cr: Ratio between shrub crown length and total height. Tree allometric coefficients (calibrated for Catalonia): a_fbt, b_fbt, c_fbt, d_fbt: Regression coefficients used to calculate foliar biomass of an individual tree from its dbh and the cummulative basal area of larger trees. a_cr, b_1cr, b_2cr, b_3cr, c_1cr, c_2cr: Regression coefficients used to calculate crown ratio of trees. a_cw, b_cw: Regression coefficients used to calculated crown width of trees. Phenology: PhenologyType: Leaf phenology type, either "oneflush-evergreen" (new leaves develop in spring-summer), "progressive-evergreen" (new leaves develop during any season), "winter-deciduous" (leaf senescence in autumn, new leaves in spring-summer) or "winter-semideciduous" (same as before, but abscission of senescent leaves occurs when new leaves are produced). LeafDuration: Leaf duration (in years). Sgdd: Degree days for leaf budburst (in Celsius). Tbgdd: Base temperature for the calculation of degree days to leaf budburst. Ssen: Degree days corresponding to leaf senescence. Phsen: Photoperiod corresponding to start counting senescence degree-days. Tbsen: Base temperature for the calculation of degree days to leaf senescence. Anatomy: SLA: Specific leaf area (mm2/mg = m2/kg). LeafDensity: Density of leaf tissue (dry weight over volume). WoodDensity: Wood density (at 0 percent humidity!). FineRootDensity: Density of fine root tissue (dry weight over volume). conduit2sapwood: Proportion of sapwood corresponding to conducive elements (vessels or tracheids) as opposed to parenchymatic tissue. r635: Ratio between the weight of leaves plus branches and the weight of leaves alone for branches of 6.35 mm. pDead: Proportion of total fine fuels that are dead Al2As: Leaf area to sapwood area ratio (in m2·m-2). LeafWidth: Leaf width (in cm). SRL: Specific root length (cm/g). RLD: Fine root length density (i.e. density of root length per soil volume) (cm·cm-3). Tissue moisture: maxFMC: Maximum fuel moisture (in percent of dry weight) minFMC: Minimum fuel moisture (in percent of dry weight) LeafPI0: Osmotic potential at full turgor of leaves (MPa). LeafEPS: Modulus of elasticity (capacity of the cell wall to resist changes in volume in response to changes in turgor) of leaves (MPa). LeafAF: Apoplastic fraction (proportion of water outside the living cells) in leaves. StemPI0: Osmotic potential at full turgor of symplastic xylem tissue (MPa). StemEPS: Modulus of elasticity (capacity of the cell wall to resist changes in volume in response to changes in turgor) of symplastic xylem tissue (Mpa). StemAF: Apoplastic fraction (proportion of water outside the living cells) in stem xylem. Flammability: LigninPercent: Percent of lignin+cutin over dry weight in leaves. ParticleDensity: Particle density (kg/m3). LeafLitterFuelType: Fuel type for leaf litter, with levels Broadleaved, LongLinear, Scale and ShortLinear. Flammability: Flammability modifier (either 1 or 2 for normal or high, respectively). SAV: Surface-area-to-volume ratio of the small fuel (1h) fraction (leaves and branches < 6.35mm) (m2/m3). HeatContent: High fuel heat content (kJ/kg). Extinction: gammaSWR: Reflectance (albedo) for SWR (gammaPAR = 0.8*gammaSWR). alphaSWR: Absorbance coefficient for SWR (alphaPAR = alphaSWR * 1.35). kPAR: Extinction coeficient for PAR (kSWR = kPAR/1.35). g: Canopy water storage capacity per LAI unit (in mm/LAI). Transpiration/Photosynthesis: Tmax_LAI: Coefficient relating LAI with the ratio of maximum transpiration over potential evapotranspiration. Tmax_LAIsq: Coefficient relating squared LAI with the ratio of maximum transpiration over potential evapotranspiration. Psi_Extract: Water potential corresponding to 50% stomatal conductance (in MPa). Psi_Critic: Water potential corresponding to 50% of stem cavitation (in MPa). WUE: Water use efficiency for carbon assimilation (gross photosynthesis) (g C /mm water). pRootDisc: Relative root conductance leading to hydraulic disconnection from a soil layer. Gswmin: Minimum stomatal conductance to water vapor per leaf area unit (in mol·s-1·m-2). Gswmax: Maximum stomatal conductance to water vapor per leaf area unit (in mol·s-1·m-2). VCleaf_kmax: Leaf hydraulic conductance (in mmol H2O·s-1·m-2·MPa-1). VCleaf_c, VCleaf_d: Parameters of the leaf vulnerability curve (VCleaf_d in MPa). Kmax_stemxylem: Sapwood-specific hydraulic conductivity of stem xylem (in kg H2O·s-1·m-1·MPa-1). VCstem_c, VCstem_d: Parameters of the stem xylem vulnerability curve (VCstem_d in MPa). Kmax_rootxylem: Sapwood-specific hydraulic conductivity of root xylem (in kg H2O·s-1·m-1·MPa-1). VCroot_c, VCroot_d: Parameters of the root xylem vulnerability curve (VCroot_d in MPa). Narea: Nitrogen contrentration per leaf area (in g·m-2). Vmax298: Maximum Rubisco carboxilation rate at 25ºC (in micromol CO2·s-1·m-2). Jmax298: Maximum electron transport rate at 25ºC (in micromol electrons·s-1·m-2). Growth: WoodC: Wood carbon content per dry weight (g C /g dry). RERleaf: Maintenance respiration rate for leaves (g gluc · g dry-1 · day-1). RERsapwood: Maintenance respiration rate for living cells of sapwood (g gluc · g dry-1 · day-1). RERfineroot: Maintenance respiration rate for fine roots (g gluc · g dry-1 · day-1). RGRleafmax: Maximum leaf area growth rate relative to sapwood area (in m2·cm-2· day-1). RGRsapwoodmax: Maximum sapwood area relative growth rate (in cm2·cm-2·day-1). RGRfinerootmax: Maximum fine root relative growth rate (in g dry·g dry-1· day-1). fHDmin: Minimum value of the height-to-diameter ratio (dimensionless). fHDmax: Maximum value of the height-to-diameter ratio (dimensionless). Recruitment: SeedProductionHeight: Minimum height (cm) for seed production. MinTempRecr: Minimum average temperature (Celsius) of the coldest month for successful recruitment. MinMoistureRecr: Minimum value of the moisture index (annual precipitation over annual PET) for successful recruitment. MinFPARRecr: Minimum percentage of PAR at the ground level for successful recruitment. RecrTreeDBH: Recruitment DBH (cm) for trees. RecrTreeHeight: Recruitment height (cm) for trees. RecrTreeDensity: Recruitment density (ind·ha-1) for trees. RecrShrubCover: Recruitment cover (%) for shrubs. RecrShrubHeight: Recruitment height (cm) for shrubs. RecrZ50: Soil depth (mm) corresponding to 50% of fine roots for recruitment. RecrZ95: Soil depth (mm) corresponding to 95% of fine roots for recruitment.

Details

See details of parameterization in De Caceres et al. (2015) and De Caceres et al. (2021).

Source

De Cáceres M, Martínez-Vilalta J, Coll L, Llorens P, Casals P, Poyatos R, Pausas JG, Brotons L. (2015) Coupling a water balance model with forest inventory data to predict drought stress: the role of forest structural changes vs. climate changes. Agricultural and Forest Meteorology (doi:10.1016/j.agrformet.2015.06.012).

De Cáceres M, Mencuccini M, Martin-StPaul N, Limousin JM, Coll L, Poyatos R, Cabon A, Granda V, Forner A, Valladares F, Martínez-Vilalta J (2021) Unravelling the effect of species mixing on water use and drought stress in holm oak forests: a modelling approach. Agricultural and Forest Meteorology 296 (doi:10.1016/j.agrformet.2020.108233).

De Cáceres M, Casals P, Gabriel E, Castro X (2019) Scaling-up individual-level allometric equations to predict stand-level fuel loading in Mediterranean shrublands. Ann For Sci 76:87 (doi:10.1007/s13595-019-0873-4).

DGCN (2005). Tercer Inventario Forestal Nacional (1997-2007): Catalunya. Dirección General de Conservación de la Naturaleza, Ministerio de Medio Ambiente, Madrid.

See also

Examples

data(SpParamsMED)