Parameter values for Mediterranean species

A data set of parameter values for Mediterranean species, resulting from bibliographic search, fit to empirical data or expert-based guesses.

data("SpParamsMED")

Format

A data frame with 89 observations (species) on the following 77 variables.

Name

A factor with taxon names (mostly species names).

IFNcodes

A factor with levels corresponding to species codes in the Third Spanish forest inventory (DGCN 2005.)

SpIndex

A numeric vector of the species index.

Group

Either "Gymnosperm" or "Angiosperm".

Order

Taxonomical order.

Family

Taxonomical family.

GrowthForm

A factor with levels Shrub, Tree and Tree/Shrub.

TreeType

A factor with levels Conifer, Deciduous, Evergreen or Shrub

Hmed

Median plant height (in cm).

Hmax

Maximum plant height (in cm).

Z50

Rooting depth (in mm) corresponding to 50% of fine roots.

Z95

Rooting depth (in mm) corresponding to 95% of fine roots.

Zmax

Maximum rooting depth (in mm).

a_ash

Regression coefficient relating the square of shrub height with shrub area.

a_bsh, b_bsh

Allometric coefficients relating phytovolume with dry weight of shrub individuals.

cr

Ratio between crown length and total height (for shrubs).

a_fbt, b_fbt, c_fbt, d_fbt

Regression coefficients used to calculate foliar biomass of an individual tree from its dbh and the cummulative basal area of larger trees.

a_cr, b_1cr, b_2cr, b_3cr, c_1cr, c_2cr

Regression coefficients used to calculate crown ratio of trees.

a_cw, b_cw

Regression coefficients used to calculated crown width of trees.

SLA

Specific leaf area (mm2/mg = m2/kg).

LeafDensity

Density of leaf tissue (dry weight over volume).

r635

Ratio between the weight of leaves plus branches and the weight of leaves alone for branches of 6.35 mm.

pDead

Proportion of total fine fuels that are dead

maxFMC

Maximum fuel moisture (in percent of dry weight)

minFMC

Minimum fuel moisture (in percent of dry weight)

LeafPI0

Osmotic potential at full turgor of leaves (MPa).

LeafEPS

Modulus of elasticity (capacity of the cell wall to resist changes in volume in response to changes in turgor) of leaves (MPa).

LeafAF

Apoplastic fraction (proportion of water outside the living cells) in leaves.

StemPI0

Osmotic potential at full turgor of symplastic xylem tissue (MPa).

StemEPS

Modulus of elasticity (capacity of the cell wall to resist changes in volume in response to changes in turgor) of symplastic xylem tissue (Mpa).

StemAF

Apoplastic fraction (proportion of water outside the living cells) in stem xylem.

Cstoragepmax

Maximum storage capacity, expressed as the fraction of C per total respiratory of C.

LeafDuration

Leaf duration (in years).

LigninPercent

Percent of lignin+cutin over dry weight in leaves.

ParticleDensity

Particle density (kg/m3).

LeafLitterFuelType

Fuel type for leaf litter, with levels Broadleaved, LongLinear, Scale and ShortLinear.

Flammability

Flammability modifier (either 1 or 2 for normal or high, respectively).

SAV

Surface-area-to-volume ratio of the small fuel (1h) fraction (leaves and branches < 6.35mm) (m2/m3).

HeatContent

High fuel heat content (kJ/kg).

fHDmin

Minimum value of the height-to-diameter ratio (dimensionless).

fHDmax

Maximum value of the height-to-diameter ratio (dimensionless).

albedo

Leaf reflectance for short wave radiation.

k

Light extinction coeficient for PAR.

g

Canopy water storage capacity per LAI unit (in mm/LAI).

Sgdd

Growth degree days for leaf budburst (in Celsius)

Al2As

Leaf area to sapwood area ratio (in m2·m-2).

WUE

Water use efficiency for carbon assimilation (g C /mm water).

WoodC

Wood carbon content per dry weight (g C /g dry).

WoodDensity

Wood density (at 0 percent humidity!).

RGRmax

Maximum relative growth rate (in basal area or sapwood area) (in cm2·cm-2).

Psi_Extract

Water potential corresponding to 50% relative conductance (in MPa).

pRootDisc

Relative root conductance leading to hydraulic disconnection from a soil layer.

Gwmin

Minimum stomatal conductance to water vapor per leaf area unit (in mol·s-1·m-2).

Gwmax

Maximum stomatal conductance to water vapor per leaf area unit (in mol·s-1·m-2).

VCleaf_kmax

Leaf hydraulic conductance (in mmol H2O·s-1·m-2·MPa-1).

VCleaf_c, VCleaf_d

Parameters of the leaf vulnerability curve (VCleaf_d in MPa).

Kmax_stemxylem

Sapwood-specific hydraulic conductivity of stem xylem (in kg H2O·s-1·m-1·MPa-1).

VCstem_c, VCstem_d

Parameters of the stem xylem vulnerability curve (VCstem_d in MPa).

Kmax_rootxylem

Sapwood-specific hydraulic conductivity of root xylem (in kg H2O·s-1·m-1·MPa-1).

VCroot_c, VCroot_d

Parameters of the root xylem vulnerability curve (VCroot_d in MPa).

LeafWidth

Leaf width (in cm).

Narea

Nitrogen contrentration per leaf area (in g·m-2).

Vmax298

Maximum Rubisco carboxilation rate at 25ºC (in micromol CO2·s-1·m-2).

Jmax298

Maximum electron transport rate at 25ºC (in micromol electrons·s-1·m-2).

Details

See details of parameterization in De Caceres et al. (2015) and De Caceres et al. (submitted).

Source

De Cáceres M, Martínez-Vilalta J, Coll L, Llorens P, Casals P, Poyatos R, Pausas JG, Brotons L. (2015) Coupling a water balance model with forest inventory data to predict drought stress: the role of forest structural changes vs. climate changes. Agricultural and Forest Meteorology (doi:10.1016/j.agrformet.2015.06.012).

DGCN (2005). Tercer Inventario Forestal Nacional (1997-2007): Catalunya. Dirección General de Conservación de la Naturaleza, Ministerio de Medio Ambiente, Madrid.

See also

Examples

data(SpParamsMED)