Default control parameters for models

Creates a list with global default parameters for simulation models.

defaultControl()

Details

The function returns a list with default parameters. Users can change those defaults that need to be set to other values and use the list as input for model functions. The relevant parameters are different for each model function.

Value

A list, with the following options:

verbose (=TRUE): Boolean flag to indicate console output during calculations.

subdailyResults (=FALSE): Boolean flag to force subdaily results to be stored (as a list called 'subdaily' of spwb_day objects, one by simulated date) in calls to spwb.

soilFunctions (="SX"): Soil water retention curve and conductivity functions, either 'SX' (for Saxton) or 'VG' (for Van Genuchten).

snowpack (=TRUE): Boolean flag to indicate the simulation of snow accumulation and melting.

drainage (=TRUE): Boolean flag to indicate the simulation of deep drainage.

leafPhenology (=TRUE): Boolean flag to indicate the simulation of leaf phenology for winter-deciduous species.

transpirationMode (="Granier"): Transpiration model (either 'Granier' or 'Sperry'). See spwbInput.

hydraulicCostFunction (= 1): Variant of the hydraulic cost function used in the stomatal regulation model of Sperry & Love (2016). Values accepted are 1 (original cost function based on the derivative of supply function), 2 (leaf vulnerability curve).

ndailysteps (= 24): Number of steps into which each day is divided for determination of stomatal conductance, transpiration and photosynthesis (24 equals 1-hour intervals).

canopyMode (="sunshade"): Indicates how crowns should be described to calculate photosynthesis. Accepted values are "sunshade" (distinguishes photosynthesis in sun leaves from shade leaves) and "multilayer" (distinguishes photosynthesis of sun leaves and shade leaves in each canopy layer).

verticalLayerSize (= 100): The size of vertical layers (in cm) for photosynthesis calculation.

nStemSegments (= 1): Number of segments within the stem.

capacitance (=FALSE): Whether capacitance is considered in simulations.

cavitationRefill (= TRUE): Whether refilling of embolized conduits is activated.

klat (= 0.1): Symplastic-apoplastic lateral conductance.

taper (= TRUE): Whether taper of xylem conduits is accounted for when calculating aboveground stem conductance from xylem conductivity.

numericParams: A list with the following elements:

maxNsteps (= 400): Maximum number of steps in supply function.

ntrial (= 200): Number of iteration trials when finding root of equation system.

psiTol (= 0.0001): Tolerance value for water potential.

ETol (= 0.0001): Tolerance value for flow.

thermalCapacityLAI (=1000000): Thermal canopy capacitance per LAI unit.

defaultWindSpeed (= 2.5): Default wind speed value (in m/s) to be used when missing from data.

Catm (=386): Atmospheric CO2 concentration (in micromol·mol-1 = ppm).

fracRootResistance (=0.40): Fraction of plant total resistance (leaf+stem+root) that corresponds to root system. If fracRootResistance = NA then root conductance is calculated from root xylem conductivity and root distribution.

averageFracRhizosphereResistance (=0.15): Fraction to total continuum (leaf+stem+root+rhizosphere) resistance that corresponds to rhizosphere (averaged across soil water potential values).

storagePool (="none"): Whether carbon storage pools are considered (either "none", "one" or "two").

See also