Extracts subdaily output

Given the result of simulations, this function extracts subdaily output corresponding to each simulated day and returns it as a data frame.

extractSubdaily(x, output = "E", dates = NULL)

Arguments

x

An object returned by simulation functions spwb, pwb or growth.

output

See options in section details.

dates

A date vector indicating the subset of simulated days for which subdaily output is desired.

Value

A data frame with a column 'datetime' and as many columns as plant cohorts.

Details

This function only works when simulations have been carried using control option 'subdailyResults = TRUE' (see defaultControl). Subdaily simulation results will then be stored as elements of the a list called 'subdaily' in the simulation output. Function extractSubdaily will assemble subdaily results from this list and return them as a data frame. Options for parameter 'output' are the following:

Functions pwb() and spwb(): "E","Ag","An","dEdP","RootPsi","StemPsi","LeafPsi","StemPLC","StemRWC","LeafRWC","StemSympRWC","LeafSympRWC","PWB", "Temperature", "ExtractionInst".

Additional options for shade and sunlit leaves in pwb() and spbw(): Either "SunlitLeaves$x" or "ShadeLeaves$x" where 'x' is one of the following: "Abs_SWR","Net_LWR","E","Ag","An","Ci","Gsw","VPD","Temp","Psi","iWUE".

Additional options for function growth(): "GrossPhotosynthesis", "MaintenanceRespiration", "GrowthCosts", "LabileCarbonBalance","SugarLeaf", "SugarSapwood", "StarchLeaf", "StarchSapwood","SugarTransport".

See also

Author

Miquel De Cáceres Ainsa, CREAF