Fire hazard

Estimates potential fire behaviour at each daily step of a simulation

fireHazard(
  x,
  SpParams,
  forest = NULL,
  standardConditions = FALSE,
  freq = "days",
  fun = "max"
)

Arguments

x

An object of class spwb, spwb_day, pwb, growth, growth_day or fordyn.

SpParams

A data frame with species parameters (see SpParamsDefinition and SpParamsMED).

forest

An object of class forest (needed if x is not of class fordyn).

standardConditions

A logical flag to indicate that standard fire weather conditions are to be used (instead of deriving fuel moisture and windspeed from x).

freq

Frequency of summary statistics (see cut.Date).

fun

Summary function (by default, maximum values).

Value

A matrix with fire behaviour variables (columns) for each simulated day (rows) or coarser time steps if summaries are requested.

Details

Live fuel moisture of shrub and canopy layers is estimated from plant water status. Dead fuel moisture is estimated following Resco-de-Dios et al. (2015).

References

Resco de Dios, V., A. W. Fellows, R. H. Nolan, M. M. Boer, R. A. Bradstock, F. Domingo, and M. L. Goulden. 2015. A semi-mechanistic model for predicting the moisture content of fine litter. Agricultural and Forest Meteorology 203:64–73.

Ruffault J, Limousin JM, Pimont F, Dupuy JL, De Cáceres M, Cochard H, Mouillot F, Blackman C, Torres-Ruiz JM, Parsons R, Moreno M, Delzon S, Jansen S, Olioso A, Choat B, Martin-StPaul N. 2023. Plant hydraulic modelling of leaf and canopy fuel moisture content reveals increasing vulnerability of a Mediterranean forest to wildfires under extreme drought. New Phytologist. (10.1111/nph.18614).

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

# \donttest{
#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (4 layers)
examplesoil <- soil(defaultSoilParams(4))

#Initialize control parameters
control <- defaultControl("Granier")

#Initialize input
x1 <- forest2spwbInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Call simulation function
S1 <- spwb(x1, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 7.05859
#> Initial soil water content (mm): 290.875
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:....................................
#> 
#> Final plant water content (mm): 7.05708
#> Final soil water content (mm): 322.866
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): -0.00151384
#> Plant water balance result (mm): -0.00151384
#> Change in soil water content (mm): 31.9909
#> Soil water balance result (mm): 31.9406
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): -7.10543e-15
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 92 Net rainfall (mm) 370
#>   Infiltration (mm) 325 Runoff (mm) 45 Deep drainage (mm) 127
#>   Soil evaporation (mm) 16  Herbaceous transpiration (mm) 14 Woody plant transpiration (mm) 187
#>   Plant extraction from soil (mm) 187  Plant water balance (mm) -0 Hydraulic redistribution (mm) 2

#Evaluate fire hazard
F1 <- fireHazard(S1, SpParamsMED, exampleforestMED)
# }