Forest dynamics

Function fordyn implements a forest dynamics model that simulates growth, mortality and recruitment for plant cohorts in a given forest stand during a period specified in the input climatic data.

fordyn(forest, soil, SpParams, meteo, control,
       latitude , elevation = NA, slope = NA, aspect = NA)

Arguments

forest
An object of class forest.

soil
An object of class soil.

SpParams
A data frame with species parameters (see SpParamsMED and SpParamsMED).

meteo
A data frame with daily meteorological data series. Row names of the data frame should correspond to date strings with format "yyyy-mm-dd" (see Date).

control
A list with default control parameters (see defaultControl).

latitude
Latitude (in degrees). Required when x$TranspirationMode = "Sperry".

elevation, slope, aspect
Elevation above sea level (in m), slope (in degrees) and aspect (in degrees from North). Required when x$TranspirationMode = "Sperry". Elevation is also required for 'Granier' if snowpack dynamics are simulated.

Details

Function fordyn simulates forest dynamics for annual time steps, building on other simulation functions. For each simulated year, the function performs the following steps:

Calls function growth to simulate daily water/carbon balance, growth and mortality processes.

Simulate recruitment (for species present in the stand).

Prepares the input of function growth for the next annual time step.

Store forest status and summaries.

Value

A list of class 'fordyn' with the following elements:

"StandSummary": A data frame with stand-level summaries (leaf area index, tree basal area, tree density, shrub cover, etc.) at the beginning of the simulation and after each simulated year.

"SpeciesSummary": A data frame with species-level summaries (leaf area index, tree basal area, tree density, shrub cover, etc.) at the beginning of the simulation and after each simulated year.

"CohortSummary": A data frame with cohort-level summaries (leaf area index, tree basal area, tree density, shrub cover, etc.) at the beginning of the simulation and after each simulated year.

"TreeTable": A data frame with tree-cohort data (species, density, diameter, height, etc.) at the beginning of the simulation (if any) and after each simulated year.

"DeadTreeTable": A data frame with dead tree-cohort data (species, density, diameter, height, etc.) at the beginning of the simulation and after each simulated year.

"ShrubTable": A data frame with shrub-cohort data (species, density, cover, height, etc.) at the beginning of the simulation and after each simulated year.

"DeadShrubTable": A data frame with dead shrub-cohort data (species, density, cover, height, etc.) at the beginning of the simulation (if any) and after each simulated year.

"ForestStructures": A list with the forest object of the stand at the beginning of the simulation and after each simulated year.

"GrowthResults": A list with the results of calling function growth for each simulated year.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

#Load example daily meteorological data
data(examplemeteo)
#Prepare a three-year meteorological data with half precipitation during 
#the second year
meteo2001 = examplemeteo
meteo2002 = examplemeteo
meteo2002$Precipitation = meteo2002$Precipitation/2
meteo2003 = examplemeteo
row.names(meteo2002) = seq(as.Date("2002-01-01"), 
                           as.Date("2002-12-31"), by="day")
row.names(meteo2003) = seq(as.Date("2003-01-01"), 
                           as.Date("2003-12-31"), by="day")
meteo_01_03 = rbind(meteo2001, meteo2002, meteo2003)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize control parameters
control = defaultControl()

#Initialize soil with default soil params (4 layers)
examplesoil = soil(defaultSoilParams(4))

#Call simulation function
fd<-fordyn(exampleforestMED, examplesoil, 
           SpParamsMED, meteo_01_03, control,
           latitude = 41.82592, elevation = 100)

#> Simulating forest dynamics for year 2001 (1/3)
#>    (a) Growth/mortality
#>    (b) Recruitment
#>        Coldest month mean temp. (Celsius): 0.05   Moisture index: 0.36   FPAR (%): 39.8
#>        Tree species with seed rain: 54,68 recruited: 54,68
#>    (c) Summaries
#> Simulating forest dynamics for year 2002 (2/3)
#>    (a) Growth/mortality
#>    (b) Recruitment
#>        Coldest month mean temp. (Celsius): 0.05   Moisture index: 0.18   FPAR (%): 38.7
#>        Tree species with seed rain: 54,68 recruited: <none>
#>    (c) Summaries
#> Simulating forest dynamics for year 2003 (3/3)
#>    (a) Growth/mortality
#>    (b) Recruitment
#>        Coldest month mean temp. (Celsius): 0.05   Moisture index: 0.36   FPAR (%): 36.8
#>        Tree species with seed rain: 54,68 recruited: 54,68
#>    (c) Summaries

#Stand-level summaries
fd$StandSummary

#>   Step LeafAreaIndex TreeDensityLive TreeBasalAreaLive ShrubCoverLive MaxHeight
#> 1    0      1.662671        552.0000          25.03330           3.75  800.0000
#> 2    1      1.926889        746.3412          26.87228           3.75  824.6774
#> 3    2      2.094240        738.6900          27.82976           3.75  838.5123
#> 4    3      2.331703        931.1174          29.55036           3.75  856.7342

#Tree table by annual steps
fd$TreeTable

#>    Step Year Cohort Species             Name         N       DBH   Height Z50
#> 1     0   NA  T1_54      54 Pinus halepensis 168.00000 37.550000 800.0000 750
#> 2     0   NA  T2_68      68     Quercus ilex 384.00000 14.600000 660.0000 750
#> 3     1 2001  T1_54      54 Pinus halepensis 166.27775 38.768543 824.6774 750
#> 4     1 2001  T2_68      68     Quercus ilex 380.06342 15.561230 685.5711 750
#> 5     1 2001  T3_54      54 Pinus halepensis 100.00000  1.000000 100.0000 100
#> 6     1 2001  T4_68      68     Quercus ilex 100.00000  1.000000 100.0000 100
#> 7     2 2002  T1_54      54 Pinus halepensis 164.57315 39.461184 838.5123 750
#> 8     2 2002  T2_68      68     Quercus ilex 376.16720 16.118512 700.2333 750
#> 9     2 2002  T3_54      54 Pinus halepensis  98.97485  1.338512 121.1477 100
#> 10    2 2002  T4_68      68     Quercus ilex  98.97485  1.273922 116.9187 100
#> 11    3 2003  T1_54      54 Pinus halepensis 162.88603 40.378721 856.7342 750
#> 12    3 2003  T2_68      68     Quercus ilex 372.31092 17.174768 728.2301 750
#> 13    3 2003  T3_54      54 Pinus halepensis  97.96021  1.908921 157.9599 100
#> 14    3 2003  T4_68      68     Quercus ilex  97.96021  1.727529 145.7910 100
#> 15    3 2003  T5_54      54 Pinus halepensis 100.00000  1.000000 100.0000 100
#> 16    3 2003  T6_68      68     Quercus ilex 100.00000  1.000000 100.0000 100
#>     Z95
#> 1  3000
#> 2  3000
#> 3  3000
#> 4  3000
#> 5  1000
#> 6  1000
#> 7  3000
#> 8  3000
#> 9  1000
#> 10 1000
#> 11 3000
#> 12 3000
#> 13 1000
#> 14 1000
#> 15 1000
#> 16 1000

#Dead tree table by annual steps
fd$DeadTreeTable

#>    Step Year Cohort Species             Name        N       DBH   Height Z50
#> 1     1 2001  T1_54      54 Pinus halepensis 1.722253 38.768543 824.6774 750
#> 2     1 2001  T2_68      68     Quercus ilex 3.936579 15.561230 685.5711 750
#> 3     2 2002  T1_54      54 Pinus halepensis 1.704598 39.461184 838.5123 750
#> 4     2 2002  T2_68      68     Quercus ilex 3.896223 16.118512 700.2333 750
#> 5     2 2002  T3_54      54 Pinus halepensis 1.025151  1.338512 121.1477 100
#> 6     2 2002  T4_68      68     Quercus ilex 1.025151  1.273922 116.9187 100
#> 7     3 2003  T1_54      54 Pinus halepensis 1.687123 40.378721 856.7342 750
#> 8     3 2003  T2_68      68     Quercus ilex 3.856281 17.174768 728.2301 750
#> 9     3 2003  T3_54      54 Pinus halepensis 1.014642  1.908921 157.9599 100
#> 10    3 2003  T4_68      68     Quercus ilex 1.014642  1.727529 145.7910 100
#>     Z95
#> 1  3000
#> 2  3000
#> 3  3000
#> 4  3000
#> 5  1000
#> 6  1000
#> 7  3000
#> 8  3000
#> 9  1000
#> 10 1000