Input for simulation models

Functions forest2spwbInput and forest2growthInput take an object of class forest and calculate input data for functions spwb, pwb and growth, respectively. Functions spwbInput and growthInput does the same from input data. Function forest2aboveground calculates aboveground variables that may be used in spwbInput and growthInput functions. Similarly, function forest2belowground calculates belowground root distribution that may be used in spwbInput and growthInput functions.

forest2aboveground(x, SpParams, gdd = NA)
forest2belowground(x, soil, SpParams)
forest2growthInput(x, soil, SpParams, control)
forest2spwbInput(x, soil, SpParams, control)
growthInput(above,  Z, V,  soil, SpParams, control)
spwbInput(above,  V,  soil, SpParams, control)

Arguments

x
An object of class forest.

SpParams
A data frame with species parameters (see SpParamsMED and SpParamsMED).

gdd
Growth degree days to account for leaf phenology effects (in Celsius). This should be left NA in most applications.

soil
An object of class soil.

control
A list with default control parameters (see defaultControl).

above
A data frame with aboveground plant information (see the return value of forest2aboveground below). In the case of spwbInput the variables should include SP, LAI_live, LAI_dead, H and CR.

Z
A numeric vector with cohort rooting depths (in cms).

V
A numeric matrix (with as many columns as soil layers and as many rows as the length as SP) containing the proportion of roots of each plant in each soil layer.

Details

Functions forest2spwbInput and forest2abovegroundInput extracts height and species identity from plant cohorts of x, and calculate leaf area index and crown ratio.forest2spwbInput also calculates the distribution of fine roots across soil. Both forest2spwbInput and spwbInput find parameter values for each plant cohort according to the parameters of its species as specified in SpParams. If control$transpirationMode = "Sperry" the functions also estimate the maximum conductance of rhizosphere, root xylem and stem xylem elements.

Value

Function forest2aboveground returns a data frame with the following columns (rows are identified as specified by function plant_ID):

SP: Species identity (an integer) (first species is 0).

N: Cohort density (ind/ha) (see function plant_density).

DBH: Tree diameter at breast height (cm).

H: Plant total height (cm).

CR: Crown ratio (crown length to total height) (between 0 and 1).

LAI_live: Live leaf area index (m2/m2) (one-side leaf area relative to plot area).

LAI_dead: Dead leaf area index (m2/m2) (one-side leaf area relative to plot area).

Functions forest2spwbInput and spwbInput return a list of class spwbInput with the following elements (rows of data frames are identified as specified by function plant_ID):

control: List with control parameters (see defaultControl).

cohorts: A data frame with cohort information, with columns SP and Name.

above: A data frame with columns H, CR and LAI (see function forest2aboveground).

below: A list. If control$transpirationMode = "Granier" it contains a single element:

V: A matrix with the proportion of fine roots of each cohort (in rows) in each soil layer (in columns).

If control$transpirationMode = "Sperry" there are the following additional elements:

VGrhizo_kmax: A matrix with maximum rhizosphere conductance values of each cohort (in rows) in each soil layer (in columns).

VGroot_kmax: A matrix with maximum root xylem conductance values of each cohort (in rows) in each soil layer (in columns).

paramsBase: A data frame with columns:

k: PAR extinction coefficient.

g: Canopy water retention capacity per LAI unit (mm/LAI).

Sgdd: Growth degree days needed for leaf budburst (for winter decideous species).

paramsTransp: A data frame with transpiration parameters. If control$transpirationMode = "Granier", columns are:

Psi_Extract: Water potential corresponding to 50% relative conductance (in MPa).

WUE: Water use efficiency for carbon assimilation (g C/mm water).

If control$transpirationMode = "Sperry" columns are:

Gwmax: Maximum stomatal conductance to water vapor (in mol H2O·m-2·s-1).

Vmax298: Maximum Rubisco carboxilation rate at 25ºC (in micromol CO2·s-1·m-2).

Jmax298: Maximum rate of electron transport at 25ºC (in micromol photons·s-1·m-2).

VCroot_c, VCroot_d: Parameters of the root xylem vulnerability curve.

xylem_kmax: Sapwood-specific hydraulic conductivity of stem xylem (in kg H2O·s-1·m-2).

VCstem_kmax: Maximum stem xylem conductance values of each cohort.

VCstem_c, VCstem_d: Parameters of the stem xylem vulnerability curve.

Transpiration: Plant cohort transpiration of the current day (mm of water = L/m2; filled with zeroes before simulations).

Photosynthesis: Plant cohort photosynthesis of the current day (g C/m2; filled with zeroes before simulations).

Functions forest2growthInput and growthInput return a list of class growthInput with the same elements as spwbInput, but with additional information.

Element above includes the following additional columns:

LA_live: Live leaf area per individual (m2/ind).

LA_dead: Dead leaf area per individual (m2/ind).

SA: Live sapwood area per individual (cm2/ind).

paramsGrowth: A data frame with columns:

SLA:Specific leaf area (mm2/mg = m2/kg).

Al2As: Leaf area to sapwood area ratio (in m2·m-2).

See also

Examples

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

# Aboveground parameters
above = forest2aboveground(exampleforest, SpParamsMED)
above

#>       SP        N   DBH Cover   H        CR   LAI_live LAI_expanded LAI_dead
#> T1_54 54  168.000 37.55    NA 800 0.7150421 0.81670117   0.81670117        0
#> T2_68 68  384.000 14.60    NA 660 0.6055642 0.79779523   0.79779523        0
#> S1_65 65 5503.456    NA  3.75  30 0.9738889 0.08913325   0.08913325        0

# Initialize soil with default soil params
examplesoil = soil(defaultSoilParams())

# Belowground parameters
below = forest2belowground(exampleforest, examplesoil, SpParamsMED)
below

#>               1          2          3          4
#> T1_54 0.1288916 0.53768380 0.24791124 0.08551338
#> T2_68 0.1288916 0.53768380 0.24791124 0.08551338
#> S1_65 0.8645807 0.09808879 0.02446089 0.01286961

# Initialize control parameters
control = defaultControl()

# Prepare spwb input
spwbInput(above, below, examplesoil,SpParamsMED, control)

#> $control
#> $control$verbose
#> [1] TRUE
#> 
#> $control$subdailyResults
#> [1] FALSE
#> 
#> $control$soilFunctions
#> [1] "SX"
#> 
#> $control$snowpack
#> [1] TRUE
#> 
#> $control$drainage
#> [1] TRUE
#> 
#> $control$leafPhenology
#> [1] TRUE
#> 
#> $control$transpirationMode
#> [1] "Granier"
#> 
#> $control$hydraulicCostFunction
#> [1] 1
#> 
#> $control$verticalLayerSize
#> [1] 100
#> 
#> $control$nStemSegments
#> [1] 1
#> 
#> $control$capacitance
#> [1] FALSE
#> 
#> $control$cavitationRefill
#> [1] TRUE
#> 
#> $control$klat
#> [1] 0.1
#> 
#> $control$taper
#> [1] TRUE
#> 
#> $control$numericParams
#> $control$numericParams$maxNsteps
#> [1] 400
#> 
#> $control$numericParams$ntrial
#> [1] 200
#> 
#> $control$numericParams$psiTol
#> [1] 1e-04
#> 
#> $control$numericParams$ETol
#> [1] 1e-07
#> 
#> 
#> $control$fracRootResistance
#> [1] 0.4
#> 
#> $control$averageFracRhizosphereResistance
#> [1] 0.15
#> 
#> $control$Catm
#> [1] 386
#> 
#> $control$ndailysteps
#> [1] 24
#> 
#> $control$thermalCapacityLAI
#> [1] 1e+06
#> 
#> $control$defaultWindSpeed
#> [1] 2.5
#> 
#> $control$storagePool
#> [1] "none"
#> 
#> 
#> $canopy
#> $canopy$gdd
#> [1] 0
#> 
#> 
#> $cohorts
#>       SP              Name
#> T1_54 54  Pinus halepensis
#> T2_68 68      Quercus ilex
#> S1_65 65 Quercus coccifera
#> 
#> $above
#>         H        CR   LAI_live LAI_expanded LAI_dead
#> T1_54 800 0.7150421 0.81670117   0.81670117        0
#> T2_68 660 0.6055642 0.79779523   0.79779523        0
#> S1_65  30 0.9738889 0.08913325   0.08913325        0
#> 
#> $below
#> $below$V
#>               1          2          3          4
#> T1_54 0.1288916 0.53768380 0.24791124 0.08551338
#> T2_68 0.1288916 0.53768380 0.24791124 0.08551338
#> S1_65 0.8645807 0.09808879 0.02446089 0.01286961
#> 
#> 
#> $paramsBase
#>       albedo    k    g Sgdd
#> T1_54   0.14 0.50 1.00    0
#> T2_68   0.18 0.55 0.50    0
#> S1_65   0.18 0.55 0.25    0
#> 
#> $paramsTransp
#>       Psi_Extract WUE pRootDisc
#> T1_54          -2   6         0
#> T2_68          -3   6         0
#> S1_65          -4   6         0
#> 
#> $Transpiration
#> T1_54 T2_68 S1_65 
#>     0     0     0 
#> 
#> $Photosynthesis
#> T1_54 T2_68 S1_65 
#>     0     0     0 
#> 
#> $PLC
#> [1] 0 0 0
#> 
#> attr(,"class")
#> [1] "spwbInput" "list"     

# When starting from an object of class 'forest' the whole process
# can be simplified:
forest2spwbInput(exampleforest, examplesoil, SpParamsMED, control)

#> $control
#> $control$verbose
#> [1] TRUE
#> 
#> $control$subdailyResults
#> [1] FALSE
#> 
#> $control$soilFunctions
#> [1] "SX"
#> 
#> $control$snowpack
#> [1] TRUE
#> 
#> $control$drainage
#> [1] TRUE
#> 
#> $control$leafPhenology
#> [1] TRUE
#> 
#> $control$transpirationMode
#> [1] "Granier"
#> 
#> $control$hydraulicCostFunction
#> [1] 1
#> 
#> $control$verticalLayerSize
#> [1] 100
#> 
#> $control$nStemSegments
#> [1] 1
#> 
#> $control$capacitance
#> [1] FALSE
#> 
#> $control$cavitationRefill
#> [1] TRUE
#> 
#> $control$klat
#> [1] 0.1
#> 
#> $control$taper
#> [1] TRUE
#> 
#> $control$numericParams
#> $control$numericParams$maxNsteps
#> [1] 400
#> 
#> $control$numericParams$ntrial
#> [1] 200
#> 
#> $control$numericParams$psiTol
#> [1] 1e-04
#> 
#> $control$numericParams$ETol
#> [1] 1e-07
#> 
#> 
#> $control$fracRootResistance
#> [1] 0.4
#> 
#> $control$averageFracRhizosphereResistance
#> [1] 0.15
#> 
#> $control$Catm
#> [1] 386
#> 
#> $control$ndailysteps
#> [1] 24
#> 
#> $control$thermalCapacityLAI
#> [1] 1e+06
#> 
#> $control$defaultWindSpeed
#> [1] 2.5
#> 
#> $control$storagePool
#> [1] "none"
#> 
#> 
#> $canopy
#> $canopy$gdd
#> [1] 0
#> 
#> 
#> $cohorts
#>       SP              Name
#> T1_54 54  Pinus halepensis
#> T2_68 68      Quercus ilex
#> S1_65 65 Quercus coccifera
#> 
#> $above
#>         H        CR   LAI_live LAI_expanded LAI_dead
#> T1_54 800 0.7150421 0.81670117   0.81670117        0
#> T2_68 660 0.6055642 0.79779523   0.79779523        0
#> S1_65  30 0.9738889 0.08913325   0.08913325        0
#> 
#> $below
#> $below$V
#>               1          2          3          4
#> T1_54 0.1288916 0.53768380 0.24791124 0.08551338
#> T2_68 0.1288916 0.53768380 0.24791124 0.08551338
#> S1_65 0.8645807 0.09808879 0.02446089 0.01286961
#> 
#> 
#> $paramsBase
#>       albedo    k    g Sgdd
#> T1_54   0.14 0.50 1.00    0
#> T2_68   0.18 0.55 0.50    0
#> S1_65   0.18 0.55 0.25    0
#> 
#> $paramsTransp
#>       Psi_Extract WUE pRootDisc
#> T1_54          -2   6         0
#> T2_68          -3   6         0
#> S1_65          -4   6         0
#> 
#> $Transpiration
#> T1_54 T2_68 S1_65 
#>     0     0     0 
#> 
#> $Photosynthesis
#> T1_54 T2_68 S1_65 
#>     0     0     0 
#> 
#> $PLC
#> [1] 0 0 0
#> 
#> attr(,"class")
#> [1] "spwbInput" "list"     

# Prepare input for complex transpiration mode
control$transpirationMode = "Sperry"
forest2spwbInput(exampleforest,examplesoil,SpParamsMED, control)

#> Warning: Soil pedotransfer functions set to Van Genuchten ('VG').

#> $control
#> $control$verbose
#> [1] TRUE
#> 
#> $control$subdailyResults
#> [1] FALSE
#> 
#> $control$soilFunctions
#> [1] "SX"
#> 
#> $control$snowpack
#> [1] TRUE
#> 
#> $control$drainage
#> [1] TRUE
#> 
#> $control$leafPhenology
#> [1] TRUE
#> 
#> $control$transpirationMode
#> [1] "Sperry"
#> 
#> $control$hydraulicCostFunction
#> [1] 1
#> 
#> $control$verticalLayerSize
#> [1] 100
#> 
#> $control$nStemSegments
#> [1] 1
#> 
#> $control$capacitance
#> [1] FALSE
#> 
#> $control$cavitationRefill
#> [1] TRUE
#> 
#> $control$klat
#> [1] 0.1
#> 
#> $control$taper
#> [1] TRUE
#> 
#> $control$numericParams
#> $control$numericParams$maxNsteps
#> [1] 400
#> 
#> $control$numericParams$ntrial
#> [1] 200
#> 
#> $control$numericParams$psiTol
#> [1] 1e-04
#> 
#> $control$numericParams$ETol
#> [1] 1e-07
#> 
#> 
#> $control$fracRootResistance
#> [1] 0.4
#> 
#> $control$averageFracRhizosphereResistance
#> [1] 0.15
#> 
#> $control$Catm
#> [1] 386
#> 
#> $control$ndailysteps
#> [1] 24
#> 
#> $control$thermalCapacityLAI
#> [1] 1e+06
#> 
#> $control$defaultWindSpeed
#> [1] 2.5
#> 
#> $control$storagePool
#> [1] "none"
#> 
#> 
#> $canopy
#> $canopy$gdd
#> [1] 0
#> 
#> $canopy$Temp
#> [1] NA
#> 
#> 
#> $cohorts
#>       SP              Name
#> T1_54 54  Pinus halepensis
#> T2_68 68      Quercus ilex
#> S1_65 65 Quercus coccifera
#> 
#> $above
#>         H        CR   LAI_live LAI_expanded LAI_dead
#> T1_54 800 0.7150421 0.81670117   0.81670117        0
#> T2_68 660 0.6055642 0.79779523   0.79779523        0
#> S1_65  30 0.9738889 0.08913325   0.08913325        0
#> 
#> $below
#> $below$V
#>               1          2          3          4
#> T1_54 0.1288916 0.53768380 0.24791124 0.08551338
#> T2_68 0.1288916 0.53768380 0.24791124 0.08551338
#> S1_65 0.8645807 0.09808879 0.02446089 0.01286961
#> 
#> $below$VGrhizo_kmax
#>               1        2       3         4
#> T1_54   2501663  3779644 1742690  601115.6
#> T2_68   3527910  5414090 2496288  861058.4
#> S1_65 498126381 21117609 5266204 2770708.0
#> 
#> $below$VCroot_kmax
#>               1         2         3         4
#> T1_54 0.4717756 0.3187333 0.3928776 0.3758143
#> T2_68 1.2693040 0.8575462 1.0570303 1.0111216
#> S1_65 0.9983568 2.7043906 2.2173668 1.2992059
#> 
#> 
#> $paramsBase
#>       albedo    k    g Sgdd
#> T1_54   0.14 0.50 1.00    0
#> T2_68   0.18 0.55 0.50    0
#> S1_65   0.18 0.55 0.25    0
#> 
#> $paramsAnatomy
#>          Al2As   SLA LeafWidth LeafDensity WoodDensity     r635
#> T1_54 1317.523 4.340       0.1         0.7    0.552530 1.964226
#> T2_68 2512.563 5.870       3.0         0.7    1.000000 1.805872
#> S1_65 2512.563 5.859       1.0         0.7    0.651891 2.289452
#> 
#> $paramsTransp
#>         Gwmin     Gwmax Vmax298  Jmax298 Kmax_stemxylem Kmax_rootxylem
#> T1_54 0.00203 0.1900000    62.5 129.5000      0.1500000       0.893000
#> T2_68 0.00450 0.1727514    66.2 130.0000      0.7735834       3.094334
#> S1_65 0.01045 0.4518345   100.0 163.6253      0.2900000       1.160000
#>       VCleaf_kmax VCleaf_c  VCleaf_d VCstem_kmax VCstem_c  VCstem_d VCroot_kmax
#> T1_54           6 2.204804 -2.448294    1.257285 2.204804 -3.672441    1.559201
#> T2_68           8 2.188265 -4.138168    4.299811 1.158507 -4.390826    4.195002
#> S1_65           8 2.931921 -5.257850   12.080826 2.931921 -7.886774    7.219320
#>       VCroot_c  VCroot_d pRootDisc Plant_kmax
#> T1_54 1.056898 -1.244767         0  0.6236803
#> T2_68 1.494098 -2.134283         0  1.6780008
#> S1_65 2.498414 -5.413810         0  2.8877280
#> 
#> $paramsWaterStorage
#>       LeafPI0 LeafEPS LeafAF        Vleaf   StemPI0  StemEPS StemAF
#> T1_54   -2.11   12.18  0.162 0.0001795440 -1.778525 10.43383    0.8
#> T2_68   -2.39   19.26  0.170 0.0001327463 -3.640000 70.78258    0.8
#> S1_65   -2.37   17.23  0.240 0.0001329955 -2.191867 16.03623    0.8
#>           Vsapwood
#> T1_54 0.0018688575
#> T2_68 0.0004421211
#> S1_65 0.0000330515
#> 
#> $Transpiration
#> T1_54 T2_68 S1_65 
#>     0     0     0 
#> 
#> $Photosynthesis
#> T1_54 T2_68 S1_65 
#>     0     0     0 
#> 
#> $PLCstem
#>       1
#> T1_54 0
#> T2_68 0
#> S1_65 0
#> 
#> $RWCsympstem
#>       1
#> T1_54 1
#> T2_68 1
#> S1_65 1
#> 
#> $RWCsympleaf
#> T1_54 T2_68 S1_65 
#>     1     1     1 
#> 
#> $Einst
#> T1_54 T2_68 S1_65 
#>     0     0     0 
#> 
#> $psiRhizo
#>       1 2 3 4
#> T1_54 0 0 0 0
#> T2_68 0 0 0 0
#> S1_65 0 0 0 0
#> 
#> $psiRoot
#> T1_54 T2_68 S1_65 
#>     0     0     0 
#> 
#> $psiStem
#>       1
#> T1_54 0
#> T2_68 0
#> S1_65 0
#> 
#> $psiLeaf
#> T1_54 T2_68 S1_65 
#>     0     0     0 
#> 
#> attr(,"class")
#> [1] "spwbInput" "list"