Forest growth

Function growth is a process-based model that performs energy, water and carbon balances; and determines changes in water/carbon pools, functional variables (leaf area, sapwood area, root area) and structural ones (tree diameter, tree height, shrub cover) for woody plant cohorts in a given forest stand during a period specified in the input climatic data.

growth(x, meteo,  latitude, elevation = NA, slope = NA, aspect = NA)

Arguments

x
An object of class growthInput.

meteo
A data frame with daily meteorological data series. Row names of the data frame should correspond to date strings with format "yyyy-mm-dd" (see Date). When x$TranspirationMode = "Granier" the following columns are required:

MeanTemperature: Mean temperature (in degrees Celsius).

Precipitation: Precipitation (in mm).

Radiation: Solar radiation (in MJ/m2/day), required only if snowpack = TRUE.

PET: Potential evapotranspiration (in mm).

WindSpeed: Wind speed (in m/s). If not available, this column can be left with NA values.

When x$TranspirationMode = "Sperry" the following columns are required:

MeanTemperature: Mean temperature(in degrees Celsius).

MinTemperature: Minimum temperature (in degrees Celsius).

MaxTemperature: Maximum temperature (in degrees Celsius).

MinRelativeHumidity: Minimum relative humidity (in percent).

MaxRelativeHumidity: Maximum relative humidity (in percent).

Precipitation: Precipitation (in mm).

Radiation: Solar radiation (in MJ/m2/day).

WindSpeed: Wind speed (in m/s). If not available, this column can be left with NA values.

latitude
Latitude (in degrees). Required when x$TranspirationMode = "Sperry".

elevation, slope, aspect
Elevation above sea level (in m), slope (in degrees) and aspect (in degrees from North). Required when x$TranspirationMode = "Sperry". Elevation is also required for 'Granier' if snowpack dynamics are simulated.

Details

Detailed model description is available in the vignettes section. Simulations using the 'Sperry' transpiration mode are computationally much more expensive than those using the simple transpiration mode.

Value

A list of class 'growth' with the following elements:

"latitude": Latitude (in degrees) given as input.

"topography": Vector with elevation, slope and aspect given as input.

"growthInput": A copy of the object x of class growthInput given as input.

"WaterBalance": A data frame where different water balance variables (see spwb).

"EnergyBalance": A data frame with the daily values of energy balance components for the soil and the canopy (only for transpirationMode = "Sperry"; see spwb).

"Temperature": A data frame with the daily values of minimum/mean/maximum temperatures for the atmosphere (input), canopy and soil (only for transpirationMode = "Sperry"; see spwb).

"Soil": A data frame where different soil variables (see spwb).

"Stand": A data frame where different stand-level variables (see spwb).

"Plants": A list of daily results for plant cohorts (see spwb).

"SunlitLeaves" and "ShadeLeaves": A list with daily results for sunlit and shade leaves (only for transpirationMode = "Sperry"; see spwb).

"PlantCarbonBalance": A list of daily carbon balance results for plant cohorts, with elements:

"GrossPhotosynthesis": Daily gross photosynthesis per dry weight of living biomass (g gluc · g dry-1).

"MaintentanceRespiration": Daily maintenance respiration per dry weight of living biomass (g gluc · g dry-1).

"GrowthCosts": Daily growth costs per dry weight of living biomass (g gluc · g dry-1).

"RootExudation": Root exudation per dry weight of living biomass (g gluc · g dry-1).

"CarbonBalance": Daily plant carbon balance (photosynthesis - maintenance respiration - growth costs - root exudation) per dry weight of living biomass (g gluc · g dry-1).

"SugarLeaf": Sugar concentration (mol·l-1) in leaves.

"StarchLeaf": Starch concentration (mol·l-1) in leaves.

"SugarSapwood": Sugar concentration (mol·l-1) in sapwood.

"StarchSapwood": Starch concentration (mol·l-1) in sapwood.

"SugarTransport": Average instantaneous rate of carbon transferred between leaves and stem compartments via floem (mol gluc·s-1).

"LeafPI0": Osmotic potential at full turgor of symplastic leaf tissue (MPa).

"StemPI0": Osmotic potential at full turgor of symplastic stem tissue (MPa).

"PlantStructure": A list of daily area and biomass values for compartments of plant cohorts, with elements:

"LeafArea": Daily amount of leaf area (in m2) for an average individual of each plant cohort.

"SapwoodArea": Daily amount of sapwood area (in cm2) for an average individual of each plant cohort.

"FineRootArea": Daily amount of fine root area (in m2) for an average individual of each plant cohort (only for transpirationMode = "Sperry").

"SapwoodBiomass": Daily amount of sapwood biomass (in g dry) for an average individual of each plant cohort.

"LeafBiomass": Daily amount of leaf biomass (in g dry) for an average individual of each plant cohort.

"FineRootBiomass": Daily amount of fine root biomass (in g dry) for an average individual of each plant cohort.

"LabileBiomass": Daily amount of labile C biomass, i.e. starch and sugars (in g dry) for an average individual of each plant cohort.

"TotalLivingBiomass": Daily amount of total living biomass, i.e. excluding heartwood, (in g dry) for an average individual of each plant cohort.

"PlantGrowth": A list of daily growth results for plant cohorts, with elements:

"LAgrowth": Leaf area growth relative to sapwood area (in m2·cm-2·day-1) for an average individual of each plant cohort.

"SAgrowth": Sapwood area relative growth rate (in cm2·cm-2·day-1) for an average individual of each plant cohort.

"FRAgrowth": Fine root area growth relative to sapwood area (in m2·cm-2·day-1) for an average individual of each plant cohort (only for transpirationMode = "Sperry").

"subdaily": A list of objects of class growth_day, one per day simulated (only if required in control parameters, see defaultControl).

Author

Miquel De Cáceres Ainsa, CREAF

Note

Objects x and soil are modified during the simulation.

See also

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize control parameters
control = defaultControl("Granier")

#Initialize soil with default soil params (4 layers)
examplesoil1 = soil(defaultSoilParams(4))

#Initialize vegetation input
x1 = forest2growthInput(exampleforestMED, examplesoil1, SpParamsMED, control)

#Call simulation function
G1<-growth(x1, examplemeteo, latitude = 41.82592, elevation = 100)

#> Initial soil water content (mm): 291.257
#> Initial snowpack content (mm): 0
#> Performing daily simulations 
#>  Year 2001:.................................... [update structural variables] done.
#> Final soil water content (mm): 271.618
#> Final snowpack content (mm): 0
#> Change in soil water content (mm): -19.6392
#> Soil water balance result (mm): -19.6392
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 85 Net rainfall (mm) 377
#>   Infiltration (mm) 419 Runoff (mm) 10 Deep drainage (mm) 108
#>   Soil evaporation (mm) 46 Transpiration (mm) 284

if (FALSE) {
#Switch to 'Sperry' transpiration mode
control = defaultControl("Sperry")

#Initialize soil with default soil params (4 layers)
examplesoil2 = soil(defaultSoilParams(4))

#Initialize vegetation input
x2 = forest2growthInput(exampleforestMED,examplesoil2, SpParamsMED, control)

#Call simulation function (11 days)
d = 100:110
G2<-growth(x2, examplemeteo[d,], latitude = 41.82592, elevation = 100)
}