Correction: Amount of water (mm) added to deep drainage to correct the water balance.

VolumeChange: Change in soil water volume (mm).

Substep: Time step of the moisture solving (seconds).

If soilDomains == "dual" the named vector contains the following additional elements:

Lateral flows: Sum of water circulating between micropores and macropores, positive when filling micropores (mm).

InfiltrationMatrix: Water infiltrated into the soil matrix (mm).

InfiltrationMacropores: Water infiltrated into the soil macropore domain (mm).

InfiltrationExcessMatrix/InfiltrationExcessMacropores: Excess infiltration in the topmost layer (mm) leading to an increase in runoff.

SaturationExcessMatrix/SaturationExcessMacropores: Excess saturation in the topmost layer (mm) leading to an increase in runoff.

DrainageMatrix: Water draining from the bottom layer of the matrix domain (mm). This quantity is corrected to close water balance in the micropore domain.

DrainageMacropores: Water draining from the bottom layer of the macropore domain (mm). This quantity is corrected to close the water balance in the macropore domain.

CorrectionMatrix: Amount of water (mm) added to deep drainage of soil matrix to correct the water balance.

CorrectionMacropores: Amount of water (mm) added to deep drainage of macropores to correct the water balance.

MatrixVolumeChange: Change in soil water volume in the soil matrix domain (mm).

MacroporeVolumeChange: Change in soil water volume in the macropore domain (mm).

Correction: Amount of water (mm) added to deep drainage to correct the water balance.

VolumeChange: Change in soil water volume (mm).

Substep: Time step of the moisture solving (seconds).

If soilDomains == "dual" the named vector contains the following additional elements:

Lateral flows: Sum of water circulating between micropores and macropores, positive when filling micropores (mm).

InfiltrationMatrix: Water infiltrated into the soil matrix (mm).

InfiltrationMacropores: Water infiltrated into the soil macropore domain (mm).

InfiltrationExcessMatrix/InfiltrationExcessMacropores: Excess infiltration in the topmost layer (mm) leading to an increase in runoff.

SaturationExcessMatrix/SaturationExcessMacropores: Excess saturation in the topmost layer (mm) leading to an increase in runoff.

DrainageMatrix: Water draining from the bottom layer of the matrix domain (mm). This quantity is corrected to close water balance in the micropore domain.

DrainageMacropores: Water draining from the bottom layer of the macropore domain (mm). This quantity is corrected to close the water balance in the macropore domain.

CorrectionMatrix: Amount of water (mm) added to deep drainage of soil matrix to correct the water balance.

CorrectionMacropores: Amount of water (mm) added to deep drainage of macropores to correct the water balance.

MatrixVolumeChange: Change in soil water volume in the soil matrix domain (mm).

MacroporeVolumeChange: Change in soil water volume in the macropore domain (mm).

Usage

hydrology_soilWaterBalance(
  soil,
  soilFunctions,
  rainfallInput,
  rainfallIntensity,
  snowmelt,
  sourceSink,
  runon = 0,
  lateralFlows = NULL,
  waterTableDepth = NA_real_,
  infiltrationMode = "GreenAmpt1911",
  infiltrationCorrection = 5,
  soilDomains = "buckets",
  nsteps = 24L,
  max_nsubsteps = 3600L,
  modifySoil = TRUE
)

References

Bonan, G. (2019). Climate change and terrestrial ecosystem modeling. Cambridge University Press, Cambridge, UK.

Jarvis, N.J., Jansson, P‐E., Dik, P.E. & Messing, I. (1991). Modelling water and solute transport in macroporous soil. I. Model description and sensitivity analysis. Journal of Soil Science, 42, 59–70.

Larsbo, M., Roulier, S., Stenemo, F., Kasteel, R. & Jarvis, N. (2005). An Improved Dual‐Permeability Model of Water Flow and Solute Transport in the Vadose Zone. Vadose Zone Journal, 4, 398–406.

Examples

# Define soil parameters
spar <- defaultSoilParams(4)

# Initializes soil hydraulic parameters
examplesoil <- soil(spar)

# Water balance in a multi-bucket model
hydrology_soilWaterBalance(examplesoil, "VG", 10, 5, 0, c(-1,-1,-1,-1), 
                           soilDomains = "buckets", modifySoil = FALSE)
#>   Local source/sinks Lateral source/sinks         Infiltration 
#>            -4.000000             0.000000            10.000000 
#>   InfiltrationExcess     SaturationExcess               Runoff 
#>             0.000000             0.000000             0.000000 
#>         DeepDrainage      CapillarityRise 
#>             4.121414             0.000000 
                           
# Water balance in a single-domain model (Richards equation)
hydrology_soilWaterBalance(examplesoil, "VG", 10, 5, 0, c(-1,-1,-1,-1), 
                           soilDomains = "single", modifySoil = FALSE)
#>   Local source/sinks Lateral source/sinks         Infiltration 
#>         -4.000000000          0.000000000         10.000000000 
#>   InfiltrationExcess     SaturationExcess               Runoff 
#>          0.000000000          0.000000000          0.000000000 
#>         DeepDrainage      CapillarityRise           Correction 
#>          0.005785380          0.000000000         -0.001286961 
#>         VolumeChange             Substeps 
#>          5.994214620         24.000000000 
                    
# Water balance in a dual-permeability model (MACRO)
hydrology_soilWaterBalance(examplesoil, "VG", 10, 5, 0, c(-1,-1,-1,-1), 
                           soilDomains = "dual", modifySoil = FALSE)
#>           Local source/sinks         Lateral source/sinks 
#>                -4.000000e+00                 0.000000e+00 
#>        Matrix-macropore flow           InfiltrationMatrix 
#>                 4.836043e+00                 4.481027e+00 
#>       InfiltrationMacropores     InfiltrationExcessMatrix 
#>                 5.518973e+00                 0.000000e+00 
#> InfiltrationExcessMacropores       SaturationExcessMatrix 
#>                 0.000000e+00                 0.000000e+00 
#>   SaturationExcessMacropores               DrainageMatrix 
#>                 0.000000e+00                 2.379449e-02 
#>           DrainageMacropores            CapillarityMatrix 
#>                 3.889382e-06                 0.000000e+00 
#>        CapillarityMacropores             CorrectionMatrix 
#>                 0.000000e+00                -4.573672e-03 
#>         CorrectionMacropores           MatrixVolumeChange 
#>                 3.889382e-06                 5.293276e+00 
#>        MacroporeVolumeChange                 Infiltration 
#>                 6.829261e-01                 1.000000e+01 
#>           InfiltrationExcess             SaturationExcess 
#>                 0.000000e+00                 0.000000e+00 
#>                       Runoff                 DeepDrainage 
#>                 0.000000e+00                 2.379838e-02 
#>              CapillarityRise                   Correction 
#>                 0.000000e+00                -4.569782e-03 
#>                 VolumeChange                     Substeps 
#>                 5.976202e+00                 2.400000e+01