Soil water input processes

Simulates soil water input processes, including canopy water interception, snow accumulation/melt, soil infiltration and percolation.

hydrology_verticalInputs(soil, soilFunctions, prec, er, tday, rad,
                         elevation, Cm, LgroundPAR, LgroundSWR, runon = 0,
                         snowpack = TRUE, drainage = TRUE, modifySoil = TRUE)

Arguments

soil
A list containing the description of the soil (see soil).

soilFunctions
Soil water retention curve and conductivity functions, either 'SX' (for Saxton) or 'VG' (for Van Genuchten).

prec
Precipitation for a given day (mm)

er
The ratio of evaporation rate to rainfall rate.

tday
Average day temperature (ºC).

rad
Solar radiation (in MJ/m2/day).

elevation
Altitude above sea level (m).

Cm
Canopy water storage capacity.

LgroundPAR
Proportion of photosynthetically-acvive radiation (PAR) reaching the ground.

LgroundSWR
Proportion of short-wave radiation (SWR) reaching the ground.

runon
Surface water amount running on the target area from upslope (in mm).

snowpack
Boolean flag to indicate the simulation of snow accumulation and melting.

drainage
Boolean flag to indicate the simulation of deep drainage.

modifySoil
Boolean flag to indicate that the input soil object should be modified during the simulation.

Details

The function simulates different vertical hydrological processes, which are described separately in other functions. If modifySoil = TRUE the function will modify the soil object (including both soil moisture and the snowpack on its surface) as a result of simulating hydrological processes.

Value

A named vector with the following elements, all in mm:

Rain

Precipitation as rainfall.

Snow

Precipitation as snow.

Interception

Rainfall water intercepted by the canopy and evaporated.

NetRain

Rainfall reaching the ground.

Snowmelt

Snow melted during the day, and added to the water infiltrated.

Runon

Surface water amount running on the target area from upslope.

Infiltration

Water infiltrated into the soil (i.e. throughfall + runon + snowmelt - runoff).

Runoff

Surface water leaving the target area.

DeepDrainage

Water leaving the target soil towards the water table.

See also