Input for simulation models

Functions forest2spwbInput and forest2growthInput take an object of class forest and create input objects for simulation functions spwb (or pwb) and growth, respectively. Function forest2aboveground calculates aboveground variables such as leaf area index. Function forest2belowground calculates belowground variables such as fine root distribution.

forest2aboveground(x, SpParams, gdd = NA_real_, loading = FALSE)

forest2belowground(x, soil, SpParams)

forest2spwbInput(x, soil, SpParams, control)

forest2growthInput(x, soil, SpParams, control)

Arguments

x

An object of class forest.

SpParams

A data frame with species parameters (see SpParamsDefinition and SpParamsMED).

gdd

Growth degree days to account for leaf phenology effects (in Celsius). This should be left NA in most applications.

loading

A logical flag to indicate that fuel loading should be included (for fire hazard calculations).

soil

An object of class soil.

control

A list with default control parameters (see defaultControl).

Value

Function forest2aboveground() returns a data frame with the following columns (rows are identified as specified by function plant_ID):

SP: Species identity (an integer) (first species is 0).

N: Cohort density (ind/ha) (see function plant_density).

DBH: Tree diameter at breast height (cm).

H: Plant total height (cm).

CR: Crown ratio (crown length to total height) (between 0 and 1).

LAI_live: Live leaf area index (m2/m2) (one-side leaf area relative to plot area), includes leaves in winter dormant buds.

LAI_expanded: Leaf area index of expanded leaves (m2/m2) (one-side leaf area relative to plot area).

LAI_dead: Dead leaf area index (m2/m2) (one-side leaf area relative to plot area).

Loading: Fine fuel loading (kg/m2), only if loading = TRUE.

Function forest2spwbInput() returns a list of class spwbInput with the following elements (rows of data frames are identified as specified by function plant_ID):

control: List with control parameters (see defaultControl).

canopy: A list of stand-level state variables.

cohorts: A data frame with cohort information, with columns SP and Name.

above: A data frame with columns H, CR and LAI (see function forest2aboveground).

below: A data frame with columns Z50, Z95. If control$transpirationMode = "Sperry" additional columns are fineRootBiomass and coarseRootSoilVolume.

belowLayers: A list. If control$transpirationMode = "Granier" it contains elements:

V: A matrix with the proportion of fine roots of each cohort (in rows) in each soil layer (in columns).

L: A matrix with the length of coarse roots of each cohort (in rows) in each soil layer (in columns).

Wpool: A matrix with the soil moisture relative to field capacity around the rhizosphere of each cohort (in rows) in each soil layer (in columns).

If control$transpirationMode = "Sperry" or control$transpirationMode = "Cochard" there are the following additional elements:

VGrhizo_kmax: A matrix with maximum rhizosphere conductance values of each cohort (in rows) in each soil layer (in columns).

VGroot_kmax: A matrix with maximum root xylem conductance values of each cohort (in rows) in each soil layer (in columns).

RhizoPsi: A matrix with the water potential around the rhizosphere of each cohort (in rows) in each soil layer (in columns).

paramsPhenology: A data frame with leaf phenology parameters:

PhenologyType: Leaf phenology type.

LeafDuration: Leaf duration (in years).

Sgdd: Degree days needed for leaf budburst (for winter decideous species).

Tbgdd: Base temperature for the calculation of degree days to leaf budburst.

Ssen: Degree days corresponding to leaf senescence.

Phsen: Photoperiod corresponding to start counting senescence degree-days.

Tbsen: Base temperature for the calculation of degree days to leaf senescence.

paramsAnatomy: A data frame with plant anatomy parameters for each cohort:

Hmax: Maximum plant height (cm).

Hmed: Median plant height (cm).

Al2As: Leaf area to sapwood area ratio (in m2·m-2).

Ar2Al: Fine root area to leaf area ratio (in m2·m-2).

SLA: Specific leaf area (mm2/mg = m2/kg).

LeafWidth: Leaf width (in cm).

LeafDensity: Density of leaf tissue (dry weight over volume).

WoodDensity: Density of wood tissue (dry weight over volume).

FineRootDensity: Density of fine root tissue (dry weight over volume).

SRL: Specific Root length (cm·g-1).

RLD: Root length density (cm·cm-3).

r635: Ratio between the weight of leaves plus branches and the weight of leaves alone for branches of 6.35 mm.

paramsInterception: A data frame with rain interception and light extinction parameters for each cohort:

kPAR: PAR extinction coefficient.

g: Canopy water retention capacity per LAI unit (mm/LAI).

If control$transpirationMode = "Sperry" or control$transpirationMode = "Cochard" additional columns are:

gammaSWR: Reflectance (albedo) coefficient for SWR .

alphaSWR: Absorbance coefficient for SWR .

paramsTranspiration: A data frame with parameters for transpiration and photosynthesis. If control$transpirationMode = "Granier", columns are:

Gswmin: Minimum stomatal conductance to water vapor (in mol H2O·m-2·s-1).

Tmax_LAI: Coefficient relating LAI with the ratio of maximum transpiration over potential evapotranspiration.

Tmax_LAIsq: Coefficient relating squared LAI with the ratio of maximum transpiration over potential evapotranspiration.

Psi_Extract: Water potential corresponding to 50% relative transpiration (in MPa).

Exp_Extract: Parameter of the Weibull function regulating transpiration reduction.

VCstem_c, VCstem_d: Parameters of the stem xylem vulnerability curve.

WUE: Daily water use efficiency (gross photosynthesis over transpiration) under no light, water or CO2 limitations and VPD = 1kPa (g C/mm water).

WUE_par: Coefficient regulating the influence of % PAR on gross photosynthesis.

WUE_par: Coefficient regulating the influence of atmospheric CO2 concentration on gross photosynthesis.

WUE_par: Coefficient regulating the influence of vapor pressure deficit (VPD) on gross photosynthesis.

If control$transpirationMode = "Sperry" columns are:

Gswmin: Minimum stomatal conductance to water vapor (in mol H2O·m-2·s-1).

Gswmax: Maximum stomatal conductance to water vapor (in mol H2O·m-2·s-1).

Vmax298: Maximum Rubisco carboxilation rate at 25ºC (in micromol CO2·s-1·m-2).

Jmax298: Maximum rate of electron transport at 25ºC (in micromol photons·s-1·m-2).

Kmax_stemxylem: Sapwood-specific hydraulic conductivity of stem xylem (in kg H2O·s-1·m-2).

Kmax_rootxylem: Sapwood-specific hydraulic conductivity of root xylem (in kg H2O·s-1·m-2).

VCleaf_kmax: Maximum leaf hydraulic conductance.

VCleaf_c, VCleaf_d: Parameters of the leaf vulnerability curve.

VCstem_kmax: Maximum stem xylem conductance.

VCstem_c, VCstem_d: Parameters of the stem xylem vulnerability curve.

VCroot_c, VCroot_d: Parameters of the root xylem vulnerability curve.

Plant_kmax: Maximum whole-plant conductance.

If control$transpirationMode = "Cochard" columns are:

Gswmin: Minimum stomatal conductance to water vapor (in mol H2O·m-2·s-1).

Gswmax: Maximum stomatal conductance to water vapor (in mol H2O·m-2·s-1).

Vmax298: Maximum Rubisco carboxilation rate at 25ºC (in micromol CO2·s-1·m-2).

Jmax298: Maximum rate of electron transport at 25ºC (in micromol photons·s-1·m-2).

Kmax_stemxylem: Sapwood-specific hydraulic conductivity of stem xylem (in kg H2O·s-1·m-2).

Kmax_rootxylem: Sapwood-specific hydraulic conductivity of root xylem (in kg H2O·s-1·m-2).

VCleaf_kmax: Maximum leaf hydraulic conductance.

VCleaf_c, VCleaf_d: Parameters of the leaf vulnerability curve.

VCstem_kmax: Maximum stem xylem conductance.

VCstem_c, VCstem_d: Parameters of the stem xylem vulnerability curve.

VCroot_c, VCroot_d: Parameters of the root xylem vulnerability curve.

Plant_kmax: Maximum whole-plant conductance.

paramsWaterStorage: A data frame with plant water storage parameters for each cohort:

LeafPI0: Osmotic potential at full turgor of leaves (MPa).

LeafEPS: Modulus of elasticity (capacity of the cell wall to resist changes in volume in response to changes in turgor) of leaves (MPa).

LeafAF: Apoplastic fraction (proportion of water outside the living cells) in leaves.

Vleaf: Storage water capacity in leaves, per leaf area (L/m2).

StemPI0: Osmotic potential at full turgor of symplastic xylem tissue (MPa).

StemEPS: Modulus of elasticity (capacity of the cell wall to resist changes in volume in response to changes in turgor) of symplastic xylem tissue (Mpa).

StemAF: Apoplastic fraction (proportion of water outside the living cells) in stem xylem.

Vstem: Storage water capacity in sapwood, per leaf area (L/m2).

internalPhenology and internalWater: data frames to store internal state variables.

internalFCCS: A data frame with fuel characteristics, according to fuel_FCCS (only if fireHazardResults = TRUE, in the control list).

Function forest2growthInput returns a list of class growthInput with the same elements as spwbInput, but with additional information.

Element above includes the following additional columns:

LA_live: Live leaf area per individual (m2/ind).

LA_dead: Dead leaf area per individual (m2/ind).

SA: Live sapwood area per individual (cm2/ind).

paramsGrowth: A data frame with growth parameters for each cohort:

RERleaf: Maintenance respiration rates (at 20ºC) for leaves (in g gluc·g dry-1·day-1).

RERsapwood: Maintenance respiration rates (at 20ºC) for sapwood (in g gluc·g dry-1·day-1).

RERfineroot: Maintenance respiration rates (at 20ºC) for fine roots (in g gluc·g dry-1·day-1).

CCleaf: Leaf construction costs (in g gluc·g dry-1).

CCsapwood: Sapwood construction costs (in g gluc·g dry-1).

CCfineroot: Fine root construction costs (in g gluc·g dry-1).

RGRleafmax: Maximum leaf relative growth rate (in m2·cm-2·day-1).

RGRsapwoodmax: Maximum sapwood relative growth rate (in cm2·cm-2·day-1).

RGRfinerootmax: Maximum fine root relative growth rate (in g dry·g dry-1·day-1).

SRsapwood: Sapwood daily senescence rate (in day-1).

SRfineroot: Fine root daily senescence rate (in day-1).

RSSG: Minimum relative starch for sapwood growth (proportion).

fHDmin: Minimum value of the height-to-diameter ratio (dimensionless).

fHDmax: Maximum value of the height-to-diameter ratio (dimensionless).

WoodC: Wood carbon content per dry weight (g C /g dry).

MortalityBaselineRate: Deterministic proportion or probability specifying the baseline reduction of cohort's density occurring in a year.

paramsAllometry: A data frame with allometric parameters for each cohort:

Aash: Regression coefficient relating the square of shrub height with shrub area.

Absh, Bbsh: Allometric coefficients relating phytovolume with dry weight of shrub individuals.

Acr, B1cr, B2cr, B3cr, C1cr, C2cr: Regression coefficients used to calculate crown ratio of trees.

Acw, Bcw: Regression coefficients used to calculated crown width of trees.

internalAllocation: A data frame with internal allocation variables for each cohort:

allocationTarget: Value of the allocation target variable.

leafAreaTarget: Target leaf area (m2) per individual.

sapwoodAreaTarget: Target sapwood area (cm2) per individual.

fineRootBiomassTarget: Target fine root biomass (g dry) per individual.

crownBudPercent: Percentage of the crown with buds.

internalCarbon: A data frame with the concentration (mol·gluc·l-1) of metabolic and storage carbon compartments for leaves and sapwood.

internalMortality: A data frame to store the cumulative mortality (density for trees and cover for shrubs) predicted during the simulation, also distinguishing mortality due to starvation or dessication.

Details

Function forest2aboveground extract height and species identity from plant cohorts of x, and calculate leaf area index and crown ratio. Functions forest2spwbInput and forest2growthInput also calculate the distribution of fine roots across soil, and finds parameter values for each plant cohort according to the parameters of its species as specified in SpParams. If control$transpirationMode = "Sperry" or control$transpirationMode = "Cochard", the forest2spwbInput and forest2growthInput also estimate the maximum conductance of rhizosphere, root xylem and stem xylem elements.

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

# Aboveground parameters
forest2aboveground(exampleforestMED, SpParamsMED)
#>         SP        N   DBH Cover   H        CR   LAI_live LAI_expanded LAI_dead
#> T1_148 148 168.0000 37.55    NA 800 0.6605196 0.84874773   0.84874773        0
#> T2_168 168 384.0000 14.60    NA 660 0.6055642 0.70557382   0.70557382        0
#> S1_165 165 749.4923    NA  3.75  80 0.8032817 0.02718804   0.02718804        0

# Initialize soil with default soil params
examplesoil <- soil(defaultSoilParams())

# Bewowground parameters (distribution of fine roots)
forest2belowground(exampleforestMED, examplesoil, SpParamsMED)
#>                1         2          3           4
#> T1_148 0.8604899 0.1194556 0.01511005 0.004944476
#> T2_168 0.5008953 0.4505941 0.04064831 0.007862284
#> S1_165 0.6799879 0.2737911 0.03567632 0.010544678

# Initialize control parameters using 'Granier' transpiration mode
control <- defaultControl("Granier")

# Prepare spwb input
forest2spwbInput(exampleforestMED, examplesoil, SpParamsMED, control)
#> $control
#> $control$fillMissingRootParams
#> [1] TRUE
#> 
#> $control$fillMissingSpParams
#> [1] TRUE
#> 
#> $control$verbose
#> [1] TRUE
#> 
#> $control$subdailyResults
#> [1] FALSE
#> 
#> $control$standResults
#> [1] TRUE
#> 
#> $control$soilResults
#> [1] TRUE
#> 
#> $control$plantResults
#> [1] TRUE
#> 
#> $control$leafResults
#> [1] TRUE
#> 
#> $control$temperatureResults
#> [1] TRUE
#> 
#> $control$fireHazardResults
#> [1] FALSE
#> 
#> $control$transpirationMode
#> [1] "Granier"
#> 
#> $control$soilFunctions
#> [1] "VG"
#> 
#> $control$defaultWindSpeed
#> [1] 2.5
#> 
#> $control$defaultCO2
#> [1] 386
#> 
#> $control$snowpack
#> [1] TRUE
#> 
#> $control$leafPhenology
#> [1] TRUE
#> 
#> $control$rockyLayerDrainage
#> [1] TRUE
#> 
#> $control$unlimitedSoilWater
#> [1] FALSE
#> 
#> $control$rhizosphereOverlap
#> [1] "total"
#> 
#> $control$unfoldingDD
#> [1] 300
#> 
#> $control$verticalLayerSize
#> [1] 100
#> 
#> $control$windMeasurementHeight
#> [1] 200
#> 
#> $control$cavitationRefill
#> [1] "total"
#> 
#> $control$hydraulicRedistributionFraction
#> [1] 0.1
#> 
#> $control$ndailysteps
#> [1] 24
#> 
#> $control$nsubsteps
#> [1] 3600
#> 
#> $control$capacitance
#> [1] FALSE
#> 
#> $control$taper
#> [1] TRUE
#> 
#> $control$multiLayerBalance
#> [1] FALSE
#> 
#> $control$maximumStemConductance
#> [1] 10
#> 
#> $control$klatstem
#> [1] 0.01
#> 
#> $control$klatleaf
#> [1] 0.01
#> 
#> $control$numericParams
#> $control$numericParams$maxNsteps
#> [1] 400
#> 
#> $control$numericParams$ntrial
#> [1] 200
#> 
#> $control$numericParams$psiTol
#> [1] 1e-04
#> 
#> $control$numericParams$ETol
#> [1] 1e-07
#> 
#> 
#> $control$fracLeafResistance
#> [1] NA
#> 
#> $control$fracRootResistance
#> [1] 0.4
#> 
#> $control$averageFracRhizosphereResistance
#> [1] 0.15
#> 
#> $control$thermalCapacityLAI
#> [1] 1e+06
#> 
#> $control$boundaryLayerSize
#> [1] 2000
#> 
#> $control$refillMaximumRate
#> [1] 0.05
#> 
#> $control$subdailyCarbonBalance
#> [1] FALSE
#> 
#> $control$allowDessication
#> [1] TRUE
#> 
#> $control$allowStarvation
#> [1] TRUE
#> 
#> $control$sinkLimitation
#> [1] TRUE
#> 
#> $control$shrubDynamics
#> [1] TRUE
#> 
#> $control$allocationStrategy
#> [1] "Al2As"
#> 
#> $control$phloemConductanceFactor
#> [1] 0.2
#> 
#> $control$nonSugarConcentration
#> [1] 0.25
#> 
#> $control$equilibriumOsmoticConcentration
#> $control$equilibriumOsmoticConcentration$leaf
#> [1] 0.8
#> 
#> $control$equilibriumOsmoticConcentration$sapwood
#> [1] 0.6
#> 
#> 
#> $control$minimumRelativeStarchForGrowth
#> [1] 0.5
#> 
#> $control$constructionCosts
#> $control$constructionCosts$leaf
#> [1] 1.5
#> 
#> $control$constructionCosts$sapwood
#> [1] 1.47
#> 
#> $control$constructionCosts$fineroot
#> [1] 1.3
#> 
#> 
#> $control$senescenceRates
#> $control$senescenceRates$sapwood
#> [1] 0.000135
#> 
#> $control$senescenceRates$fineroot
#> [1] 0.001897231
#> 
#> 
#> $control$maximumRelativeGrowthRates
#> $control$maximumRelativeGrowthRates$leaf
#> [1] 0.03
#> 
#> $control$maximumRelativeGrowthRates$cambium
#> [1] 0.0025
#> 
#> $control$maximumRelativeGrowthRates$sapwood
#> [1] 0.002
#> 
#> $control$maximumRelativeGrowthRates$fineroot
#> [1] 0.1
#> 
#> 
#> $control$mortalityMode
#> [1] "density/deterministic"
#> 
#> $control$mortalityBaselineRate
#> [1] 0.0015
#> 
#> $control$mortalityRelativeSugarThreshold
#> [1] 0.3
#> 
#> $control$mortalityRWCThreshold
#> [1] 0.3
#> 
#> $control$recrTreeDBH
#> [1] 1
#> 
#> $control$recrTreeDensity
#> [1] 1000
#> 
#> $control$ingrowthTreeDBH
#> [1] 7.5
#> 
#> $control$ingrowthTreeDensity
#> [1] 127
#> 
#> $control$allowRecruitment
#> [1] TRUE
#> 
#> $control$recruitmentMode
#> [1] "stochastic"
#> 
#> $control$allowResprouting
#> [1] TRUE
#> 
#> $control$removeEmptyCohorts
#> [1] TRUE
#> 
#> $control$minimumCohortDensity
#> [1] 1
#> 
#> $control$seedRain
#> NULL
#> 
#> $control$seedProductionTreeHeight
#> [1] 300
#> 
#> $control$seedProductionShrubHeight
#> [1] 30
#> 
#> $control$probRecr
#> [1] 0.05
#> 
#> $control$minTempRecr
#> [1] 0
#> 
#> $control$minMoistureRecr
#> [1] 0.3
#> 
#> $control$minFPARRecr
#> [1] 10
#> 
#> $control$recrTreeHeight
#> [1] 620
#> 
#> $control$recrShrubCover
#> [1] 1
#> 
#> $control$recrShrubHeight
#> [1] 25
#> 
#> $control$recrTreeZ50
#> [1] 100
#> 
#> $control$recrShrubZ50
#> [1] 50
#> 
#> $control$recrTreeZ95
#> [1] 1000
#> 
#> $control$recrShrubZ95
#> [1] 500
#> 
#> 
#> $soil
#> Soil depth (mm): 4000 
#> 
#> Layer  1  [ 0  to  300 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 25 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 32 Vol. FC (mm): 68 Vol. SAT (mm): 111 Vol. current (mm): 68 
#>     Temperature (Celsius): NA 
#> 
#> Layer  2  [ 300  to  1000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 45 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 55 Vol. FC (mm): 117 Vol. SAT (mm): 190 Vol. current (mm): 117 
#>     Temperature (Celsius): NA 
#> 
#> Layer  3  [ 1000  to  2000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 75 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 36 Vol. FC (mm): 76 Vol. SAT (mm): 123 Vol. current (mm): 76 
#>     Temperature (Celsius): NA 
#> 
#> Layer  4  [ 2000  to  4000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 95 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 14 Vol. FC (mm): 30 Vol. SAT (mm): 49 Vol. current (mm): 30 
#>     Temperature (Celsius): NA 
#> 
#> Total soil saturated capacity (mm): 473 
#> Total soil water holding capacity (mm): 291 
#> Total soil extractable water (mm): 183 
#> Total soil current Volume (mm): 291 
#> 
#> Snow pack water equivalent (mm): 0 
#> Soil water table depth (mm): 4000 
#> 
#> 
#> $canopy
#> data frame with 0 columns and 0 rows
#> 
#> $herbLAI
#> [1] 0.252
#> 
#> $cohorts
#>         SP              Name
#> T1_148 148  Pinus halepensis
#> T2_168 168      Quercus ilex
#> S1_165 165 Quercus coccifera
#> 
#> $above
#>          H        CR        N   LAI_live LAI_expanded LAI_dead
#> T1_148 800 0.6605196 168.0000 0.84874773   0.84874773        0
#> T2_168 660 0.6055642 384.0000 0.70557382   0.70557382        0
#> S1_165  80 0.8032817 749.4923 0.02718804   0.02718804        0
#> 
#> $below
#>        Z50  Z95 fineRootBiomass coarseRootSoilVolume
#> T1_148 100  600     1381.890945             4.381122
#> T2_168 300 1000      546.693144             7.526346
#> S1_165 200 1000        9.397368             5.544086
#> 
#> $belowLayers
#> $belowLayers$V
#>                1         2          3           4
#> T1_148 0.8604899 0.1194556 0.01511005 0.004944476
#> T2_168 0.5008953 0.4505941 0.04064831 0.007862284
#> S1_165 0.6799879 0.2737911 0.03567632 0.010544678
#> 
#> $belowLayers$L
#>           1        2        3        4
#> T1_148 2150 1137.834 1645.161 3058.717
#> T2_168 2150 1891.826 1812.060 3097.046
#> S1_165 2150 1480.808 1750.917 3096.459
#> 
#> $belowLayers$Wpool
#>        1 2 3 4
#> T1_148 1 1 1 1
#> T2_168 1 1 1 1
#> S1_165 1 1 1 1
#> 
#> 
#> $paramsPhenology
#>             PhenologyType LeafDuration t0gdd  Sgdd Tbgdd  Ssen Phsen Tbsen xsen
#> T1_148 oneflush-evergreen     2.536875  50.0 200.0  0.00  8268  12.5  28.5    2
#> T2_168 oneflush-evergreen     2.183837  54.5 240.7  4.34 10178  12.5  28.5    2
#> S1_165 oneflush-evergreen     1.250000  54.5 240.7  4.34 10178  12.5  28.5    2
#>        ysen
#> T1_148    2
#> T2_168    2
#> S1_165    2
#> 
#> $paramsAnatomy
#>           Al2As Ar2Al      SLA LeafDensity WoodDensity FineRootDensity      SRL
#> T1_148 1317.523     1 5.140523   0.2982842   0.6077016       0.2982842 3172.572
#> T2_168 3908.823     1 6.340000   0.4893392   0.9008264       0.4893392 4398.812
#> S1_165 4189.325     1 4.980084   0.3709679   0.4389106       0.3709679 4398.812
#>        RLD     r635
#> T1_148  10 1.964226
#> T2_168  10 1.805872
#> S1_165  10 2.289452
#> 
#> $paramsInterception
#>        kPAR   g
#> T1_148 0.50 1.0
#> T2_168 0.55 0.5
#> S1_165 0.55 0.5
#> 
#> $paramsTranspiration
#>             Gswmin   Tmax_LAI   Tmax_LAIsq Psi_Extract Exp_Extract  VCstem_c
#> T1_148 0.003086667 0.13847869 -0.006200539  -0.8507809     1.47061 12.710000
#> T2_168 0.004473333 0.09146279 -0.004095349  -1.6598896     1.06530  3.560000
#> S1_165 0.010455247 0.13400000 -0.006000000  -2.1210726     1.30000  3.537784
#>         VCstem_d      WUE   WUE_par     WUE_co2    WUE_vpd
#> T1_148 -5.290000 8.523012 0.6843513 0.002517798 -0.3035192
#> T2_168 -7.720000 8.447722 0.2523021 0.002721234 -0.5791330
#> S1_165 -4.126512 7.900000 0.3643000 0.002757000 -0.4636000
#> 
#> $paramsWaterStorage
#>           maxFMC   LeafPI0   LeafEPS LeafAF     Vleaf   StemPI0   StemEPS
#> T1_148 126.03063 -1.591429  8.918571 0.3525 0.5258525 -2.008039 13.256355
#> T2_168  93.15304 -1.483333 19.260000 0.1700 0.2199087 -3.227438 46.420610
#> S1_165  96.53441 -2.370000 17.230000 0.2400 0.4108968 -1.305868  6.297155
#>           StemAF Vsapwood
#> T1_148 0.9236406 6.440145
#> T2_168 0.6238125 1.339581
#> S1_165 0.6238125 1.163264
#> 
#> $internalPhenology
#>        gdd sen budFormation leafUnfolding leafSenescence leafDormancy phi
#> T1_148   0   0        FALSE         FALSE          FALSE        FALSE   0
#> T2_168   0   0        FALSE         FALSE          FALSE        FALSE   0
#> S1_165   0   0        FALSE         FALSE          FALSE        FALSE   0
#> 
#> $internalWater
#>        PlantPsi StemPLC
#> T1_148   -0.033       0
#> T2_168   -0.033       0
#> S1_165   -0.033       0
#> 
#> $internalFCCS
#> data frame with 0 columns and 0 rows
#> 
#> attr(,"class")
#> [1] "spwbInput" "list"     
                
# Prepare input for 'Sperry' transpiration mode
control <- defaultControl("Sperry")
forest2spwbInput(exampleforestMED,examplesoil,SpParamsMED, control)
#> $control
#> $control$fillMissingRootParams
#> [1] TRUE
#> 
#> $control$fillMissingSpParams
#> [1] TRUE
#> 
#> $control$verbose
#> [1] TRUE
#> 
#> $control$subdailyResults
#> [1] FALSE
#> 
#> $control$standResults
#> [1] TRUE
#> 
#> $control$soilResults
#> [1] TRUE
#> 
#> $control$plantResults
#> [1] TRUE
#> 
#> $control$leafResults
#> [1] TRUE
#> 
#> $control$temperatureResults
#> [1] TRUE
#> 
#> $control$fireHazardResults
#> [1] FALSE
#> 
#> $control$transpirationMode
#> [1] "Sperry"
#> 
#> $control$soilFunctions
#> [1] "VG"
#> 
#> $control$defaultWindSpeed
#> [1] 2.5
#> 
#> $control$defaultCO2
#> [1] 386
#> 
#> $control$snowpack
#> [1] TRUE
#> 
#> $control$leafPhenology
#> [1] TRUE
#> 
#> $control$rockyLayerDrainage
#> [1] TRUE
#> 
#> $control$unlimitedSoilWater
#> [1] FALSE
#> 
#> $control$rhizosphereOverlap
#> [1] "total"
#> 
#> $control$unfoldingDD
#> [1] 300
#> 
#> $control$verticalLayerSize
#> [1] 100
#> 
#> $control$windMeasurementHeight
#> [1] 200
#> 
#> $control$cavitationRefill
#> [1] "total"
#> 
#> $control$hydraulicRedistributionFraction
#> [1] 0.1
#> 
#> $control$ndailysteps
#> [1] 24
#> 
#> $control$nsubsteps
#> [1] 3600
#> 
#> $control$capacitance
#> [1] FALSE
#> 
#> $control$taper
#> [1] TRUE
#> 
#> $control$multiLayerBalance
#> [1] FALSE
#> 
#> $control$maximumStemConductance
#> [1] 10
#> 
#> $control$klatstem
#> [1] 0.01
#> 
#> $control$klatleaf
#> [1] 0.01
#> 
#> $control$numericParams
#> $control$numericParams$maxNsteps
#> [1] 400
#> 
#> $control$numericParams$ntrial
#> [1] 200
#> 
#> $control$numericParams$psiTol
#> [1] 1e-04
#> 
#> $control$numericParams$ETol
#> [1] 1e-07
#> 
#> 
#> $control$fracLeafResistance
#> [1] NA
#> 
#> $control$fracRootResistance
#> [1] 0.4
#> 
#> $control$averageFracRhizosphereResistance
#> [1] 0.15
#> 
#> $control$thermalCapacityLAI
#> [1] 1e+06
#> 
#> $control$boundaryLayerSize
#> [1] 2000
#> 
#> $control$refillMaximumRate
#> [1] 0.05
#> 
#> $control$subdailyCarbonBalance
#> [1] FALSE
#> 
#> $control$allowDessication
#> [1] TRUE
#> 
#> $control$allowStarvation
#> [1] TRUE
#> 
#> $control$sinkLimitation
#> [1] TRUE
#> 
#> $control$shrubDynamics
#> [1] TRUE
#> 
#> $control$allocationStrategy
#> [1] "Al2As"
#> 
#> $control$phloemConductanceFactor
#> [1] 0.2
#> 
#> $control$nonSugarConcentration
#> [1] 0.25
#> 
#> $control$equilibriumOsmoticConcentration
#> $control$equilibriumOsmoticConcentration$leaf
#> [1] 0.8
#> 
#> $control$equilibriumOsmoticConcentration$sapwood
#> [1] 0.6
#> 
#> 
#> $control$minimumRelativeStarchForGrowth
#> [1] 0.5
#> 
#> $control$constructionCosts
#> $control$constructionCosts$leaf
#> [1] 1.5
#> 
#> $control$constructionCosts$sapwood
#> [1] 1.47
#> 
#> $control$constructionCosts$fineroot
#> [1] 1.3
#> 
#> 
#> $control$senescenceRates
#> $control$senescenceRates$sapwood
#> [1] 0.000135
#> 
#> $control$senescenceRates$fineroot
#> [1] 0.001897231
#> 
#> 
#> $control$maximumRelativeGrowthRates
#> $control$maximumRelativeGrowthRates$leaf
#> [1] 0.03
#> 
#> $control$maximumRelativeGrowthRates$cambium
#> [1] 0.0025
#> 
#> $control$maximumRelativeGrowthRates$sapwood
#> [1] 0.002
#> 
#> $control$maximumRelativeGrowthRates$fineroot
#> [1] 0.1
#> 
#> 
#> $control$mortalityMode
#> [1] "density/deterministic"
#> 
#> $control$mortalityBaselineRate
#> [1] 0.0015
#> 
#> $control$mortalityRelativeSugarThreshold
#> [1] 0.3
#> 
#> $control$mortalityRWCThreshold
#> [1] 0.3
#> 
#> $control$recrTreeDBH
#> [1] 1
#> 
#> $control$recrTreeDensity
#> [1] 1000
#> 
#> $control$ingrowthTreeDBH
#> [1] 7.5
#> 
#> $control$ingrowthTreeDensity
#> [1] 127
#> 
#> $control$allowRecruitment
#> [1] TRUE
#> 
#> $control$recruitmentMode
#> [1] "stochastic"
#> 
#> $control$allowResprouting
#> [1] TRUE
#> 
#> $control$removeEmptyCohorts
#> [1] TRUE
#> 
#> $control$minimumCohortDensity
#> [1] 1
#> 
#> $control$seedRain
#> NULL
#> 
#> $control$seedProductionTreeHeight
#> [1] 300
#> 
#> $control$seedProductionShrubHeight
#> [1] 30
#> 
#> $control$probRecr
#> [1] 0.05
#> 
#> $control$minTempRecr
#> [1] 0
#> 
#> $control$minMoistureRecr
#> [1] 0.3
#> 
#> $control$minFPARRecr
#> [1] 10
#> 
#> $control$recrTreeHeight
#> [1] 620
#> 
#> $control$recrShrubCover
#> [1] 1
#> 
#> $control$recrShrubHeight
#> [1] 25
#> 
#> $control$recrTreeZ50
#> [1] 100
#> 
#> $control$recrShrubZ50
#> [1] 50
#> 
#> $control$recrTreeZ95
#> [1] 1000
#> 
#> $control$recrShrubZ95
#> [1] 500
#> 
#> 
#> $soil
#> Soil depth (mm): 4000 
#> 
#> Layer  1  [ 0  to  300 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 25 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 32 Vol. FC (mm): 68 Vol. SAT (mm): 111 Vol. current (mm): 68 
#>     Temperature (Celsius): NA 
#> 
#> Layer  2  [ 300  to  1000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 45 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 55 Vol. FC (mm): 117 Vol. SAT (mm): 190 Vol. current (mm): 117 
#>     Temperature (Celsius): NA 
#> 
#> Layer  3  [ 1000  to  2000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 75 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 36 Vol. FC (mm): 76 Vol. SAT (mm): 123 Vol. current (mm): 76 
#>     Temperature (Celsius): NA 
#> 
#> Layer  4  [ 2000  to  4000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 95 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 14 Vol. FC (mm): 30 Vol. SAT (mm): 49 Vol. current (mm): 30 
#>     Temperature (Celsius): NA 
#> 
#> Total soil saturated capacity (mm): 473 
#> Total soil water holding capacity (mm): 291 
#> Total soil extractable water (mm): 183 
#> Total soil current Volume (mm): 291 
#> 
#> Snow pack water equivalent (mm): 0 
#> Soil water table depth (mm): 4000 
#> 
#> 
#> $canopy
#>    zlow zmid  zup Tair Cair VPair
#> 1     0   50  100   NA   NA    NA
#> 2   100  150  200   NA   NA    NA
#> 3   200  250  300   NA   NA    NA
#> 4   300  350  400   NA   NA    NA
#> 5   400  450  500   NA   NA    NA
#> 6   500  550  600   NA   NA    NA
#> 7   600  650  700   NA   NA    NA
#> 8   700  750  800   NA   NA    NA
#> 9   800  850  900   NA   NA    NA
#> 10  900  950 1000   NA   NA    NA
#> 11 1000 1050 1100   NA   NA    NA
#> 12 1100 1150 1200   NA   NA    NA
#> 13 1200 1250 1300   NA   NA    NA
#> 14 1300 1350 1400   NA   NA    NA
#> 15 1400 1450 1500   NA   NA    NA
#> 16 1500 1550 1600   NA   NA    NA
#> 17 1600 1650 1700   NA   NA    NA
#> 18 1700 1750 1800   NA   NA    NA
#> 19 1800 1850 1900   NA   NA    NA
#> 20 1900 1950 2000   NA   NA    NA
#> 21 2000 2050 2100   NA   NA    NA
#> 22 2100 2150 2200   NA   NA    NA
#> 23 2200 2250 2300   NA   NA    NA
#> 24 2300 2350 2400   NA   NA    NA
#> 25 2400 2450 2500   NA   NA    NA
#> 26 2500 2550 2600   NA   NA    NA
#> 27 2600 2650 2700   NA   NA    NA
#> 28 2700 2750 2800   NA   NA    NA
#> 
#> $herbLAI
#> [1] 0.252
#> 
#> $cohorts
#>         SP              Name
#> T1_148 148  Pinus halepensis
#> T2_168 168      Quercus ilex
#> S1_165 165 Quercus coccifera
#> 
#> $above
#>          H        CR        N   LAI_live LAI_expanded LAI_dead
#> T1_148 800 0.6605196 168.0000 0.84874773   0.84874773        0
#> T2_168 660 0.6055642 384.0000 0.70557382   0.70557382        0
#> S1_165  80 0.8032817 749.4923 0.02718804   0.02718804        0
#> 
#> $below
#>        Z50  Z95 fineRootBiomass coarseRootSoilVolume
#> T1_148 100  600       1612.7328                16.75
#> T2_168 300 1000       1590.6433                13.25
#> S1_165 200 1000        152.1546                 0.25
#> 
#> $belowLayers
#> $belowLayers$V
#>                1         2          3           4
#> T1_148 0.8604899 0.1194556 0.01511005 0.004944476
#> T2_168 0.5008953 0.4505941 0.04064831 0.007862284
#> S1_165 0.6799879 0.2737911 0.03567632 0.010544678
#> 
#> $belowLayers$L
#>                1        2        3        4
#> T1_148 4815.5858 2286.191 2567.669 4513.443
#> T2_168 3364.1867 3221.751 2828.103 4575.847
#> S1_165  790.4044 1237.889 2106.565 4091.603
#> 
#> $belowLayers$VGrhizo_kmax
#>                1        2         3          4
#> T1_148  27535774  1321053  167101.2   54680.68
#> T2_168  67483373 22684370 2046368.0  395812.89
#> S1_165 404209828 58944700 7680780.7 2270171.56
#> 
#> $belowLayers$VCroot_kmax
#>                1         2          3           4
#> T1_148 1.1305113 0.3305767 0.03723103 0.006930914
#> T2_168 0.8450863 0.7938304 0.08157959 0.009752397
#> S1_165 3.6420477 0.9363356 0.07169666 0.010910223
#> 
#> $belowLayers$Wpool
#>        1 2 3 4
#> T1_148 1 1 1 1
#> T2_168 1 1 1 1
#> S1_165 1 1 1 1
#> 
#> $belowLayers$RhizoPsi
#>             1      2      3      4
#> T1_148 -0.033 -0.033 -0.033 -0.033
#> T2_168 -0.033 -0.033 -0.033 -0.033
#> S1_165 -0.033 -0.033 -0.033 -0.033
#> 
#> 
#> $paramsPhenology
#>             PhenologyType LeafDuration t0gdd  Sgdd Tbgdd  Ssen Phsen Tbsen xsen
#> T1_148 oneflush-evergreen     2.536875  50.0 200.0  0.00  8268  12.5  28.5    2
#> T2_168 oneflush-evergreen     2.183837  54.5 240.7  4.34 10178  12.5  28.5    2
#> S1_165 oneflush-evergreen     1.250000  54.5 240.7  4.34 10178  12.5  28.5    2
#>        ysen
#> T1_148    2
#> T2_168    2
#> S1_165    2
#> 
#> $paramsAnatomy
#>        Hmed    Al2As      SLA LeafWidth LeafDensity WoodDensity FineRootDensity
#> T1_148  850 1317.523 5.140523 0.1384772   0.2982842   0.6077016       0.2982842
#> T2_168  500 3908.823 6.340000 1.7674359   0.4893392   0.9008264       0.4893392
#> S1_165   80 4189.325 4.980084 1.3761085   0.3709679   0.4389106       0.3709679
#>        conduit2sapwood      SRL RLD     r635
#> T1_148       0.9236406 3172.572  10 1.964226
#> T2_168       0.6238125 4398.812  10 1.805872
#> S1_165       0.6238125 4398.812  10 2.289452
#> 
#> $paramsInterception
#>        alphaSWR gammaSWR kPAR   g
#> T1_148      0.7     0.14 0.50 1.0
#> T2_168      0.7     0.18 0.55 0.5
#> S1_165      0.7     0.18 0.55 0.5
#> 
#> $paramsTranspiration
#>             Gswmin    Gswmax  Vmax298  Jmax298 Kmax_stemxylem Kmax_rootxylem
#> T1_148 0.003086667 0.2850000 72.19617 124.1687           0.15           0.60
#> T2_168 0.004473333 0.2007222 68.51600 118.7863           0.40           1.60
#> S1_165 0.010455247 0.2830167 62.78100 118.4486           0.29           1.16
#>        VCleaf_kmax  VCleaf_c  VCleaf_d VCstem_kmax  VCstem_c  VCstem_d
#> T1_148    4.000000 11.137051 -2.380849    1.339563 12.710000 -5.290000
#> T2_168    4.000000  1.339370 -2.582279    1.620936  3.560000 -7.720000
#> S1_165    9.579077  1.844224 -3.030130    4.599269  3.537784 -4.126512
#>        VCroot_kmax  VCroot_c  VCroot_d VGrhizo_kmax Plant_kmax
#> T1_148    1.505250 11.137051 -3.065569     29078608  0.6021000
#> T2_168    1.730249  1.339370 -2.214081     92609924  0.6920994
#> S1_165    4.660990  1.760307 -2.797092    473105480  1.8643961
#> 
#> $paramsWaterStorage
#>           maxFMC   LeafPI0   LeafEPS LeafAF     Vleaf   StemPI0   StemEPS
#> T1_148 126.03063 -1.591429  8.918571 0.3525 0.5258525 -2.008039 13.256355
#> T2_168  93.15304 -1.483333 19.260000 0.1700 0.2199087 -3.227438 46.420610
#> S1_165  96.53441 -2.370000 17.230000 0.2400 0.4108968 -1.305868  6.297155
#>           StemAF Vsapwood
#> T1_148 0.9236406 7.701237
#> T2_168 0.6238125 1.631003
#> S1_165 0.6238125 1.631676
#> 
#> $internalPhenology
#>        gdd sen budFormation leafUnfolding leafSenescence leafDormancy phi
#> T1_148   0   0        FALSE         FALSE          FALSE        FALSE   0
#> T2_168   0   0        FALSE         FALSE          FALSE        FALSE   0
#> S1_165   0   0        FALSE         FALSE          FALSE        FALSE   0
#> 
#> $internalWater
#>        Einst RootCrownPsi Stem1Psi Stem2Psi LeafPsi StemSympPsi LeafSympPsi
#> T1_148     0       -0.033   -0.033   -0.033  -0.033      -0.033      -0.033
#> T2_168     0       -0.033   -0.033   -0.033  -0.033      -0.033      -0.033
#> S1_165     0       -0.033   -0.033   -0.033  -0.033      -0.033      -0.033
#>        StemPLC
#> T1_148       0
#> T2_168       0
#> S1_165       0
#> 
#> $internalFCCS
#> data frame with 0 columns and 0 rows
#> 
#> attr(,"class")
#> [1] "spwbInput" "list"     

# Prepare input for 'Cochard' transpiration mode
control <- defaultControl("Cochard")
forest2spwbInput(exampleforestMED,examplesoil,SpParamsMED, control)
#> $control
#> $control$fillMissingRootParams
#> [1] TRUE
#> 
#> $control$fillMissingSpParams
#> [1] TRUE
#> 
#> $control$verbose
#> [1] TRUE
#> 
#> $control$subdailyResults
#> [1] FALSE
#> 
#> $control$standResults
#> [1] TRUE
#> 
#> $control$soilResults
#> [1] TRUE
#> 
#> $control$plantResults
#> [1] TRUE
#> 
#> $control$leafResults
#> [1] TRUE
#> 
#> $control$temperatureResults
#> [1] TRUE
#> 
#> $control$fireHazardResults
#> [1] FALSE
#> 
#> $control$transpirationMode
#> [1] "Cochard"
#> 
#> $control$soilFunctions
#> [1] "VG"
#> 
#> $control$defaultWindSpeed
#> [1] 2.5
#> 
#> $control$defaultCO2
#> [1] 386
#> 
#> $control$snowpack
#> [1] TRUE
#> 
#> $control$leafPhenology
#> [1] TRUE
#> 
#> $control$rockyLayerDrainage
#> [1] TRUE
#> 
#> $control$unlimitedSoilWater
#> [1] FALSE
#> 
#> $control$rhizosphereOverlap
#> [1] "total"
#> 
#> $control$unfoldingDD
#> [1] 300
#> 
#> $control$verticalLayerSize
#> [1] 100
#> 
#> $control$windMeasurementHeight
#> [1] 200
#> 
#> $control$cavitationRefill
#> [1] "total"
#> 
#> $control$hydraulicRedistributionFraction
#> [1] 0.1
#> 
#> $control$ndailysteps
#> [1] 24
#> 
#> $control$nsubsteps
#> [1] 3600
#> 
#> $control$capacitance
#> [1] FALSE
#> 
#> $control$taper
#> [1] TRUE
#> 
#> $control$multiLayerBalance
#> [1] FALSE
#> 
#> $control$maximumStemConductance
#> [1] 10
#> 
#> $control$klatstem
#> [1] 0.01
#> 
#> $control$klatleaf
#> [1] 0.01
#> 
#> $control$numericParams
#> $control$numericParams$maxNsteps
#> [1] 400
#> 
#> $control$numericParams$ntrial
#> [1] 200
#> 
#> $control$numericParams$psiTol
#> [1] 1e-04
#> 
#> $control$numericParams$ETol
#> [1] 1e-07
#> 
#> 
#> $control$fracLeafResistance
#> [1] NA
#> 
#> $control$fracRootResistance
#> [1] 0.4
#> 
#> $control$averageFracRhizosphereResistance
#> [1] 0.15
#> 
#> $control$thermalCapacityLAI
#> [1] 1e+06
#> 
#> $control$boundaryLayerSize
#> [1] 2000
#> 
#> $control$refillMaximumRate
#> [1] 0.05
#> 
#> $control$subdailyCarbonBalance
#> [1] FALSE
#> 
#> $control$allowDessication
#> [1] TRUE
#> 
#> $control$allowStarvation
#> [1] TRUE
#> 
#> $control$sinkLimitation
#> [1] TRUE
#> 
#> $control$shrubDynamics
#> [1] TRUE
#> 
#> $control$allocationStrategy
#> [1] "Al2As"
#> 
#> $control$phloemConductanceFactor
#> [1] 0.2
#> 
#> $control$nonSugarConcentration
#> [1] 0.25
#> 
#> $control$equilibriumOsmoticConcentration
#> $control$equilibriumOsmoticConcentration$leaf
#> [1] 0.8
#> 
#> $control$equilibriumOsmoticConcentration$sapwood
#> [1] 0.6
#> 
#> 
#> $control$minimumRelativeStarchForGrowth
#> [1] 0.5
#> 
#> $control$constructionCosts
#> $control$constructionCosts$leaf
#> [1] 1.5
#> 
#> $control$constructionCosts$sapwood
#> [1] 1.47
#> 
#> $control$constructionCosts$fineroot
#> [1] 1.3
#> 
#> 
#> $control$senescenceRates
#> $control$senescenceRates$sapwood
#> [1] 0.000135
#> 
#> $control$senescenceRates$fineroot
#> [1] 0.001897231
#> 
#> 
#> $control$maximumRelativeGrowthRates
#> $control$maximumRelativeGrowthRates$leaf
#> [1] 0.03
#> 
#> $control$maximumRelativeGrowthRates$cambium
#> [1] 0.0025
#> 
#> $control$maximumRelativeGrowthRates$sapwood
#> [1] 0.002
#> 
#> $control$maximumRelativeGrowthRates$fineroot
#> [1] 0.1
#> 
#> 
#> $control$mortalityMode
#> [1] "density/deterministic"
#> 
#> $control$mortalityBaselineRate
#> [1] 0.0015
#> 
#> $control$mortalityRelativeSugarThreshold
#> [1] 0.3
#> 
#> $control$mortalityRWCThreshold
#> [1] 0.3
#> 
#> $control$recrTreeDBH
#> [1] 1
#> 
#> $control$recrTreeDensity
#> [1] 1000
#> 
#> $control$ingrowthTreeDBH
#> [1] 7.5
#> 
#> $control$ingrowthTreeDensity
#> [1] 127
#> 
#> $control$allowRecruitment
#> [1] TRUE
#> 
#> $control$recruitmentMode
#> [1] "stochastic"
#> 
#> $control$allowResprouting
#> [1] TRUE
#> 
#> $control$removeEmptyCohorts
#> [1] TRUE
#> 
#> $control$minimumCohortDensity
#> [1] 1
#> 
#> $control$seedRain
#> NULL
#> 
#> $control$seedProductionTreeHeight
#> [1] 300
#> 
#> $control$seedProductionShrubHeight
#> [1] 30
#> 
#> $control$probRecr
#> [1] 0.05
#> 
#> $control$minTempRecr
#> [1] 0
#> 
#> $control$minMoistureRecr
#> [1] 0.3
#> 
#> $control$minFPARRecr
#> [1] 10
#> 
#> $control$recrTreeHeight
#> [1] 620
#> 
#> $control$recrShrubCover
#> [1] 1
#> 
#> $control$recrShrubHeight
#> [1] 25
#> 
#> $control$recrTreeZ50
#> [1] 100
#> 
#> $control$recrShrubZ50
#> [1] 50
#> 
#> $control$recrTreeZ95
#> [1] 1000
#> 
#> $control$recrShrubZ95
#> [1] 500
#> 
#> 
#> $soil
#> Soil depth (mm): 4000 
#> 
#> Layer  1  [ 0  to  300 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 25 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 32 Vol. FC (mm): 68 Vol. SAT (mm): 111 Vol. current (mm): 68 
#>     Temperature (Celsius): NA 
#> 
#> Layer  2  [ 300  to  1000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 45 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 55 Vol. FC (mm): 117 Vol. SAT (mm): 190 Vol. current (mm): 117 
#>     Temperature (Celsius): NA 
#> 
#> Layer  3  [ 1000  to  2000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 75 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 36 Vol. FC (mm): 76 Vol. SAT (mm): 123 Vol. current (mm): 76 
#>     Temperature (Celsius): NA 
#> 
#> Layer  4  [ 2000  to  4000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 95 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 14 Vol. FC (mm): 30 Vol. SAT (mm): 49 Vol. current (mm): 30 
#>     Temperature (Celsius): NA 
#> 
#> Total soil saturated capacity (mm): 473 
#> Total soil water holding capacity (mm): 291 
#> Total soil extractable water (mm): 183 
#> Total soil current Volume (mm): 291 
#> 
#> Snow pack water equivalent (mm): 0 
#> Soil water table depth (mm): 4000 
#> 
#> 
#> $canopy
#>    zlow zmid  zup Tair Cair VPair
#> 1     0   50  100   NA   NA    NA
#> 2   100  150  200   NA   NA    NA
#> 3   200  250  300   NA   NA    NA
#> 4   300  350  400   NA   NA    NA
#> 5   400  450  500   NA   NA    NA
#> 6   500  550  600   NA   NA    NA
#> 7   600  650  700   NA   NA    NA
#> 8   700  750  800   NA   NA    NA
#> 9   800  850  900   NA   NA    NA
#> 10  900  950 1000   NA   NA    NA
#> 11 1000 1050 1100   NA   NA    NA
#> 12 1100 1150 1200   NA   NA    NA
#> 13 1200 1250 1300   NA   NA    NA
#> 14 1300 1350 1400   NA   NA    NA
#> 15 1400 1450 1500   NA   NA    NA
#> 16 1500 1550 1600   NA   NA    NA
#> 17 1600 1650 1700   NA   NA    NA
#> 18 1700 1750 1800   NA   NA    NA
#> 19 1800 1850 1900   NA   NA    NA
#> 20 1900 1950 2000   NA   NA    NA
#> 21 2000 2050 2100   NA   NA    NA
#> 22 2100 2150 2200   NA   NA    NA
#> 23 2200 2250 2300   NA   NA    NA
#> 24 2300 2350 2400   NA   NA    NA
#> 25 2400 2450 2500   NA   NA    NA
#> 26 2500 2550 2600   NA   NA    NA
#> 27 2600 2650 2700   NA   NA    NA
#> 28 2700 2750 2800   NA   NA    NA
#> 
#> $herbLAI
#> [1] 0.252
#> 
#> $cohorts
#>         SP              Name
#> T1_148 148  Pinus halepensis
#> T2_168 168      Quercus ilex
#> S1_165 165 Quercus coccifera
#> 
#> $above
#>          H        CR        N   LAI_live LAI_expanded LAI_dead
#> T1_148 800 0.6605196 168.0000 0.84874773   0.84874773        0
#> T2_168 660 0.6055642 384.0000 0.70557382   0.70557382        0
#> S1_165  80 0.8032817 749.4923 0.02718804   0.02718804        0
#> 
#> $below
#>        Z50  Z95 fineRootBiomass coarseRootSoilVolume
#> T1_148 100  600       1612.7328                16.75
#> T2_168 300 1000       1590.6433                13.25
#> S1_165 200 1000        152.1546                 0.25
#> 
#> $belowLayers
#> $belowLayers$V
#>                1         2          3           4
#> T1_148 0.8604899 0.1194556 0.01511005 0.004944476
#> T2_168 0.5008953 0.4505941 0.04064831 0.007862284
#> S1_165 0.6799879 0.2737911 0.03567632 0.010544678
#> 
#> $belowLayers$L
#>                1        2        3        4
#> T1_148 4815.5858 2286.191 2567.669 4513.443
#> T2_168 3364.1867 3221.751 2828.103 4575.847
#> S1_165  790.4044 1237.889 2106.565 4091.603
#> 
#> $belowLayers$VGrhizo_kmax
#>                1        2         3          4
#> T1_148  27535774  1321053  167101.2   54680.68
#> T2_168  67483373 22684370 2046368.0  395812.89
#> S1_165 404209828 58944700 7680780.7 2270171.56
#> 
#> $belowLayers$VCroot_kmax
#>                1         2          3           4
#> T1_148 1.1305113 0.3305767 0.03723103 0.006930914
#> T2_168 0.8450863 0.7938304 0.08157959 0.009752397
#> S1_165 3.6420477 0.9363356 0.07169666 0.010910223
#> 
#> $belowLayers$Wpool
#>        1 2 3 4
#> T1_148 1 1 1 1
#> T2_168 1 1 1 1
#> S1_165 1 1 1 1
#> 
#> $belowLayers$RhizoPsi
#>             1      2      3      4
#> T1_148 -0.033 -0.033 -0.033 -0.033
#> T2_168 -0.033 -0.033 -0.033 -0.033
#> S1_165 -0.033 -0.033 -0.033 -0.033
#> 
#> 
#> $paramsPhenology
#>             PhenologyType LeafDuration t0gdd  Sgdd Tbgdd  Ssen Phsen Tbsen xsen
#> T1_148 oneflush-evergreen     2.536875  50.0 200.0  0.00  8268  12.5  28.5    2
#> T2_168 oneflush-evergreen     2.183837  54.5 240.7  4.34 10178  12.5  28.5    2
#> S1_165 oneflush-evergreen     1.250000  54.5 240.7  4.34 10178  12.5  28.5    2
#>        ysen
#> T1_148    2
#> T2_168    2
#> S1_165    2
#> 
#> $paramsAnatomy
#>        Hmed    Al2As      SLA LeafWidth LeafDensity WoodDensity FineRootDensity
#> T1_148  850 1317.523 5.140523 0.1384772   0.2982842   0.6077016       0.2982842
#> T2_168  500 3908.823 6.340000 1.7674359   0.4893392   0.9008264       0.4893392
#> S1_165   80 4189.325 4.980084 1.3761085   0.3709679   0.4389106       0.3709679
#>        conduit2sapwood      SRL RLD     r635
#> T1_148       0.9236406 3172.572  10 1.964226
#> T2_168       0.6238125 4398.812  10 1.805872
#> S1_165       0.6238125 4398.812  10 2.289452
#> 
#> $paramsInterception
#>        alphaSWR gammaSWR kPAR   g
#> T1_148      0.7     0.14 0.50 1.0
#> T2_168      0.7     0.18 0.55 0.5
#> S1_165      0.7     0.18 0.55 0.5
#> 
#> $paramsTranspiration
#>             Gswmin    Gswmax  Vmax298  Jmax298 Kmax_stemxylem Kmax_rootxylem
#> T1_148 0.003086667 0.2850000 72.19617 124.1687           0.15           0.60
#> T2_168 0.004473333 0.2007222 68.51600 118.7863           0.40           1.60
#> S1_165 0.010455247 0.2830167 62.78100 118.4486           0.29           1.16
#>        VCleaf_kmax VCleaf_slope VCleaf_P50  VCleaf_c  VCleaf_d VCstem_kmax
#> T1_148    4.000000           30  -2.303772 11.137051 -2.380849    1.339563
#> T2_168    4.000000           30  -1.964085  1.339370 -2.582279    1.620936
#> S1_165    9.579077           30  -2.483999  1.844224 -3.030130    4.599269
#>        VCstem_slope VCstem_P50  VCstem_c  VCstem_d VCroot_kmax VCroot_slope
#> T1_148           30  -5.139633 12.710000 -5.290000    1.505250           30
#> T2_168           30  -6.964747  3.560000 -7.720000    1.730249           30
#> S1_165           30  -3.720407  3.537784 -4.126512    4.660990           30
#>        VCroot_P50  VCroot_c  VCroot_d VGrhizo_kmax Plant_kmax
#> T1_148  -2.966325 11.137051 -3.065569     29078608  0.6021000
#> T2_168  -1.684034  1.339370 -2.214081     92609924  0.6920994
#> S1_165  -2.271342  1.760307 -2.797092    473105480  1.8643961
#> 
#> $paramsWaterStorage
#>           maxFMC   LeafPI0   LeafEPS LeafAF     Vleaf   StemPI0   StemEPS
#> T1_148 126.03063 -1.591429  8.918571 0.3525 0.5258525 -2.008039 13.256355
#> T2_168  93.15304 -1.483333 19.260000 0.1700 0.2199087 -3.227438 46.420610
#> S1_165  96.53441 -2.370000 17.230000 0.2400 0.4108968 -1.305868  6.297155
#>           StemAF Vsapwood
#> T1_148 0.9236406 7.701237
#> T2_168 0.6238125 1.631003
#> S1_165 0.6238125 1.631676
#> 
#> $internalPhenology
#>        gdd sen budFormation leafUnfolding leafSenescence leafDormancy phi
#> T1_148   0   0        FALSE         FALSE          FALSE        FALSE   0
#> T2_168   0   0        FALSE         FALSE          FALSE        FALSE   0
#> S1_165   0   0        FALSE         FALSE          FALSE        FALSE   0
#> 
#> $internalWater
#>        Einst Elim Emin_L Emin_S RootCrownPsi StemPsi LeafPsi StemSympPsi
#> T1_148     0    0      0      0       -0.033  -0.033  -0.033      -0.033
#> T2_168     0    0      0      0       -0.033  -0.033  -0.033      -0.033
#> S1_165     0    0      0      0       -0.033  -0.033  -0.033      -0.033
#>        LeafPLC StemPLC
#> T1_148       0       0
#> T2_168       0       0
#> S1_165       0       0
#> 
#> $internalFCCS
#> data frame with 0 columns and 0 rows
#> 
#> attr(,"class")
#> [1] "spwbInput" "list"