Photosynthesis submodel functions

Set of functions used in the calculation of photosynthesis.

photo_GammaTemp(leaf_temp)
photo_KmTemp(leaf_temp, Oi = 209)
photo_VmaxTemp(Vmax298, leaf_temp)
photo_JmaxTemp(Jmax298, leaf_temp)
photo_electronLimitedPhotosynthesis(Q, Ci, GT, Jmax)
photo_rubiscoLimitedPhotosynthesis(Ci, GT, Km, Vmax)
photo_photosynthesis(Q, Catm, Gc, leaf_temp, Vmax298, Jmax298, verbose = FALSE)
photo_leafPhotosynthesisFunction(E, Catm, Patm, Tair, vpa, u,
                             absRad, Q, Vmax298, Jmax298, Gwmin, Gwmax,
                             leafWidth = 1.0, refLeafArea = 1, verbose = FALSE)
photo_sunshadePhotosynthesisFunction(E, Catm, Patm, Tair, vpa,
                                     SLarea, SHarea, u,
                                     absRadSL, absRadSH, QSL, QSH,
                                     Vmax298SL, Vmax298SH, Jmax298SL, Jmax298SH,
                                     Gwmin, Gwmax, leafWidth = 1.0, verbose = FALSE)
photo_multilayerPhotosynthesisFunction(E, Catm, Patm, Tair, vpa,
                                      SLarea, SHarea, u,
                                      absRadSL, absRadSH, QSL, QSH,
                                      Vmax298, Jmax298, Gwmin, Gwmax, leafWidth = 1.0,
                                      verbose = FALSE)

Arguments

leaf_temp
Leaf temperature (in ºC).

Oi
Oxigen concentration (mmol*mol-1).

Vmax298, Vmax298SL, Vmax298SH
Maximum Rubisco carboxylation rate per leaf area at 298ºK (i.e. 25 ºC) (micromol*s-1*m-2) (for each canopy layer in the case of photo_multilayerPhotosynthesisFunction). 'SH' stands for shade leaves, whereas 'SL' stands for sunlit leaves.

Jmax298, Jmax298SL, Jmax298SH
Maximum electron transport rate per leaf area at 298ºK (i.e. 25 ºC) (micromol*s-1*m-2) (for each canopy layer in the case of photo_multilayerPhotosynthesisFunction). 'SH' stands for shade leaves, whereas 'SL' stands for sunlit leaves.

Q
Active photon flux density (micromol * s-1 * m-2).

Ci
CO2 internal concentration (micromol * mol-1).

GT
CO2 saturation point corrected by temperature (micromol * mol-1).

Jmax
Maximum electron transport rate per leaf area (micromol*s-1*m-2).

Km
Km = Kc*(1.0+(Oi/Ko)) - Michaelis-Menten term corrected by temperature (in micromol * mol-1).

Vmax
Maximum Rubisco carboxylation rate per leaf area (micromol*s-1*m-2).

Catm
CO2 air concentration (micromol * mol-1).

Gc
CO2 leaf (stomatal) conductance (mol * s-1 * m-2).

E
Transpiration flow rate per leaf area (mmol*s-1*m-2).

Patm
Atmospheric air pressure (in kPa).

Tair
Air temperature (in ºC).

vpa
Vapour pressure deficit (in kPa).

u
Wind speed above the leaf boundary (in m/s) (for each canopy layer in the case of photo_multilayerPhotosynthesisFunction).

absRad
Absorbed long- and short-wave radiation (in W*m^-2).

leafWidth
Leaf width (in cm).

refLeafArea
Leaf reference area.

Gwmin, Gwmax
Minimum and maximum leaf water conductance (i.e. cuticular and maximum stomatal conductance) (mol * s-1 * m-2).

verbose
Boolean flag to indicate console output.

SLarea, SHarea
Leaf area index of sunlit/shade leaves (for each canopy layer in the case of photo_multilayerPhotosynthesisFunction).

absRadSL, absRadSH
Instantaneous absorbed radiation (W·m-2) per unit of sunlit/shade leaf area (for each canopy layer in the case of photo_multilayerPhotosynthesisFunction).

QSL, QSH
Active photon flux density (micromol * s-1 * m-2) per unit of sunlit/shade leaf area (for each canopy layer in the case of photo_multilayerPhotosynthesisFunction).

Details

Details of the photosynthesis submodel are given in a vignette.

Value

Values returned for each function are:

photo_GammaTemp: CO2 compensation concentration (micromol * mol-1).

photo_KmTemp: Michaelis-Menten coefficients of Rubisco for Carbon (micromol * mol-1) and Oxigen (mmol * mol-1).

photo_VmaxTemp: Temperature correction of Vmax298.

photo_JmaxTemp: Temperature correction of Jmax298.

photo_electronLimitedPhotosynthesis: Electron-limited photosynthesis (micromol*s-1*m-2) following Farquhar et al. (1980).

photo_rubiscoLimitedPhotosynthesis: Rubisco-limited photosynthesis (micromol*s-1*m-2) following Farquhar et al. (1980).

photo_photosynthesis: Calculates gross photosynthesis (micromol*s-1*m-2) following (Farquhar et al. (1980) and Collatz et al (1991).

photo_leafPhotosynthesisFunction: Returns a data frame with the following columns:

LeafTemperature: Leaf temperature (ºC).

LeafVPD: Leaf vapor pressure deficit (kPa).

WaterVaporConductance: Leaf vapor conductance (mol * s-1 * m-2).

Photosynthesis: Gross photosynthesis (micromol*s-1*m-2).

NetPhotosynthesis: Net photosynthesis, after discounting autotrophic respiration (micromol*s-1*m-2).

photo_sunshadePhotosynthesisFunction and photo_multilayerPhotosynthesisFunction: Return a data frame with the following columns:

Photosynthesis: Gross photosynthesis (micromol*s-1*m-2).

NetPhotosynthesis: Net photosynthesis, after discounting autotrophic respiration (micromol*s-1*m-2).

References

Bernacchi, C. J., E. L. Singsaas, C. Pimentel, A. R. Portis, and S. P. Long. 2001. Improved temperature response functions for models of Rubisco-limited photosynthesis. Plant, Cell and Environment 24:253–259.

Collatz, G. J., J. T. Ball, C. Grivet, and J. A. Berry. 1991. Physiological and environmental regulation of stomatal conductance, photosynthesis and transpiration: a model that includes a laminar boundary layer. Agricultural and Forest Meteorology 54:107–136.

Farquhar, G. D., S. von Caemmerer, and J. A. Berry. 1980. A biochemical model of photosynthetic CO2 assimilation in leaves of C3 species. Planta 149:78–90.

Leuning, R. 2002. Temperature dependence of two parameters in a photosynthesis model. Plant, Cell and Environment 25:1205–1210.

Sperry, J. S., M. D. Venturas, W. R. L. Anderegg, M. Mencuccini, D. S. Mackay, Y. Wang, and D. M. Love. 2016. Predicting stomatal responses to the environment from the optimization of photosynthetic gain and hydraulic cost. Plant Cell and Environment.

See also