Plant description functions

Functions to calculate attributes of plants in a forest object.

plant_basalArea(x)
plant_largerTreeBasalArea(x)
plant_characterParameter(x, SpParams, parName)
plant_cover(x)
plant_crownBaseHeight(x, SpParams)
plant_crownLength(x, SpParams)
plant_crownRatio(x, SpParams)
plant_density(x, SpParams)
plant_equilibriumLeafLitter(x, SpParams, AET = 800)
plant_equilibriumSmallBranchLitter(x, SpParams,
                                   smallBranchDecompositionRate = 0.81)
plant_foliarBiomass(x, SpParams, gdd = NA)
plant_fuel(x, SpParams, gdd = NA, includeDead = TRUE)
plant_height(x)
plant_ID(x)
plant_LAI(x, SpParams, gdd = NA)
plant_parameter(x, SpParams, parName)
plant_phytovolume(x, SpParams)
plant_species(x)
plant_speciesName(x, SpParams)

Arguments

x
An object of class forest.

SpParams
A data frame with species parameters (see SpParamsMED).

parName
A string with a parameter name.

gdd
Growth degree days (to account for leaf phenology effects).

AET
Actual annual evapotranspiration (in mm).

smallBranchDecompositionRate
Decomposition rate of small branches.

includeDead
A flag to indicate that standing dead fuels (dead branches) are included.

Value

A vector with values for each plant of the input forest object:

plant_basalArea: Tree basal area (m2/ha).

plant_largerTreeBasalArea: Basal area (m2/ha) of trees larger (in diameter) than the tree. Half of the trees of the same record are included.

plant_characterParameter: The parameter values of each plant, as strings.

plant_cover: Shrub cover (in percent).

plant_crownBaseHeight: The height corresponding to the start of the crown (in cm).

plant_crownLength: The difference between crown base height and total height (in cm).

plant_crownRatio: The ratio between crown length and total height (between 0 and 1).

plant_density: Plant density (ind/ha). Tree density is directly taken from the forest object, while the shrub density is estimated from cover and height by calculating the area of a single individual.

plant_equilibriumLeafLitter: Litter biomass of leaves at equilibrium (in kg/m2).

plant_equilibriumSmallBranchLitter: Litter biomass of small branches (< 6.35 mm diameter) at equilibrium (in kg/m2).

plant_foliarBiomass: Standing biomass of leaves (in kg/m2).

plant_fuel: Fine fuel load (in kg/m2).

plant_height: Total height (in cm).

plant_ID: Cohort coding for simulation functions (concatenation of 'T' (Trees) or 'S' (Shrub), cohort index and species index).

plant_LAI: Leaf area index (m2/m2).

plant_parameter: The parameter values of each plant, as numeric.

plant_phytovolume: Shrub phytovolume (m3/m2).

plant_species: Species identity integer (indices start with 0).

plant_speciesName: String with species taxonomic name (or a functional group).

See also

Examples

#Default species parameterization
data(SpParamsMED)

#Load example plot
data(exampleforest)

#A short way to obtain total basal area
sum(plant_basalArea(exampleforest), na.rm=TRUE)

#> [1] 25.0333

#The same forest level function for LAI
sum(plant_LAI(exampleforest, SpParamsMED))

#> [1] 1.70363

#The same forest level function for fuel loading
sum(plant_fuel(exampleforest, SpParamsMED))

#> [1] 0.7033713

#Summary function for 'forest' objects can be also used
summary(exampleforest, SpParamsMED)

#> Forest summary:
#> 
#>   Tree density (ind/ha): 552 
#>   BA (m2/ha): 25.0333016 
#>   Shrub crown phytovolume (m3/m2): 0.0109563 
#>   LAI (m2/m2) total: 1.7036297  trees: 1.6144964  shrubs: 0.0891333 
#>   Live fine fuel (kg/m2) total: 0.7033713  trees: 0.6642999  shrubs: 0.0390714 

plant_speciesName(exampleforest, SpParamsMED)

#>               T1_54               T2_68               S1_65 
#>  "Pinus halepensis"      "Quercus ilex" "Quercus coccifera" 

plant_ID(exampleforest)

#> [1] "T1_54" "T2_68" "S1_65"