Example forest inventory data

Example data to illustrate the creation of forest objects from inventory data, coming from a forest inventory survey, used to illustrate the general function forest_mapTreeTable:

poblet_trees - Data frame with example tree plot data from Poblet, Catalonia (717 observations and 4 variables).

Plot.Code - Plot ID (character)

Indv.Ref - Tree individual (integer)

Species - Species name (character)

Diameter.cm - Tree diameter at breast height (cm)

Source

Data table poblet_trees corresponds to field data sampled by the Catalan Forest Ownership Center (Centre de la Propietat Forestal; CPF).

See also