Plant recruitment

Annual plant recruitment observed in a forest stand

recruitment(forest, SpParams, control, minMonthTemp, moistureIndex)

Arguments

forest

An object of class forest.

SpParams

A data frame with species parameters (see SpParamsMED and SpParamsDefinition).

control

A list with default control parameters (see defaultControl).

minMonthTemp

Minimum month temperature.

moistureIndex

Moisture index (annual precipitation over annual potential evapotranspiration).

Details

Species can recruit if adults (sufficiently tall individuals) are present (seed rain can also be specified in a control parameter). Minimum month temperature and moisture index values are used to determine if recruitment was successful. Species also require a minimum amount of light at the ground level.

Value

An object of class forest with the new plant cohorts.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize control parameters
control = defaultControl("Granier")

#Recruitment limits
plant_parameter(exampleforestMED, SpParamsMED, "MinTempRecr")
#>    T1_148    T2_168    S1_165 
#>  0.571485 -3.386079  1.669536 
plant_parameter(exampleforestMED, SpParamsMED, "MinMoistureRecr")
#>    T1_148    T2_168    S1_165 
#> 0.1468244 0.1580670 0.2230189 

#Compare recruitment outcomes
recruitment(exampleforestMED, SpParamsMED, control, 0, 0.25)
#> $ID
#> [1] ""
#> 
#> $patchsize
#> [1] 10000
#> 
#> $treeData
#> [1] Species DBH     Height  N       Z50     Z95    
#> <0 rows> (or 0-length row.names)
#> 
#> $shrubData
#> [1] Species Height  Cover   Z50     Z95    
#> <0 rows> (or 0-length row.names)
#> 
#> $herbCover
#> [1] 0
#> 
#> $herbHeight
#> [1] 0
#> 
#> attr(,"class")
#> [1] "forest" "list"  
recruitment(exampleforestMED, SpParamsMED, control, 3, 0.25)
#> $ID
#> [1] ""
#> 
#> $patchsize
#> [1] 10000
#> 
#> $treeData
#>   Species DBH   Height   N Z50  Z95
#> 1     148 7.5 569.2412 231 100 1000
#> 
#> $shrubData
#> [1] Species Height  Cover   Z50     Z95    
#> <0 rows> (or 0-length row.names)
#> 
#> $herbCover
#> [1] 0
#> 
#> $herbHeight
#> [1] 0
#> 
#> attr(,"class")
#> [1] "forest" "list"