Distribution of fine roots

Functions to calculate the distribution of fine roots within the soil, given root system parameters and soil layer definition (layer widths).

root_conicDistribution(Zcone, d)
root_ldrDistribution(Z50, Z95, d)
root_xylemConductanceProportions(v, d, depthWidthRatio = 1)
root_rootLengths(v, d, depthWidthRatio = 1.0)

Arguments

Z50
A vector of depths (in mm) corresponding to 50% of roots.

Z95
A vector of depths (in mm) corresponding to 95% of roots.

Zcone
A vector of depths (in mm) corresponding to the root cone tip.

d
The width (in mm) corresponding to each soil layer.

v
Proportions of fine roots, as returned by functions root_conicDistribution or root_ldrDistribution.

depthWidthRatio
Ratio between radius of the soil layer with the largest radius and maximum rooting depth.

Details

Function root_conicDistribution assumes a conic distribution of fine roots, whereas function root_ldrDistribution distributes fine roots according to the linear dose response model of Schenck & Jackson (2002). Function root_xylemConductanceProportions calculates the proportion of total root xylem conductance that can be attributed to each layer, according to layer widths and the proportion of fine roots (Sperry et al. 2016).

Value

Functions root_conicDistribution and root_ldrDistribution return a matrix with as many rows as elements in Z (or Z50) and as many columns as soil layers. Values in all cases correspond to the proportion of fine roots in each soil layer. Function root_xylemConductanceProportions returns a vector of proportions of the same length as the number of layers.

References

Schenk, H., Jackson, R., 2002. The global biogeography of roots. Ecol. Monogr. 72, 311–328.

Sperry, J. S., Y. Wang, B. T. Wolfe, D. S. Mackay, W. R. L. Anderegg, N. G. Mcdowell, and W. T. Pockman. 2016. Pragmatic hydraulic theory predicts stomatal responses to climatic water deficits. New Phytologist 212, 577–589.

See also

Examples

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

ntree = nrow(exampleforest$treeData)

#Initialize soil with default soil params
S = soil(defaultSoilParams())

#Calculate conic root system for trees
V1 = root_conicDistribution(Z=rep(2000,ntree), S$dVec)
print(V1)

#>          [,1]     [,2]  [,3] [,4]
#> [1,] 0.385875 0.489125 0.125    0
#> [2,] 0.385875 0.489125 0.125    0

#Calculate LDR root system for trees (Schenck & Jackson 2002)
V2 = root_ldrDistribution(Z50 = rep(200,ntree),
          Z95 = rep(1000,ntree), S$dVec)
print(V2)

#>           [,1]      [,2]       [,3]       [,4]
#> [1,] 0.6799879 0.2737911 0.03567632 0.01054468
#> [2,] 0.6799879 0.2737911 0.03567632 0.01054468

#Equal xylem conductance proportions for a cone distribution
#(assuming depth-width ratio 1)
root_xylemConductanceProportions(V1[1,], S$dVec)

#> [1] 0.3306134 0.3346441 0.3347425 0.0000000

#Xylem conductance proportions for LDR distribution
root_xylemConductanceProportions(V2[1,], S$dVec)

#> [1] 0.1621237 0.2910575 0.3360984 0.2107205