Soil initialization

Initializes soil parameters and state variables for its use in simulations.

soil(SoilParams, VG_PTF = "Toth", W = as.numeric(c(1)), SWE = 0)
# S3 method for soil
print(x, model="SX", ...)

Arguments

SoilParams
A list of soil parameters (see defaultSoilParams).

VG_PTF
Pedotransfer functions to obtain parameters for the van Genuchten-Mualem equations. Either "Carsel" (Carsel & Parrish 1988) or "Toth" (Toth et al. 2015).

W
A numerical vector with the initial relative water content of each soil layer.

SWE
Initial snow water equivalent of the snow pack on the soil surface (mm).

x
An object of class soil.

model
Either 'SX' or 'VG' for Saxton or Van Genuchten pedotransfer models.

...
Additional parameters to print.

Details

Function print prompts a description of soil characteristics and state variables (water content and temperature) according to a water retention curve (either Saxton's or Van Genuchten's). Volume at field capacity is calculated assuming a soil water potential equal to -0.033 MPa. Parameter Temp is initialized as missing for all soil layers.

Value

An list of class soil with the following elements:

SoilDepth: Soil depth (in mm).

W: State variable with relative water content of each layer (in as proportion relative to FC).

Temp: State variable with temperature (in ºC) of each layer.

Ksoil: Kappa parameter for infiltration.

Gsoil: Gamma parameter for infiltration.

dVec: Width of soil layers (in mm).

sand: Sand percentage for each layer (in percent volume).

clay: Clay percentage for each layer (in percent volume).

om: Organic matter percentage for each layer (in percent volume).

usda_Type: USDA texture type.

VG_alpha, VG_n, VG_theta_res, VG_theta_sat: Parameters for van Genuchten's pedotransfer functions, for each layer, corresponding to the USDA texture type.

macro: Macroporosity for each layer (estimated using Stolf et al. 2011).

rfc: Percentage of rock fragment content for each layer.

References

Carsel, R.F., & Parrish, R.S. 1988. Developing joint probability distributions of soil water retention characteristics. Water Resources Research 24: 755–769.

Tóth, B., Weynants, M., Nemes, A., Makó, A., Bilas, G., & Tóth, G. 2015. New generation of hydraulic pedotransfer functions for Europe. European Journal of Soil Science 66: 226–238.

Stolf, R., Thurler, A., Oliveira, O., Bacchi, S., Reichardt, K., 2011. Method to estimate soil macroporosity and microporosity based on sand content and bulk density. Rev. Bras. Ciencias do Solo 35, 447–459.

See also

Examples

# Initializes soil
s = soil(defaultSoilParams())

# Prints soil characteristics according to Saxton's water retention curve
print(s, model="SX")

#> Soil depth (mm): 4000 
#> 
#> Layer  1  [ 0  to  300 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 20 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 34 Vol. FC (mm): 73 Vol. SAT (mm): 118 Vol. current (mm): 73 
#>     Temperature (Celsius): NA 
#> 
#> Layer  2  [ 300  to  1000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 40 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 60 Vol. FC (mm): 127 Vol. SAT (mm): 207 Vol. current (mm): 127 
#>     Temperature (Celsius): NA 
#> 
#> Layer  3  [ 1000  to  2000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 70 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 43 Vol. FC (mm): 91 Vol. SAT (mm): 148 Vol. current (mm): 91 
#>     Temperature (Celsius): NA 
#> 
#> Layer  4  [ 2000  to  4000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 90 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 29 Vol. FC (mm): 61 Vol. SAT (mm): 98 Vol. current (mm): 61 
#>     Temperature (Celsius): NA 
#> 
#> Total soil saturated capacity (mm): 571 
#> Total soil water holding capacity (mm): 352 
#> Total soil extractable water (mm): 221 
#> Total soil current Volume (mm): 352 
#> Water table depth (mm): 4000 
#> 
#> Snow pack water equivalent (mm): 0 

# Prints soil characteristics according to Van Genuchten's water retention curve
print(s, model="VG")

#> Soil depth (mm): 4000 
#> 
#> Layer  1  [ 0  to  300 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 20 Macroporosity (%): 5 
#>     Theta WP (%): 13 Theta FC (%): 29 Theta SAT (%): 42 Theta current (%) 29 
#>     Vol. WP (mm): 32 Vol. FC (mm): 68 Vol. SAT (mm): 102 Vol. current (mm): 68 
#>     Temperature (Celsius): NA 
#> 
#> Layer  2  [ 300  to  1000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 40 Macroporosity (%): 5 
#>     Theta WP (%): 13 Theta FC (%): 30 Theta SAT (%): 42 Theta current (%) 30 
#>     Vol. WP (mm): 54 Vol. FC (mm): 127 Vol. SAT (mm): 178 Vol. current (mm): 127 
#>     Temperature (Celsius): NA 
#> 
#> Layer  3  [ 1000  to  2000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 70 Macroporosity (%): 5 
#>     Theta WP (%): 13 Theta FC (%): 30 Theta SAT (%): 42 Theta current (%) 30 
#>     Vol. WP (mm): 38 Vol. FC (mm): 91 Vol. SAT (mm): 127 Vol. current (mm): 91 
#>     Temperature (Celsius): NA 
#> 
#> Layer  4  [ 2000  to  4000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 90 Macroporosity (%): 5 
#>     Theta WP (%): 13 Theta FC (%): 30 Theta SAT (%): 42 Theta current (%) 30 
#>     Vol. WP (mm): 25 Vol. FC (mm): 61 Vol. SAT (mm): 85 Vol. current (mm): 61 
#>     Temperature (Celsius): NA 
#> 
#> Total soil saturated capacity (mm): 492 
#> Total soil water holding capacity (mm): 347 
#> Total soil extractable water (mm): 229 
#> Total soil current Volume (mm): 347 
#> Water table depth (mm): 4000 
#> 
#> Snow pack water equivalent (mm): 0