Species description functions

Functions to calculate attributes of a forest object by species or to extract species parameters from a species parameter table (SpParamsMED).

species_basalArea(x, SpParams)
species_cover(x, SpParams)
species_density(x, SpParams, mode = "MED")
species_foliarBiomass(x, SpParams, gdd = NA, mode = "MED")
species_fuel(x, SpParams, gdd = NA, includeDead = TRUE, mode = "MED")
species_LAI(x, SpParams, gdd = NA, mode = "MED")
species_phytovolume(x, SpParams)
species_parameter(SP, SpParams, parName, fillMissing = TRUE)
species_characterParameter(SP, SpParams, parName)

Arguments

x

An object of class forest.

SpParams

A data frame with species parameters (see SpParamsMED).

gdd

Growth degree days (to account for leaf phenology effects).

includeDead

A flag to indicate that standing dead fuels (dead branches) are included.

mode

Calculation mode, either "MED" or "US".

SP

An integer vector of species codes.

parName

A string with a parameter name.

fillMissing

A boolean flag to try imputation on missing values.

Value

A vector with values for each species in SpParams:

species_basalArea: Species basal area (m2/ha).

species_cover: Shrub cover (in percent).

species_density: Plant density (ind/ha). Tree density is directly taken from the forest object, while the shrub density is estimated from cover and height by calculating the area of a single individual.

species_foliarBiomass: Standing biomass of leaves (in kg/m2).

species_fuel: Fine fuel load (in kg/m2).

species_LAI: Leaf area index (m2/m2).

species_phytovolume: Shrub phytovolume (m3/m2).

species_parameter: A numeric vector with the parameter values of each input species.

species_characterParameter: A character vector with the parameter values of each input species.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

# Default species parameterization
data(SpParamsMED)

# Load example plot
data(exampleforestMED)

# Species basal area in the forest plot
species_basalArea(exampleforestMED, SpParamsMED)
#>  Pinus halepensis      Quercus ilex Quercus coccifera 
#>         18.604547          6.428755                NA 

# Value of parameter `Psi_Extract` for species 157 (Pinus halepensis)
# and 176 (Quercus ilex)
species_parameter(c(157,176), SpParamsMED, "Psi_Extract")
#> [1] -3.00000 -1.86753