Calibration of root distribution

The function spwb_ldrCalibration calibrates the species root distribution within spwb, given the arguments x, meteo, soil, psi_crit, obs and calibVar. This calibration is based on reference measured values. These reference measured values can be Soil water content, Total tranpiration or Transpiration by cohort. Return the calibrated root distribution for each tree species (no shrub calibration is done), expressed as parameters of the function root_ldrDistribution.

spwb_ldrCalibration(x, soil, meteo, calibVar, obs,
                   RZmin = 301, RZmax = 4000,
                   V1min = 0.01, V1max = 0.94, resolution = 20, heat_stop = 0,
                   transformation = "identity", verbose = FALSE)

Arguments

x
An object of class spwbInput.

soil
A list containing the description of the soil (see soil).

meteo
A data frame with daily meteorological data series. When using the 'Granier' transpiration mode the following columns are required:

DOY: Day of the year (Julian day).

Precipitation: Precipitation (in mm).

MeanTemperature: Mean temperature (in degrees Celsius).

PET: Potential evapotranspiration (in mm).

calibVar
A character string indicating the calibration variable to be used. It can be one of the following: SWC, Eplanttot or Cohorts.

obs
Measured calibration variable. Depending on the value of calibVar it can be a numeric vector with the measured SWC values (if calibVar = "SWC"), or a data frame with the first column containing the measured total transpiration (named Eplanttot) and the following columns containing the cohorts transpiration.

RZmin
The minimum value of RZ (the rooting depth) to be explored (in mm)

RZmax
The maximum value of RZ (the rooting depth) to be explored (in mm)

V1min
The minimum value of V1 (the root proportion in the first soil layer) to be explored

V1max
The maximum value of V1 (the root proportion in the first soil layer) to be explored

resolution
An integer defining the number of values to obtain by discretization of the root parameters RZ and V1. The number of parameter combinations and therefore the computation cost increases increase with the square of resolution

transformation
Function to modify the size of Z intervals to be explored (by default, bins are equal).

heat_stop
An integer defining the number of days during to discard from the calculation of the optimal root distribution. Usefull if the soil water content initialization is not certain

verbose
A logical value. Print the internal messages of the function?

Details

This function performs three different kinds of calibration, selecting those root distribution parameters that minimize the MAE between the predicted values and the measured values provided in obs argument. If calibVar = "SWC" different V1 values are tested running spwb maintaining the total soil depth provided in x and assuming that value is also the depth containing 95 percent of the roots. If calibVar = "Eplanttot" or calibVar = 'Cohorts' different combinations of RZ and V1 values are tested for each tree cohort and the root paramters are selected based on the MAE between the total transpiration or the cohort transpiration.

Value

THe function returns a data frame containing the species index used in medfate, calibrated values for Z50, Z95 and V1 and the MAE value for that combination.

See also