Soil-plant resistances

Calculates and draws rhizosphere, root, stem and leaf resistances for simulation time steps

spwb_resistances(x, cohort = 1, relative = FALSE, draw = FALSE,
                 cumulative = FALSE, xlab = NULL, ylab = NULL)

Arguments

x
An object of class spwb. The function only works with the result of simulations with transpirationMode = "Sperry".

cohort
An integer index indicating the cohort for which resistances are desired (by default the first cohort).

relative
A boolean flag to indicate that relative percentages are desired as output

draw
A boolean flag to indicate that a plot should be drawn.

cumulative
A flag to indicate that drawn series should be cumulative.

xlab
x-axis label.

ylab
y-axis label.

Details

The function makes internal calls to hydraulics_soilPlantResistances.

Value

A data frame with dates in rows and resistance segments in columns (Rhizosphere, Root, Stem and Leaf). Values depend on whether relative = TRUE (percentages) or relative = FALSE (absolute resistance values). If draw = TRUE then a plot object is returned.

See also