Summary of forest structure

Displays a summary of forest structure

Usage

# S3 method for forest
summary(object, SpParams, ...)

# S3 method for summary.forest
print(x, digits = getOption("digits"), ...)

Arguments

object

An object of class forest

SpParams

A data frame with species parameters (see SpParamsMED).

...

Additional parameters for functions summary and print.

x

The object returned by summary.forest.

digits

Minimal number of significant digits.

Value

Function summary.forest returns a list with several structural attributes, such as the basal area and LAI of the forest.

Details

Function summary.forest can be used to summarize a forest object in the console.

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

# Summary of example forest
summary(exampleforest, SpParamsMED)
#> Tree BA (m2/ha): 25.0333016  adult trees: 25.0333016  saplings: 0 
#> Density (ind/ha) adult trees: 552  saplings: 0  shrubs (estimated): 749.4923076 
#> Cover (%) adult trees: 100  saplings: 0  shrubs: 3.75  herbs: 10 
#> LAI (m2/m2) total: 1.7585845  adult trees: 1.5543216  saplings: 0  shrubs: 0.030626  herbs: 0.1736369 
#> Fuel loading (kg/m2) total: 0.5588728  adult trees: 0.5255004  saplings: 0  shrubs: 0.0140795  herbs: 0.019293 
#> PAR ground (%): NA  SWR ground (%): NA