Parameter average values

Internal data set with parameter averages for taxonomic families. This is used by input initialization functions to provide suitable parameter values when missing from species parameter tables.

Format

Data frame trait_family_means has taxonomic families in rows and parameter names as columns.

Source

Same sources as SpParamsMED

See also

Examples

medfate:::trait_family_means
#>                    LeafDensity WoodDensity   LeafPI0   LeafEPS LeafAF    Gswmax
#> Acanthaceae         0.31359059  0.56758803 -3.395000 23.230000 0.1270 0.2500000
#> Achariaceae         0.25524948  0.59247961        NA        NA     NA        NA
#> Acoraceae           0.10000000          NA        NA        NA     NA        NA
#> Actinidiaceae       0.44861072  0.40923199        NA        NA     NA        NA
#> Adoxaceae           0.40799269  0.51384234 -1.560000 12.790000     NA        NA
#> Aextoxicaceae               NA  0.56666667        NA        NA     NA        NA
#> Aizoaceae           0.05526858          NA        NA        NA     NA        NA
#> Akaniaceae                  NA  0.55478250        NA        NA     NA        NA
#> Alismataceae        0.19142078          NA        NA        NA     NA        NA
#> Alstroemeriaceae            NA          NA        NA        NA     NA        NA
#> Altingiaceae        0.68339710  0.60109480        NA        NA     NA 0.5000000
#> Amaranthaceae       0.21187621  0.59891031 -2.250000        NA     NA 0.0750000
#> Amaryllidaceae      0.12367381          NA        NA        NA     NA        NA
#> Amborellaceae               NA          NA        NA        NA     NA        NA
#> Amphorogynaceae             NA  0.60974000        NA        NA     NA        NA
#> Anacampserotaceae   0.02500475          NA        NA        NA     NA        NA
#> Anacardiaceae       0.45624702  0.57106817 -1.700000 12.760000     NA 0.3148333
#> Anisophylleaceae            NA  0.67347799        NA        NA     NA        NA
#> Annonaceae          0.37651502  0.56259675 -2.160000 23.710000     NA        NA
#> Aphloiaceae         0.46270602  0.62052000        NA        NA     NA        NA
#> Apiaceae            0.28055342  0.15098836        NA        NA     NA        NA
#> Apocynaceae         0.29798311  0.56790984 -2.390000 20.940000     NA        NA
#> Aptandraceae        0.33184918  0.70767562        NA        NA     NA        NA
#> Aquifoliaceae       0.48216313  0.55793939 -2.300000 20.700000 0.4000 0.2600000
#> Araceae             0.15917916          NA        NA        NA     NA        NA
#> Araliaceae          0.28116351  0.41507448 -1.528503 11.231462 0.4065        NA
#> Araucariaceae       0.34479018  0.46414559        NA        NA     NA        NA
#> Arecaceae           0.41472385  0.59090347 -3.400000 73.400000 0.2000        NA
#> Aristolochiaceae    0.26814343  0.26000000        NA        NA     NA        NA
#> Asparagaceae        0.21607312  0.42542584        NA        NA     NA        NA
#> Asphodelaceae       0.40325275  0.39519900        NA        NA     NA        NA
#> Aspleniaceae        0.23694373          NA -1.240000 35.300000     NA        NA
#> Asteraceae          0.25025214  0.44342039 -1.471389 14.491429 0.2435 0.1275000
#> Asteropeiaceae              NA  0.75548618        NA        NA     NA        NA
#> Atherospermataceae  0.24513566  0.47676207 -1.340000  8.380000     NA        NA
#> Athyriaceae         0.22706154          NA        NA        NA     NA        NA
#> Aulacomniaceae              NA          NA        NA        NA     NA        NA
#> Austrobaileyaceae   0.26209917          NA        NA        NA     NA        NA
#> Balanopaceae                NA  0.73489760        NA        NA     NA        NA
#> Balsaminaceae       0.40450133  0.09108341        NA        NA     NA        NA
#> Begoniaceae         0.10939929          NA        NA        NA     NA        NA
#> Berberidaceae       0.35994527  0.68526616        NA        NA     NA        NA
#> Betulaceae          0.44493906  0.53788183 -1.201884  5.498667     NA 0.3965000
#> Bignoniaceae        0.34461106  0.62506774 -1.990000 17.610000 0.1770        NA
#> Bixaceae            0.44000000  0.35463573        NA        NA     NA        NA
#> Blechnaceae         0.22361115          NA        NA        NA     NA        NA
#> Bonnetiaceae        0.27978289  0.84000000        NA        NA     NA        NA
#> Boraginaceae        0.28483641  0.45253642 -1.140000        NA     NA        NA
#> Brassicaceae        0.24137149  0.39095103 -1.485000  7.710000 0.2320        NA
#> Bromeliaceae        0.15934340          NA        NA        NA     NA        NA
#> Brunelliaceae               NA  0.31125000        NA        NA     NA        NA
#> Bruniaceae                  NA  0.56365000        NA        NA     NA        NA
#> Burseraceae         0.43321183  0.51975514 -1.435000 14.980000     NA        NA
#> Butomaceae                  NA          NA        NA        NA     NA        NA
#> Buxaceae            0.27062945  0.70292030        NA        NA     NA        NA
#> Cabombaceae                 NA          NA        NA        NA     NA        NA
#> Cactaceae           0.20564854  0.61875000        NA  8.700000     NA        NA
#> Calophyllaceae      0.39682160  0.60678553        NA        NA     NA 0.1350000
#> Calycanthaceae      0.33567851  0.65005500        NA        NA     NA        NA
#> Calyceraceae        0.16893665          NA        NA        NA     NA        NA
#> Campanulaceae       0.23880328  0.67766295        NA        NA     NA        NA
#> Canellaceae                 NA  0.68082545        NA        NA     NA        NA
#>                          Gswmin     Nleaf  Nsapwood Nfineroot     WoodC
#> Acanthaceae                  NA 27.122403        NA  5.580000 0.4242167
#> Achariaceae                  NA 22.162815        NA        NA        NA
#> Acoraceae                    NA 18.000000        NA        NA        NA
#> Actinidiaceae                NA 20.183321        NA        NA        NA
#> Adoxaceae                    NA 20.439735        NA 11.013750        NA
#> Aextoxicaceae                NA  9.621429        NA        NA        NA
#> Aizoaceae                    NA 14.800000        NA        NA        NA
#> Akaniaceae                   NA        NA        NA        NA        NA
#> Alismataceae                 NA 27.287834        NA        NA        NA
#> Alstroemeriaceae             NA 20.014375        NA        NA        NA
#> Altingiaceae                 NA 15.934294        NA  7.550000 0.4435000
#> Amaranthaceae                NA 24.488182        NA 12.411441 0.4189176
#> Amaryllidaceae     1.804000e-02 29.057091        NA 11.440000 0.4508400
#> Amborellaceae                NA        NA        NA        NA        NA
#> Amphorogynaceae              NA 25.236766        NA        NA        NA
#> Anacampserotaceae            NA        NA        NA        NA        NA
#> Anacardiaceae      1.222994e-02 17.906247        NA 10.539736 0.4566867
#> Anisophylleaceae             NA        NA        NA        NA        NA
#> Annonaceae                   NA 23.528495        NA 23.830391 0.4726000
#> Aphloiaceae                  NA        NA        NA        NA        NA
#> Apiaceae           1.587302e-03 24.267326        NA 10.787434 0.4413076
#> Apocynaceae                  NA 21.871073        NA 16.845574 0.4924750
#> Aptandraceae                 NA 28.435524        NA        NA        NA
#> Aquifoliaceae      5.740000e-04 14.471632        NA 14.324450 0.4580000
#> Araceae                      NA 21.957940        NA        NA        NA
#> Araliaceae         3.902458e-04 18.262158        NA 20.500000 0.4540667
#> Araucariaceae      1.878580e-03 11.647458        NA 13.000000        NA
#> Arecaceae          4.500000e-04 18.363663        NA 13.216667 0.4556000
#> Aristolochiaceae   3.589556e-03 31.726525        NA        NA        NA
#> Asparagaceae                 NA 23.961878        NA        NA 0.4631000
#> Asphodelaceae                NA 12.009650        NA        NA        NA
#> Aspleniaceae       8.200000e-03 28.260000        NA        NA        NA
#> Asteraceae         1.052882e-02 21.974744        NA 10.170307 0.4371519
#> Asteropeiaceae               NA        NA        NA        NA        NA
#> Atherospermataceae           NA 17.869722        NA 23.200000        NA
#> Athyriaceae                  NA 26.901037        NA        NA        NA
#> Aulacomniaceae               NA  8.000000        NA        NA        NA
#> Austrobaileyaceae            NA        NA        NA        NA        NA
#> Balanopaceae                 NA        NA        NA        NA        NA
#> Balsaminaceae      1.265730e-02 32.122977        NA        NA        NA
#> Begoniaceae                  NA 34.200000        NA        NA        NA
#> Berberidaceae      1.650000e-03 20.510037        NA 21.087500        NA
#> Betulaceae         2.973057e-03 24.459193 16.357256 14.491831 0.4721962
#> Bignoniaceae                 NA 23.545324  6.575973 19.915047 0.4618077
#> Bixaceae                     NA 25.531043        NA        NA        NA
#> Blechnaceae                  NA 11.872377        NA        NA        NA
#> Bonnetiaceae                 NA  9.800000        NA        NA        NA
#> Boraginaceae       5.740212e-03 23.132595        NA        NA 0.4153500
#> Brassicaceae       1.066000e-02 33.888909        NA 18.501306 0.4272792
#> Bromeliaceae                 NA  9.175618        NA        NA        NA
#> Brunelliaceae                NA 21.006960        NA        NA        NA
#> Bruniaceae                   NA  7.781667        NA        NA        NA
#> Burseraceae                  NA 19.006440        NA 10.964427 0.4548544
#> Butomaceae                   NA 42.600000        NA        NA        NA
#> Buxaceae                     NA 22.691830        NA        NA        NA
#> Cabombaceae                  NA 19.500000        NA        NA        NA
#> Cactaceae          2.433862e-05 19.923915        NA        NA        NA
#> Calophyllaceae               NA 12.820646        NA        NA 0.4663000
#> Calycanthaceae               NA 17.400000        NA        NA        NA
#> Calyceraceae                 NA 43.000000        NA        NA        NA
#> Campanulaceae                NA 27.204986        NA  5.721013 0.4498475
#> Canellaceae                  NA        NA        NA        NA        NA
#>                    Kmax_stemxylem       WUE        P50       Al2As
#> Acanthaceae                    NA        NA         NA  3070.43202
#> Achariaceae                    NA  3.662630         NA          NA
#> Acoraceae                      NA        NA         NA          NA
#> Actinidiaceae                  NA        NA         NA          NA
#> Adoxaceae              4.05355750  3.117284 -3.0384833  4889.57295
#> Aextoxicaceae                  NA        NA         NA          NA
#> Aizoaceae                      NA        NA         NA          NA
#> Akaniaceae                     NA        NA         NA          NA
#> Alismataceae                   NA        NA         NA          NA
#> Alstroemeriaceae               NA        NA         NA          NA
#> Altingiaceae           0.50500000        NA -2.0370147  5129.90090
#> Amaranthaceae          0.08190000 28.856717 -2.4252402   973.13500
#> Amaryllidaceae                 NA 24.140939         NA          NA
#> Amborellaceae          0.54000000        NA -3.0000000  4255.31915
#> Amphorogynaceae                NA        NA         NA          NA
#> Anacampserotaceae              NA        NA         NA          NA
#> Anacardiaceae          4.07720162 15.074736 -2.6535235 17392.13509
#> Anisophylleaceae               NA        NA         NA          NA
#> Annonaceae             5.27066667        NA -2.5276068 10266.65654
#> Aphloiaceae                    NA        NA         NA          NA
#> Apiaceae               0.51500000  3.099757 -5.7000000    81.22547
#> Apocynaceae            2.56516667 29.952870 -2.4864334 19766.80134
#> Aptandraceae                   NA        NA         NA          NA
#> Aquifoliaceae          0.22557572        NA -3.6437782  5890.90789
#> Araceae                        NA        NA         NA          NA
#> Araliaceae             1.68090114        NA -1.6530859  3591.08704
#> Araucariaceae          0.73225000        NA -2.6183226  3846.15385
#> Arecaceae                      NA  3.962648 -1.8100000  9428.35821
#> Aristolochiaceae               NA        NA         NA          NA
#> Asparagaceae                   NA 21.356961 -1.6960000          NA
#> Asphodelaceae                  NA        NA         NA          NA
#> Aspleniaceae                   NA        NA         NA          NA
#> Asteraceae             0.49865714  3.896510 -3.2565860  2423.12470
#> Asteropeiaceae                 NA        NA         NA          NA
#> Atherospermataceae             NA        NA -3.0063333  2435.95630
#> Athyriaceae                    NA        NA         NA          NA
#> Aulacomniaceae                 NA        NA         NA          NA
#> Austrobaileyaceae      2.30000000        NA -0.4990000 15384.61538
#> Balanopaceae                   NA        NA         NA          NA
#> Balsaminaceae                  NA  2.842391         NA          NA
#> Begoniaceae                    NA        NA         NA          NA
#> Berberidaceae          0.08733333  4.039735 -4.5000000   570.09407
#> Betulaceae             2.87333740        NA -2.1017591  7591.16581
#> Bignoniaceae           2.10149000        NA -0.8616667 12439.80827
#> Bixaceae                       NA        NA         NA 12274.01424
#> Blechnaceae                    NA        NA         NA          NA
#> Bonnetiaceae                   NA        NA         NA          NA
#> Boraginaceae                   NA  2.819216 -3.5677066          NA
#> Brassicaceae                   NA  3.632075         NA          NA
#> Bromeliaceae                   NA        NA         NA          NA
#> Brunelliaceae                  NA        NA         NA          NA
#> Bruniaceae             0.25150000        NA -3.3883558          NA
#> Burseraceae            3.40500000  4.390857 -1.3054970 12432.79483
#> Butomaceae                     NA        NA         NA          NA
#> Buxaceae                       NA        NA -8.0000000          NA
#> Cabombaceae                    NA        NA         NA          NA
#> Cactaceae              1.86880952        NA -1.2875000  2555.33971
#> Calophyllaceae         0.89828571        NA -1.5400000  2662.69155
#> Calycanthaceae                 NA        NA -1.2808475          NA
#> Calyceraceae                   NA        NA         NA          NA
#> Campanulaceae                  NA  4.790244         NA          NA
#> Canellaceae                    NA        NA -0.2320000          NA
#>                    conduit2sapwood
#> Acanthaceae              0.6300000
#> Achariaceae                     NA
#> Acoraceae                       NA
#> Actinidiaceae                   NA
#> Adoxaceae                       NA
#> Aextoxicaceae                   NA
#> Aizoaceae                       NA
#> Akaniaceae                      NA
#> Alismataceae                    NA
#> Alstroemeriaceae                NA
#> Altingiaceae             0.8226667
#> Amaranthaceae                   NA
#> Amaryllidaceae                  NA
#> Amborellaceae                   NA
#> Amphorogynaceae                 NA
#> Anacampserotaceae               NA
#> Anacardiaceae            0.7155889
#> Anisophylleaceae                NA
#> Annonaceae               0.5685000
#> Aphloiaceae                     NA
#> Apiaceae                        NA
#> Apocynaceae              0.7112500
#> Aptandraceae                    NA
#> Aquifoliaceae            0.6528500
#> Araceae                         NA
#> Araliaceae               0.7785000
#> Araucariaceae            0.9375000
#> Arecaceae                       NA
#> Aristolochiaceae                NA
#> Asparagaceae                    NA
#> Asphodelaceae                   NA
#> Aspleniaceae                    NA
#> Asteraceae               0.7219423
#> Asteropeiaceae                  NA
#> Atherospermataceae       0.7560000
#> Athyriaceae                     NA
#> Aulacomniaceae                  NA
#> Austrobaileyaceae               NA
#> Balanopaceae                    NA
#> Balsaminaceae                   NA
#> Begoniaceae                     NA
#> Berberidaceae                   NA
#> Betulaceae               0.8444000
#> Bignoniaceae             0.6360476
#> Bixaceae                        NA
#> Blechnaceae                     NA
#> Bonnetiaceae                    NA
#> Boraginaceae                    NA
#> Brassicaceae                    NA
#> Bromeliaceae                    NA
#> Brunelliaceae                   NA
#> Bruniaceae                      NA
#> Burseraceae              0.8204286
#> Butomaceae                      NA
#> Buxaceae                 0.8330000
#> Cabombaceae                     NA
#> Cactaceae                0.3636905
#> Calophyllaceae           0.7335000
#> Calycanthaceae           0.6535500
#> Calyceraceae                    NA
#> Campanulaceae                   NA
#> Canellaceae                     NA
#>  [ reached 'max' / getOption("max.print") -- omitted 320 rows ]