Parameter average values

Internal data set with parameter averages for taxonomic families. This is used by input initialization functions to provide suitable parameter values when missing from species parameter tables.

Format

Data frame trait_family_means has taxonomic families in rows and parameter names as columns.

Source

Same sources as SpParamsMED

See also

Examples

medfate:::trait_family_means
#>                    LeafDensity WoodDensity   LeafPI0   LeafEPS    LeafAF
#> Acanthaceae         0.31359059  0.56758803 -3.395000 23.230000 0.1270000
#> Achariaceae         0.25524948  0.59247961        NA        NA        NA
#> Acoraceae           0.10000000          NA        NA        NA        NA
#> Actinidiaceae       0.44861072  0.40923199        NA        NA        NA
#> Adoxaceae           0.40799269  0.51384234 -1.560000 12.790000        NA
#> Aextoxicaceae               NA  0.56666667        NA        NA        NA
#> Aizoaceae           0.05526858          NA        NA        NA        NA
#> Akaniaceae                  NA  0.55478250        NA        NA        NA
#> Alismataceae        0.19142078          NA        NA        NA        NA
#> Alstroemeriaceae            NA          NA        NA        NA        NA
#> Altingiaceae        0.68339710  0.60109480        NA        NA        NA
#> Amaranthaceae       0.21187621  0.59891031 -2.250000        NA        NA
#> Amaryllidaceae      0.12367381          NA        NA        NA        NA
#> Amborellaceae               NA          NA        NA        NA        NA
#> Amphorogynaceae             NA          NA        NA        NA        NA
#> Anacampserotaceae   0.02500475          NA        NA        NA        NA
#> Anacardiaceae       0.45624702  0.57106817 -1.700000 12.760000        NA
#> Ancistrocladaceae           NA          NA        NA        NA        NA
#> Anisophylleaceae            NA  0.67347799        NA        NA        NA
#> Annonaceae          0.37651502  0.56259675 -2.160000 23.710000        NA
#> Aphloiaceae         0.46270602  0.62052000        NA        NA        NA
#> Apiaceae            0.27609204  0.20604993        NA        NA        NA
#> Apocynaceae         0.29798311  0.56790984 -2.390000 20.940000        NA
#> Aptandraceae        0.33184918  0.70767562        NA        NA        NA
#> Aquifoliaceae       0.48216313  0.55793939 -2.300000 20.700000 0.4000000
#> Araceae             0.15917916          NA        NA        NA        NA
#> Araliaceae          0.28116351  0.40982320 -1.528503 11.231462 0.4065000
#> Araucariaceae       0.34479018  0.46414559        NA        NA        NA
#> Arecaceae           0.41472385  0.59090347 -3.400000 73.400000 0.2000000
#> Aristolochiaceae    0.26814343  0.26000000        NA        NA        NA
#> Asparagaceae        0.21607312  0.42542584        NA        NA        NA
#> Asphodelaceae       0.40325275  0.39519900        NA        NA        NA
#> Aspleniaceae        0.23694373          NA -1.240000 35.300000        NA
#> Asteraceae          0.25025214  0.44342039 -1.471389 14.491429 0.2435000
#> Asteropeiaceae              NA  0.84169641        NA        NA        NA
#> Atherospermataceae  0.24513566  0.47676207 -1.340000  8.380000        NA
#> Athyriaceae         0.22706154          NA        NA        NA        NA
#> Austrobaileyaceae   0.26209917          NA        NA        NA        NA
#> Balanopaceae                NA  0.73489760        NA        NA        NA
#> Balsaminaceae       0.40450133  0.09108341        NA        NA        NA
#> Begoniaceae         0.10939929          NA        NA        NA        NA
#> Berberidaceae       0.35994527  0.68526616        NA        NA        NA
#> Betulaceae          0.44493906  0.53788183 -1.201884  5.498667        NA
#> Bignoniaceae        0.34563784  0.62506774 -1.990000 17.610000 0.1770000
#> Bixaceae            0.44000000  0.35463573        NA        NA        NA
#> Blechnaceae         0.22361115          NA        NA        NA        NA
#> Bonnetiaceae        0.27978289  0.84000000        NA        NA        NA
#> Boraginaceae        0.28483641  0.45253642 -1.140000        NA        NA
#> Brassicaceae        0.24137149  0.39095103 -1.485000  7.710000 0.2320000
#> Bromeliaceae        0.15934340          NA        NA        NA        NA
#> Brunelliaceae               NA  0.31125000        NA        NA        NA
#> Bruniaceae                  NA  0.56365000        NA        NA        NA
#> Burseraceae         0.43321183  0.51975514 -1.435000 14.980000        NA
#> Buxaceae            0.27062945  0.70292030        NA        NA        NA
#> Cactaceae           0.20564854  0.61875000        NA  8.700000        NA
#> Calophyllaceae      0.39682160  0.60678553        NA        NA        NA
#> Calycanthaceae      0.33567851  0.65005500        NA        NA        NA
#> Calyceraceae        0.16893665          NA        NA        NA        NA
#> Campanulaceae       0.23880328  0.67766295        NA        NA        NA
#> Canellaceae                 NA  0.68082545        NA        NA        NA
#> Cannabaceae         0.40555174  0.54869223 -1.580000  5.180000        NA
#> Cannaceae                   NA          NA        NA        NA        NA
#> Capparaceae         0.36252813  0.63512564 -2.840000 14.750000 0.1723333
#> Caprifoliaceae      0.32823425  0.43347916        NA        NA        NA
#> Cardiopteridaceae           NA  0.61757948        NA        NA        NA
#> Caricaceae          0.21453541  0.19103921        NA        NA        NA
#> Caryocaraceae       0.46085582  0.67581652 -1.710000 11.340000        NA
#> Caryophyllaceae     0.29864198  0.32634381        NA        NA        NA
#> Casuarinaceae       0.50739895  0.82376618        NA        NA        NA
#> Celastraceae        0.41742039  0.64146387 -2.600000 19.060000        NA
#> Centroplacaceae             NA  0.65500000        NA        NA        NA
#>                       Gswmax       Gswmin       Narea     WoodC Kmax_stemxylem
#> Acanthaceae        0.2500000           NA  1.52359901 0.4242167             NA
#> Achariaceae               NA           NA  1.72887846        NA             NA
#> Acoraceae                 NA           NA          NA        NA             NA
#> Actinidiaceae             NA           NA  1.34845869        NA             NA
#> Adoxaceae                 NA           NA  1.21864401        NA     4.05355750
#> Aextoxicaceae             NA           NA  1.27323648        NA             NA
#> Aizoaceae                 NA           NA          NA        NA             NA
#> Akaniaceae                NA           NA          NA        NA             NA
#> Alismataceae              NA           NA  1.73301226        NA             NA
#> Alstroemeriaceae          NA           NA  2.47380000        NA             NA
#> Altingiaceae       0.5000000           NA  0.95592334 0.4435000     0.50500000
#> Amaranthaceae      0.0750000           NA  2.37451130 0.4154472     0.08190000
#> Amaryllidaceae            NA 1.804000e-02  1.83883420 0.4508400             NA
#> Amborellaceae             NA           NA          NA        NA     0.54000000
#> Amphorogynaceae           NA           NA  1.58472456        NA             NA
#> Anacampserotaceae         NA           NA          NA        NA             NA
#> Anacardiaceae      0.3148333 1.222994e-02  1.94736953 0.4566867     4.07720162
#> Ancistrocladaceae         NA           NA  0.08628395        NA             NA
#> Anisophylleaceae          NA           NA          NA        NA             NA
#> Annonaceae                NA           NA  1.63339662 0.4726000     5.27066667
#> Aphloiaceae               NA           NA          NA        NA             NA
#> Apiaceae                  NA 1.587302e-03  1.92894126 0.4374312     0.51500000
#> Apocynaceae               NA           NA  1.72454178 0.4924750     2.56516667
#> Aptandraceae              NA           NA  1.90689008        NA             NA
#> Aquifoliaceae      0.2600000 5.740000e-04  1.52957276 0.4317500     0.22557572
#> Araceae                   NA           NA  1.03570140        NA             NA
#> Araliaceae                NA 3.902458e-04  1.53457119 0.4540667     1.68090114
#> Araucariaceae             NA 1.878580e-03  2.74761196        NA     0.73225000
#> Arecaceae                 NA 4.500000e-04  1.59176336 0.4556000             NA
#> Aristolochiaceae          NA 3.589556e-03  1.23375321        NA             NA
#> Asparagaceae              NA           NA  1.85161314 0.4631000             NA
#> Asphodelaceae             NA           NA  3.30552890        NA             NA
#> Aspleniaceae              NA 8.200000e-03  1.21331333        NA             NA
#> Asteraceae         0.1275000 1.052882e-02  1.65579844 0.4365142     0.49865714
#> Asteropeiaceae            NA           NA          NA        NA             NA
#> Atherospermataceae        NA           NA  1.77630822        NA             NA
#> Athyriaceae               NA           NA  1.21836904        NA             NA
#> Austrobaileyaceae         NA           NA          NA        NA     2.30000000
#> Balanopaceae              NA           NA          NA        NA             NA
#> Balsaminaceae             NA 1.265730e-02  0.72395054        NA             NA
#> Begoniaceae               NA           NA  0.85842000        NA             NA
#> Berberidaceae             NA 1.650000e-03  1.42858611        NA     0.08733333
#> Betulaceae         0.3965000 2.973057e-03  1.33151572 0.4721962     2.87333740
#> Bignoniaceae              NA           NA  1.94172003 0.4618077     2.10149000
#> Bixaceae                  NA           NA  1.60716298        NA             NA
#> Blechnaceae               NA           NA  1.01470163        NA             NA
#> Bonnetiaceae              NA           NA          NA        NA             NA
#> Boraginaceae              NA 5.740212e-03  1.29599181 0.4153500             NA
#> Brassicaceae              NA 1.066000e-02  1.76141359 0.4268611             NA
#> Bromeliaceae              NA           NA  1.54925814        NA             NA
#> Brunelliaceae             NA           NA  2.37454346        NA             NA
#> Bruniaceae                NA           NA          NA        NA     0.25150000
#> Burseraceae               NA           NA  1.54976941 0.4516289     3.40500000
#> Buxaceae                  NA           NA  4.21163486        NA             NA
#> Cactaceae                 NA 2.433862e-05 17.58494099        NA     1.86880952
#> Calophyllaceae     0.1350000           NA  1.48246658 0.4663000     0.89828571
#> Calycanthaceae            NA           NA  0.84472534        NA             NA
#> Calyceraceae              NA           NA          NA        NA             NA
#> Campanulaceae             NA           NA  1.68582166 0.4498475             NA
#> Canellaceae               NA           NA          NA        NA             NA
#> Cannabaceae        0.3300000           NA  1.94823496 0.4683583     4.33961052
#> Cannaceae                 NA           NA  0.92648775        NA             NA
#> Capparaceae               NA           NA  4.45229986        NA     0.55726667
#> Caprifoliaceae            NA           NA  1.13656877 0.4475552     0.25210100
#> Cardiopteridaceae         NA           NA  3.20932215        NA             NA
#> Caricaceae                NA           NA  1.16511468        NA             NA
#> Caryocaraceae             NA           NA  1.51507526        NA     1.75872727
#> Caryophyllaceae           NA 1.851852e-03  1.34387670 0.4163613             NA
#> Casuarinaceae             NA           NA  4.01268806 0.4200000     1.89540479
#> Celastraceae              NA           NA  1.75243918 0.4900000             NA
#> Centroplacaceae           NA           NA  1.49156102        NA             NA
#>                          WUE        P50       Al2As conduit2sapwood
#> Acanthaceae               NA         NA  3070.43202       0.6300000
#> Achariaceae         3.662630         NA          NA              NA
#> Acoraceae                 NA         NA          NA              NA
#> Actinidiaceae             NA         NA          NA              NA
#> Adoxaceae           2.807424 -3.0384833  4889.57295              NA
#> Aextoxicaceae             NA         NA          NA              NA
#> Aizoaceae                 NA         NA          NA              NA
#> Akaniaceae                NA         NA          NA              NA
#> Alismataceae              NA         NA          NA              NA
#> Alstroemeriaceae          NA         NA          NA              NA
#> Altingiaceae              NA -2.0370147  5129.90090       0.8226667
#> Amaranthaceae      19.004079 -2.4252402   973.13500              NA
#> Amaryllidaceae     11.450560         NA          NA              NA
#> Amborellaceae             NA -3.0000000  4255.31915              NA
#> Amphorogynaceae           NA         NA          NA              NA
#> Anacampserotaceae         NA         NA          NA              NA
#> Anacardiaceae      15.042810 -2.6535235 17392.13509       0.7155889
#> Ancistrocladaceae         NA         NA          NA              NA
#> Anisophylleaceae          NA         NA          NA              NA
#> Annonaceae                NA -2.5276068 10266.65654       0.5685000
#> Aphloiaceae               NA         NA          NA              NA
#> Apiaceae            2.515830 -5.7000000    81.22547              NA
#> Apocynaceae        28.924811 -2.4864334 19766.80134       0.7112500
#> Aptandraceae              NA         NA          NA              NA
#> Aquifoliaceae             NA -3.6437782  5890.90789       0.6528500
#> Araceae                   NA         NA          NA              NA
#> Araliaceae                NA -1.6530859  3591.08704       0.7785000
#> Araucariaceae             NA -2.6183226  3846.15385       0.9375000
#> Arecaceae           3.558320 -1.8100000  9428.35821              NA
#> Aristolochiaceae          NA         NA          NA              NA
#> Asparagaceae       14.254229 -1.6960000          NA              NA
#> Asphodelaceae             NA         NA          NA              NA
#> Aspleniaceae              NA         NA          NA              NA
#> Asteraceae          2.726027 -3.2565860  2423.12470       0.7219423
#> Asteropeiaceae            NA         NA          NA              NA
#> Atherospermataceae        NA -3.0063333  2435.95630       0.7560000
#> Athyriaceae               NA         NA          NA              NA
#> Austrobaileyaceae         NA -0.4990000 15384.61538              NA
#> Balanopaceae              NA         NA          NA              NA
#> Balsaminaceae       1.818998         NA          NA              NA
#> Begoniaceae               NA         NA          NA              NA
#> Berberidaceae       3.436421 -4.5000000   570.09407              NA
#> Betulaceae                NA -2.1017591  7591.16581       0.8444000
#> Bignoniaceae              NA -0.8616667 12439.80827       0.6360476
#> Bixaceae                  NA         NA 12274.01424              NA
#> Blechnaceae               NA         NA          NA              NA
#> Bonnetiaceae              NA         NA          NA              NA
#> Boraginaceae        2.519816 -3.5677066          NA              NA
#> Brassicaceae        2.881295         NA          NA              NA
#> Bromeliaceae              NA         NA          NA              NA
#> Brunelliaceae             NA         NA          NA              NA
#> Bruniaceae                NA -3.3883558          NA              NA
#> Burseraceae         4.390857 -1.3054970 12432.79483       0.8204286
#> Buxaceae                  NA -8.0000000          NA       0.8330000
#> Cactaceae                 NA -1.2875000  2555.33971       0.3636905
#> Calophyllaceae            NA -1.5400000  2662.69155       0.7335000
#> Calycanthaceae            NA -1.2808475          NA       0.6535500
#> Calyceraceae              NA         NA          NA              NA
#> Campanulaceae       2.804326         NA          NA              NA
#> Canellaceae               NA -0.2320000          NA              NA
#> Cannabaceae        31.385044 -1.5325304 29402.79875       0.7598125
#> Cannaceae                 NA         NA          NA              NA
#> Capparaceae               NA -2.3615234 26604.55487       0.7005000
#> Caprifoliaceae      3.439840 -5.5144833  7568.54261              NA
#> Cardiopteridaceae         NA         NA 11764.70588              NA
#> Caricaceae         27.909979         NA          NA              NA
#> Caryocaraceae             NA -1.6766667  5579.67761              NA
#> Caryophyllaceae           NA         NA          NA              NA
#> Casuarinaceae             NA -2.1750000  3647.15784              NA
#> Celastraceae        2.418337 -3.4679167  8199.20056              NA
#> Centroplacaceae           NA         NA          NA              NA
#>  [ reached 'max' / getOption("max.print") -- omitted 291 rows ]