Physical and biophysical utility functions

Utility functions for the calculation of biophysical variables.

biophysics_irradianceToPhotonFlux(I, lambda)
biophysics_leafTemperature(absRad, airTemperature, u, E, leafWidth = 1.0)
biophysics_leafTemperature2(SWRabs, LWRnet, airTemperature, u, E, leafWidth = 1.0)
biophysics_leafVapourPressure(leafTemp, leafPsi)
biophysics_radiationDiurnalPattern(t, daylength)
biophysics_temperatureDiurnalPattern(t, tmin, tmax, 
                                     tminPrev, tmaxPrev, tminNext, daylength)
biophysics_waterDynamicViscosity(temp)

Arguments

I
Irradiance (in W*m-2).

lambda
Wavelength (in nm).

u
Wind speed above the leaf boundary layer (in m/s).

airTemperature
Air temperature (in ºC).

tmin, tmax
Minimum and maximum daily temperature (ºC).

tminPrev, tmaxPrev, tminNext
Maximum and minimum daily temperatures of the previous and following day (ºC).

absRad
Absorbed long- and short-wave radiation (in W·m-2).

SWRabs
Absorbed short-wave radiation (in W·m-2).

LWRnet
Net long-wave radiation balance (in W·m-2).

E
Transpiration flow (in mmol H20·m-2·s-1) per one sided leaf area basis.

leafWidth
Leaf width (in cm).

t
Time of the day (in seconds).

daylength
Day length (in seconds).

temp
Temperature (ºC).

leafTemp
Leaf temperature (ºC).

leafPsi
Leaf water potential (MPa).

Details

Function biophysics_leafTemperature calculates leaf temperature according to energy balance equation given in Campbell and Norman (1988). Function biophysics_radiationDiurnalPattern follows the equations given in Liu and Jordan (1960). Function biophysics_temperatureDiurnalPattern determines diurnal temperature pattern assuming a sinusoidal pattern with T = Tmin at sunrise and T = (Tmin+Tmax)/2 at sunset and a linear change in temperature between sunset and Tmin of the day after (McMurtrie et al. 1990). Function biophysics_waterDynamicViscosity calculates water dynamic viscosity following the Vogel (1921) equation.

Value

Values returned for each function are:

biophysics_leafTemperature: leaf temperature (in ºC)

biophysics_leafVapourPressure: leaf vapour pressure (in kPa)

biophysics_radiationDiurnalPattern: the proportion of daily radiation corresponding to the input time in seconds after sunrise.

biophysics_temperatureDiurnalPattern: diurnal pattern of temperature.

biophysics_waterDynamicViscosity: Water dynamic viscosity relative to 20ºC.

References

Campbell, G. S., and J. M. Norman. 1998. An introduction to environmental biophysics: 2nd edition. (eqns. 14.1 & 14.3)

B. Y. H. Liu and R. C. Jordan, “The interrelationship and characteristic distribution of direct, diffuse and total solar radiation,” Solar Energy, vol. 4, no. 3, pp. 1–19, 1960.

McMurtrie, R. E., D. A. Rook, and F. M. Kelliher. 1990. Modelling the yield of Pinus radiata on a site limited by water and nitrogen. Forest Ecology and Management 30:381–413.

H. Vogel, "Das Temperaturabhangigkeitsgesetz der Viskositat von Flussigkeiten", Physikalische Zeitschrift, vol. 22, pp. 645–646, 1921.

Author

Miquel De Cáceres Ainsa, CREAF

See also