Carbon-related functions

Set of functions used in the calculation of carbon balance.

carbon_leafStarchCapacity(LAI, N, SLA, leafDensity)
carbon_leafStructuralBiomass(LAI, N, SLA)
carbon_sapwoodStarchCapacity(SA, H, L, V, woodDensity, conduit2sapwood)
carbon_sapwoodStructuralBiomass(SA, H, L, V, woodDensity)
carbon_sapwoodStructuralLivingBiomass(SA, H, L, V, woodDensity, conduit2sapwood)
carbon_sugarConcentration(osmoticWP, temp, nonSugarConc)
carbon_osmoticWaterPotential(sugarConc, temp, nonSugarConc)
carbon_relativeSapViscosity(sugarConc, temp)
carbon_sugarStarchDynamicsLeaf(sugarConc, starchConc, eqSugarConc)

Arguments

LAI
Leaf area index.

N
Density (ind·ha-1).

SLA
Specific leaf area (mm2/mg = m2/kg).

leafDensity
Density of leaf tissue (dry weight over volume).

SA
Sapwood area (cm2).

H
Plant height (cm).

L
Coarse root length (mm) for each soil layer.

V
Proportion of fine roots in each soil layer.

woodDensity
Wood density (dry weight over volume).

conduit2sapwood
Proportion of sapwood corresponding to conducive elements (vessels or tracheids) as opposed to parenchymatic tissue.

osmoticWP
Osmotic water potential (MPa).

temp
Temperature (degrees Celsius).

nonSugarConc
Concentration of inorganic solutes (mol/l).

sugarConc
Concentration of soluble sugars (mol/l).

starchConc
Concentration of starch (mol/l)

eqSugarConc
Equilibrium concentration of soluble sugars (mol/l).

Value

Values returned for each function are:

carbon_leafStarchCapacity: Capacity of storing starch in the leaf compartment (mol glucose).

carbon_sapwoodStarchCapacity: Capacity of storing starch in the sapwood compartment (mol glucose).

Author

Miquel De Cáceres Ainsa, CREAF

See also