Default simulation control parameters for models

Creates a list with global default parameters for simulation models.

defaultControl(transpirationMode = "Granier")

Arguments

transpirationMode
Transpiration model (either 'Granier' or 'Sperry'). See spwbInput.

Details

The function returns a list with default parameters. Users can change those defaults that need to be set to other values and use the list as input for model functions. The relevant parameters are different for each model function.

Value

A list, with the following options (default values in parentheses):

General:

verbose (=TRUE): Boolean flag to indicate console output during calculations. In function fordyn verbose is always set to FALSE.

modifyInput (=TRUE): Boolean flag to indicate that simulations will modify input object. If set to FALSE, simulations will not modify the input R object but return the current (modified) state variables within the output. In function fordyn modifyInput is always set to FALSE.

fillMissingSpParams (=TRUE): Boolean flag to indicate that functions spwbInput and growthInput should provide estimates for functional parameters if these are lacking in the species parameter table SpParams. Note that if fillMissingSpParams is set to FALSE then simulations may fail if the user does not provide values for required parameters.

subdailyResults (=FALSE): Boolean flag to force subdaily results to be stored (as a list called 'subdaily' of spwb_day objects, one by simulated date) in calls to spwb. In function fordyn subdailyResults is always set to FALSE.

Water balance:

transpirationMode (="Granier"): Transpiration model (either 'Granier' or 'Sperry'). See spwbInput.

soilFunctions (="SX"): Soil water retention curve and conductivity functions, either 'SX' (for Saxton) or 'VG' (for Van Genuchten). If transpirationMode = "Sperry" then soilFunctions is set by default to 'VG'.

defaultWindSpeed (= 2.5): Default wind speed value (in m/s) to be used when missing from data.

snowpack (=TRUE): Boolean flag to indicate the simulation of snow accumulation and melting.

leafPhenology (=TRUE): Boolean flag to indicate the simulation of leaf phenology for winter-deciduous species.

rockyLayerDrainage (=TRUE): Boolean flag to indicate the simulation of drainage from rocky layers (> 95% of rocks).

unlimitedSoilWater (=FALSE): Boolean flag to indicate the simulation of plant transpiration assuming that soil water is always at field capacity.

unfoldingDD (=300): Degree-days for complete leaf unfolding after budburst has occurred.

plantWaterPools (=FALSE): Boolean flag to indicate the simulation of water balance assuming that each cohort has its own water pools.

verticalLayerSize (= 100): Size of vertical layers (in cm) for the calculation of light extinction (and photosynthesis).

windMeasurementHeight (=200): Height (in cm) over the canopy corresponding to wind measurements.

cavitationRefill (= "total"): A string indicating how refilling of embolized conduits is done:

"none" - no refilling.

"annual" - every first day of the year.

"rate" - following a rate of new sapwood formation (only available if transpirationMode = "Sperry")

"total" - instantaneous complete refilling.

Water balance with 'Sperry':

ndailysteps (= 24): Number of steps into which each day is divided for determination of stomatal conductance, transpiration and photosynthesis (24 equals 1-hour intervals).

nsubsteps (= 3600): Number of substeps into which each step is divided for multi-layer canopy energy balance solving.

capacitance (=FALSE): Whether the effect of plant water compartments is considered in simulations.

multiLayerBalance (=FALSE): Flag to indicate multiple canopy energy balance. If FALSE, canopy is considered a single layer for energy balance.

cochard (=FALSE): Boolean flag to indicate the stomatal closure when leaf water potential is below turgor loss point so that transpiration depends on cuticular properties (Martin-StPaul et al. 2017).

taper (= TRUE): Whether taper of xylem conduits is accounted for when calculating aboveground stem conductance from xylem conductivity.

gainModifier, costModifier (= 1): Modifiers (exponents) of the gain and cost functions defined in Sperry et al. (2016).

klatstem (= 0.01): Stem symplastic-apoplastic lateral conductance (in mmol·s-1·m-2·MPa-1). Only relevant when capacitance = TRUE.

klatleaf (= 0.01): Leaf symplastic-apoplastic lateral conductance (in mmol·s-1·m-2·MPa-1). Only relevant when capacitance = TRUE.

numericParams: A list with the following elements:

maxNsteps (= 400): Maximum number of steps in supply function.

ntrial (= 200): Number of iteration trials when finding root of equation system.

psiTol (= 0.0001): Tolerance value for water potential.

ETol (= 0.0001): Tolerance value for flow.

thermalCapacityLAI (=1000000): Thermal canopy capacitance per LAI unit.

Catm (=386): Atmospheric CO2 concentration (in micromol·mol-1 = ppm).

fracLeafResistance (= NA): Fraction of plant total resistance (leaf+stem+root) that corresponds to leaves. This fraction is used if VCleaf_kmax = NA.

fracRootResistance (=0.40): Fraction of plant total resistance (leaf+stem+root) that corresponds to root system.

averageFracRhizosphereResistance (=0.15): Fraction to total continuum (leaf+stem+root+rhizosphere) resistance that corresponds to rhizosphere (averaged across soil water potential values).

boundaryLayerSize (= 2000): Size of the boundary layer (in cm) over the canopy (relevant for multi-layer canopy energy balance).

refillMaximumRate (= 0.05): Maximum rate of daily refilling of embolized conduits as sapwood area per leaf area (in cm2·m-2·day-1).

Growth/mortality:

allowDessication (=TRUE): Boolean flag to indicate that mortality by dessication is allowed.

allowStarvation (=TRUE): Boolean flag to indicate that mortality by starvation is allowed.

allowDefoliation (=TRUE): Boolean flag to indicate that complete drought-driven defoliation is allowed.

sinkLimitation (=TRUE): Boolean flag to indicate that temperature and turgor limitations to growth are applied.

shrubDynamics [= FALSE]: Boolean flag to allow the application of demographic processes to shrubs.

allocationStrategy (="Plant_kmax"): Strategy for allocation (either "Plant_kmax", for constant maximum plant conductance, or "Al2As" for constant Huber value).

nonStomatalPhotosynthesisLimitation (=TRUE): A flag to apply limitations of pho-tosynthesis when leaf sugar levels become too high.

phloemConductanceFactor (= 0.2)): Factor to transform stem xylem conductance to stem phloem conductance (only for transpirationMode = "Sperry").

nonSugarConcentration (= 0.25): Non-sugar (inorganic) solute concentration (mol·l-1) in cells.

equilibriumOsmoticConcentration [= c(leaf = 0.8, sapwood = 0.6)]: Equilibrium osmotic concentrations (mol·l-1) for leaf and sapwood cells. The difference between leaf and sapwood values helps maintaining phloem transport. The equilibrium sugar concentration is equilibriumOsmoticConcentration - nonSugarConcentration defaults to [= c(leaf = 0.55, sapwood = 0.35)].

minimumRelativeSugarForGrowth [= 0.5]: Minimum concentration of metabolic sugar relative to equilibrium sugar concentration for growth to occur in different tissues. This value (by default 50%) should always be larger than mortalityRelativeSugarThreshold [= 0.3], so that growth stops before starvation, and smaller than 1 (i.e. 100%), so that growth occurs under normal metabolic status.

respirationRates [= c(leaf = 0.00260274, sapwood = 6.849315e-05, fineroot = 0.002054795)]: Maintenance respiration rates for different tissues (g gluc · g dry-1 · day-1). Sapwood respiration rates apply to parenchymatic tissue only. Values for sapwood and fine roots are used for all species, whereas the value for leaves is the default used when leaf respiration is not specified via SpParams (RERleaf).

senescenceRates [= c(sapwood = 0.0001261398, fineroot = 0.001897231)]: Senescence rates (day-1) for sapwood and fineroots. Default are equivalent to 4.5% and 50% annual turnover for sapwood and fine roots, respectively.

constructionCosts [= c(leaf = 1.5, sapwood = 1.47, fineroot = 1.30)]: Construction costs, including respiration and structural carbon, per dry weight of new tissue (g gluc · g dry -1).

maximumRelativeGrowthRates [= c(leaf = 0.01, sapwood = 0.002, fineroot = 0.1)]: Maximum relative growth rates for leaves (m2 leaf ·cm-2 sapwood· day-1), sapwood (cm2 sapwood ·cm-2 sapwood· day-1) and fine roots (g dw · g dw -1 · day -1). Values for leaves and fine roots are used for all species, whereas the value for sapwood is the default used when relative growth rate is not specified via SpParams (RGRsapwoodmax).

mortalityMode [= "density/deterministic"]: String describing how mortality is applied. Current accepted values are combinations of "cohort" vs "density" (for whole-cohort mortality vs reduction of stem density) and "deterministic" vs. "stochastic".

mortalityBaselineRate [= 0.01]: Deterministic proportion or probability specifying the baseline reduction of cohort's density occurring in a year (for mortalityMode = "density/deterministic" or "density/stochastic").

mortalityRelativeSugarThreshold [= 0.3]: Threshold of stem sugar concentration relative to equilibrium values (by default 30%), resulting in an increased mortality rate/probability whenever levels are below.

mortalityRWCThreshold [= 0.3]: Threshold of stem relative water content resulting in increased mortality rate/probability whenever levels are below.

Forest dynamics:

recruitmentMode [= "deterministic"]: String describing how recruitment is applied. Current accepted values are "deterministic" or "stochastic".

removeDeadCohorts [= TRUE]: Boolean flag to indicate the removal of cohorts whose density is too low.

minimumCohortDensity [= 1]: Threshold of density resulting in cohort removal.

seedRain [= NULL]: Vector of species codes whose seed rain is to be simulated. If NULL the species identity of seed rain is taken from species currently present in the forest stand and with minimum size (see below).

seedProductionTreeHeight [= 300]: Default minimum tree height for producing seeds (when species parameter SeedProductionHeight is missing).

seedProductionShrubHeight [= 30]: Default minimum shrub height for producing seeds (when species parameter SeedProductionHeight is missing).

minTempRecr [= 0]: Default threshold of minimum average temperature of the coldest month necessary for recruiting (when species parameter MinTempRecr is missing).

minMoistureRecr [= 0.3]: Default threshold of minimum moisture index (annual precipitation over annual ETP) necessary for recruiting (when species parameter MinMoistureRecr is missing).

minFPARRecr [= 10]: Default threshold of minimum fraction of PAR (in %) reaching the ground necessary for recruiting (when species parameter MinFPARRecr is missing).

recrTreeDBH [= 1]: Default DBH (cm) for recruited trees (when species parameter RecrTreeDBH is missing).

recrTreeDensity [= 100]: Default density (ind·ha-1) for recruited trees (when species parameter RecrTreeDensity is missing).

recrTreeHeight [= 100]: Default height (cm) for recruited trees (when species parameter RecrTreeHeight is missing).

recrShrubCover [= 1]: Default cover (%) for recruited shrubs (when species parameter RecrShrubCover is missing).

recrShrubHeight [= 100]: Default height (cm) for recruited shrubs (when species parameter RecrShrubHeight is missing).

recrTreeZ50 [= 100]: Default value for Z50 (mm) in recruited trees (when species parameter RecrZ50 is missing).

recrShrubZ50 [= 50]: Default value for Z50 (mm) in recruited shrubs (when species parameter RecrZ50 is missing).

recrTreeZ95 [= 1000]: Default value for Z95 (mm) in recruited trees (when species parameter RecrZ50 is missing).

recrShrubZ50 [= 500]: Default value for Z95 (mm) in recruited shrubs (when species parameter RecrZ50 is missing).

Author

Miquel De Cáceres Ainsa, CREAF

See also