Control parameters for simulation models

Creates a list control parameters default values for simulations.

defaultControl(transpirationMode = "Granier")

Arguments

transpirationMode

Transpiration model (either 'Granier' or 'Sperry'). See spwbInput.

Details

The function returns a list with default parameters. Users can change those defaults that need to be set to other values and use the list as input for model functions. The relevant parameters are different for each model function.

Value

A list, with the following options (default values in parentheses):

General:

verbose (=TRUE): Boolean flag to indicate console output during calculations. In function fordyn verbose is always set to FALSE.

fillMissingRootParams (=TRUE): Boolean flag to indicate that functions spwbInput and growthInput should provide estimates for Z50 and Z95 if these are lacking in the forest data. Note that if fillMissingRootParams is set to FALSE then simulations may fail if the user does not provide values for Z50 and Z95 in tree or shrub data.

fillMissingSpParams (=TRUE): Boolean flag to indicate that functions spwbInput and growthInput should provide estimates for functional parameters if these are lacking in the species parameter table SpParams. Note that if fillMissingSpParams is set to FALSE then simulations may fail if the user does not provide values for required parameters.

subdailyResults (=FALSE): Boolean flag to force subdaily results to be stored (as a list called 'subdaily' of spwb_day objects, one by simulated date) in calls to spwb. In function fordyn subdailyResults is always set to FALSE.

Water balance (functions spwb, pwb or spwb_day):

transpirationMode (="Granier"): Transpiration model (either 'Granier' or 'Sperry'). See spwbInput.

soilFunctions (="SX"): Soil water retention curve and conductivity functions, either 'SX' (for Saxton) or 'VG' (for Van Genuchten). If transpirationMode = "Sperry" then soilFunctions is set by default to 'VG'.

defaultWindSpeed (= 2.5): Default wind speed value (in m/s) to be used when missing from data.

defaultCO2 (= 386): Default atmospheric (abovecanopy) CO2 concentration (in micromol·mol-1 = ppm). This value will be used whenever CO2 concentration is not specified in the weather input.

snowpack (=TRUE): Boolean flag to indicate the simulation of snow accumulation and melting.

leafPhenology (=TRUE): Boolean flag to indicate the simulation of leaf phenology for winter-deciduous species.

rockyLayerDrainage (=TRUE): Boolean flag to indicate the simulation of drainage from rocky layers (> 95% of rocks).

unlimitedSoilWater (=FALSE): Boolean flag to indicate the simulation of plant transpiration assuming that soil water is always at field capacity.

unfoldingDD (=300): Degree-days for complete leaf unfolding after budburst has occurred.

rhizosphereOverlap (="total"): A string indicating the degree of rhizosphere spatial overlap between plant cohorts:

"none" - no overlap (independent water pools).

"partial" - partial overlap determined by coarse root volume.

"total" - total overlap (plants extract from common soil pools).

verticalLayerSize (= 100): Size of vertical layers (in cm) for the calculation of light extinction (and photosynthesis).

windMeasurementHeight (=200): Height (in cm) over the canopy corresponding to wind measurements.

cavitationRefill (= "total"): A string indicating how refilling of embolized conduits is done:

"none" - no refilling.

"annual" - every first day of the year.

"rate" - following a rate of new sapwood formation.

"total" - instantaneous complete refilling.

Water balance (functions spwb, pwb or spwb_day when traspirationMode = "Granier"):

hydraulicRedistributionFraction (= 0.1): Fraction of plant transpiration corresponding to hydraulic redistribution.

Water balance (functions spwb, pwb or spwb_day when traspirationMode = "Sperry"):

ndailysteps (= 24): Number of steps into which each day is divided for determination of stomatal conductance, transpiration and photosynthesis (24 equals 1-hour intervals).

nsubsteps (= 3600): Number of substeps into which each step is divided for multi-layer canopy energy balance solving.

capacitance (=FALSE): Whether the effect of plant water compartments is considered in simulations.

multiLayerBalance (=FALSE): Flag to indicate multiple canopy energy balance. If FALSE, canopy is considered a single layer for energy balance.

taper (= TRUE): Whether taper of xylem conduits is accounted for when calculating aboveground stem conductance from xylem conductivity.

maximumStemConductance (= 10): Maximum allowed value for the stem maximum hydraulic conductance (in mmol·s-1·m-2·MPa-1). Introduced to avoid excessive hydraulic redistribution caused by species with small size (i.e. very large stem conductance).

klatstem (= 0.01): Stem symplastic-apoplastic lateral conductance (in mmol·s-1·m-2·MPa-1). Only relevant when capacitance = TRUE.

klatleaf (= 0.01): Leaf symplastic-apoplastic lateral conductance (in mmol·s-1·m-2·MPa-1). Only relevant when capacitance = TRUE.

numericParams: A list with the following elements:

maxNsteps (= 400): Maximum number of steps in supply function.

ntrial (= 200): Number of iteration trials when finding root of equation system.

psiTol (= 0.0001): Tolerance value for water potential.

ETol (= 0.0001): Tolerance value for flow.

thermalCapacityLAI (=1000000): Thermal canopy capacitance per LAI unit.

fracLeafResistance (= NA): Fraction of plant total resistance (leaf+stem+root) that corresponds to leaves. This fraction is used if VCleaf_kmax = NA.

fracRootResistance (= 0.40): Fraction of plant total resistance (leaf+stem+root) that corresponds to root system.

averageFracRhizosphereResistance (=0.15): Fraction to total continuum (leaf+stem+root+rhizosphere) resistance that corresponds to rhizosphere (averaged across soil water potential values).

boundaryLayerSize (= 2000): Size of the boundary layer (in cm) over the canopy (relevant for multi-layer canopy energy balance).

refillMaximumRate (= 0.05): Maximum rate of daily refilling of embolized conduits as sapwood area per leaf area (in cm2·m-2·day-1).

Forest growth (functions growth or growth_day):

subdailyCarbonBalance (=FALSE): Boolean flag to indicate that labile carbon balance should be conducted at sub-daily steps (applies only to transpirationMode = "Sperry").

allowDessication (=TRUE): Boolean flag to indicate that mortality by dessication is allowed.

allowStarvation (=TRUE): Boolean flag to indicate that mortality by starvation is allowed.

sinkLimitation (=TRUE): Boolean flag to indicate that temperature and turgor limitations to growth are applied.

shrubDynamics [= TRUE]: Boolean flag to allow the application of demographic processes to shrubs.

allocationStrategy (="Al2As"): Strategy for allocation (either "Plant_kmax", for constant maximum plant conductance, or "Al2As" for constant Huber value).

phloemConductanceFactor (= 0.2)): Factor to transform stem xylem conductance to stem phloem conductance (only for transpirationMode = "Sperry").

nonSugarConcentration (= 0.25): Non-sugar (inorganic) solute concentration (mol·l-1) in cells.

equilibriumOsmoticConcentration [= c(leaf = 0.8, sapwood = 0.6)]: Equilibrium osmotic concentrations (mol·l-1) for leaf and sapwood cells. The difference between leaf and sapwood values helps maintaining phloem transport. The equilibrium sugar concentration is equilibriumOsmoticConcentration - nonSugarConcentration defaults to [= c(leaf = 0.55, sapwood = 0.35)].

minimumRelativeStarchForGrowth [= 0.50]: Default minimum concentration of storage carbon (starch), relative to the maximum storage capacity, for sapwood growth to occur, when not specified via SpParams (RSSG).

constructionCosts [= c(leaf = 1.5, sapwood = 1.47, fineroot = 1.30)]: Default construction costs, including respiration and structural carbon, per dry weight of new tissue (g gluc · g dry -1) when not specified via SpParams (CCleaf, CCsapwood and CCfineroot).

senescenceRates [= c(sapwood = 0.0001261398, fineroot = 0.001897231)]: Default senescence rates (day-1) for sapwood and fineroots when not specified via SpParams (SRsapwood and SRfineroot). Defaults are equivalent to 9%, 5% and 50% annual turnover for gymnosperm sapwood, angiosperm sapwood and fine roots, respectively.

maximumRelativeGrowthRates [= c(leaf = 0.03, cambium = 0.005, sapwood = 0.002, fineroot = 0.1)]: Default maximum relative growth rates for leaves (m2 leaf ·cm-2 sapwood· day-1), tree sapwood (cm2 sapwood· cm-1 cambium · day-1), shrub sapwood (cm2 sapwood ·cm-2 sapwood· day-1) and fine roots (g dw · g dw -1 · day -1) when not specified via SpParams (RGRleafmax, RGRcambiummax , RGRsapwoodmax and RGRfinerootmax, respectively).

mortalityMode [= "density/deterministic"]: String describing how mortality is applied. Current accepted values are combinations of "cohort" vs "density" (for whole-cohort mortality vs reduction of stem density) and "deterministic" vs. "stochastic".

mortalityBaselineRate [= 0.0015]: Default deterministic proportion or probability specifying the baseline reduction of cohort's density occurring in a year (for mortalityMode = "density/deterministic" or "density/stochastic").

mortalityRelativeSugarThreshold [= 0.3]: Threshold of stem sugar concentration, in relation to the equilibrium sugar concentration, resulting in an increased mortality rate/probability whenever levels are below.

mortalityRWCThreshold [= 0.3]: Threshold of stem relative water content resulting in increased mortality rate/probability whenever levels are below.

Forest dynamics (function fordyn):

allowRecruitment (=TRUE): Boolean flag to indicate that recruitment is allowed.

recruitmentMode [= "stochastic"]: String describing how recruitment is applied. Current accepted values are "deterministic" or "stochastic".

removeEmptyCohorts [= TRUE]: Boolean flag to indicate the removal of cohorts whose density is too low.

minimumCohortDensity [= 1]: Threshold of density resulting in cohort removal.

seedRain [= NULL]: Vector of species codes whose seed rain is to be simulated. If NULL the species identity of seed rain is taken from species currently present in the forest stand and with minimum size (see below).

seedProductionTreeHeight [= 300]: Default minimum tree height for producing seeds (when species parameter SeedProductionHeight is missing).

seedProductionShrubHeight [= 30]: Default minimum shrub height for producing seeds (when species parameter SeedProductionHeight is missing).

probRecr = [0.05]: Default annual probability of recruitment (when species parameter ProbRecr is missing).

minTempRecr [= 0]: Default threshold of minimum average temperature of the coldest month necessary for recruiting (when species parameter MinTempRecr is missing).

minMoistureRecr [= 0.3]: Default threshold of minimum moisture index (annual precipitation over annual ETP) necessary for recruiting (when species parameter MinMoistureRecr is missing).

minFPARRecr [= 10]: Default threshold of minimum fraction of PAR (in %) reaching the ground necessary for recruiting (when species parameter MinFPARRecr is missing).

recrTreeDBH [= 1]: Default DBH (cm) for recruited trees (when species parameter RecrTreeDBH is missing).

recrTreeDensity [= 100]: Default density (ind·ha-1) for recruited trees (when species parameter RecrTreeDensity is missing).

recrTreeHeight [= 100]: Default height (cm) for recruited trees (when species parameter RecrTreeHeight is missing).

recrShrubCover [= 1]: Default cover (%) for recruited shrubs (when species parameter RecrShrubCover is missing).

recrShrubHeight [= 25]: Default height (cm) for recruited shrubs (when species parameter RecrShrubHeight is missing).

recrTreeZ50 [= 100]: Default value for Z50 (mm) in recruited trees (when species parameter RecrZ50 is missing).

recrShrubZ50 [= 50]: Default value for Z50 (mm) in recruited shrubs (when species parameter RecrZ50 is missing).

recrTreeZ95 [= 1000]: Default value for Z95 (mm) in recruited trees (when species parameter RecrZ50 is missing).

recrShrubZ50 [= 500]: Default value for Z95 (mm) in recruited shrubs (when species parameter RecrZ50 is missing).

Author

Miquel De Cáceres Ainsa, CREAF

See also