Default soil parameters

Creates a data frame with default soil physical description for model functions.

defaultSoilParams(n = 4)

Arguments

n

An integer with the number of soil layers (between two and five).

Details

The function returns a data frame with default physical soil description, with soil layers in rows. Users can change those that need to be set to other values and use the list as input for function soil.

Note

While this function is limited to five soil layers, user defined data frames can discretize soils using an unlimited number of soil layers.

Value

A data frame with layers in rows and the following columns (and default values):

widths (= c(300,700,1000,2000): Width of soil layers (in mm).

clay (= 25): Clay percentage for each layer (in %).

sand (= 25): Sand percentage for each layer (in %).

om (= NA): Organic matter percentage for each layer (in %).

bd (= 1.5): Bulk density for each layer (in g/cm3).

rfc (= c(20,40,60,85)): Percentage of rock fragment content (volume basis) for each layer.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

defaultSoilParams(4)
#>   widths clay sand om  bd rfc
#> 1    300   25   25 NA 1.5  25
#> 2    700   25   25 NA 1.5  45
#> 3   1000   25   25 NA 1.5  75
#> 4   2000   25   25 NA 1.5  95