Extracts model outputs

Function extract() extracts daily or subdaily output and returns it as a tidy data frame.

Usage

extract(
  x,
  level = "forest",
  output = NULL,
  vars = NULL,
  dates = NULL,
  subdaily = FALSE,
  addunits = FALSE
)

Arguments

x

An object returned by simulation functions spwb, aspwb, pwb, growth or fordyn.

level

Level of simulation output, either "forest" (stand-level results), "soillayer" (soil layer-level results), "cohort" (cohort-level results), "sunlitleaf" or "shadeleaf" (leaf-level results)

output

Section of the model output to be explored. See details.

vars

Variables to be extracted (by default, all of them).

dates

A date vector indicating the subset of simulated days for which output is desired.

subdaily

A flag to indicate that subdaily values are desired (see details).

addunits

A flag to indicate that variable units should be added whenever possible.

Value

Function extract() returns a data frame:

If level = "forest", columns are "date" and variable names.

If level = "soillayer", columns are "date", "soillayer" and variable names.

If level = "cohort", level = "sunlitleaf" or level = "shadeleaf", columns are "date", "cohorts", "species" and variable names.

If subdaily = TRUE, columns are "datetime", "cohorts", "species" and variable names.

Details

When subdaily = FALSE, parameter output is used to restrict the section in x where variables are located. For example output = "Plants" will correspond to variables "LAI", "LAIlive", "Transpiration", "StemPLC",... as returned by a call names(x$Plants).

Option subdaily = TRUE only works when simulations have been carried using control option 'subdailyResults = TRUE' (see defaultControl). When using subdaily = TRUE, parameter output is not taken into account, and options for parameter vars are the following:

Variables for level = "forest" or level = "soillayer": Not allowed. An error is raised.

Variables for level = "cohort": "E","Ag","An","dEdP","RootPsi","StemPsi","LeafPsi","StemPLC","StemRWC","LeafRWC","StemSympRWC","LeafSympRWC","PWB".

Variables for level = "shadeleaf" and level="sunlitleaf": "Abs_SWR","Abs_PAR","Net_LWR","E","Ag","An","Ci","Gsw","VPD","Temp","Psi","iWUE".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

#Define soil with default soil params (4 layers)
examplesoil <- defaultSoilParams(4)

#Initialize control parameters
control <- defaultControl("Granier")

#Initialize input
x <- spwbInput(exampleforest,examplesoil, SpParamsMED, control)

#Call simulation function (ten days)
S1<-spwb(x, examplemeteo[1:10, ], latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 6.27649
#> Initial soil water content (mm): 290.875
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:
#> 
#> Final plant water content (mm): 6.27638
#> Final soil water content (mm): 294.384
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): -0.000104731
#> Plant water balance result (mm): -0.000104731
#> Change in soil water content (mm): 3.50949
#> Soil water balance result (mm): 3.50949
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 39 Rain (mm) 34 Snow (mm) 5
#>   Interception (mm) 9 Net rainfall (mm) 24
#>   Infiltration (mm) 30 Infiltration excess (mm) 0 Saturation excess (mm) 0 Capillarity rise (mm) 0
#>   Soil evaporation (mm) 2  Herbaceous transpiration (mm) 0  Woody plant transpiration (mm) 3  Mistletoe transpiration (mm) 0
#>   Plant extraction from soil (mm) 3  Plant water balance (mm) -0 Hydraulic redistribution (mm) 0
#>   Runoff (mm) 0 Deep drainage (mm) 21

#Extracts daily forest-level output as a data frame
extract(S1, level = "forest", addunits = TRUE)
#>          date               PET     Precipitation              Rain
#> 1  2001-01-01 0.8828475 [L/m^2]  4.869109 [L/m^2]  4.869109 [L/m^2]
#> 2  2001-01-02 1.6375337 [L/m^2]  2.498292 [L/m^2]  2.498292 [L/m^2]
#> 3  2001-01-03 1.3017026 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 4  2001-01-04 0.5690790 [L/m^2]  5.796973 [L/m^2]  5.796973 [L/m^2]
#> 5  2001-01-05 1.6760567 [L/m^2]  1.884401 [L/m^2]  1.884401 [L/m^2]
#> 6  2001-01-06 1.2077028 [L/m^2] 13.359801 [L/m^2] 13.359801 [L/m^2]
#> 7  2001-01-07 0.6368075 [L/m^2]  5.377400 [L/m^2]  0.000000 [L/m^2]
#> 8  2001-01-08 0.8318207 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 9  2001-01-09 1.9823797 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 10 2001-01-10 0.8286873 [L/m^2]  5.117768 [L/m^2]  5.117768 [L/m^2]
#>              Snow            NetRain       Snowmelt       Infiltration
#> 1  0.0000 [L/m^2]  3.2972315 [L/m^2] 0.0000 [L/m^2]  3.2972315 [L/m^2]
#> 2  0.0000 [L/m^2]  0.9721473 [L/m^2] 0.0000 [L/m^2]  0.9721473 [L/m^2]
#> 3  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 4  0.0000 [L/m^2]  4.2353558 [L/m^2] 0.0000 [L/m^2]  4.2353558 [L/m^2]
#> 5  0.0000 [L/m^2]  0.7332674 [L/m^2] 0.0000 [L/m^2]  0.7332674 [L/m^2]
#> 6  0.0000 [L/m^2] 11.5940646 [L/m^2] 0.0000 [L/m^2] 11.5940646 [L/m^2]
#> 7  5.3774 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 8  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 9  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 10 0.0000 [L/m^2]  3.5457087 [L/m^2] 5.3774 [L/m^2]  8.9231083 [L/m^2]
#>    InfiltrationExcess SaturationExcess    Runoff      DeepDrainage
#> 1           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 2.6175309 [L/m^2]
#> 2           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.1881079 [L/m^2]
#> 3           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.0000000 [L/m^2]
#> 4           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 3.0529310 [L/m^2]
#> 5           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.1206369 [L/m^2]
#> 6           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#> 7           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#> 8           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 2.3794465 [L/m^2]
#> 9           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.0000000 [L/m^2]
#> 10          0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#>    CapillarityRise Evapotranspiration     Interception   SoilEvaporation
#> 1        0 [L/m^2]  2.2515780 [L/m^2] 1.571877 [L/m^2] 0.4387811 [L/m^2]
#> 2        0 [L/m^2]  2.4730095 [L/m^2] 1.526144 [L/m^2] 0.5000000 [L/m^2]
#> 3        0 [L/m^2]  0.7182154 [L/m^2] 0.000000 [L/m^2] 0.3630184 [L/m^2]
#> 4        0 [L/m^2]  1.8630006 [L/m^2] 1.561617 [L/m^2] 0.1461396 [L/m^2]
#> 5        0 [L/m^2]  2.1085118 [L/m^2] 1.151134 [L/m^2] 0.5000000 [L/m^2]
#> 6        0 [L/m^2]  2.3578418 [L/m^2] 1.765736 [L/m^2] 0.2625835 [L/m^2]
#> 7        0 [L/m^2]  0.1737800 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 8        0 [L/m^2]  0.2269834 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 9        0 [L/m^2]  0.5408987 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 10       0 [L/m^2]  1.9575383 [L/m^2] 1.572059 [L/m^2] 0.1594079 [L/m^2]
#>    HerbTranspiration   PlantExtraction     Transpiration MistletoeTranspiration
#> 1          0 [L/m^2] 0.2409196 [L/m^2] 0.2409196 [L/m^2]              0 [L/m^2]
#> 2          0 [L/m^2] 0.4468653 [L/m^2] 0.4468653 [L/m^2]              0 [L/m^2]
#> 3          0 [L/m^2] 0.3551970 [L/m^2] 0.3551970 [L/m^2]              0 [L/m^2]
#> 4          0 [L/m^2] 0.1552438 [L/m^2] 0.1552438 [L/m^2]              0 [L/m^2]
#> 5          0 [L/m^2] 0.4573778 [L/m^2] 0.4573778 [L/m^2]              0 [L/m^2]
#> 6          0 [L/m^2] 0.3295223 [L/m^2] 0.3295223 [L/m^2]              0 [L/m^2]
#> 7          0 [L/m^2] 0.1737800 [L/m^2] 0.1737800 [L/m^2]              0 [L/m^2]
#> 8          0 [L/m^2] 0.2269834 [L/m^2] 0.2269834 [L/m^2]              0 [L/m^2]
#> 9          0 [L/m^2] 0.5408987 [L/m^2] 0.5408987 [L/m^2]              0 [L/m^2]
#> 10         0 [L/m^2] 0.2260713 [L/m^2] 0.2260713 [L/m^2]              0 [L/m^2]
#>    HydraulicRedistribution                LAI     LAIherb            LAIlive
#> 1      0.000000000 [L/m^2] 1.826041 [m^2/m^2] 0 [m^2/m^2] 1.826041 [m^2/m^2]
#> 2      0.000000000 [L/m^2] 1.826041 [m^2/m^2] 0 [m^2/m^2] 1.826041 [m^2/m^2]
#> 3      0.000000000 [L/m^2] 1.826041 [m^2/m^2] 0 [m^2/m^2] 1.826041 [m^2/m^2]
#> 4      0.000000000 [L/m^2] 1.826041 [m^2/m^2] 0 [m^2/m^2] 1.826041 [m^2/m^2]
#> 5      0.000000000 [L/m^2] 1.826041 [m^2/m^2] 0 [m^2/m^2] 1.826041 [m^2/m^2]
#> 6      0.000000000 [L/m^2] 1.826041 [m^2/m^2] 0 [m^2/m^2] 1.826041 [m^2/m^2]
#> 7      0.003861640 [L/m^2] 1.826041 [m^2/m^2] 0 [m^2/m^2] 1.826041 [m^2/m^2]
#> 8      0.004609261 [L/m^2] 1.826041 [m^2/m^2] 0 [m^2/m^2] 1.826041 [m^2/m^2]
#> 9      0.000000000 [L/m^2] 1.826041 [m^2/m^2] 0 [m^2/m^2] 1.826041 [m^2/m^2]
#> 10     0.000000000 [L/m^2] 1.826041 [m^2/m^2] 0 [m^2/m^2] 1.826041 [m^2/m^2]
#>           LAIexpanded     LAIdead LAImistletoe               Cm   LgroundPAR
#> 1  1.826041 [m^2/m^2] 0 [m^2/m^2]            0 1.517698 [L/m^2] 38.91249 [%]
#> 2  1.826041 [m^2/m^2] 0 [m^2/m^2]            0 1.517698 [L/m^2] 38.91249 [%]
#> 3  1.826041 [m^2/m^2] 0 [m^2/m^2]            0 1.517698 [L/m^2] 38.91249 [%]
#> 4  1.826041 [m^2/m^2] 0 [m^2/m^2]            0 1.517698 [L/m^2] 38.91249 [%]
#> 5  1.826041 [m^2/m^2] 0 [m^2/m^2]            0 1.517698 [L/m^2] 38.91249 [%]
#> 6  1.826041 [m^2/m^2] 0 [m^2/m^2]            0 1.517698 [L/m^2] 38.91249 [%]
#> 7  1.826041 [m^2/m^2] 0 [m^2/m^2]            0 1.517698 [L/m^2] 38.91249 [%]
#> 8  1.826041 [m^2/m^2] 0 [m^2/m^2]            0 1.517698 [L/m^2] 38.91249 [%]
#> 9  1.826041 [m^2/m^2] 0 [m^2/m^2]            0 1.517698 [L/m^2] 38.91249 [%]
#> 10 1.826041 [m^2/m^2] 0 [m^2/m^2]            0 1.517698 [L/m^2] 38.91249 [%]
#>      LgroundSWR            SWE
#> 1  49.70066 [%] 0.0000 [L/m^2]
#> 2  49.70066 [%] 0.0000 [L/m^2]
#> 3  49.70066 [%] 0.0000 [L/m^2]
#> 4  49.70066 [%] 0.0000 [L/m^2]
#> 5  49.70066 [%] 0.0000 [L/m^2]
#> 6  49.70066 [%] 0.0000 [L/m^2]
#> 7  49.70066 [%] 5.3774 [L/m^2]
#> 8  49.70066 [%] 5.3774 [L/m^2]
#> 9  49.70066 [%] 5.3774 [L/m^2]
#> 10 49.70066 [%] 0.0000 [L/m^2]

#Extracts daily soil layer-level output as a data frame
extract(S1, level = "soillayer", addunits = TRUE)
#>          date soillayer       SWC       RWC       REW                ML
#> 1  2001-01-01         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 2  2001-01-02         1 0.3022711 0.9976116 0.9964181  68.01099 [L/m^2]
#> 3  2001-01-03         1 0.2995179 0.9885253 0.9827911  67.39154 [L/m^2]
#> 4  2001-01-04         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 5  2001-01-05         1 0.3014625 0.9949431 0.9924161  67.82907 [L/m^2]
#> 6  2001-01-06         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 7  2001-01-07         1 0.3023870 0.9979944 0.9969922  68.03709 [L/m^2]
#> 8  2001-01-08         1 0.3015967 0.9953860 0.9930803  67.85926 [L/m^2]
#> 9  2001-01-09         1 0.2998639 0.9896669 0.9845032  67.46937 [L/m^2]
#> 10 2001-01-10         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 11 2001-01-01         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 12 2001-01-02         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 13 2001-01-03         2 0.3028143 0.9994046 0.9991070 116.58351 [L/m^2]
#> 14 2001-01-04         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 15 2001-01-05         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 16 2001-01-06         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 17 2001-01-07         2 0.3029382 0.9998134 0.9997202 116.63120 [L/m^2]
#> 18 2001-01-08         2 0.3028616 0.9995605 0.9993408 116.60170 [L/m^2]
#> 19 2001-01-09         2 0.3025978 0.9986900 0.9980354 116.50016 [L/m^2]
#> 20 2001-01-10         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 21 2001-01-01         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 22 2001-01-02         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 23 2001-01-03         3 0.3029286 0.9997818 0.9996728  75.73215 [L/m^2]
#> 24 2001-01-04         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 25 2001-01-05         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 26 2001-01-06         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 27 2001-01-07         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 28 2001-01-08         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 29 2001-01-09         3 0.3028843 0.9996355 0.9994533  75.72107 [L/m^2]
#> 30 2001-01-10         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 31 2001-01-01         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 32 2001-01-02         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 33 2001-01-03         4 0.3028670 0.9995783 0.9993676  30.28670 [L/m^2]
#> 34 2001-01-04         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 35 2001-01-05         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 36 2001-01-06         4 0.3683527 1.2157066 1.3235001  36.83527 [L/m^2]
#> 37 2001-01-07         4 0.3269857 1.0791796 1.1187473  32.69857 [L/m^2]
#> 38 2001-01-08         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 39 2001-01-09         4 0.3027762 0.9992788 0.9989183  30.27762 [L/m^2]
#> 40 2001-01-10         4 0.3380896 1.1158268 1.1737080  33.80896 [L/m^2]
#> 41 2001-01-01   Overall 0.3029947 1.0000000 1.0000000 290.87494 [L/m^2]
#> 42 2001-01-02   Overall 0.3028251 0.9994402 0.9991605 290.71211 [L/m^2]
#> 43 2001-01-03   Overall 0.3020770 0.9969711 0.9954574 289.99390 [L/m^2]
#> 44 2001-01-04   Overall 0.3029947 1.0000000 1.0000000 290.87494 [L/m^2]
#> 45 2001-01-05   Overall 0.3026356 0.9988148 0.9982225 290.53019 [L/m^2]
#> 46 2001-01-06   Overall 0.3098029 1.0224694 1.0336979 297.41074 [L/m^2]
#> 47 2001-01-07   Overall 0.3053287 1.0077030 1.0115523 293.11554 [L/m^2]
#> 48 2001-01-08   Overall 0.3026137 0.9987423 0.9981138 290.50911 [L/m^2]
#> 49 2001-01-09   Overall 0.3020502 0.9968828 0.9953250 289.96822 [L/m^2]
#> 50 2001-01-10   Overall 0.3066504 1.0120653 1.0180946 294.38443 [L/m^2]
#>                  Psi              PlantExt      HydraulicInput
#> 1  -0.03300000 [MPa]  0.1983726725 [L/m^2] 0.000000000 [L/m^2]
#> 2  -0.03337552 [MPa]  0.3679479453 [L/m^2] 0.000000000 [L/m^2]
#> 3  -0.03484341 [MPa]  0.2564321880 [L/m^2] 0.000000000 [L/m^2]
#> 4  -0.03300000 [MPa]  0.1112605361 [L/m^2] 0.000000000 [L/m^2]
#> 5  -0.03380010 [MPa]  0.3766039244 [L/m^2] 0.000000000 [L/m^2]
#> 6  -0.03300000 [MPa]  0.2373576762 [L/m^2] 0.000000000 [L/m^2]
#> 7  -0.03331505 [MPa]  0.1367284850 [L/m^2] 0.000000000 [L/m^2]
#> 8  -0.03372926 [MPa]  0.1778246865 [L/m^2] 0.000000000 [L/m^2]
#> 9  -0.03465549 [MPa]  0.3898945850 [L/m^2] 0.000000000 [L/m^2]
#> 10 -0.03300000 [MPa]  0.1622757075 [L/m^2] 0.000000000 [L/m^2]
#> 11 -0.03300000 [MPa]  0.0387602762 [L/m^2] 0.000000000 [L/m^2]
#> 12 -0.03300000 [MPa]  0.0718937937 [L/m^2] 0.000000000 [L/m^2]
#> 13 -0.03309323 [MPa]  0.0694594280 [L/m^2] 0.000000000 [L/m^2]
#> 14 -0.03300000 [MPa]  0.0307368274 [L/m^2] 0.000000000 [L/m^2]
#> 15 -0.03300000 [MPa]  0.0735850959 [L/m^2] 0.000000000 [L/m^2]
#> 16 -0.03300000 [MPa]  0.0649329291 [L/m^2] 0.000000000 [L/m^2]
#> 17 -0.03302919 [MPa]  0.0217662492 [L/m^2] 0.000000000 [L/m^2]
#> 18 -0.03306879 [MPa]  0.0295065465 [L/m^2] 0.000000000 [L/m^2]
#> 19 -0.03320544 [MPa]  0.1015382875 [L/m^2] 0.000000000 [L/m^2]
#> 20 -0.03300000 [MPa]  0.0433759302 [L/m^2] 0.000000000 [L/m^2]
#> 21 -0.03300000 [MPa]  0.0031558555 [L/m^2] 0.000000000 [L/m^2]
#> 22 -0.03300000 [MPa]  0.0058535812 [L/m^2] 0.000000000 [L/m^2]
#> 23 -0.03303413 [MPa]  0.0165283845 [L/m^2] 0.000000000 [L/m^2]
#> 24 -0.03300000 [MPa]  0.0076703051 [L/m^2] 0.000000000 [L/m^2]
#> 25 -0.03300000 [MPa]  0.0059912868 [L/m^2] 0.000000000 [L/m^2]
#> 26 -0.03300000 [MPa]  0.0153708189 [L/m^2] 0.000000000 [L/m^2]
#> 27 -0.03300000 [MPa] -0.0035537249 [L/m^2] 0.003861640 [L/m^2]
#> 28 -0.03300000 [MPa] -0.0039719816 [L/m^2] 0.004609261 [L/m^2]
#> 29 -0.03305704 [MPa]  0.0276124702 [L/m^2] 0.000000000 [L/m^2]
#> 30 -0.03300000 [MPa]  0.0120198008 [L/m^2] 0.000000000 [L/m^2]
#> 31 -0.03300000 [MPa]  0.0006307718 [L/m^2] 0.000000000 [L/m^2]
#> 32 -0.03300000 [MPa]  0.0011699756 [L/m^2] 0.000000000 [L/m^2]
#> 33 -0.03306600 [MPa]  0.0127769910 [L/m^2] 0.000000000 [L/m^2]
#> 34 -0.03300000 [MPa]  0.0055761451 [L/m^2] 0.000000000 [L/m^2]
#> 35 -0.03300000 [MPa]  0.0011974993 [L/m^2] 0.000000000 [L/m^2]
#> 36 -0.01082795 [MPa]  0.0118608435 [L/m^2] 0.000000000 [L/m^2]
#> 37 -0.02258481 [MPa]  0.0188389967 [L/m^2] 0.000000000 [L/m^2]
#> 38 -0.03300000 [MPa]  0.0236241226 [L/m^2] 0.000000000 [L/m^2]
#> 39 -0.03311296 [MPa]  0.0218533541 [L/m^2] 0.000000000 [L/m^2]
#> 40 -0.01882523 [MPa]  0.0083998431 [L/m^2] 0.000000000 [L/m^2]
#> 41 -0.03300000 [MPa]  0.2409195760 [L/m^2] 0.000000000 [L/m^2]
#> 42 -0.03308801 [MPa]  0.4468652958 [L/m^2] 0.000000000 [L/m^2]
#> 43 -0.03348520 [MPa]  0.3551969915 [L/m^2] 0.000000000 [L/m^2]
#> 44 -0.03300000 [MPa]  0.1552438137 [L/m^2] 0.000000000 [L/m^2]
#> 45 -0.03318752 [MPa]  0.4573778063 [L/m^2] 0.000000000 [L/m^2]
#> 46 -0.03069041 [MPa]  0.3295222677 [L/m^2] 0.000000000 [L/m^2]
#> 47 -0.03200063 [MPa]  0.1737800060 [L/m^2] 0.003861640 [L/m^2]
#> 48 -0.03319851 [MPa]  0.2269833740 [L/m^2] 0.004609261 [L/m^2]
#> 49 -0.03349702 [MPa]  0.5408986968 [L/m^2] 0.000000000 [L/m^2]
#> 50 -0.03152346 [MPa]  0.2260712817 [L/m^2] 0.000000000 [L/m^2]

#Extracts daily cohort-level output as a data frame
extract(S1, level = "cohort", addunits = TRUE)
#>          date cohort           species                  LAI
#> 1  2001-01-01 T1_148  Pinus halepensis 1.20935408 [m^2/m^2]
#> 2  2001-01-02 T1_148  Pinus halepensis 1.20935408 [m^2/m^2]
#> 3  2001-01-03 T1_148  Pinus halepensis 1.20935408 [m^2/m^2]
#> 4  2001-01-04 T1_148  Pinus halepensis 1.20935408 [m^2/m^2]
#> 5  2001-01-05 T1_148  Pinus halepensis 1.20935408 [m^2/m^2]
#> 6  2001-01-06 T1_148  Pinus halepensis 1.20935408 [m^2/m^2]
#> 7  2001-01-07 T1_148  Pinus halepensis 1.20935408 [m^2/m^2]
#> 8  2001-01-08 T1_148  Pinus halepensis 1.20935408 [m^2/m^2]
#> 9  2001-01-09 T1_148  Pinus halepensis 1.20935408 [m^2/m^2]
#> 10 2001-01-10 T1_148  Pinus halepensis 1.20935408 [m^2/m^2]
#> 11 2001-01-01 T2_168      Quercus ilex 0.56790878 [m^2/m^2]
#> 12 2001-01-02 T2_168      Quercus ilex 0.56790878 [m^2/m^2]
#> 13 2001-01-03 T2_168      Quercus ilex 0.56790878 [m^2/m^2]
#> 14 2001-01-04 T2_168      Quercus ilex 0.56790878 [m^2/m^2]
#> 15 2001-01-05 T2_168      Quercus ilex 0.56790878 [m^2/m^2]
#> 16 2001-01-06 T2_168      Quercus ilex 0.56790878 [m^2/m^2]
#> 17 2001-01-07 T2_168      Quercus ilex 0.56790878 [m^2/m^2]
#> 18 2001-01-08 T2_168      Quercus ilex 0.56790878 [m^2/m^2]
#> 19 2001-01-09 T2_168      Quercus ilex 0.56790878 [m^2/m^2]
#> 20 2001-01-10 T2_168      Quercus ilex 0.56790878 [m^2/m^2]
#> 21 2001-01-01 S1_165 Quercus coccifera 0.04877848 [m^2/m^2]
#> 22 2001-01-02 S1_165 Quercus coccifera 0.04877848 [m^2/m^2]
#> 23 2001-01-03 S1_165 Quercus coccifera 0.04877848 [m^2/m^2]
#> 24 2001-01-04 S1_165 Quercus coccifera 0.04877848 [m^2/m^2]
#> 25 2001-01-05 S1_165 Quercus coccifera 0.04877848 [m^2/m^2]
#> 26 2001-01-06 S1_165 Quercus coccifera 0.04877848 [m^2/m^2]
#> 27 2001-01-07 S1_165 Quercus coccifera 0.04877848 [m^2/m^2]
#> 28 2001-01-08 S1_165 Quercus coccifera 0.04877848 [m^2/m^2]
#> 29 2001-01-09 S1_165 Quercus coccifera 0.04877848 [m^2/m^2]
#> 30 2001-01-10 S1_165 Quercus coccifera 0.04877848 [m^2/m^2]
#>                 LAIlive         FPAR AbsorbedSWRFraction       Transpiration
#> 1  1.20935408 [m^2/m^2] 88.57701 [%]           44.432640 0.169631434 [L/m^2]
#> 2  1.20935408 [m^2/m^2] 88.57701 [%]           44.432640 0.314637782 [L/m^2]
#> 3  1.20935408 [m^2/m^2] 88.57701 [%]           44.432640 0.250091911 [L/m^2]
#> 4  1.20935408 [m^2/m^2] 88.57701 [%]           44.432640 0.109302523 [L/m^2]
#> 5  1.20935408 [m^2/m^2] 88.57701 [%]           44.432640 0.322039638 [L/m^2]
#> 6  1.20935408 [m^2/m^2] 88.57701 [%]           44.432640 0.232012134 [L/m^2]
#> 7  1.20935408 [m^2/m^2] 88.57701 [%]           44.432640 0.122357110 [L/m^2]
#> 8  1.20935408 [m^2/m^2] 88.57701 [%]           44.432640 0.159817006 [L/m^2]
#> 9  1.20935408 [m^2/m^2] 88.57701 [%]           44.432640 0.380840716 [L/m^2]
#> 10 1.20935408 [m^2/m^2] 88.57701 [%]           44.432640 0.159171238 [L/m^2]
#> 11 0.56790878 [m^2/m^2] 69.05317 [%]           24.578915 0.062675617 [L/m^2]
#> 12 0.56790878 [m^2/m^2] 69.05317 [%]           24.578915 0.116252729 [L/m^2]
#> 13 0.56790878 [m^2/m^2] 69.05317 [%]           24.578915 0.092407399 [L/m^2]
#> 14 0.56790878 [m^2/m^2] 69.05317 [%]           24.578915 0.040391773 [L/m^2]
#> 15 0.56790878 [m^2/m^2] 69.05317 [%]           24.578915 0.118987576 [L/m^2]
#> 16 0.56790878 [m^2/m^2] 69.05317 [%]           24.578915 0.085730377 [L/m^2]
#> 17 0.56790878 [m^2/m^2] 69.05317 [%]           24.578915 0.045210220 [L/m^2]
#> 18 0.56790878 [m^2/m^2] 69.05317 [%]           24.578915 0.059051908 [L/m^2]
#> 19 0.56790878 [m^2/m^2] 69.05317 [%]           24.578915 0.140721974 [L/m^2]
#> 20 0.56790878 [m^2/m^2] 69.05317 [%]           24.578915 0.058818662 [L/m^2]
#> 21 0.04877848 [m^2/m^2] 39.89406 [%]            3.156504 0.008612525 [L/m^2]
#> 22 0.04877848 [m^2/m^2] 39.89406 [%]            3.156504 0.015974785 [L/m^2]
#> 23 0.04877848 [m^2/m^2] 39.89406 [%]            3.156504 0.012697682 [L/m^2]
#> 24 0.04877848 [m^2/m^2] 39.89406 [%]            3.156504 0.005549517 [L/m^2]
#> 25 0.04877848 [m^2/m^2] 39.89406 [%]            3.156504 0.016350592 [L/m^2]
#> 26 0.04877848 [m^2/m^2] 39.89406 [%]            3.156504 0.011779757 [L/m^2]
#> 27 0.04877848 [m^2/m^2] 39.89406 [%]            3.156504 0.006212676 [L/m^2]
#> 28 0.04877848 [m^2/m^2] 39.89406 [%]            3.156504 0.008114461 [L/m^2]
#> 29 0.04877848 [m^2/m^2] 39.89406 [%]            3.156504 0.019336007 [L/m^2]
#> 30 0.04877848 [m^2/m^2] 39.89406 [%]            3.156504 0.008081382 [L/m^2]
#>    GrossPhotosynthesis          PlantPsi      LeafPLC StemPLC
#> 1   1.61242254 [L/m^2] -0.03300000 [MPa] 0.0000000000       0
#> 2   2.79595687 [L/m^2] -0.03300000 [MPa] 0.0000000000       0
#> 3   2.35779525 [L/m^2] -0.03335674 [MPa] 0.0000000000       0
#> 4   1.08068388 [L/m^2] -0.03475656 [MPa] 0.0000000000       0
#> 5   2.90251792 [L/m^2] -0.03300000 [MPa] 0.0000000000       0
#> 6   2.38003980 [L/m^2] -0.03376019 [MPa] 0.0000000000       0
#> 7   1.36568545 [L/m^2] -0.03299503 [MPa] 0.0000000000       0
#> 8   1.61771289 [L/m^2] -0.03329804 [MPa] 0.0000000000       0
#> 9   2.77613711 [L/m^2] -0.03369614 [MPa] 0.0000000000       0
#> 10  1.54244904 [L/m^2] -0.03458319 [MPa] 0.0000000000       0
#> 11  0.75011437 [L/m^2] -0.03300000 [MPa] 0.0009945146       0
#> 12  1.36999277 [L/m^2] -0.03300000 [MPa] 0.0009945146       0
#> 13  1.10382046 [L/m^2] -0.03318817 [MPa] 0.0010169423       0
#> 14  0.48798768 [L/m^2] -0.03396887 [MPa] 0.0011104407       0
#> 15  1.40669023 [L/m^2] -0.03300000 [MPa] 0.0009945146       0
#> 16  1.04517492 [L/m^2] -0.03340110 [MPa] 0.0010423718       0
#> 17  0.56328471 [L/m^2] -0.03283652 [MPa] 0.0009750646       0
#> 18  0.71754569 [L/m^2] -0.03309120 [MPa] 0.0010053793       0
#> 19  1.59046701 [L/m^2] -0.03339652 [MPa] 0.0010418233       0
#> 20  0.70718101 [L/m^2] -0.03392612 [MPa] 0.0011053024       0
#> 21  0.06854474 [L/m^2] -0.03300000 [MPa] 0.0000000000       0
#> 22  0.12303404 [L/m^2] -0.03300000 [MPa] 0.0000000000       0
#> 23  0.10064778 [L/m^2] -0.03325548 [MPa] 0.0000000000       0
#> 24  0.04504606 [L/m^2] -0.03428379 [MPa] 0.0000000000       0
#> 25  0.12678781 [L/m^2] -0.03300000 [MPa] 0.0000000000       0
#> 26  0.09743214 [L/m^2] -0.03354466 [MPa] 0.0000000000       0
#> 27  0.05371475 [L/m^2] -0.03279444 [MPa] 0.0000000000       0
#> 28  0.06665803 [L/m^2] -0.03311813 [MPa] 0.0000000000       0
#> 29  0.13603484 [L/m^2] -0.03351515 [MPa] 0.0000000000       0
#> 30  0.06493816 [L/m^2] -0.03418719 [MPa] 0.0000000000       0
#>        PlantWaterBalance       LeafRWC       StemRWC          LFMC PlantStress
#> 1  -8.881784e-16 [L/m^2] 0.9981628 [%] 0.9998346 [%] 122.02311 [%] 0.004613739
#> 2   0.000000e+00 [L/m^2] 0.9981628 [%] 0.9998346 [%] 122.02311 [%] 0.004613739
#> 3  -2.218448e-05 [L/m^2] 0.9981429 [%] 0.9998329 [%] 122.02159 [%] 0.004688806
#> 4  -8.705002e-05 [L/m^2] 0.9980650 [%] 0.9998258 [%] 122.01559 [%] 0.004987215
#> 5   1.092345e-04 [L/m^2] 0.9981628 [%] 0.9998346 [%] 122.02311 [%] 0.004613739
#> 6  -4.727401e-05 [L/m^2] 0.9981205 [%] 0.9998308 [%] 122.01986 [%] 0.004774186
#> 7   4.758305e-05 [L/m^2] 0.9981631 [%] 0.9998347 [%] 122.02313 [%] 0.004612696
#> 8  -1.884324e-05 [L/m^2] 0.9981462 [%] 0.9998332 [%] 122.02184 [%] 0.004676427
#> 9  -2.475704e-05 [L/m^2] 0.9981240 [%] 0.9998312 [%] 122.02013 [%] 0.004760599
#> 10 -5.516260e-05 [L/m^2] 0.9980747 [%] 0.9998267 [%] 122.01634 [%] 0.004949928
#> 11  0.000000e+00 [L/m^2] 0.9991208 [%] 0.9997405 [%] 109.14738 [%] 0.006282261
#> 12  0.000000e+00 [L/m^2] 0.9991208 [%] 0.9997405 [%] 109.14738 [%] 0.006282261
#> 13 -1.072842e-06 [L/m^2] 0.9991158 [%] 0.9997390 [%] 109.14691 [%] 0.006323310
#> 14 -4.451059e-06 [L/m^2] 0.9990950 [%] 0.9997329 [%] 109.14494 [%] 0.006493968
#> 15  5.523901e-06 [L/m^2] 0.9991208 [%] 0.9997405 [%] 109.14738 [%] 0.006282261
#> 16 -2.286844e-06 [L/m^2] 0.9991101 [%] 0.9997373 [%] 109.14637 [%] 0.006369800
#> 17  3.218894e-06 [L/m^2] 0.9991251 [%] 0.9997418 [%] 109.14779 [%] 0.006246626
#> 18 -1.452022e-06 [L/m^2] 0.9991184 [%] 0.9997398 [%] 109.14715 [%] 0.006302152
#> 19 -1.740712e-06 [L/m^2] 0.9991102 [%] 0.9997374 [%] 109.14638 [%] 0.006368798
#> 20 -3.019489e-06 [L/m^2] 0.9990961 [%] 0.9997332 [%] 109.14505 [%] 0.006484609
#> 21  0.000000e+00 [L/m^2] 0.9987189 [%] 0.9995007 [%]  98.29725 [%] 0.007310922
#> 22  0.000000e+00 [L/m^2] 0.9987189 [%] 0.9995007 [%]  98.29725 [%] 0.007310922
#> 23 -2.144826e-07 [L/m^2] 0.9987089 [%] 0.9994967 [%]  98.29652 [%] 0.007384316
#> 24 -8.634388e-07 [L/m^2] 0.9986689 [%] 0.9994810 [%]  98.29356 [%] 0.007681368
#> 25  1.077921e-06 [L/m^2] 0.9987189 [%] 0.9995007 [%]  98.29725 [%] 0.007310922
#> 26 -4.572688e-07 [L/m^2] 0.9986977 [%] 0.9994923 [%]  98.29569 [%] 0.007467586
#> 27  6.298313e-07 [L/m^2] 0.9987268 [%] 0.9995038 [%]  98.29784 [%] 0.007251989
#> 28 -2.717307e-07 [L/m^2] 0.9987143 [%] 0.9994989 [%]  98.29691 [%] 0.007344837
#> 29 -3.333209e-07 [L/m^2] 0.9986988 [%] 0.9994928 [%]  98.29577 [%] 0.007459078
#> 30 -5.643111e-07 [L/m^2] 0.9986727 [%] 0.9994825 [%]  98.29384 [%] 0.007653352

#Select the output tables/variables to be extracted
extract(S1, level ="cohort", output="Plants", vars = c("PlantStress", "StemPLC"))
#>          date cohort           species PlantStress StemPLC
#> 1  2001-01-01 T1_148  Pinus halepensis 0.004613739       0
#> 2  2001-01-02 T1_148  Pinus halepensis 0.004613739       0
#> 3  2001-01-03 T1_148  Pinus halepensis 0.004688806       0
#> 4  2001-01-04 T1_148  Pinus halepensis 0.004987215       0
#> 5  2001-01-05 T1_148  Pinus halepensis 0.004613739       0
#> 6  2001-01-06 T1_148  Pinus halepensis 0.004774186       0
#> 7  2001-01-07 T1_148  Pinus halepensis 0.004612696       0
#> 8  2001-01-08 T1_148  Pinus halepensis 0.004676427       0
#> 9  2001-01-09 T1_148  Pinus halepensis 0.004760599       0
#> 10 2001-01-10 T1_148  Pinus halepensis 0.004949928       0
#> 11 2001-01-01 T2_168      Quercus ilex 0.006282261       0
#> 12 2001-01-02 T2_168      Quercus ilex 0.006282261       0
#> 13 2001-01-03 T2_168      Quercus ilex 0.006323310       0
#> 14 2001-01-04 T2_168      Quercus ilex 0.006493968       0
#> 15 2001-01-05 T2_168      Quercus ilex 0.006282261       0
#> 16 2001-01-06 T2_168      Quercus ilex 0.006369800       0
#> 17 2001-01-07 T2_168      Quercus ilex 0.006246626       0
#> 18 2001-01-08 T2_168      Quercus ilex 0.006302152       0
#> 19 2001-01-09 T2_168      Quercus ilex 0.006368798       0
#> 20 2001-01-10 T2_168      Quercus ilex 0.006484609       0
#> 21 2001-01-01 S1_165 Quercus coccifera 0.007310922       0
#> 22 2001-01-02 S1_165 Quercus coccifera 0.007310922       0
#> 23 2001-01-03 S1_165 Quercus coccifera 0.007384316       0
#> 24 2001-01-04 S1_165 Quercus coccifera 0.007681368       0
#> 25 2001-01-05 S1_165 Quercus coccifera 0.007310922       0
#> 26 2001-01-06 S1_165 Quercus coccifera 0.007467586       0
#> 27 2001-01-07 S1_165 Quercus coccifera 0.007251989       0
#> 28 2001-01-08 S1_165 Quercus coccifera 0.007344837       0
#> 29 2001-01-09 S1_165 Quercus coccifera 0.007459078       0
#> 30 2001-01-10 S1_165 Quercus coccifera 0.007653352       0