Extracts model outputs

Function extract() extracts daily or subdaily output and returns it as a tidy data frame.

Usage

extract(
  x,
  level = "forest",
  output = NULL,
  vars = NULL,
  dates = NULL,
  subdaily = FALSE,
  addunits = FALSE
)

Arguments

x

An object returned by simulation functions spwb, aspwb, pwb, growth or fordyn.

level

Level of simulation output, either "forest" (stand-level results), "soillayer" (soil layer-level results), "cohort" (cohort-level results), "sunlitleaf" or "shadeleaf" (leaf-level results)

output

Section of the model output to be explored. See details.

vars

Variables to be extracted (by default, all of them).

dates

A date vector indicating the subset of simulated days for which output is desired.

subdaily

A flag to indicate that subdaily values are desired (see details).

addunits

A flag to indicate that variable units should be added whenever possible.

Value

Function extract() returns a data frame:

If level = "forest", columns are "date" and variable names.

If level = "soillayer", columns are "date", "soillayer" and variable names.

If level = "cohort", level = "sunlitleaf" or level = "shadeleaf", columns are "date", "cohorts", "species" and variable names.

If subdaily = TRUE, columns are "datetime", "cohorts", "species" and variable names.

Details

When subdaily = FALSE, parameter output is used to restrict the section in x where variables are located. For example output = "Plants" will correspond to variables "LAI", "LAIlive", "Transpiration", "StemPLC",... as returned by a call names(x$Plants).

Option subdaily = TRUE only works when simulations have been carried using control option 'subdailyResults = TRUE' (see defaultControl). When using subdaily = TRUE, parameter output is not taken into account, and options for parameter vars are the following:

Variables for level = "forest" or level = "soillayer": Not allowed. An error is raised.

Variables for level = "cohort": "E","Ag","An","dEdP","RootPsi","StemPsi","LeafPsi","StemPLC","StemRWC","LeafRWC","StemSympRWC","LeafSympRWC","PWB".

Variables for level = "shadeleaf" and level="sunlitleaf": "Abs_SWR","Abs_PAR","Net_LWR","E","Ag","An","Ci","Gsw","VPD","Temp","Psi","iWUE".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

#Define soil with default soil params (4 layers)
examplesoil <- defaultSoilParams(4)

#Initialize control parameters
control <- defaultControl("Granier")

#Initialize input
x <- spwbInput(exampleforest,examplesoil, SpParamsMED, control)

#Call simulation function (ten days)
S1<-spwb(x, examplemeteo[1:10, ], latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 4.20666
#> Initial soil water content (mm): 290.875
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:
#> 
#> Final plant water content (mm): 4.20661
#> Final soil water content (mm): 295.017
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): -4.9763e-05
#> Plant water balance result (mm): -4.9763e-05
#> Change in soil water content (mm): 4.14183
#> Soil water balance result (mm): 4.14183
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 39 Rain (mm) 34 Snow (mm) 5
#>   Interception (mm) 7 Net rainfall (mm) 27
#>   Infiltration (mm) 32 Infiltration excess (mm) 0 Saturation excess (mm) 0 Capillarity rise (mm) 0
#>   Soil evaporation (mm) 3  Herbaceous transpiration (mm) 0  Woody plant transpiration (mm) 2  Mistletoe transpiration (mm) 0
#>   Plant extraction from soil (mm) 2  Plant water balance (mm) -0 Hydraulic redistribution (mm) 0
#>   Runoff (mm) 0 Deep drainage (mm) 23

#Extracts daily forest-level output as a data frame
extract(S1, level = "forest", addunits = TRUE)
#>          date               PET     Precipitation              Rain
#> 1  2001-01-01 0.8828475 [L/m^2]  4.869109 [L/m^2]  4.869109 [L/m^2]
#> 2  2001-01-02 1.6375337 [L/m^2]  2.498292 [L/m^2]  2.498292 [L/m^2]
#> 3  2001-01-03 1.3017026 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 4  2001-01-04 0.5690790 [L/m^2]  5.796973 [L/m^2]  5.796973 [L/m^2]
#> 5  2001-01-05 1.6760567 [L/m^2]  1.884401 [L/m^2]  1.884401 [L/m^2]
#> 6  2001-01-06 1.2077028 [L/m^2] 13.359801 [L/m^2] 13.359801 [L/m^2]
#> 7  2001-01-07 0.6368075 [L/m^2]  5.377400 [L/m^2]  0.000000 [L/m^2]
#> 8  2001-01-08 0.8318207 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 9  2001-01-09 1.9823797 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 10 2001-01-10 0.8286873 [L/m^2]  5.117768 [L/m^2]  5.117768 [L/m^2]
#>              Snow            NetRain       Snowmelt       Infiltration
#> 1  0.0000 [L/m^2]  3.7161954 [L/m^2] 0.0000 [L/m^2]  3.7161954 [L/m^2]
#> 2  0.0000 [L/m^2]  1.3602886 [L/m^2] 0.0000 [L/m^2]  1.3602886 [L/m^2]
#> 3  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 4  0.0000 [L/m^2]  4.6539798 [L/m^2] 0.0000 [L/m^2]  4.6539798 [L/m^2]
#> 5  0.0000 [L/m^2]  0.8806518 [L/m^2] 0.0000 [L/m^2]  0.8806518 [L/m^2]
#> 6  0.0000 [L/m^2] 12.0368347 [L/m^2] 0.0000 [L/m^2] 12.0368347 [L/m^2]
#> 7  5.3774 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 8  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 9  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 10 0.0000 [L/m^2]  3.9648643 [L/m^2] 5.3774 [L/m^2]  9.3422640 [L/m^2]
#>    InfiltrationExcess SaturationExcess    Runoff      DeepDrainage
#> 1           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 3.0295684 [L/m^2]
#> 2           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.5141268 [L/m^2]
#> 3           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.0000000 [L/m^2]
#> 4           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 3.6009039 [L/m^2]
#> 5           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.1632798 [L/m^2]
#> 6           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#> 7           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#> 8           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 2.9913325 [L/m^2]
#> 9           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.0000000 [L/m^2]
#> 10          0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#>    CapillarityRise Evapotranspiration     Interception   SoilEvaporation
#> 1        0 [L/m^2]  1.8395405 [L/m^2] 1.152913 [L/m^2] 0.5000000 [L/m^2]
#> 2        0 [L/m^2]  1.9841648 [L/m^2] 1.138003 [L/m^2] 0.5000000 [L/m^2]
#> 3        0 [L/m^2]  0.7751697 [L/m^2] 0.000000 [L/m^2] 0.5000000 [L/m^2]
#> 4        0 [L/m^2]  1.4208993 [L/m^2] 1.142993 [L/m^2] 0.1576421 [L/m^2]
#> 5        0 [L/m^2]  1.8580549 [L/m^2] 1.003750 [L/m^2] 0.5000000 [L/m^2]
#> 6        0 [L/m^2]  1.9597298 [L/m^2] 1.322966 [L/m^2] 0.3814785 [L/m^2]
#> 7        0 [L/m^2]  0.1346184 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 8        0 [L/m^2]  0.1758350 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 9        0 [L/m^2]  0.4190207 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 10       0 [L/m^2]  1.5314260 [L/m^2] 1.152903 [L/m^2] 0.2033817 [L/m^2]
#>    HerbTranspiration   PlantExtraction     Transpiration MistletoeTranspiration
#> 1          0 [L/m^2] 0.1866270 [L/m^2] 0.1866270 [L/m^2]              0 [L/m^2]
#> 2          0 [L/m^2] 0.3461618 [L/m^2] 0.3461618 [L/m^2]              0 [L/m^2]
#> 3          0 [L/m^2] 0.2751697 [L/m^2] 0.2751697 [L/m^2]              0 [L/m^2]
#> 4          0 [L/m^2] 0.1202641 [L/m^2] 0.1202641 [L/m^2]              0 [L/m^2]
#> 5          0 [L/m^2] 0.3543052 [L/m^2] 0.3543052 [L/m^2]              0 [L/m^2]
#> 6          0 [L/m^2] 0.2552853 [L/m^2] 0.2552853 [L/m^2]              0 [L/m^2]
#> 7          0 [L/m^2] 0.1346184 [L/m^2] 0.1346184 [L/m^2]              0 [L/m^2]
#> 8          0 [L/m^2] 0.1758350 [L/m^2] 0.1758350 [L/m^2]              0 [L/m^2]
#> 9          0 [L/m^2] 0.4190207 [L/m^2] 0.4190207 [L/m^2]              0 [L/m^2]
#> 10         0 [L/m^2] 0.1751409 [L/m^2] 0.1751409 [L/m^2]              0 [L/m^2]
#>    HydraulicRedistribution                LAI     LAIherb            LAIlive
#> 1      0.000000000 [L/m^2] 1.451663 [m^2/m^2] 0 [m^2/m^2] 1.451663 [m^2/m^2]
#> 2      0.000000000 [L/m^2] 1.451663 [m^2/m^2] 0 [m^2/m^2] 1.451663 [m^2/m^2]
#> 3      0.000000000 [L/m^2] 1.451663 [m^2/m^2] 0 [m^2/m^2] 1.451663 [m^2/m^2]
#> 4      0.000000000 [L/m^2] 1.451663 [m^2/m^2] 0 [m^2/m^2] 1.451663 [m^2/m^2]
#> 5      0.000000000 [L/m^2] 1.451663 [m^2/m^2] 0 [m^2/m^2] 1.451663 [m^2/m^2]
#> 6      0.000000000 [L/m^2] 1.451663 [m^2/m^2] 0 [m^2/m^2] 1.451663 [m^2/m^2]
#> 7      0.002540942 [L/m^2] 1.451663 [m^2/m^2] 0 [m^2/m^2] 1.451663 [m^2/m^2]
#> 8      0.003148008 [L/m^2] 1.451663 [m^2/m^2] 0 [m^2/m^2] 1.451663 [m^2/m^2]
#> 9      0.000000000 [L/m^2] 1.451663 [m^2/m^2] 0 [m^2/m^2] 1.451663 [m^2/m^2]
#> 10     0.000000000 [L/m^2] 1.451663 [m^2/m^2] 0 [m^2/m^2] 1.451663 [m^2/m^2]
#>           LAIexpanded     LAIdead LAImistletoe               Cm   LgroundPAR
#> 1  1.451663 [m^2/m^2] 0 [m^2/m^2]            0 1.102945 [L/m^2] 46.73377 [%]
#> 2  1.451663 [m^2/m^2] 0 [m^2/m^2]            0 1.102945 [L/m^2] 46.73377 [%]
#> 3  1.451663 [m^2/m^2] 0 [m^2/m^2]            0 1.102945 [L/m^2] 46.73377 [%]
#> 4  1.451663 [m^2/m^2] 0 [m^2/m^2]            0 1.102945 [L/m^2] 46.73377 [%]
#> 5  1.451663 [m^2/m^2] 0 [m^2/m^2]            0 1.102945 [L/m^2] 46.73377 [%]
#> 6  1.451663 [m^2/m^2] 0 [m^2/m^2]            0 1.102945 [L/m^2] 46.73377 [%]
#> 7  1.451663 [m^2/m^2] 0 [m^2/m^2]            0 1.102945 [L/m^2] 46.73377 [%]
#> 8  1.451663 [m^2/m^2] 0 [m^2/m^2]            0 1.102945 [L/m^2] 46.73377 [%]
#> 9  1.451663 [m^2/m^2] 0 [m^2/m^2]            0 1.102945 [L/m^2] 46.73377 [%]
#> 10 1.451663 [m^2/m^2] 0 [m^2/m^2]            0 1.102945 [L/m^2] 46.73377 [%]
#>      LgroundSWR            SWE
#> 1  56.92225 [%] 0.0000 [L/m^2]
#> 2  56.92225 [%] 0.0000 [L/m^2]
#> 3  56.92225 [%] 0.0000 [L/m^2]
#> 4  56.92225 [%] 0.0000 [L/m^2]
#> 5  56.92225 [%] 0.0000 [L/m^2]
#> 6  56.92225 [%] 0.0000 [L/m^2]
#> 7  56.92225 [%] 5.3774 [L/m^2]
#> 8  56.92225 [%] 5.3774 [L/m^2]
#> 9  56.92225 [%] 5.3774 [L/m^2]
#> 10 56.92225 [%] 0.0000 [L/m^2]

#Extracts daily soil layer-level output as a data frame
extract(S1, level = "soillayer", addunits = TRUE)
#>          date soillayer       SWC       RWC       REW                ML
#> 1  2001-01-01         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 2  2001-01-02         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 3  2001-01-03         1 0.2998209 0.9895251 0.9842905  67.45970 [L/m^2]
#> 4  2001-01-04         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 5  2001-01-05         1 0.3023861 0.9979914 0.9969877  68.03688 [L/m^2]
#> 6  2001-01-06         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 7  2001-01-07         1 0.3025521 0.9985391 0.9978090  68.07422 [L/m^2]
#> 8  2001-01-08         1 0.3019739 0.9966310 0.9949474  67.94414 [L/m^2]
#> 9  2001-01-09         1 0.3007151 0.9924763 0.9887166  67.66090 [L/m^2]
#> 10 2001-01-10         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 11 2001-01-01         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 12 2001-01-02         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 13 2001-01-03         2 0.3028509 0.9995252 0.9992880 116.59759 [L/m^2]
#> 14 2001-01-04         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 15 2001-01-05         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 16 2001-01-06         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 17 2001-01-07         2 0.3029373 0.9998104 0.9997156 116.63085 [L/m^2]
#> 18 2001-01-08         2 0.3028602 0.9995560 0.9993341 116.60117 [L/m^2]
#> 19 2001-01-09         2 0.3026159 0.9987497 0.9981249 116.50712 [L/m^2]
#> 20 2001-01-10         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 21 2001-01-01         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 22 2001-01-02         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 23 2001-01-03         3 0.3029758 0.9999377 0.9999065  75.74396 [L/m^2]
#> 24 2001-01-04         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 25 2001-01-05         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 26 2001-01-06         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 27 2001-01-07         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 28 2001-01-08         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 29 2001-01-09         3 0.3028997 0.9996862 0.9995294  75.72491 [L/m^2]
#> 30 2001-01-10         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 31 2001-01-01         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 32 2001-01-02         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 33 2001-01-03         4 0.3029852 0.9999686 0.9999529  30.29852 [L/m^2]
#> 34 2001-01-04         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 35 2001-01-05         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 36 2001-01-06         4 0.3744120 1.2357046 1.3534914  37.44120 [L/m^2]
#> 37 2001-01-07         4 0.3330688 1.0992560 1.1488565  33.30688 [L/m^2]
#> 38 2001-01-08         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 39 2001-01-09         4 0.3028151 0.9994072 0.9991110  30.28151 [L/m^2]
#> 40 2001-01-10         4 0.3444130 1.1366965 1.2050068  34.44130 [L/m^2]
#> 41 2001-01-01   Overall 0.3029947 1.0000000 1.0000000 290.87494 [L/m^2]
#> 42 2001-01-02   Overall 0.3029947 1.0000000 1.0000000 290.87494 [L/m^2]
#> 43 2001-01-03   Overall 0.3021873 0.9973350 0.9960033 290.09977 [L/m^2]
#> 44 2001-01-04   Overall 0.3029947 1.0000000 1.0000000 290.87494 [L/m^2]
#> 45 2001-01-05   Overall 0.3028521 0.9995292 0.9992940 290.73801 [L/m^2]
#> 46 2001-01-06   Overall 0.3104340 1.0245526 1.0368220 298.01666 [L/m^2]
#> 47 2001-01-07   Overall 0.3060007 1.0099207 1.0148783 293.76063 [L/m^2]
#> 48 2001-01-08   Overall 0.3027015 0.9990323 0.9985487 290.59346 [L/m^2]
#> 49 2001-01-09   Overall 0.3022650 0.9975918 0.9963883 290.17444 [L/m^2]
#> 50 2001-01-10   Overall 0.3073091 1.0142392 1.0213549 295.01677 [L/m^2]
#>                   Psi              PlantExt      HydraulicInput
#> 1  -0.033000000 [MPa]  0.1452178891 [L/m^2] 0.000000000 [L/m^2]
#> 2  -0.033000000 [MPa]  0.2693547618 [L/m^2] 0.000000000 [L/m^2]
#> 3  -0.034678774 [MPa]  0.2141145523 [L/m^2] 0.000000000 [L/m^2]
#> 4  -0.033000000 [MPa]  0.0806662947 [L/m^2] 0.000000000 [L/m^2]
#> 5  -0.033315525 [MPa]  0.2756913352 [L/m^2] 0.000000000 [L/m^2]
#> 6  -0.033000000 [MPa]  0.1727400047 [L/m^2] 0.000000000 [L/m^2]
#> 7  -0.033229197 [MPa]  0.0995962794 [L/m^2] 0.000000000 [L/m^2]
#> 8  -0.033530924 [MPa]  0.1300814150 [L/m^2] 0.000000000 [L/m^2]
#> 9  -0.034197357 [MPa]  0.2832393861 [L/m^2] 0.000000000 [L/m^2]
#> 10 -0.033000000 [MPa]  0.1179513642 [L/m^2] 0.000000000 [L/m^2]
#> 11 -0.033000000 [MPa]  0.0375620887 [L/m^2] 0.000000000 [L/m^2]
#> 12 -0.033000000 [MPa]  0.0696713573 [L/m^2] 0.000000000 [L/m^2]
#> 13 -0.033074312 [MPa]  0.0553829135 [L/m^2] 0.000000000 [L/m^2]
#> 14 -0.033000000 [MPa]  0.0285517659 [L/m^2] 0.000000000 [L/m^2]
#> 15 -0.033000000 [MPa]  0.0713103767 [L/m^2] 0.000000000 [L/m^2]
#> 16 -0.033000000 [MPa]  0.0602308704 [L/m^2] 0.000000000 [L/m^2]
#> 17 -0.033029657 [MPa]  0.0221176881 [L/m^2] 0.000000000 [L/m^2]
#> 18 -0.033069497 [MPa]  0.0296804815 [L/m^2] 0.000000000 [L/m^2]
#> 19 -0.033196058 [MPa]  0.0940534092 [L/m^2] 0.000000000 [L/m^2]
#> 20 -0.033000000 [MPa]  0.0399689074 [L/m^2] 0.000000000 [L/m^2]
#> 21 -0.033000000 [MPa]  0.0032018826 [L/m^2] 0.000000000 [L/m^2]
#> 22 -0.033000000 [MPa]  0.0059389537 [L/m^2] 0.000000000 [L/m^2]
#> 23 -0.033009746 [MPa]  0.0047209724 [L/m^2] 0.000000000 [L/m^2]
#> 24 -0.033000000 [MPa]  0.0063493367 [L/m^2] 0.000000000 [L/m^2]
#> 25 -0.033000000 [MPa]  0.0060786676 [L/m^2] 0.000000000 [L/m^2]
#> 26 -0.033000000 [MPa]  0.0129168408 [L/m^2] 0.000000000 [L/m^2]
#> 27 -0.033000000 [MPa] -0.0022236416 [L/m^2] 0.002540942 [L/m^2]
#> 28 -0.033000000 [MPa] -0.0025632028 [L/m^2] 0.003148008 [L/m^2]
#> 29 -0.033049093 [MPa]  0.0237673721 [L/m^2] 0.000000000 [L/m^2]
#> 30 -0.033000000 [MPa]  0.0103247721 [L/m^2] 0.000000000 [L/m^2]
#> 31 -0.033000000 [MPa]  0.0006451879 [L/m^2] 0.000000000 [L/m^2]
#> 32 -0.033000000 [MPa]  0.0011967150 [L/m^2] 0.000000000 [L/m^2]
#> 33 -0.033004909 [MPa]  0.0009512885 [L/m^2] 0.000000000 [L/m^2]
#> 34 -0.033000000 [MPa]  0.0046967237 [L/m^2] 0.000000000 [L/m^2]
#> 35 -0.033000000 [MPa]  0.0012248677 [L/m^2] 0.000000000 [L/m^2]
#> 36 -0.009516594 [MPa]  0.0093976224 [L/m^2] 0.000000000 [L/m^2]
#> 37 -0.020457352 [MPa]  0.0151280676 [L/m^2] 0.000000000 [L/m^2]
#> 38 -0.033000000 [MPa]  0.0186363371 [L/m^2] 0.000000000 [L/m^2]
#> 39 -0.033092808 [MPa]  0.0179605062 [L/m^2] 0.000000000 [L/m^2]
#> 40 -0.016912327 [MPa]  0.0068958348 [L/m^2] 0.000000000 [L/m^2]
#> 41 -0.033000000 [MPa]  0.1866270482 [L/m^2] 0.000000000 [L/m^2]
#> 42 -0.033000000 [MPa]  0.3461617878 [L/m^2] 0.000000000 [L/m^2]
#> 43 -0.033426314 [MPa]  0.2751697267 [L/m^2] 0.000000000 [L/m^2]
#> 44 -0.033000000 [MPa]  0.1202641210 [L/m^2] 0.000000000 [L/m^2]
#> 45 -0.033073951 [MPa]  0.3543052472 [L/m^2] 0.000000000 [L/m^2]
#> 46 -0.030553812 [MPa]  0.2552853383 [L/m^2] 0.000000000 [L/m^2]
#> 47 -0.031759086 [MPa]  0.1346183935 [L/m^2] 0.002540942 [L/m^2]
#> 48 -0.033152307 [MPa]  0.1758350309 [L/m^2] 0.003148008 [L/m^2]
#> 49 -0.033381710 [MPa]  0.4190206736 [L/m^2] 0.000000000 [L/m^2]
#> 50 -0.031324201 [MPa]  0.1751408785 [L/m^2] 0.000000000 [L/m^2]

#Extracts daily cohort-level output as a data frame
extract(S1, level = "cohort", addunits = TRUE)
#>          date cohort           species                  LAI
#> 1  2001-01-01 T1_148  Pinus halepensis 0.75422783 [m^2/m^2]
#> 2  2001-01-02 T1_148  Pinus halepensis 0.75422783 [m^2/m^2]
#> 3  2001-01-03 T1_148  Pinus halepensis 0.75422783 [m^2/m^2]
#> 4  2001-01-04 T1_148  Pinus halepensis 0.75422783 [m^2/m^2]
#> 5  2001-01-05 T1_148  Pinus halepensis 0.75422783 [m^2/m^2]
#> 6  2001-01-06 T1_148  Pinus halepensis 0.75422783 [m^2/m^2]
#> 7  2001-01-07 T1_148  Pinus halepensis 0.75422783 [m^2/m^2]
#> 8  2001-01-08 T1_148  Pinus halepensis 0.75422783 [m^2/m^2]
#> 9  2001-01-09 T1_148  Pinus halepensis 0.75422783 [m^2/m^2]
#> 10 2001-01-10 T1_148  Pinus halepensis 0.75422783 [m^2/m^2]
#> 11 2001-01-01 T2_168      Quercus ilex 0.64411380 [m^2/m^2]
#> 12 2001-01-02 T2_168      Quercus ilex 0.64411380 [m^2/m^2]
#> 13 2001-01-03 T2_168      Quercus ilex 0.64411380 [m^2/m^2]
#> 14 2001-01-04 T2_168      Quercus ilex 0.64411380 [m^2/m^2]
#> 15 2001-01-05 T2_168      Quercus ilex 0.64411380 [m^2/m^2]
#> 16 2001-01-06 T2_168      Quercus ilex 0.64411380 [m^2/m^2]
#> 17 2001-01-07 T2_168      Quercus ilex 0.64411380 [m^2/m^2]
#> 18 2001-01-08 T2_168      Quercus ilex 0.64411380 [m^2/m^2]
#> 19 2001-01-09 T2_168      Quercus ilex 0.64411380 [m^2/m^2]
#> 20 2001-01-10 T2_168      Quercus ilex 0.64411380 [m^2/m^2]
#> 21 2001-01-01 S1_165 Quercus coccifera 0.05332129 [m^2/m^2]
#> 22 2001-01-02 S1_165 Quercus coccifera 0.05332129 [m^2/m^2]
#> 23 2001-01-03 S1_165 Quercus coccifera 0.05332129 [m^2/m^2]
#> 24 2001-01-04 S1_165 Quercus coccifera 0.05332129 [m^2/m^2]
#> 25 2001-01-05 S1_165 Quercus coccifera 0.05332129 [m^2/m^2]
#> 26 2001-01-06 S1_165 Quercus coccifera 0.05332129 [m^2/m^2]
#> 27 2001-01-07 S1_165 Quercus coccifera 0.05332129 [m^2/m^2]
#> 28 2001-01-08 S1_165 Quercus coccifera 0.05332129 [m^2/m^2]
#> 29 2001-01-09 S1_165 Quercus coccifera 0.05332129 [m^2/m^2]
#> 30 2001-01-10 S1_165 Quercus coccifera 0.05332129 [m^2/m^2]
#>                 LAIlive         FPAR AbsorbedSWRFraction       Transpiration
#> 1  0.75422783 [m^2/m^2] 92.71417 [%]           32.805782 0.111616077 [L/m^2]
#> 2  0.75422783 [m^2/m^2] 92.71417 [%]           32.805782 0.207029051 [L/m^2]
#> 3  0.75422783 [m^2/m^2] 92.71417 [%]           32.805782 0.164570814 [L/m^2]
#> 4  0.75422783 [m^2/m^2] 92.71417 [%]           32.805782 0.071922602 [L/m^2]
#> 5  0.75422783 [m^2/m^2] 92.71417 [%]           32.805782 0.211899411 [L/m^2]
#> 6  0.75422783 [m^2/m^2] 92.71417 [%]           32.805782 0.152677001 [L/m^2]
#> 7  0.75422783 [m^2/m^2] 92.71417 [%]           32.805782 0.080509967 [L/m^2]
#> 8  0.75422783 [m^2/m^2] 92.71417 [%]           32.805782 0.105160090 [L/m^2]
#> 9  0.75422783 [m^2/m^2] 92.71417 [%]           32.805782 0.250600080 [L/m^2]
#> 10 0.75422783 [m^2/m^2] 92.71417 [%]           32.805782 0.104743181 [L/m^2]
#> 11 0.64411380 [m^2/m^2] 76.59611 [%]           29.007524 0.065920231 [L/m^2]
#> 12 0.64411380 [m^2/m^2] 76.59611 [%]           29.007524 0.122270941 [L/m^2]
#> 13 0.64411380 [m^2/m^2] 76.59611 [%]           29.007524 0.097195192 [L/m^2]
#> 14 0.64411380 [m^2/m^2] 76.59611 [%]           29.007524 0.042483657 [L/m^2]
#> 15 0.64411380 [m^2/m^2] 76.59611 [%]           29.007524 0.125147366 [L/m^2]
#> 16 0.64411380 [m^2/m^2] 76.59611 [%]           29.007524 0.090173312 [L/m^2]
#> 17 0.64411380 [m^2/m^2] 76.59611 [%]           29.007524 0.047550769 [L/m^2]
#> 18 0.64411380 [m^2/m^2] 76.59611 [%]           29.007524 0.062109726 [L/m^2]
#> 19 0.64411380 [m^2/m^2] 76.59611 [%]           29.007524 0.148010144 [L/m^2]
#> 20 0.64411380 [m^2/m^2] 76.59611 [%]           29.007524 0.061866806 [L/m^2]
#> 21 0.05332129 [m^2/m^2] 48.02393 [%]            3.738549 0.009090741 [L/m^2]
#> 22 0.05332129 [m^2/m^2] 48.02393 [%]            3.738549 0.016861796 [L/m^2]
#> 23 0.05332129 [m^2/m^2] 48.02393 [%]            3.738549 0.013403720 [L/m^2]
#> 24 0.05332129 [m^2/m^2] 48.02393 [%]            3.738549 0.005857862 [L/m^2]
#> 25 0.05332129 [m^2/m^2] 48.02393 [%]            3.738549 0.017258470 [L/m^2]
#> 26 0.05332129 [m^2/m^2] 48.02393 [%]            3.738549 0.012435026 [L/m^2]
#> 27 0.05332129 [m^2/m^2] 48.02393 [%]            3.738549 0.006557657 [L/m^2]
#> 28 0.05332129 [m^2/m^2] 48.02393 [%]            3.738549 0.008565215 [L/m^2]
#> 29 0.05332129 [m^2/m^2] 48.02393 [%]            3.738549 0.020410450 [L/m^2]
#> 30 0.05332129 [m^2/m^2] 48.02393 [%]            3.738549 0.008530891 [L/m^2]
#>    GrossPhotosynthesis          PlantPsi      LeafPLC StemPLC
#> 1   1.08664061 [L/m^2] -0.03300000 [MPa] 0.0000000000       0
#> 2   1.88424572 [L/m^2] -0.03300000 [MPa] 0.0000000000       0
#> 3   1.58908055 [L/m^2] -0.03300000 [MPa] 0.0000000000       0
#> 4   0.72831714 [L/m^2] -0.03459905 [MPa] 0.0000000000       0
#> 5   1.95605913 [L/m^2] -0.03300000 [MPa] 0.0000000000       0
#> 6   1.60410869 [L/m^2] -0.03329973 [MPa] 0.0000000000       0
#> 7   0.92036008 [L/m^2] -0.03299483 [MPa] 0.0000000000       0
#> 8   1.09022479 [L/m^2] -0.03321602 [MPa] 0.0000000000       0
#> 9   1.87096384 [L/m^2] -0.03350771 [MPa] 0.0000000000       0
#> 10  1.03958174 [L/m^2] -0.03414719 [MPa] 0.0000000000       0
#> 11  0.80058242 [L/m^2] -0.03300000 [MPa] 0.0009945146       0
#> 12  1.46216652 [L/m^2] -0.03300000 [MPa] 0.0009945146       0
#> 13  1.17813467 [L/m^2] -0.03300000 [MPa] 0.0009945146       0
#> 14  0.52083035 [L/m^2] -0.03387617 [MPa] 0.0010993002       0
#> 15  1.50133300 [L/m^2] -0.03300000 [MPa] 0.0009945146       0
#> 16  1.11555433 [L/m^2] -0.03315810 [MPa] 0.0010133550       0
#> 17  0.60118391 [L/m^2] -0.03282782 [MPa] 0.0009740307       0
#> 18  0.76583233 [L/m^2] -0.03303264 [MPa] 0.0009984023       0
#> 19  1.69751135 [L/m^2] -0.03329737 [MPa] 0.0010299765       0
#> 20  0.75479949 [L/m^2] -0.03369136 [MPa] 0.0010771228       0
#> 21  0.07740823 [L/m^2] -0.03300000 [MPa] 0.0000000000       0
#> 22  0.13894350 [L/m^2] -0.03300000 [MPa] 0.0000000000       0
#> 23  0.11367089 [L/m^2] -0.03300000 [MPa] 0.0000000000       0
#> 24  0.05087271 [L/m^2] -0.03416471 [MPa] 0.0000000000       0
#> 25  0.14318268 [L/m^2] -0.03300000 [MPa] 0.0000000000       0
#> 26  0.11004157 [L/m^2] -0.03321465 [MPa] 0.0000000000       0
#> 27  0.06066075 [L/m^2] -0.03278465 [MPa] 0.0000000000       0
#> 28  0.07527924 [L/m^2] -0.03304000 [MPa] 0.0000000000       0
#> 29  0.15363144 [L/m^2] -0.03338026 [MPa] 0.0000000000       0
#> 30  0.07334203 [L/m^2] -0.03387158 [MPa] 0.0000000000       0
#>        PlantWaterBalance       LeafRWC       StemRWC          LFMC PlantStress
#> 1  -4.440892e-16 [L/m^2] 0.9981628 [%] 0.9998346 [%] 122.02311 [%] 0.004613739
#> 2   0.000000e+00 [L/m^2] 0.9981628 [%] 0.9998346 [%] 122.02311 [%] 0.004613739
#> 3   0.000000e+00 [L/m^2] 0.9981628 [%] 0.9998346 [%] 122.02311 [%] 0.004613739
#> 4  -6.201676e-05 [L/m^2] 0.9980738 [%] 0.9998266 [%] 122.01627 [%] 0.004953334
#> 5   6.201676e-05 [L/m^2] 0.9981628 [%] 0.9998346 [%] 122.02311 [%] 0.004613739
#> 6  -1.162484e-05 [L/m^2] 0.9981461 [%] 0.9998331 [%] 122.02183 [%] 0.004676784
#> 7   1.182541e-05 [L/m^2] 0.9981631 [%] 0.9998347 [%] 122.02314 [%] 0.004612654
#> 8  -8.578641e-06 [L/m^2] 0.9981508 [%] 0.9998336 [%] 122.02219 [%] 0.004659147
#> 9  -1.131298e-05 [L/m^2] 0.9981345 [%] 0.9998321 [%] 122.02094 [%] 0.004720696
#> 10 -2.480140e-05 [L/m^2] 0.9980989 [%] 0.9998289 [%] 122.01820 [%] 0.004856564
#> 11  0.000000e+00 [L/m^2] 0.9991208 [%] 0.9997405 [%] 109.14738 [%] 0.006282261
#> 12  0.000000e+00 [L/m^2] 0.9991208 [%] 0.9997405 [%] 109.14738 [%] 0.006282261
#> 13  0.000000e+00 [L/m^2] 0.9991208 [%] 0.9997405 [%] 109.14738 [%] 0.006282261
#> 14 -5.665647e-06 [L/m^2] 0.9990975 [%] 0.9997336 [%] 109.14518 [%] 0.006473673
#> 15  5.665647e-06 [L/m^2] 0.9991208 [%] 0.9997405 [%] 109.14738 [%] 0.006282261
#> 16 -1.022328e-06 [L/m^2] 0.9991166 [%] 0.9997392 [%] 109.14698 [%] 0.006316747
#> 17  2.135686e-06 [L/m^2] 0.9991254 [%] 0.9997418 [%] 109.14781 [%] 0.006244731
#> 18 -1.324447e-06 [L/m^2] 0.9991199 [%] 0.9997402 [%] 109.14730 [%] 0.006289380
#> 19 -1.711825e-06 [L/m^2] 0.9991129 [%] 0.9997381 [%] 109.14663 [%] 0.006347146
#> 20 -2.547715e-06 [L/m^2] 0.9991024 [%] 0.9997350 [%] 109.14564 [%] 0.006433241
#> 21  0.000000e+00 [L/m^2] 0.9987189 [%] 0.9995007 [%]  98.29725 [%] 0.007310922
#> 22  0.000000e+00 [L/m^2] 0.9987189 [%] 0.9995007 [%]  98.29725 [%] 0.007310922
#> 23  0.000000e+00 [L/m^2] 0.9987189 [%] 0.9995007 [%]  98.29725 [%] 0.007310922
#> 24 -1.068995e-06 [L/m^2] 0.9986736 [%] 0.9994828 [%]  98.29391 [%] 0.007646834
#> 25  1.068995e-06 [L/m^2] 0.9987189 [%] 0.9995007 [%]  98.29725 [%] 0.007310922
#> 26 -1.969807e-07 [L/m^2] 0.9987105 [%] 0.9994974 [%]  98.29664 [%] 0.007372574
#> 27  3.945951e-07 [L/m^2] 0.9987272 [%] 0.9995039 [%]  98.29787 [%] 0.007249185
#> 28 -2.343208e-07 [L/m^2] 0.9987173 [%] 0.9995000 [%]  98.29714 [%] 0.007322402
#> 29 -3.122650e-07 [L/m^2] 0.9987041 [%] 0.9994948 [%]  98.29616 [%] 0.007420220
#> 30 -4.509445e-07 [L/m^2] 0.9986850 [%] 0.9994873 [%]  98.29475 [%] 0.007561974

#Select the output tables/variables to be extracted
extract(S1, level ="cohort", output="Plants", vars = c("PlantStress", "StemPLC"))
#>          date cohort           species PlantStress StemPLC
#> 1  2001-01-01 T1_148  Pinus halepensis 0.004613739       0
#> 2  2001-01-02 T1_148  Pinus halepensis 0.004613739       0
#> 3  2001-01-03 T1_148  Pinus halepensis 0.004613739       0
#> 4  2001-01-04 T1_148  Pinus halepensis 0.004953334       0
#> 5  2001-01-05 T1_148  Pinus halepensis 0.004613739       0
#> 6  2001-01-06 T1_148  Pinus halepensis 0.004676784       0
#> 7  2001-01-07 T1_148  Pinus halepensis 0.004612654       0
#> 8  2001-01-08 T1_148  Pinus halepensis 0.004659147       0
#> 9  2001-01-09 T1_148  Pinus halepensis 0.004720696       0
#> 10 2001-01-10 T1_148  Pinus halepensis 0.004856564       0
#> 11 2001-01-01 T2_168      Quercus ilex 0.006282261       0
#> 12 2001-01-02 T2_168      Quercus ilex 0.006282261       0
#> 13 2001-01-03 T2_168      Quercus ilex 0.006282261       0
#> 14 2001-01-04 T2_168      Quercus ilex 0.006473673       0
#> 15 2001-01-05 T2_168      Quercus ilex 0.006282261       0
#> 16 2001-01-06 T2_168      Quercus ilex 0.006316747       0
#> 17 2001-01-07 T2_168      Quercus ilex 0.006244731       0
#> 18 2001-01-08 T2_168      Quercus ilex 0.006289380       0
#> 19 2001-01-09 T2_168      Quercus ilex 0.006347146       0
#> 20 2001-01-10 T2_168      Quercus ilex 0.006433241       0
#> 21 2001-01-01 S1_165 Quercus coccifera 0.007310922       0
#> 22 2001-01-02 S1_165 Quercus coccifera 0.007310922       0
#> 23 2001-01-03 S1_165 Quercus coccifera 0.007310922       0
#> 24 2001-01-04 S1_165 Quercus coccifera 0.007646834       0
#> 25 2001-01-05 S1_165 Quercus coccifera 0.007310922       0
#> 26 2001-01-06 S1_165 Quercus coccifera 0.007372574       0
#> 27 2001-01-07 S1_165 Quercus coccifera 0.007249185       0
#> 28 2001-01-08 S1_165 Quercus coccifera 0.007322402       0
#> 29 2001-01-09 S1_165 Quercus coccifera 0.007420220       0
#> 30 2001-01-10 S1_165 Quercus coccifera 0.007561974       0