Extracts model outputs

Function extract() extracts daily or subdaily output and returns it as a tidy data frame.

extract(
  x,
  level = "forest",
  output = NULL,
  vars = NULL,
  dates = NULL,
  subdaily = FALSE
)

Arguments

x

An object returned by simulation functions spwb, pwb or growth.

level

Level of simulation output, either "forest" (stand-level results), "soillayer" (soil layer-level results), "cohort" (cohort-level results), "sunlitleaf" or "shadeleaf" (leaf-level results)

output

Section of the model output to be explored. See details.

vars

Variables to be extracted (by default, all of them).

dates

A date vector indicating the subset of simulated days for which output is desired.

subdaily

A flag to indicate that subdaily values are desired (see details).

Value

Function extract() returns a data frame:

If level = "forest", columns are "date" and variable names.

If level = "soillayer", columns are "date", "soillayer" and variable names.

If level = "cohort", level = "sunlitleaf" or level = "shadeleaf", columns are "date", "cohorts", "species" and variable names.

If subdaily = TRUE, columns are "datetime", "cohorts", "species" and variable names.

Details

When subdaily = FALSE, parameter output is used to restrict the section in x where variables are located. For example output = "Plants" will correspond to variables "LAI", "LAIlive", "Transpiration", "StemPLC",... as returned by a call names(x$Plants).

Option subdaily = TRUE only works when simulations have been carried using control option 'subdailyResults = TRUE' (see defaultControl). When using subdaily = TRUE, parameter output is not taken into account, and options for parameter vars are the following:

Variables for level = "forest" or level = "soillayer": Not allowed. An error is raised.

Variables for level = "cohort": "E","Ag","An","dEdP","RootPsi","StemPsi","LeafPsi","StemPLC","StemRWC","LeafRWC","StemSympRWC","LeafSympRWC","PWB".

Variables for level = "shadeleaf" and level="sunlitleaf": "Abs_SWR","Abs_PAR","Net_LWR","E","Ag","An","Ci","Gsw","VPD","Temp","Psi","iWUE".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

#Define soil with default soil params (4 layers)
examplesoil <- defaultSoilParams(4)

#Initialize control parameters
control <- defaultControl("Granier")

#Initialize input
x <- spwbInput(exampleforest,examplesoil, SpParamsMED, control)

#Call simulation function (ten days)
S1<-spwb(x, examplemeteo[1:10, ], latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 7.05859
#> Initial soil water content (mm): 290.875
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:
#> 
#> Final plant water content (mm): 7.05979
#> Final soil water content (mm): 318.295
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): 0.00119168
#> Plant water balance result (mm): 0.00119168
#> Change in soil water content (mm): 27.42
#> Soil water balance result (mm): 27.42
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 39 Rain (mm) 34 Snow (mm) 5
#>   Interception (mm) 9 Net rainfall (mm) 25
#>   Infiltration (mm) 30 Infiltration excess (mm) 0 Saturation excess (mm) 0 Capillarity rise (mm) 0
#>   Soil evaporation (mm) 1  Herbaceous transpiration (mm) 0 Woody plant transpiration (mm) 2
#>   Plant extraction from soil (mm) 2  Plant water balance (mm) 0 Hydraulic redistribution (mm) 0
#>   Runoff (mm) 0 Deep drainage (mm) 0

#Extracts daily forest-level output as a data frame
extract(S1, level = "forest")
#>          date       PET Precipitation      Rain   Snow    NetRain Snowmelt
#> 1  2001-01-01 0.8828475      4.869109  4.869109 0.0000  3.4241795   0.0000
#> 2  2001-01-02 1.6375337      2.498292  2.498292 0.0000  1.0717469   0.0000
#> 3  2001-01-03 1.3017026      0.000000  0.000000 0.0000  0.0000000   0.0000
#> 4  2001-01-04 0.5690790      5.796973  5.796973 0.0000  4.3625616   0.0000
#> 5  2001-01-05 1.6760567      1.884401  1.884401 0.0000  0.7539027   0.0000
#> 6  2001-01-06 1.2077028     13.359801 13.359801 0.0000 11.7240275   0.0000
#> 7  2001-01-07 0.6368075      5.377400  0.000000 5.3774  0.0000000   0.0000
#> 8  2001-01-08 0.8318207      0.000000  0.000000 0.0000  0.0000000   0.0000
#> 9  2001-01-09 1.9823797      0.000000  0.000000 0.0000  0.0000000   0.0000
#> 10 2001-01-10 0.8286873      5.117768  5.117768 0.0000  3.6727363   5.3774
#>    Infiltration InfiltrationExcess SaturationExcess Runoff DeepDrainage
#> 1     3.4241795                  0                0      0   0.02167253
#> 2     1.0717469                  0                0      0   0.02166566
#> 3     0.0000000                  0                0      0   0.02160841
#> 4     4.3625616                  0                0      0   0.02075411
#> 5     0.7539027                  0                0      0   0.02742466
#> 6    11.7240275                  0                0      0   0.02130238
#> 7     0.0000000                  0                0      0   0.00000000
#> 8     0.0000000                  0                0      0   0.00000000
#> 9     0.0000000                  0                0      0   0.00000000
#> 10    9.0501360                  0                0      0   0.04149520
#>    CapillarityRise Evapotranspiration Interception SoilEvaporation
#> 1     5.999741e-06          2.0551277     1.444929     0.447894750
#> 2     5.741773e-06          1.7785732     1.426545     0.050694125
#> 3     5.596839e-06          0.2853870     0.000000     0.045824404
#> 4     5.572850e-06          1.6005976     1.434411     0.061477811
#> 5     5.727266e-06          1.4615643     1.130499     0.022325301
#> 6     3.770931e-06          1.8824594     1.635773     0.024234290
#> 7     2.962100e-06          0.1175216     0.000000     0.000000000
#> 8     2.967549e-06          0.1535085     0.000000     0.000000000
#> 9     3.076676e-06          0.3658307     0.000000     0.000000000
#> 10    4.581836e-06          1.6072699     1.445031     0.009319572
#>    HerbTranspiration PlantExtraction Transpiration HydraulicRedistribution
#> 1        0.011023432      0.15128020    0.15128020             0.000000000
#> 2        0.020448591      0.28088577    0.28088577             0.010721158
#> 3        0.016255064      0.22330751    0.22330751             0.008215229
#> 4        0.007106228      0.09760227    0.09760227             0.002838363
#> 5        0.020931441      0.28780874    0.28780874             0.012742884
#> 6        0.015082251      0.20736968    0.20736968             0.009561415
#> 7        0.007953664      0.10956797    0.10956797             0.006931282
#> 8        0.010389352      0.14311911    0.14311911             0.009041278
#> 9        0.024759685      0.34107097    0.34107097             0.021507155
#> 10       0.010350177      0.14256876    0.14256876             0.008949768
#>         LAI   LAIherb  LAIlive LAIexpanded LAIdead       Cm LgroundPAR
#> 1  1.758585 0.1736369 1.584948    1.584948       0 1.390485   40.00754
#> 2  1.758585 0.1736369 1.584948    1.584948       0 1.390485   40.00754
#> 3  1.758585 0.1736369 1.584948    1.584948       0 1.390485   40.00754
#> 4  1.758585 0.1736369 1.584948    1.584948       0 1.390485   40.00754
#> 5  1.758585 0.1736369 1.584948    1.584948       0 1.390485   40.00754
#> 6  1.758585 0.1736369 1.584948    1.584948       0 1.390485   40.00754
#> 7  1.758585 0.1736369 1.584948    1.584948       0 1.390485   40.00754
#> 8  1.758585 0.1736369 1.584948    1.584948       0 1.390485   40.00754
#> 9  1.758585 0.1736369 1.584948    1.584948       0 1.390485   40.00754
#> 10 1.758585 0.1736369 1.584948    1.584948       0 1.390485   40.00754
#>    LgroundSWR    SWE
#> 1    50.73297 0.0000
#> 2    50.73297 0.0000
#> 3    50.73297 0.0000
#> 4    50.73297 0.0000
#> 5    50.73297 0.0000
#> 6    50.73297 0.0000
#> 7    50.73297 5.3774
#> 8    50.73297 5.3774
#> 9    50.73297 5.3774
#> 10   50.73297 0.0000

#Extracts daily soil layer-level output as a data frame
extract(S1, level = "soillayer")
#>          date soillayer       SWC      RWC      REW        ML         Psi
#> 1  2001-01-01         1 0.3138410 1.035797 1.053686  70.61424 -0.02783926
#> 2  2001-01-02         1 0.3150165 1.039676 1.059504  70.87871 -0.02732832
#> 3  2001-01-03         1 0.3118948 1.029374 1.044052  70.17633 -0.02870483
#> 4  2001-01-04         1 0.3278297 1.081965 1.122925  73.76168 -0.02227916
#> 5  2001-01-05         1 0.3258652 1.075482 1.113201  73.31968 -0.02299605
#> 6  2001-01-06         1 0.3653612 1.205833 1.308693  82.20626 -0.01150463
#> 7  2001-01-07         1 0.3648483 1.204141 1.306155  82.09088 -0.01162271
#> 8  2001-01-08         1 0.3641790 1.201932 1.302842  81.94028 -0.01177776
#> 9  2001-01-09         1 0.3625857 1.196673 1.294955  81.58179 -0.01215123
#> 10 2001-01-10         1 0.3707174 1.223511 1.335205  83.41141 -0.01030711
#> 11 2001-01-01         2 0.3033553 1.001190 1.001785 116.79179 -0.03281445
#> 12 2001-01-02         2 0.3039073 1.003012 1.004517 117.00433 -0.03253238
#> 13 2001-01-03         2 0.3043846 1.004587 1.006880 117.18809 -0.03229041
#> 14 2001-01-04         2 0.3053500 1.007773 1.011658 117.55973 -0.03180642
#> 15 2001-01-05         2 0.3068314 1.012662 1.018990 118.13008 -0.03107738
#> 16 2001-01-06         2 0.3127294 1.032128 1.048184 120.40084 -0.02833060
#> 17 2001-01-07         2 0.3127063 1.032052 1.048069 120.39191 -0.02834093
#> 18 2001-01-08         2 0.3126757 1.031951 1.047918 120.38015 -0.02835456
#> 19 2001-01-09         2 0.3126019 1.031707 1.047552 120.35174 -0.02838748
#> 20 2001-01-10         2 0.3296572 1.087996 1.131970 126.91800 -0.02162917
#> 21 2001-01-01         3 0.3035010 1.001671 1.002506  75.87526 -0.03273975
#> 22 2001-01-02         3 0.3040208 1.003386 1.005079  76.00520 -0.03247470
#> 23 2001-01-03         3 0.3045235 1.005046 1.007567  76.13089 -0.03222033
#> 24 2001-01-04         3 0.3050522 1.006790 1.010184  76.26304 -0.03195496
#> 25 2001-01-05         3 0.3057049 1.008945 1.013415  76.42623 -0.03163024
#> 26 2001-01-06         3 0.3064843 1.011517 1.017272  76.62109 -0.03124670
#> 27 2001-01-07         3 0.3064974 1.011560 1.017337  76.62434 -0.03124032
#> 28 2001-01-08         3 0.3065144 1.011616 1.017421  76.62860 -0.03123200
#> 29 2001-01-09         3 0.3065549 1.011750 1.017622  76.63872 -0.03121221
#> 30 2001-01-10         3 0.3079617 1.016393 1.024585  76.99043 -0.03053202
#> 31 2001-01-01         4 0.3038596 1.002855 1.004281  30.38596 -0.03255666
#> 32 2001-01-02         4 0.3047708 1.005862 1.008791  30.47708 -0.03209594
#> 33 2001-01-03         4 0.3056302 1.008698 1.013045  30.56302 -0.03166724
#> 34 2001-01-04         4 0.3064950 1.011552 1.017325  30.64950 -0.03124150
#> 35 2001-01-05         4 0.3075338 1.014981 1.022467  30.75338 -0.03073739
#> 36 2001-01-06         4 0.3085722 1.018408 1.027607  30.85722 -0.03024130
#> 37 2001-01-07         4 0.3086076 1.018525 1.027782  30.86076 -0.03022456
#> 38 2001-01-08         4 0.3086536 1.018677 1.028010  30.86536 -0.03020275
#> 39 2001-01-09         4 0.3087631 1.019038 1.028551  30.87631 -0.03015098
#> 40 2001-01-10         4 0.3097512 1.022299 1.033442  30.97512 -0.02968740
#> 41 2001-01-01   Overall 0.3059034 1.009600 1.014397 293.66725 -0.03160209
#> 42 2001-01-02   Overall 0.3066305 1.012000 1.017996 294.36531 -0.03125219
#> 43 2001-01-03   Overall 0.3063108 1.010944 1.016413 294.05832 -0.03136687
#> 44 2001-01-04   Overall 0.3106604 1.025300 1.037942 298.23395 -0.02955331
#> 45 2001-01-05   Overall 0.3110723 1.026659 1.039981 298.62937 -0.02929188
#> 46 2001-01-06   Overall 0.3230056 1.066044 1.099047 310.08541 -0.02534544
#> 47 2001-01-07   Overall 0.3228832 1.065640 1.098441 309.96789 -0.02537386
#> 48 2001-01-08   Overall 0.3227233 1.065112 1.097650 309.81439 -0.02541122
#> 49 2001-01-09   Overall 0.3223422 1.063854 1.095764 309.44856 -0.02550141
#> 50 2001-01-10   Overall 0.3315573 1.094267 1.141375 318.29497 -0.02213341
#>         PlantExt HydraulicInput
#> 1   0.1101121597    0.000000000
#> 2   0.2454061071    0.000000000
#> 3   0.1956134822    0.000000000
#> 4   0.0840506863    0.000000000
#> 5   0.2636754886    0.000000000
#> 6   0.1875466338    0.000000000
#> 7   0.1081409219    0.000000000
#> 8   0.1411291223    0.000000000
#> 9   0.3359328945    0.000000000
#> 10  0.1400195775    0.000000000
#> 11  0.0364746151    0.000000000
#> 12  0.0461328545    0.000000000
#> 13  0.0359092529    0.000000000
#> 14  0.0163899505    0.000000000
#> 15  0.0368761313    0.000000000
#> 16  0.0293844626    0.000000000
#> 17  0.0083583345    0.000000000
#> 18  0.0110312678    0.000000000
#> 19  0.0266452270    0.000000000
#> 20  0.0114989480    0.000000000
#> 21  0.0037576026    0.000000000
#> 22 -0.0038537232    0.003921687
#> 23 -0.0035360586    0.003536059
#> 24 -0.0018307621    0.001830762
#> 25 -0.0059051442    0.005905144
#> 26 -0.0048320365    0.004832037
#> 27 -0.0033571965    0.003357197
#> 28 -0.0043819270    0.004381927
#> 29 -0.0104324975    0.010432498
#> 30 -0.0043508297    0.004350830
#> 31  0.0009358264    0.000000000
#> 32 -0.0067994712    0.006799471
#> 33 -0.0046791706    0.004679171
#> 34 -0.0010076006    0.001007601
#> 35 -0.0068377402    0.006837740
#> 36 -0.0047293786    0.004729379
#> 37 -0.0035740857    0.003574086
#> 38 -0.0046593512    0.004659351
#> 39 -0.0110746574    0.011074657
#> 40 -0.0045989388    0.004598939
#> 41  0.1512802038    0.000000000
#> 42  0.2808857673    0.010721158
#> 43  0.2233075059    0.008215229
#> 44  0.0976022741    0.002838363
#> 45  0.2878087355    0.012742884
#> 46  0.2073696812    0.009561415
#> 47  0.1095679741    0.006931282
#> 48  0.1431191119    0.009041278
#> 49  0.3410709667    0.021507155
#> 50  0.1425687569    0.008949768

#Extracts daily cohort-level output as a data frame
extract(S1, level = "cohort")
#>          date cohort           species        LAI    LAIlive     FPAR
#> 1  2001-01-01 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 2  2001-01-02 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 3  2001-01-03 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 4  2001-01-04 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 5  2001-01-05 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 6  2001-01-06 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 7  2001-01-07 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 8  2001-01-08 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 9  2001-01-09 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 10 2001-01-10 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 11 2001-01-01 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 12 2001-01-02 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 13 2001-01-03 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 14 2001-01-04 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 15 2001-01-05 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 16 2001-01-06 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 17 2001-01-07 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 18 2001-01-08 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 19 2001-01-09 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 20 2001-01-10 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 21 2001-01-01 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 22 2001-01-02 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 23 2001-01-03 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 24 2001-01-04 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 25 2001-01-05 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 26 2001-01-06 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 27 2001-01-07 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 28 2001-01-08 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 29 2001-01-09 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 30 2001-01-10 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#>    AbsorbedSWRFraction Transpiration GrossPhotosynthesis    PlantPsi
#> 1            35.076344   0.092472560          0.89344066 -0.03300000
#> 2            35.076344   0.171712682          1.56154413 -0.02855440
#> 3            35.076344   0.136512980          1.30921526 -0.02807734
#> 4            35.076344   0.059664354          0.59733741 -0.02921278
#> 5            35.076344   0.175950322          1.62051413 -0.02370991
#> 6            35.076344   0.126769149          1.31231012 -0.02420353
#> 7            35.076344   0.066975457          0.74879671 -0.01443139
#> 8            35.076344   0.087484308          0.89472720 -0.01452427
#> 9            35.076344   0.208486477          1.59053189 -0.01464649
#> 10           35.076344   0.087148295          0.85678701 -0.01494199
#> 11           30.444383   0.052754988          0.62572744 -0.03300000
#> 12           30.444383   0.097941652          1.14626945 -0.03032293
#> 13           30.444383   0.077866017          0.92203082 -0.02992612
#> 14           30.444383   0.034034936          0.40699100 -0.03048953
#> 15           30.444383   0.100357443          1.17777424 -0.02705955
#> 16           30.444383   0.072313594          0.87155230 -0.02706646
#> 17           30.444383   0.038219101          0.46940802 -0.02014677
#> 18           30.444383   0.049922133          0.59995572 -0.02020831
#> 19           30.444383   0.118970291          1.34297955 -0.02028919
#> 20           30.444383   0.049729501          0.59218173 -0.02048424
#> 21            2.366131   0.006052657          0.05005525 -0.03300000
#> 22            2.366131   0.011231433          0.08988462 -0.02945259
#> 23            2.366131   0.008928509          0.07353916 -0.02901651
#> 24            2.366131   0.003902984          0.03291989 -0.02985675
#> 25            2.366131   0.011500971          0.09266985 -0.02543988
#> 26            2.366131   0.008286939          0.07122296 -0.02567983
#> 27            2.366131   0.004373416          0.03929124 -0.01754495
#> 28            2.366131   0.005712671          0.04876317 -0.01761859
#> 29            2.366131   0.013614198          0.09952573 -0.01771548
#> 30            2.366131   0.005690961          0.04751819 -0.01794979
#>         LeafPLC      StemPLC PlantWaterBalance   LeafRWC   StemRWC      LFMC
#> 1  3.713826e-03 1.320577e-11      0.000000e+00 0.9979679 0.9998350 125.89003
#> 2  3.713826e-03 1.320577e-11      2.436250e-04 0.9982415 0.9998572 125.90897
#> 3  3.713826e-03 1.320577e-11      2.614456e-05 0.9982709 0.9998596 125.91100
#> 4  3.713826e-03 1.320577e-11     -6.222653e-05 0.9982010 0.9998539 125.90616
#> 5  3.713826e-03 1.320577e-11      3.015949e-04 0.9985398 0.9998814 125.92960
#> 6  3.713826e-03 1.320577e-11     -2.705538e-05 0.9985094 0.9998789 125.92750
#> 7  3.713826e-03 1.320577e-11      5.356686e-04 0.9991111 0.9999278 125.96913
#> 8  3.713826e-03 1.320577e-11     -5.092210e-06 0.9991054 0.9999274 125.96873
#> 9  3.713826e-03 1.320577e-11     -6.700399e-06 0.9990978 0.9999267 125.96821
#> 10 3.713826e-03 1.320577e-11     -1.619983e-05 0.9990797 0.9999253 125.96695
#> 11 3.473351e-05 1.691266e-06      0.000000e+00 0.9986794 0.9997489  93.07448
#> 12 3.473351e-05 1.691266e-06      3.600254e-05 0.9987866 0.9997694  93.08086
#> 13 3.473351e-05 1.691266e-06      5.333817e-06 0.9988025 0.9997724  93.08181
#> 14 3.473351e-05 1.691266e-06     -7.573469e-06 0.9987799 0.9997681  93.08047
#> 15 3.473351e-05 1.691266e-06      4.608669e-05 0.9989172 0.9997944  93.08864
#> 16 3.473351e-05 1.691266e-06     -9.274178e-08 0.9989169 0.9997943  93.08862
#> 17 3.473351e-05 1.691266e-06      9.284263e-05 0.9991939 0.9998471  93.10510
#> 18 3.473351e-05 1.691266e-06     -8.249954e-07 0.9991914 0.9998466  93.10495
#> 19 3.473351e-05 1.691266e-06     -1.084282e-06 0.9991882 0.9998460  93.10476
#> 20 3.473351e-05 1.691266e-06     -2.614898e-06 0.9991803 0.9998445  93.10430
#> 21 3.473351e-05 4.394058e-08      0.000000e+00 0.9987207 0.9983684  96.39176
#> 22 3.473351e-05 4.394058e-08      7.974930e-06 0.9988582 0.9985437  96.40709
#> 23 3.473351e-05 4.394058e-08      9.804294e-07 0.9988751 0.9985652  96.40897
#> 24 3.473351e-05 4.394058e-08     -1.889066e-06 0.9988425 0.9985237  96.40534
#> 25 3.473351e-05 4.394058e-08      9.930824e-06 0.9990137 0.9987420  96.42443
#> 26 3.473351e-05 4.394058e-08     -5.395342e-07 0.9990044 0.9987301  96.42339
#> 27 3.473351e-05 4.394058e-08      1.829435e-05 0.9993198 0.9991322  96.45855
#> 28 3.473351e-05 4.394058e-08     -1.656131e-07 0.9993169 0.9991286  96.45823
#> 29 3.473351e-05 4.394058e-08     -2.179452e-07 0.9993131 0.9991238  96.45782
#> 30 3.473351e-05 4.394058e-08     -5.269941e-07 0.9993041 0.9991122  96.45680
#>    PlantStress
#> 1  0.005808769
#> 2  0.004697988
#> 3  0.004583281
#> 4  0.004857759
#> 5  0.003576142
#> 6  0.003685964
#> 7  0.001724734
#> 8  0.001741070
#> 9  0.001762639
#> 10 0.001815137
#> 11 0.010612890
#> 12 0.009702662
#> 13 0.009568109
#> 14 0.009759184
#> 15 0.008599094
#> 16 0.008601422
#> 17 0.006287482
#> 18 0.006307879
#> 19 0.006334692
#> 20 0.006399379
#> 21 0.003088161
#> 22 0.002664309
#> 23 0.002613207
#> 24 0.002711871
#> 25 0.002202897
#> 26 0.002229915
#> 27 0.001359583
#> 28 0.001367000
#> 29 0.001376775
#> 30 0.001400477

#Select the output tables/variables to be extracted
extract(S1, level ="cohort", output="Plants", vars = c("PlantStress", "StemPLC"))
#>          date cohort           species PlantStress      StemPLC
#> 1  2001-01-01 T1_148  Pinus halepensis 0.005808769 1.320577e-11
#> 2  2001-01-02 T1_148  Pinus halepensis 0.004697988 1.320577e-11
#> 3  2001-01-03 T1_148  Pinus halepensis 0.004583281 1.320577e-11
#> 4  2001-01-04 T1_148  Pinus halepensis 0.004857759 1.320577e-11
#> 5  2001-01-05 T1_148  Pinus halepensis 0.003576142 1.320577e-11
#> 6  2001-01-06 T1_148  Pinus halepensis 0.003685964 1.320577e-11
#> 7  2001-01-07 T1_148  Pinus halepensis 0.001724734 1.320577e-11
#> 8  2001-01-08 T1_148  Pinus halepensis 0.001741070 1.320577e-11
#> 9  2001-01-09 T1_148  Pinus halepensis 0.001762639 1.320577e-11
#> 10 2001-01-10 T1_148  Pinus halepensis 0.001815137 1.320577e-11
#> 11 2001-01-01 T2_168      Quercus ilex 0.010612890 1.691266e-06
#> 12 2001-01-02 T2_168      Quercus ilex 0.009702662 1.691266e-06
#> 13 2001-01-03 T2_168      Quercus ilex 0.009568109 1.691266e-06
#> 14 2001-01-04 T2_168      Quercus ilex 0.009759184 1.691266e-06
#> 15 2001-01-05 T2_168      Quercus ilex 0.008599094 1.691266e-06
#> 16 2001-01-06 T2_168      Quercus ilex 0.008601422 1.691266e-06
#> 17 2001-01-07 T2_168      Quercus ilex 0.006287482 1.691266e-06
#> 18 2001-01-08 T2_168      Quercus ilex 0.006307879 1.691266e-06
#> 19 2001-01-09 T2_168      Quercus ilex 0.006334692 1.691266e-06
#> 20 2001-01-10 T2_168      Quercus ilex 0.006399379 1.691266e-06
#> 21 2001-01-01 S1_165 Quercus coccifera 0.003088161 4.394058e-08
#> 22 2001-01-02 S1_165 Quercus coccifera 0.002664309 4.394058e-08
#> 23 2001-01-03 S1_165 Quercus coccifera 0.002613207 4.394058e-08
#> 24 2001-01-04 S1_165 Quercus coccifera 0.002711871 4.394058e-08
#> 25 2001-01-05 S1_165 Quercus coccifera 0.002202897 4.394058e-08
#> 26 2001-01-06 S1_165 Quercus coccifera 0.002229915 4.394058e-08
#> 27 2001-01-07 S1_165 Quercus coccifera 0.001359583 4.394058e-08
#> 28 2001-01-08 S1_165 Quercus coccifera 0.001367000 4.394058e-08
#> 29 2001-01-09 S1_165 Quercus coccifera 0.001376775 4.394058e-08
#> 30 2001-01-10 S1_165 Quercus coccifera 0.001400477 4.394058e-08