Extracts model outputs

Function extract() extracts daily or subdaily output and returns it as a tidy data frame.

Usage

extract(
  x,
  level = "forest",
  output = NULL,
  vars = NULL,
  dates = NULL,
  subdaily = FALSE,
  addunits = FALSE
)

Arguments

x

An object returned by simulation functions spwb, aspwb, pwb, growth or fordyn.

level

Level of simulation output, either "forest" (stand-level results), "soillayer" (soil layer-level results), "cohort" (cohort-level results), "sunlitleaf" or "shadeleaf" (leaf-level results)

output

Section of the model output to be explored. See details.

vars

Variables to be extracted (by default, all of them).

dates

A date vector indicating the subset of simulated days for which output is desired.

subdaily

A flag to indicate that subdaily values are desired (see details).

addunits

A flag to indicate that variable units should be added whenever possible.

Value

Function extract() returns a data frame:

If level = "forest", columns are "date" and variable names.

If level = "soillayer", columns are "date", "soillayer" and variable names.

If level = "cohort", level = "sunlitleaf" or level = "shadeleaf", columns are "date", "cohorts", "species" and variable names.

If subdaily = TRUE, columns are "datetime", "cohorts", "species" and variable names.

Details

When subdaily = FALSE, parameter output is used to restrict the section in x where variables are located. For example output = "Plants" will correspond to variables "LAI", "LAIlive", "Transpiration", "StemPLC",... as returned by a call names(x$Plants).

Option subdaily = TRUE only works when simulations have been carried using control option 'subdailyResults = TRUE' (see defaultControl). When using subdaily = TRUE, parameter output is not taken into account, and options for parameter vars are the following:

Variables for level = "forest" or level = "soillayer": Not allowed. An error is raised.

Variables for level = "cohort": "E","Ag","An","dEdP","RootPsi","StemPsi","LeafPsi","StemPLC","StemRWC","LeafRWC","StemSympRWC","LeafSympRWC","PWB".

Variables for level = "shadeleaf" and level="sunlitleaf": "Abs_SWR","Abs_PAR","Net_LWR","E","Ag","An","Ci","Gsw","VPD","Temp","Psi","iWUE".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

#Define soil with default soil params (4 layers)
examplesoil <- defaultSoilParams(4)

#Initialize control parameters
control <- defaultControl("Granier")

#Initialize input
x <- spwbInput(exampleforest,examplesoil, SpParamsMED, control)

#Call simulation function (ten days)
S1<-spwb(x, examplemeteo[1:10, ], latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 4.69853
#> Initial soil water content (mm): 290.875
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:
#> 
#> Final plant water content (mm): 4.69847
#> Final soil water content (mm): 294.836
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): -6.55262e-05
#> Plant water balance result (mm): -6.55126e-05
#> Change in soil water content (mm): 3.96115
#> Soil water balance result (mm): 3.96115
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 39 Rain (mm) 34 Snow (mm) 5
#>   Interception (mm) 8 Net rainfall (mm) 26
#>   Infiltration (mm) 31 Infiltration excess (mm) 0 Saturation excess (mm) 0 Capillarity rise (mm) 0
#>   Soil evaporation (mm) 3  Herbaceous transpiration (mm) 0 Woody plant transpiration (mm) 3
#>   Plant extraction from soil (mm) 3  Plant water balance (mm) -0 Hydraulic redistribution (mm) 0
#>   Runoff (mm) 0 Deep drainage (mm) 22

#Extracts daily forest-level output as a data frame
extract(S1, level = "forest", addunits = TRUE)
#>          date               PET     Precipitation              Rain
#> 1  2001-01-01 0.8828475 [L/m^2]  4.869109 [L/m^2]  4.869109 [L/m^2]
#> 2  2001-01-02 1.6375337 [L/m^2]  2.498292 [L/m^2]  2.498292 [L/m^2]
#> 3  2001-01-03 1.3017026 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 4  2001-01-04 0.5690790 [L/m^2]  5.796973 [L/m^2]  5.796973 [L/m^2]
#> 5  2001-01-05 1.6760567 [L/m^2]  1.884401 [L/m^2]  1.884401 [L/m^2]
#> 6  2001-01-06 1.2077028 [L/m^2] 13.359801 [L/m^2] 13.359801 [L/m^2]
#> 7  2001-01-07 0.6368075 [L/m^2]  5.377400 [L/m^2]  0.000000 [L/m^2]
#> 8  2001-01-08 0.8318207 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 9  2001-01-09 1.9823797 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 10 2001-01-10 0.8286873 [L/m^2]  5.117768 [L/m^2]  5.117768 [L/m^2]
#>              Snow            NetRain       Snowmelt       Infiltration
#> 1  0.0000 [L/m^2]  3.6000026 [L/m^2] 0.0000 [L/m^2]  3.6000026 [L/m^2]
#> 2  0.0000 [L/m^2]  1.2465550 [L/m^2] 0.0000 [L/m^2]  1.2465550 [L/m^2]
#> 3  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 4  0.0000 [L/m^2]  4.5392067 [L/m^2] 0.0000 [L/m^2]  4.5392067 [L/m^2]
#> 5  0.0000 [L/m^2]  0.8232431 [L/m^2] 0.0000 [L/m^2]  0.8232431 [L/m^2]
#> 6  0.0000 [L/m^2] 11.9122560 [L/m^2] 0.0000 [L/m^2] 11.9122560 [L/m^2]
#> 7  5.3774 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 8  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 9  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 10 0.0000 [L/m^2]  3.8497453 [L/m^2] 5.3774 [L/m^2]  9.2271449 [L/m^2]
#>    InfiltrationExcess SaturationExcess    Runoff      DeepDrainage
#> 1           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 2.9188167 [L/m^2]
#> 2           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.3694012 [L/m^2]
#> 3           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.0000000 [L/m^2]
#> 4           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 3.4543368 [L/m^2]
#> 5           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.1420391 [L/m^2]
#> 6           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#> 7           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#> 8           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 2.8197798 [L/m^2]
#> 9           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.0000000 [L/m^2]
#> 10          0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#>    CapillarityRise Evapotranspiration     Interception   SoilEvaporation
#> 1        0 [L/m^2]  1.9502921 [L/m^2] 1.269106 [L/m^2] 0.4776452 [L/m^2]
#> 2        0 [L/m^2]  2.1288904 [L/m^2] 1.251737 [L/m^2] 0.5000000 [L/m^2]
#> 3        0 [L/m^2]  0.7998058 [L/m^2] 0.000000 [L/m^2] 0.5000000 [L/m^2]
#> 4        0 [L/m^2]  1.5428303 [L/m^2] 1.257766 [L/m^2] 0.1540330 [L/m^2]
#> 5        0 [L/m^2]  1.9471677 [L/m^2] 1.061158 [L/m^2] 0.5000000 [L/m^2]
#> 6        0 [L/m^2]  2.0611808 [L/m^2] 1.447545 [L/m^2] 0.3355142 [L/m^2]
#> 7        0 [L/m^2]  0.1466640 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 8        0 [L/m^2]  0.1915683 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 9        0 [L/m^2]  0.4565126 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 10       0 [L/m^2]  1.6490675 [L/m^2] 1.268022 [L/m^2] 0.1902350 [L/m^2]
#>    HerbTranspiration   PlantExtraction     Transpiration
#> 1          0 [L/m^2] 0.2035406 [L/m^2] 0.2035406 [L/m^2]
#> 2          0 [L/m^2] 0.3771538 [L/m^2] 0.3771538 [L/m^2]
#> 3          0 [L/m^2] 0.2998058 [L/m^2] 0.2998058 [L/m^2]
#> 4          0 [L/m^2] 0.1310311 [L/m^2] 0.1310311 [L/m^2]
#> 5          0 [L/m^2] 0.3860093 [L/m^2] 0.3860093 [L/m^2]
#> 6          0 [L/m^2] 0.2781219 [L/m^2] 0.2781219 [L/m^2]
#> 7          0 [L/m^2] 0.1466640 [L/m^2] 0.1466640 [L/m^2]
#> 8          0 [L/m^2] 0.1915683 [L/m^2] 0.1915683 [L/m^2]
#> 9          0 [L/m^2] 0.4565126 [L/m^2] 0.4565126 [L/m^2]
#> 10         0 [L/m^2] 0.1908100 [L/m^2] 0.1908100 [L/m^2]
#>    HydraulicRedistribution                LAI     LAIherb            LAIlive
#> 1      0.000000000 [L/m^2] 1.584948 [m^2/m^2] 0 [m^2/m^2] 1.584948 [m^2/m^2]
#> 2      0.000000000 [L/m^2] 1.582896 [m^2/m^2] 0 [m^2/m^2] 1.584948 [m^2/m^2]
#> 3      0.000000000 [L/m^2] 1.582896 [m^2/m^2] 0 [m^2/m^2] 1.584948 [m^2/m^2]
#> 4      0.000000000 [L/m^2] 1.582896 [m^2/m^2] 0 [m^2/m^2] 1.584948 [m^2/m^2]
#> 5      0.000000000 [L/m^2] 1.582821 [m^2/m^2] 0 [m^2/m^2] 1.584948 [m^2/m^2]
#> 6      0.000000000 [L/m^2] 1.582821 [m^2/m^2] 0 [m^2/m^2] 1.584948 [m^2/m^2]
#> 7      0.002845017 [L/m^2] 1.582821 [m^2/m^2] 0 [m^2/m^2] 1.584948 [m^2/m^2]
#> 8      0.003493923 [L/m^2] 1.582821 [m^2/m^2] 0 [m^2/m^2] 1.584948 [m^2/m^2]
#> 9      0.000000000 [L/m^2] 1.582821 [m^2/m^2] 0 [m^2/m^2] 1.584948 [m^2/m^2]
#> 10     0.000000000 [L/m^2] 1.582821 [m^2/m^2] 0 [m^2/m^2] 1.584948 [m^2/m^2]
#>           LAIexpanded     LAIdead               Cm   LgroundPAR   LgroundSWR
#> 1  1.584948 [m^2/m^2] 0 [m^2/m^2] 1.216848 [L/m^2] 43.63617 [%] 54.10280 [%]
#> 2  1.582896 [m^2/m^2] 0 [m^2/m^2] 1.215822 [L/m^2] 43.68543 [%] 54.14803 [%]
#> 3  1.582896 [m^2/m^2] 0 [m^2/m^2] 1.215822 [L/m^2] 43.68543 [%] 54.14803 [%]
#> 4  1.582896 [m^2/m^2] 0 [m^2/m^2] 1.215822 [L/m^2] 43.68543 [%] 54.14803 [%]
#> 5  1.582821 [m^2/m^2] 0 [m^2/m^2] 1.215784 [L/m^2] 43.68725 [%] 54.14970 [%]
#> 6  1.582821 [m^2/m^2] 0 [m^2/m^2] 1.215784 [L/m^2] 43.68725 [%] 54.14970 [%]
#> 7  1.582821 [m^2/m^2] 0 [m^2/m^2] 1.215784 [L/m^2] 43.68725 [%] 54.14970 [%]
#> 8  1.582821 [m^2/m^2] 0 [m^2/m^2] 1.215784 [L/m^2] 43.68725 [%] 54.14970 [%]
#> 9  1.582821 [m^2/m^2] 0 [m^2/m^2] 1.215784 [L/m^2] 43.68725 [%] 54.14970 [%]
#> 10 1.582821 [m^2/m^2] 0 [m^2/m^2] 1.215784 [L/m^2] 43.68725 [%] 54.14970 [%]
#>               SWE
#> 1  0.0000 [L/m^2]
#> 2  0.0000 [L/m^2]
#> 3  0.0000 [L/m^2]
#> 4  0.0000 [L/m^2]
#> 5  0.0000 [L/m^2]
#> 6  0.0000 [L/m^2]
#> 7  5.3774 [L/m^2]
#> 8  5.3774 [L/m^2]
#> 9  5.3774 [L/m^2]
#> 10 0.0000 [L/m^2]

#Extracts daily soil layer-level output as a data frame
extract(S1, level = "soillayer", addunits = TRUE)
#>          date soillayer       SWC       RWC       REW                ML
#> 1  2001-01-01         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 2  2001-01-02         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 3  2001-01-03         1 0.2997303 0.9892261 0.9838421  67.43932 [L/m^2]
#> 4  2001-01-04         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 5  2001-01-05         1 0.3020845 0.9969958 0.9954946  67.96901 [L/m^2]
#> 6  2001-01-06         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 7  2001-01-07         1 0.3025095 0.9983986 0.9975983  68.06464 [L/m^2]
#> 8  2001-01-08         1 0.3018765 0.9963093 0.9944650  67.92221 [L/m^2]
#> 9  2001-01-09         1 0.3004971 0.9917569 0.9876376  67.61185 [L/m^2]
#> 10 2001-01-10         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 11 2001-01-01         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 12 2001-01-02         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 13 2001-01-03         2 0.3028406 0.9994913 0.9992370 116.59363 [L/m^2]
#> 14 2001-01-04         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 15 2001-01-05         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 16 2001-01-06         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 17 2001-01-07         2 0.3029333 0.9997973 0.9996960 116.62932 [L/m^2]
#> 18 2001-01-08         2 0.3028508 0.9995249 0.9992876 116.59755 [L/m^2]
#> 19 2001-01-09         2 0.3025881 0.9986579 0.9979872 116.49641 [L/m^2]
#> 20 2001-01-10         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 21 2001-01-01         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 22 2001-01-02         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 23 2001-01-03         3 0.3029748 0.9999343 0.9999014  75.74370 [L/m^2]
#> 24 2001-01-04         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 25 2001-01-05         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 26 2001-01-06         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 27 2001-01-07         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 28 2001-01-08         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 29 2001-01-09         3 0.3028924 0.9996622 0.9994934  75.72310 [L/m^2]
#> 30 2001-01-10         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 31 2001-01-01         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 32 2001-01-02         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 33 2001-01-03         4 0.3029849 0.9999675 0.9999513  30.29849 [L/m^2]
#> 34 2001-01-04         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 35 2001-01-05         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 36 2001-01-06         4 0.3727187 1.2301162 1.3451105  37.27187 [L/m^2]
#> 37 2001-01-07         4 0.3313662 1.0936369 1.1404293  33.13662 [L/m^2]
#> 38 2001-01-08         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 39 2001-01-09         4 0.3028005 0.9993589 0.9990385  30.28005 [L/m^2]
#> 40 2001-01-10         4 0.3426062 1.1307332 1.1960636  34.26062 [L/m^2]
#> 41 2001-01-01   Overall 0.3029947 1.0000000 1.0000000 290.87494 [L/m^2]
#> 42 2001-01-02   Overall 0.3029947 1.0000000 1.0000000 290.87494 [L/m^2]
#> 43 2001-01-03   Overall 0.3021616 0.9972503 0.9958763 290.07513 [L/m^2]
#> 44 2001-01-04   Overall 0.3029947 1.0000000 1.0000000 290.87494 [L/m^2]
#> 45 2001-01-05   Overall 0.3027814 0.9992959 0.9989440 290.67013 [L/m^2]
#> 46 2001-01-06   Overall 0.3102576 1.0239704 1.0359490 297.84734 [L/m^2]
#> 47 2001-01-07   Overall 0.3058117 1.0092972 1.0139432 293.57926 [L/m^2]
#> 48 2001-01-08   Overall 0.3026749 0.9989445 0.9984170 290.56791 [L/m^2]
#> 49 2001-01-09   Overall 0.3021994 0.9973750 0.9960633 290.11140 [L/m^2]
#> 50 2001-01-10   Overall 0.3071209 1.0136180 1.0204233 294.83609 [L/m^2]
#>                   Psi              PlantExt      HydraulicInput
#> 1  -0.033000000 [MPa]  0.1591605366 [L/m^2] 0.000000000 [L/m^2]
#> 2  -0.033000000 [MPa]  0.2949969778 [L/m^2] 0.000000000 [L/m^2]
#> 3  -0.034727924 [MPa]  0.2344979736 [L/m^2] 0.000000000 [L/m^2]
#> 4  -0.033000000 [MPa]  0.0884138280 [L/m^2] 0.000000000 [L/m^2]
#> 5  -0.033473022 [MPa]  0.3019234831 [L/m^2] 0.000000000 [L/m^2]
#> 6  -0.033000000 [MPa]  0.1891290311 [L/m^2] 0.000000000 [L/m^2]
#> 7  -0.033251327 [MPa]  0.1091765935 [L/m^2] 0.000000000 [L/m^2]
#> 8  -0.033582060 [MPa]  0.1424320474 [L/m^2] 0.000000000 [L/m^2]
#> 9  -0.034314095 [MPa]  0.3103557369 [L/m^2] 0.000000000 [L/m^2]
#> 10 -0.033000000 [MPa]  0.1292083039 [L/m^2] 0.000000000 [L/m^2]
#> 11 -0.033000000 [MPa]  0.0403279820 [L/m^2] 0.000000000 [L/m^2]
#> 12 -0.033000000 [MPa]  0.0746563013 [L/m^2] 0.000000000 [L/m^2]
#> 13 -0.033079635 [MPa]  0.0593455279 [L/m^2] 0.000000000 [L/m^2]
#> 14 -0.033000000 [MPa]  0.0306565358 [L/m^2] 0.000000000 [L/m^2]
#> 15 -0.033000000 [MPa]  0.0764092252 [L/m^2] 0.000000000 [L/m^2]
#> 16 -0.033000000 [MPa]  0.0648688739 [L/m^2] 0.000000000 [L/m^2]
#> 17 -0.033031711 [MPa]  0.0236487045 [L/m^2] 0.000000000 [L/m^2]
#> 18 -0.033074357 [MPa]  0.0317678128 [L/m^2] 0.000000000 [L/m^2]
#> 19 -0.033210500 [MPa]  0.1011449996 [L/m^2] 0.000000000 [L/m^2]
#> 20 -0.033000000 [MPa]  0.0430212580 [L/m^2] 0.000000000 [L/m^2]
#> 21 -0.033000000 [MPa]  0.0033836661 [L/m^2] 0.000000000 [L/m^2]
#> 22 -0.033000000 [MPa]  0.0062632055 [L/m^2] 0.000000000 [L/m^2]
#> 23 -0.033010278 [MPa]  0.0049787256 [L/m^2] 0.000000000 [L/m^2]
#> 24 -0.033000000 [MPa]  0.0068547664 [L/m^2] 0.000000000 [L/m^2]
#> 25 -0.033000000 [MPa]  0.0064102651 [L/m^2] 0.000000000 [L/m^2]
#> 26 -0.033000000 [MPa]  0.0139323961 [L/m^2] 0.000000000 [L/m^2]
#> 27 -0.033000000 [MPa] -0.0024988843 [L/m^2] 0.002845017 [L/m^2]
#> 28 -0.033000000 [MPa] -0.0028451154 [L/m^2] 0.003493923 [L/m^2]
#> 29 -0.033052853 [MPa]  0.0255859552 [L/m^2] 0.000000000 [L/m^2]
#> 30 -0.033000000 [MPa]  0.0111204547 [L/m^2] 0.000000000 [L/m^2]
#> 31 -0.033000000 [MPa]  0.0006684070 [L/m^2] 0.000000000 [L/m^2]
#> 32 -0.033000000 [MPa]  0.0012372923 [L/m^2] 0.000000000 [L/m^2]
#> 33 -0.033005076 [MPa]  0.0009835441 [L/m^2] 0.000000000 [L/m^2]
#> 34 -0.033000000 [MPa]  0.0051059711 [L/m^2] 0.000000000 [L/m^2]
#> 35 -0.033000000 [MPa]  0.0012663438 [L/m^2] 0.000000000 [L/m^2]
#> 36 -0.009875488 [MPa]  0.0101916343 [L/m^2] 0.000000000 [L/m^2]
#> 37 -0.021035520 [MPa]  0.0163376079 [L/m^2] 0.000000000 [L/m^2]
#> 38 -0.033000000 [MPa]  0.0202135646 [L/m^2] 0.000000000 [L/m^2]
#> 39 -0.033100391 [MPa]  0.0194259109 [L/m^2] 0.000000000 [L/m^2]
#> 40 -0.017443648 [MPa]  0.0074599410 [L/m^2] 0.000000000 [L/m^2]
#> 41 -0.033000000 [MPa]  0.2035405917 [L/m^2] 0.000000000 [L/m^2]
#> 42 -0.033000000 [MPa]  0.3771537769 [L/m^2] 0.000000000 [L/m^2]
#> 43 -0.033440125 [MPa]  0.2998057712 [L/m^2] 0.000000000 [L/m^2]
#> 44 -0.033000000 [MPa]  0.1310311013 [L/m^2] 0.000000000 [L/m^2]
#> 45 -0.033110865 [MPa]  0.3860093172 [L/m^2] 0.000000000 [L/m^2]
#> 46 -0.030591197 [MPa]  0.2781219354 [L/m^2] 0.000000000 [L/m^2]
#> 47 -0.031825322 [MPa]  0.1466640215 [L/m^2] 0.000000000 [L/m^2]
#> 48 -0.033166241 [MPa]  0.1915683094 [L/m^2] 0.000000000 [L/m^2]
#> 49 -0.033416631 [MPa]  0.4565126026 [L/m^2] 0.000000000 [L/m^2]
#> 50 -0.031379547 [MPa]  0.1908099577 [L/m^2] 0.000000000 [L/m^2]

#Extracts daily cohort-level output as a data frame
extract(S1, level = "cohort", addunits = TRUE)
#>          date cohort           species                  LAI
#> 1  2001-01-01 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 2  2001-01-02 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 3  2001-01-03 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 4  2001-01-04 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 5  2001-01-05 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 6  2001-01-06 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 7  2001-01-07 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 8  2001-01-08 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 9  2001-01-09 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 10 2001-01-10 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 11 2001-01-01 T2_168      Quercus ilex 0.70557382 [m^2/m^2]
#> 12 2001-01-02 T2_168      Quercus ilex 0.70352348 [m^2/m^2]
#> 13 2001-01-03 T2_168      Quercus ilex 0.70352348 [m^2/m^2]
#> 14 2001-01-04 T2_168      Quercus ilex 0.70352348 [m^2/m^2]
#> 15 2001-01-05 T2_168      Quercus ilex 0.70344807 [m^2/m^2]
#> 16 2001-01-06 T2_168      Quercus ilex 0.70344807 [m^2/m^2]
#> 17 2001-01-07 T2_168      Quercus ilex 0.70344807 [m^2/m^2]
#> 18 2001-01-08 T2_168      Quercus ilex 0.70344807 [m^2/m^2]
#> 19 2001-01-09 T2_168      Quercus ilex 0.70344807 [m^2/m^2]
#> 20 2001-01-10 T2_168      Quercus ilex 0.70344807 [m^2/m^2]
#> 21 2001-01-01 S1_165 Quercus coccifera 0.03062604 [m^2/m^2]
#> 22 2001-01-02 S1_165 Quercus coccifera 0.03062498 [m^2/m^2]
#> 23 2001-01-03 S1_165 Quercus coccifera 0.03062498 [m^2/m^2]
#> 24 2001-01-04 S1_165 Quercus coccifera 0.03062498 [m^2/m^2]
#> 25 2001-01-05 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 26 2001-01-06 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 27 2001-01-07 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 28 2001-01-08 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 29 2001-01-09 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 30 2001-01-10 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#>                 LAIlive         FPAR AbsorbedSWRFraction       Transpiration
#> 1  0.84874773 [m^2/m^2] 92.18285 [%]           35.076344 0.125013855 [L/m^2]
#> 2  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.231814863 [L/m^2]
#> 3  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.184273466 [L/m^2]
#> 4  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.080532461 [L/m^2]
#> 5  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.237257858 [L/m^2]
#> 6  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.170942753 [L/m^2]
#> 7  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.090144763 [L/m^2]
#> 8  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.117744289 [L/m^2]
#> 9  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.280587005 [L/m^2]
#> 10 0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.117275206 [L/m^2]
#> 11 0.70557382 [m^2/m^2] 72.36365 [%]           30.444383 0.072474080 [L/m^2]
#> 12 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.134110679 [L/m^2]
#> 13 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.106606795 [L/m^2]
#> 14 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.046597155 [L/m^2]
#> 15 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.137259587 [L/m^2]
#> 16 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.098898900 [L/m^2]
#> 17 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.052152879 [L/m^2]
#> 18 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.068120690 [L/m^2]
#> 19 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.162334115 [L/m^2]
#> 20 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.067853526 [L/m^2]
#> 21 0.03062604 [m^2/m^2] 44.32407 [%]            2.366131 0.006052657 [L/m^2]
#> 22 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.011228235 [L/m^2]
#> 23 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.008925509 [L/m^2]
#> 24 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.003901485 [L/m^2]
#> 25 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.011491873 [L/m^2]
#> 26 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.008280282 [L/m^2]
#> 27 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.004366379 [L/m^2]
#> 28 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.005703331 [L/m^2]
#> 29 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.013591483 [L/m^2]
#> 30 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.005681225 [L/m^2]
#>    GrossPhotosynthesis          PlantPsi      LeafPLC      StemPLC
#> 1   1.21341578 [L/m^2] -0.03300000 [MPa] 0.000000e+00 0.000000e+00
#> 2   2.10348703 [L/m^2] -0.03300000 [MPa] 0.000000e+00 0.000000e+00
#> 3   1.77397793 [L/m^2] -0.03300000 [MPa] 0.000000e+00 0.000000e+00
#> 4   0.81305219 [L/m^2] -0.03464604 [MPa] 0.000000e+00 0.000000e+00
#> 5   2.18356009 [L/m^2] -0.03300000 [MPa] 0.000000e+00 0.000000e+00
#> 6   1.79061896 [L/m^2] -0.03344937 [MPa] 0.000000e+00 0.000000e+00
#> 7   1.02740325 [L/m^2] -0.03299488 [MPa] 0.000000e+00 0.000000e+00
#> 8   1.21701875 [L/m^2] -0.03323727 [MPa] 0.000000e+00 0.000000e+00
#> 9   2.08854605 [L/m^2] -0.03355653 [MPa] 0.000000e+00 0.000000e+00
#> 10  1.16046327 [L/m^2] -0.03425885 [MPa] 0.000000e+00 0.000000e+00
#> 11  0.87313959 [L/m^2] -0.03300000 [MPa] 2.905929e-03 3.681226e-09
#> 12  1.59099654 [L/m^2] -0.03300000 [MPa] 2.905929e-03 3.681226e-09
#> 13  1.28193894 [L/m^2] -0.03300000 [MPa] 2.905929e-03 3.681226e-09
#> 14  0.56671681 [L/m^2] -0.03390328 [MPa] 3.012795e-03 4.052691e-09
#> 15  1.63354198 [L/m^2] -0.03300000 [MPa] 3.012795e-03 4.052691e-09
#> 16  1.21377018 [L/m^2] -0.03323705 [MPa] 3.012795e-03 4.052691e-09
#> 17  0.65412447 [L/m^2] -0.03283019 [MPa] 3.012795e-03 4.052691e-09
#> 18  0.83326937 [L/m^2] -0.03304892 [MPa] 3.012795e-03 4.052691e-09
#> 19  1.84698471 [L/m^2] -0.03332520 [MPa] 3.012795e-03 4.052691e-09
#> 20  0.82125613 [L/m^2] -0.03375667 [MPa] 3.012795e-03 4.052691e-09
#> 21  0.05005525 [L/m^2] -0.03300000 [MPa] 3.473351e-05 4.394058e-08
#> 22  0.08989595 [L/m^2] -0.03300000 [MPa] 3.473351e-05 4.394058e-08
#> 23  0.07354466 [L/m^2] -0.03300000 [MPa] 3.473351e-05 4.394058e-08
#> 24  0.03292077 [L/m^2] -0.03419979 [MPa] 3.764626e-05 4.907662e-08
#> 25  0.09263459 [L/m^2] -0.03300000 [MPa] 3.764626e-05 4.907662e-08
#> 26  0.07119499 [L/m^2] -0.03332187 [MPa] 3.764626e-05 4.907662e-08
#> 27  0.03924413 [L/m^2] -0.03278764 [MPa] 3.764626e-05 4.907662e-08
#> 28  0.04870345 [L/m^2] -0.03306107 [MPa] 3.764626e-05 4.907662e-08
#> 29  0.09940050 [L/m^2] -0.03341641 [MPa] 3.764626e-05 4.907662e-08
#> 30  0.04745640 [L/m^2] -0.03395535 [MPa] 3.764626e-05 4.907662e-08
#>        PlantWaterBalance       LeafRWC       StemRWC          LFMC PlantStress
#> 1   0.000000e+00 [L/m^2] 0.9979679 [%] 0.9998350 [%] 125.86544 [%] 0.004613739
#> 2   0.000000e+00 [L/m^2] 0.9979679 [%] 0.9998350 [%] 125.86544 [%] 0.004613739
#> 3   0.000000e+00 [L/m^2] 0.9979679 [%] 0.9998350 [%] 125.86544 [%] 0.004613739
#> 4  -7.408455e-05 [L/m^2] 0.9978666 [%] 0.9998267 [%] 125.85720 [%] 0.004963434
#> 5   7.408455e-05 [L/m^2] 0.9979679 [%] 0.9998350 [%] 125.86544 [%] 0.004613739
#> 6  -2.022563e-05 [L/m^2] 0.9979402 [%] 0.9998327 [%] 125.86319 [%] 0.004708365
#> 7   2.045597e-05 [L/m^2] 0.9979682 [%] 0.9998350 [%] 125.86546 [%] 0.004612665
#> 8  -1.090943e-05 [L/m^2] 0.9979533 [%] 0.9998338 [%] 125.86425 [%] 0.004663622
#> 9  -1.436955e-05 [L/m^2] 0.9979337 [%] 0.9998322 [%] 125.86265 [%] 0.004731024
#> 10 -3.160991e-05 [L/m^2] 0.9978904 [%] 0.9998287 [%] 125.85914 [%] 0.004880419
#> 11  0.000000e+00 [L/m^2] 0.9986797 [%] 0.9997500 [%]  93.05186 [%] 0.006282261
#> 12  0.000000e+00 [L/m^2] 0.9986797 [%] 0.9997500 [%]  92.91408 [%] 0.006282261
#> 13  0.000000e+00 [L/m^2] 0.9986797 [%] 0.9997500 [%]  92.91408 [%] 0.006282261
#> 14 -9.484628e-06 [L/m^2] 0.9986436 [%] 0.9997431 [%]  92.91131 [%] 0.006479608
#> 15  9.484029e-06 [L/m^2] 0.9986797 [%] 0.9997500 [%]  92.90901 [%] 0.006282261
#> 16 -2.488960e-06 [L/m^2] 0.9986702 [%] 0.9997482 [%]  92.90828 [%] 0.006333979
#> 17  4.271906e-06 [L/m^2] 0.9986865 [%] 0.9997513 [%]  92.90953 [%] 0.006245246
#> 18 -2.296537e-06 [L/m^2] 0.9986778 [%] 0.9997496 [%]  92.90886 [%] 0.006292929
#> 19 -2.900836e-06 [L/m^2] 0.9986667 [%] 0.9997475 [%]  92.90801 [%] 0.006353223
#> 20 -4.530255e-06 [L/m^2] 0.9986495 [%] 0.9997442 [%]  92.90669 [%] 0.006447526
#> 21  3.469447e-18 [L/m^2] 0.9987207 [%] 0.9983684 [%]  96.38420 [%] 0.003088161
#> 22  0.000000e+00 [L/m^2] 0.9987207 [%] 0.9983684 [%]  96.38552 [%] 0.003088161
#> 23  0.000000e+00 [L/m^2] 0.9987207 [%] 0.9983684 [%]  96.38552 [%] 0.003088161
#> 24 -1.141335e-06 [L/m^2] 0.9986742 [%] 0.9983091 [%]  96.38007 [%] 0.003234673
#> 25  1.141333e-06 [L/m^2] 0.9987207 [%] 0.9983684 [%]  96.38563 [%] 0.003088161
#> 26 -3.061870e-07 [L/m^2] 0.9987082 [%] 0.9983525 [%]  96.38417 [%] 0.003127313
#> 27  5.082007e-07 [L/m^2] 0.9987289 [%] 0.9983789 [%]  96.38660 [%] 0.003062391
#> 28 -2.601082e-07 [L/m^2] 0.9987183 [%] 0.9983654 [%]  96.38536 [%] 0.003095581
#> 29 -3.380329e-07 [L/m^2] 0.9987045 [%] 0.9983478 [%]  96.38374 [%] 0.003138836
#> 30 -5.126799e-07 [L/m^2] 0.9986836 [%] 0.9983212 [%]  96.38129 [%] 0.003204698

#Select the output tables/variables to be extracted
extract(S1, level ="cohort", output="Plants", vars = c("PlantStress", "StemPLC"))
#>          date cohort           species PlantStress      StemPLC
#> 1  2001-01-01 T1_148  Pinus halepensis 0.004613739 0.000000e+00
#> 2  2001-01-02 T1_148  Pinus halepensis 0.004613739 0.000000e+00
#> 3  2001-01-03 T1_148  Pinus halepensis 0.004613739 0.000000e+00
#> 4  2001-01-04 T1_148  Pinus halepensis 0.004963434 0.000000e+00
#> 5  2001-01-05 T1_148  Pinus halepensis 0.004613739 0.000000e+00
#> 6  2001-01-06 T1_148  Pinus halepensis 0.004708365 0.000000e+00
#> 7  2001-01-07 T1_148  Pinus halepensis 0.004612665 0.000000e+00
#> 8  2001-01-08 T1_148  Pinus halepensis 0.004663622 0.000000e+00
#> 9  2001-01-09 T1_148  Pinus halepensis 0.004731024 0.000000e+00
#> 10 2001-01-10 T1_148  Pinus halepensis 0.004880419 0.000000e+00
#> 11 2001-01-01 T2_168      Quercus ilex 0.006282261 3.681226e-09
#> 12 2001-01-02 T2_168      Quercus ilex 0.006282261 3.681226e-09
#> 13 2001-01-03 T2_168      Quercus ilex 0.006282261 3.681226e-09
#> 14 2001-01-04 T2_168      Quercus ilex 0.006479608 4.052691e-09
#> 15 2001-01-05 T2_168      Quercus ilex 0.006282261 4.052691e-09
#> 16 2001-01-06 T2_168      Quercus ilex 0.006333979 4.052691e-09
#> 17 2001-01-07 T2_168      Quercus ilex 0.006245246 4.052691e-09
#> 18 2001-01-08 T2_168      Quercus ilex 0.006292929 4.052691e-09
#> 19 2001-01-09 T2_168      Quercus ilex 0.006353223 4.052691e-09
#> 20 2001-01-10 T2_168      Quercus ilex 0.006447526 4.052691e-09
#> 21 2001-01-01 S1_165 Quercus coccifera 0.003088161 4.394058e-08
#> 22 2001-01-02 S1_165 Quercus coccifera 0.003088161 4.394058e-08
#> 23 2001-01-03 S1_165 Quercus coccifera 0.003088161 4.394058e-08
#> 24 2001-01-04 S1_165 Quercus coccifera 0.003234673 4.907662e-08
#> 25 2001-01-05 S1_165 Quercus coccifera 0.003088161 4.907662e-08
#> 26 2001-01-06 S1_165 Quercus coccifera 0.003127313 4.907662e-08
#> 27 2001-01-07 S1_165 Quercus coccifera 0.003062391 4.907662e-08
#> 28 2001-01-08 S1_165 Quercus coccifera 0.003095581 4.907662e-08
#> 29 2001-01-09 S1_165 Quercus coccifera 0.003138836 4.907662e-08
#> 30 2001-01-10 S1_165 Quercus coccifera 0.003204698 4.907662e-08