Extracts model outputs

Function extract() extracts daily or subdaily output and returns it as a tidy data frame.

Usage

extract(
  x,
  level = "forest",
  output = NULL,
  vars = NULL,
  dates = NULL,
  subdaily = FALSE,
  addunits = FALSE
)

Arguments

x

An object returned by simulation functions spwb, aspwb, pwb, growth or fordyn.

level

Level of simulation output, either "forest" (stand-level results), "soillayer" (soil layer-level results), "cohort" (cohort-level results), "sunlitleaf" or "shadeleaf" (leaf-level results)

output

Section of the model output to be explored. See details.

vars

Variables to be extracted (by default, all of them).

dates

A date vector indicating the subset of simulated days for which output is desired.

subdaily

A flag to indicate that subdaily values are desired (see details).

addunits

A flag to indicate that variable units should be added whenever possible.

Value

Function extract() returns a data frame:

If level = "forest", columns are "date" and variable names.

If level = "soillayer", columns are "date", "soillayer" and variable names.

If level = "cohort", level = "sunlitleaf" or level = "shadeleaf", columns are "date", "cohorts", "species" and variable names.

If subdaily = TRUE, columns are "datetime", "cohorts", "species" and variable names.

Details

When subdaily = FALSE, parameter output is used to restrict the section in x where variables are located. For example output = "Plants" will correspond to variables "LAI", "LAIlive", "Transpiration", "StemPLC",... as returned by a call names(x$Plants).

Option subdaily = TRUE only works when simulations have been carried using control option 'subdailyResults = TRUE' (see defaultControl). When using subdaily = TRUE, parameter output is not taken into account, and options for parameter vars are the following:

Variables for level = "forest" or level = "soillayer": Not allowed. An error is raised.

Variables for level = "cohort": "E","Ag","An","dEdP","RootPsi","StemPsi","LeafPsi","StemPLC","StemRWC","LeafRWC","StemSympRWC","LeafSympRWC","PWB".

Variables for level = "shadeleaf" and level="sunlitleaf": "Abs_SWR","Abs_PAR","Net_LWR","E","Ag","An","Ci","Gsw","VPD","Temp","Psi","iWUE".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

#Define soil with default soil params (4 layers)
examplesoil <- defaultSoilParams(4)

#Initialize control parameters
control <- defaultControl("Granier")

#Initialize input
x <- spwbInput(exampleforest,examplesoil, SpParamsMED, control)

#Call simulation function (ten days)
S1<-spwb(x, examplemeteo[1:10, ], latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 4.69853
#> Initial soil water content (mm): 290.875
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:
#> 
#> Final plant water content (mm): 4.69846
#> Final soil water content (mm): 294.616
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): -7.01146e-05
#> Plant water balance result (mm): -7.01001e-05
#> Change in soil water content (mm): 3.74152
#> Soil water balance result (mm): 3.74152
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 39 Rain (mm) 34 Snow (mm) 5
#>   Interception (mm) 9 Net rainfall (mm) 25
#>   Infiltration (mm) 30 Infiltration excess (mm) 0 Saturation excess (mm) 0 Capillarity rise (mm) 0
#>   Soil evaporation (mm) 3  Herbaceous transpiration (mm) 0 Woody plant transpiration (mm) 3
#>   Plant extraction from soil (mm) 3  Plant water balance (mm) -0 Hydraulic redistribution (mm) 0
#>   Runoff (mm) 0 Deep drainage (mm) 21

#Extracts daily forest-level output as a data frame
extract(S1, level = "forest", addunits = TRUE)
#>          date               PET     Precipitation              Rain
#> 1  2001-01-01 0.8828475 [L/m^2]  4.869109 [L/m^2]  4.869109 [L/m^2]
#> 2  2001-01-02 1.6375337 [L/m^2]  2.498292 [L/m^2]  2.498292 [L/m^2]
#> 3  2001-01-03 1.3017026 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 4  2001-01-04 0.5690790 [L/m^2]  5.796973 [L/m^2]  5.796973 [L/m^2]
#> 5  2001-01-05 1.6760567 [L/m^2]  1.884401 [L/m^2]  1.884401 [L/m^2]
#> 6  2001-01-06 1.2077028 [L/m^2] 13.359801 [L/m^2] 13.359801 [L/m^2]
#> 7  2001-01-07 0.6368075 [L/m^2]  5.377400 [L/m^2]  0.000000 [L/m^2]
#> 8  2001-01-08 0.8318207 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 9  2001-01-09 1.9823797 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 10 2001-01-10 0.8286873 [L/m^2]  5.117768 [L/m^2]  5.117768 [L/m^2]
#>              Snow            NetRain       Snowmelt       Infiltration
#> 1  0.0000 [L/m^2]  3.4241795 [L/m^2] 0.0000 [L/m^2]  3.4241795 [L/m^2]
#> 2  0.0000 [L/m^2]  1.0728002 [L/m^2] 0.0000 [L/m^2]  1.0728002 [L/m^2]
#> 3  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 4  0.0000 [L/m^2]  4.3636223 [L/m^2] 0.0000 [L/m^2]  4.3636223 [L/m^2]
#> 5  0.0000 [L/m^2]  0.7547848 [L/m^2] 0.0000 [L/m^2]  0.7547848 [L/m^2]
#> 6  0.0000 [L/m^2] 11.7252823 [L/m^2] 0.0000 [L/m^2] 11.7252823 [L/m^2]
#> 7  5.3774 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 8  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 9  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 10 0.0000 [L/m^2]  3.6738422 [L/m^2] 5.3774 [L/m^2]  9.0512419 [L/m^2]
#>    InfiltrationExcess SaturationExcess    Runoff      DeepDrainage
#> 1           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 2.7617207 [L/m^2]
#> 2           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.2106989 [L/m^2]
#> 3           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.0000000 [L/m^2]
#> 4           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 3.2516535 [L/m^2]
#> 5           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.1213759 [L/m^2]
#> 6           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#> 7           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#> 8           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 2.5878264 [L/m^2]
#> 9           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.0000000 [L/m^2]
#> 10          0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#>    CapillarityRise Evapotranspiration     Interception   SoilEvaporation
#> 1        0 [L/m^2]  2.1073881 [L/m^2] 1.444929 [L/m^2] 0.4478948 [L/m^2]
#> 2        0 [L/m^2]  2.3231089 [L/m^2] 1.425491 [L/m^2] 0.5000000 [L/m^2]
#> 3        0 [L/m^2]  0.7818123 [L/m^2] 0.000000 [L/m^2] 0.4657437 [L/m^2]
#> 4        0 [L/m^2]  1.7279909 [L/m^2] 1.433351 [L/m^2] 0.1564975 [L/m^2]
#> 5        0 [L/m^2]  2.0365719 [L/m^2] 1.129617 [L/m^2] 0.5000000 [L/m^2]
#> 6        0 [L/m^2]  2.2241156 [L/m^2] 1.634518 [L/m^2] 0.2963901 [L/m^2]
#> 7        0 [L/m^2]  0.1546223 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 8        0 [L/m^2]  0.2019630 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 9        0 [L/m^2]  0.4812834 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 10       0 [L/m^2]  1.8258517 [L/m^2] 1.443926 [L/m^2] 0.1807633 [L/m^2]
#>      HerbTranspiration   PlantExtraction     Transpiration
#> 1  0.011023432 [L/m^2] 0.2035406 [L/m^2] 0.2035406 [L/m^2]
#> 2  0.020463709 [L/m^2] 0.3771538 [L/m^2] 0.3771538 [L/m^2]
#> 3  0.016266913 [L/m^2] 0.2998017 [L/m^2] 0.2998017 [L/m^2]
#> 4  0.007111348 [L/m^2] 0.1310315 [L/m^2] 0.1310315 [L/m^2]
#> 5  0.020945756 [L/m^2] 0.3860095 [L/m^2] 0.3860095 [L/m^2]
#> 6  0.015092520 [L/m^2] 0.2781147 [L/m^2] 0.2781147 [L/m^2]
#> 7  0.007958211 [L/m^2] 0.1466640 [L/m^2] 0.1466640 [L/m^2]
#> 8  0.010395244 [L/m^2] 0.1915678 [L/m^2] 0.1915678 [L/m^2]
#> 9  0.024773572 [L/m^2] 0.4565098 [L/m^2] 0.4565098 [L/m^2]
#> 10 0.010355839 [L/m^2] 0.1908070 [L/m^2] 0.1908070 [L/m^2]
#>    HydraulicRedistribution                LAI             LAIherb
#> 1      0.000000000 [L/m^2] 1.758585 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 2      0.000000000 [L/m^2] 1.756533 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 3      0.000000000 [L/m^2] 1.756533 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 4      0.000000000 [L/m^2] 1.756530 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 5      0.000000000 [L/m^2] 1.756458 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 6      0.000000000 [L/m^2] 1.756458 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 7      0.002845012 [L/m^2] 1.756458 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 8      0.003458333 [L/m^2] 1.756458 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 9      0.000000000 [L/m^2] 1.756458 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 10     0.000000000 [L/m^2] 1.756458 [m^2/m^2] 0.1736369 [m^2/m^2]
#>               LAIlive        LAIexpanded     LAIdead               Cm
#> 1  1.584948 [m^2/m^2] 1.584948 [m^2/m^2] 0 [m^2/m^2] 1.390485 [L/m^2]
#> 2  1.584948 [m^2/m^2] 1.582896 [m^2/m^2] 0 [m^2/m^2] 1.389459 [L/m^2]
#> 3  1.584948 [m^2/m^2] 1.582896 [m^2/m^2] 0 [m^2/m^2] 1.389459 [L/m^2]
#> 4  1.584948 [m^2/m^2] 1.582893 [m^2/m^2] 0 [m^2/m^2] 1.389457 [L/m^2]
#> 5  1.584948 [m^2/m^2] 1.582821 [m^2/m^2] 0 [m^2/m^2] 1.389421 [L/m^2]
#> 6  1.584948 [m^2/m^2] 1.582821 [m^2/m^2] 0 [m^2/m^2] 1.389421 [L/m^2]
#> 7  1.584948 [m^2/m^2] 1.582821 [m^2/m^2] 0 [m^2/m^2] 1.389421 [L/m^2]
#> 8  1.584948 [m^2/m^2] 1.582821 [m^2/m^2] 0 [m^2/m^2] 1.389421 [L/m^2]
#> 9  1.584948 [m^2/m^2] 1.582821 [m^2/m^2] 0 [m^2/m^2] 1.389421 [L/m^2]
#> 10 1.584948 [m^2/m^2] 1.582821 [m^2/m^2] 0 [m^2/m^2] 1.389421 [L/m^2]
#>      LgroundPAR   LgroundSWR            SWE
#> 1  40.00754 [%] 50.73297 [%] 0.0000 [L/m^2]
#> 2  40.05271 [%] 50.77539 [%] 0.0000 [L/m^2]
#> 3  40.05271 [%] 50.77539 [%] 0.0000 [L/m^2]
#> 4  40.05278 [%] 50.77546 [%] 0.0000 [L/m^2]
#> 5  40.05435 [%] 50.77693 [%] 0.0000 [L/m^2]
#> 6  40.05435 [%] 50.77693 [%] 0.0000 [L/m^2]
#> 7  40.05435 [%] 50.77693 [%] 5.3774 [L/m^2]
#> 8  40.05435 [%] 50.77693 [%] 5.3774 [L/m^2]
#> 9  40.05435 [%] 50.77693 [%] 5.3774 [L/m^2]
#> 10 40.05435 [%] 50.77693 [%] 0.0000 [L/m^2]

#Extracts daily soil layer-level output as a data frame
extract(S1, level = "soillayer", addunits = TRUE)
#>          date soillayer       SWC       RWC       REW                ML
#> 1  2001-01-01         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 2  2001-01-02         1 0.3028369 0.9994790 0.9992187  68.13830 [L/m^2]
#> 3  2001-01-03         1 0.2997926 0.9894316 0.9841504  67.45333 [L/m^2]
#> 4  2001-01-04         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 5  2001-01-05         1 0.3017790 0.9959875 0.9939824  67.90027 [L/m^2]
#> 6  2001-01-06         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 7  2001-01-07         1 0.3024771 0.9982916 0.9974379  68.05735 [L/m^2]
#> 8  2001-01-08         1 0.3018024 0.9960648 0.9940983  67.90554 [L/m^2]
#> 9  2001-01-09         1 0.3003230 0.9911824 0.9867760  67.57268 [L/m^2]
#> 10 2001-01-10         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 11 2001-01-01         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 12 2001-01-02         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 13 2001-01-03         2 0.3028114 0.9993948 0.9990924 116.58238 [L/m^2]
#> 14 2001-01-04         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 15 2001-01-05         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 16 2001-01-06         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 17 2001-01-07         2 0.3029318 0.9997922 0.9996883 116.62873 [L/m^2]
#> 18 2001-01-08         2 0.3028471 0.9995129 0.9992695 116.59615 [L/m^2]
#> 19 2001-01-09         2 0.3025792 0.9986285 0.9979431 116.49298 [L/m^2]
#> 20 2001-01-10         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 21 2001-01-01         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 22 2001-01-02         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 23 2001-01-03         3 0.3029340 0.9997996 0.9996994  75.73350 [L/m^2]
#> 24 2001-01-04         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 25 2001-01-05         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 26 2001-01-06         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 27 2001-01-07         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 28 2001-01-08         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 29 2001-01-09         3 0.3028915 0.9996594 0.9994892  75.72288 [L/m^2]
#> 30 2001-01-10         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 31 2001-01-01         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 32 2001-01-02         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 33 2001-01-03         4 0.3028841 0.9996349 0.9994524  30.28841 [L/m^2]
#> 34 2001-01-04         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 35 2001-01-05         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 36 2001-01-06         4 0.3704020 1.2224700 1.3336433  37.04020 [L/m^2]
#> 37 2001-01-07         4 0.3290487 1.0859882 1.1289584  32.90487 [L/m^2]
#> 38 2001-01-08         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 39 2001-01-09         4 0.3028002 0.9993578 0.9990369  30.28002 [L/m^2]
#> 40 2001-01-10         4 0.3404099 1.1234846 1.1851926  34.04099 [L/m^2]
#> 41 2001-01-01   Overall 0.3029947 1.0000000 1.0000000 290.87494 [L/m^2]
#> 42 2001-01-02   Overall 0.3029577 0.9998779 0.9998169 290.83942 [L/m^2]
#> 43 2001-01-03   Overall 0.3021433 0.9971901 0.9957859 290.05761 [L/m^2]
#> 44 2001-01-04   Overall 0.3029947 1.0000000 1.0000000 290.87494 [L/m^2]
#> 45 2001-01-05   Overall 0.3027098 0.9990596 0.9985896 290.60139 [L/m^2]
#> 46 2001-01-06   Overall 0.3100163 1.0231740 1.0347545 297.61566 [L/m^2]
#> 47 2001-01-07   Overall 0.3055621 1.0084734 1.0127077 293.33963 [L/m^2]
#> 48 2001-01-08   Overall 0.3026561 0.9988823 0.9983238 290.54984 [L/m^2]
#> 49 2001-01-09   Overall 0.3021547 0.9972277 0.9958424 290.06856 [L/m^2]
#> 50 2001-01-10   Overall 0.3068921 1.0128630 1.0192909 294.61646 [L/m^2]
#>                  Psi             PlantExt      HydraulicInput
#> 1  -0.03300000 [MPa]  0.159160537 [L/m^2] 0.000000000 [L/m^2]
#> 2  -0.03308155 [MPa]  0.294996978 [L/m^2] 0.000000000 [L/m^2]
#> 3  -0.03469413 [MPa]  0.204403447 [L/m^2] 0.000000000 [L/m^2]
#> 4  -0.03300000 [MPa]  0.088443349 [L/m^2] 0.000000000 [L/m^2]
#> 5  -0.03363329 [MPa]  0.301923615 [L/m^2] 0.000000000 [L/m^2]
#> 6  -0.03300000 [MPa]  0.188923107 [L/m^2] 0.000000000 [L/m^2]
#> 7  -0.03326817 [MPa]  0.109214028 [L/m^2] 0.000000000 [L/m^2]
#> 8  -0.03362098 [MPa]  0.142340008 [L/m^2] 0.000000000 [L/m^2]
#> 9  -0.03440760 [MPa]  0.310277436 [L/m^2] 0.000000000 [L/m^2]
#> 10 -0.03300000 [MPa]  0.129130799 [L/m^2] 0.000000000 [L/m^2]
#> 11 -0.03300000 [MPa]  0.040327982 [L/m^2] 0.000000000 [L/m^2]
#> 12 -0.03300000 [MPa]  0.074656301 [L/m^2] 0.000000000 [L/m^2]
#> 13 -0.03309475 [MPa]  0.069439709 [L/m^2] 0.000000000 [L/m^2]
#> 14 -0.03300000 [MPa]  0.030769449 [L/m^2] 0.000000000 [L/m^2]
#> 15 -0.03300000 [MPa]  0.076409259 [L/m^2] 0.000000000 [L/m^2]
#> 16 -0.03300000 [MPa]  0.065050815 [L/m^2] 0.000000000 [L/m^2]
#> 17 -0.03303251 [MPa]  0.023677141 [L/m^2] 0.000000000 [L/m^2]
#> 18 -0.03307625 [MPa]  0.031841157 [L/m^2] 0.000000000 [L/m^2]
#> 19 -0.03321513 [MPa]  0.101412039 [L/m^2] 0.000000000 [L/m^2]
#> 20 -0.03300000 [MPa]  0.043163843 [L/m^2] 0.000000000 [L/m^2]
#> 21 -0.03300000 [MPa]  0.003383666 [L/m^2] 0.000000000 [L/m^2]
#> 22 -0.03300000 [MPa]  0.006263206 [L/m^2] 0.000000000 [L/m^2]
#> 23 -0.03303135 [MPa]  0.014980079 [L/m^2] 0.000000000 [L/m^2]
#> 24 -0.03300000 [MPa]  0.006961533 [L/m^2] 0.000000000 [L/m^2]
#> 25 -0.03300000 [MPa]  0.006410268 [L/m^2] 0.000000000 [L/m^2]
#> 26 -0.03300000 [MPa]  0.013946518 [L/m^2] 0.000000000 [L/m^2]
#> 27 -0.03300000 [MPa] -0.002496258 [L/m^2] 0.002845012 [L/m^2]
#> 28 -0.03300000 [MPa] -0.002784699 [L/m^2] 0.003458333 [L/m^2]
#> 29 -0.03305330 [MPa]  0.025492583 [L/m^2] 0.000000000 [L/m^2]
#> 30 -0.03300000 [MPa]  0.011050187 [L/m^2] 0.000000000 [L/m^2]
#> 31 -0.03300000 [MPa]  0.000668407 [L/m^2] 0.000000000 [L/m^2]
#> 32 -0.03300000 [MPa]  0.001237292 [L/m^2] 0.000000000 [L/m^2]
#> 33 -0.03305714 [MPa]  0.010978508 [L/m^2] 0.000000000 [L/m^2]
#> 34 -0.03300000 [MPa]  0.004857156 [L/m^2] 0.000000000 [L/m^2]
#> 35 -0.03300000 [MPa]  0.001266344 [L/m^2] 0.000000000 [L/m^2]
#> 36 -0.01037589 [MPa]  0.010194227 [L/m^2] 0.000000000 [L/m^2]
#> 37 -0.02184381 [MPa]  0.016269133 [L/m^2] 0.000000000 [L/m^2]
#> 38 -0.03300000 [MPa]  0.020171289 [L/m^2] 0.000000000 [L/m^2]
#> 39 -0.03310056 [MPa]  0.019327761 [L/m^2] 0.000000000 [L/m^2]
#> 40 -0.01810570 [MPa]  0.007462187 [L/m^2] 0.000000000 [L/m^2]
#> 41 -0.03300000 [MPa]  0.203540592 [L/m^2] 0.000000000 [L/m^2]
#> 42 -0.03301911 [MPa]  0.377153777 [L/m^2] 0.000000000 [L/m^2]
#> 43 -0.03344918 [MPa]  0.299801744 [L/m^2] 0.000000000 [L/m^2]
#> 44 -0.03300000 [MPa]  0.131031487 [L/m^2] 0.000000000 [L/m^2]
#> 45 -0.03314843 [MPa]  0.386009486 [L/m^2] 0.000000000 [L/m^2]
#> 46 -0.03064332 [MPa]  0.278114668 [L/m^2] 0.000000000 [L/m^2]
#> 47 -0.03191379 [MPa]  0.146664044 [L/m^2] 0.000000000 [L/m^2]
#> 48 -0.03317612 [MPa]  0.191567755 [L/m^2] 0.000000000 [L/m^2]
#> 49 -0.03344054 [MPa]  0.456509819 [L/m^2] 0.000000000 [L/m^2]
#> 50 -0.03144851 [MPa]  0.190807016 [L/m^2] 0.000000000 [L/m^2]

#Extracts daily cohort-level output as a data frame
extract(S1, level = "cohort", addunits = TRUE)
#>          date cohort           species                  LAI
#> 1  2001-01-01 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 2  2001-01-02 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 3  2001-01-03 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 4  2001-01-04 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 5  2001-01-05 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 6  2001-01-06 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 7  2001-01-07 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 8  2001-01-08 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 9  2001-01-09 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 10 2001-01-10 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 11 2001-01-01 T2_168      Quercus ilex 0.70557382 [m^2/m^2]
#> 12 2001-01-02 T2_168      Quercus ilex 0.70352348 [m^2/m^2]
#> 13 2001-01-03 T2_168      Quercus ilex 0.70352348 [m^2/m^2]
#> 14 2001-01-04 T2_168      Quercus ilex 0.70352008 [m^2/m^2]
#> 15 2001-01-05 T2_168      Quercus ilex 0.70344882 [m^2/m^2]
#> 16 2001-01-06 T2_168      Quercus ilex 0.70344882 [m^2/m^2]
#> 17 2001-01-07 T2_168      Quercus ilex 0.70344882 [m^2/m^2]
#> 18 2001-01-08 T2_168      Quercus ilex 0.70344882 [m^2/m^2]
#> 19 2001-01-09 T2_168      Quercus ilex 0.70344882 [m^2/m^2]
#> 20 2001-01-10 T2_168      Quercus ilex 0.70344882 [m^2/m^2]
#> 21 2001-01-01 S1_165 Quercus coccifera 0.03062604 [m^2/m^2]
#> 22 2001-01-02 S1_165 Quercus coccifera 0.03062498 [m^2/m^2]
#> 23 2001-01-03 S1_165 Quercus coccifera 0.03062498 [m^2/m^2]
#> 24 2001-01-04 S1_165 Quercus coccifera 0.03062497 [m^2/m^2]
#> 25 2001-01-05 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 26 2001-01-06 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 27 2001-01-07 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 28 2001-01-08 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 29 2001-01-09 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 30 2001-01-10 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#>                 LAIlive         FPAR AbsorbedSWRFraction       Transpiration
#> 1  0.84874773 [m^2/m^2] 92.18285 [%]           35.076344 0.125013855 [L/m^2]
#> 2  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.231814863 [L/m^2]
#> 3  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.184270455 [L/m^2]
#> 4  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.080532847 [L/m^2]
#> 5  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.237257962 [L/m^2]
#> 6  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.170937295 [L/m^2]
#> 7  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.090144801 [L/m^2]
#> 8  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.117743936 [L/m^2]
#> 9  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.280584915 [L/m^2]
#> 10 0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.117273011 [L/m^2]
#> 11 0.70557382 [m^2/m^2] 72.36365 [%]           30.444383 0.072474080 [L/m^2]
#> 12 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.134110679 [L/m^2]
#> 13 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.106605840 [L/m^2]
#> 14 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.046597154 [L/m^2]
#> 15 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.137259647 [L/m^2]
#> 16 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.098897197 [L/m^2]
#> 17 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.052152864 [L/m^2]
#> 18 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.068120498 [L/m^2]
#> 19 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.162333459 [L/m^2]
#> 20 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.067852823 [L/m^2]
#> 21 0.03062604 [m^2/m^2] 44.32407 [%]            2.366131 0.006052657 [L/m^2]
#> 22 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.011228235 [L/m^2]
#> 23 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.008925449 [L/m^2]
#> 24 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.003901486 [L/m^2]
#> 25 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.011491878 [L/m^2]
#> 26 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.008280175 [L/m^2]
#> 27 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.004366379 [L/m^2]
#> 28 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.005703320 [L/m^2]
#> 29 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.013591444 [L/m^2]
#> 30 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.005681182 [L/m^2]
#>    GrossPhotosynthesis          PlantPsi      LeafPLC      StemPLC
#> 1   1.21341578 [L/m^2] -0.03300000 [MPa] 0.000000e+00 0.000000e+00
#> 2   2.10348703 [L/m^2] -0.03300000 [MPa] 0.000000e+00 0.000000e+00
#> 3   1.77394894 [L/m^2] -0.03307746 [MPa] 0.000000e+00 0.000000e+00
#> 4   0.81305608 [L/m^2] -0.03461466 [MPa] 0.000000e+00 0.000000e+00
#> 5   2.18356104 [L/m^2] -0.03300000 [MPa] 0.000000e+00 0.000000e+00
#> 6   1.79056179 [L/m^2] -0.03360166 [MPa] 0.000000e+00 0.000000e+00
#> 7   1.02740368 [L/m^2] -0.03299496 [MPa] 0.000000e+00 0.000000e+00
#> 8   1.21701511 [L/m^2] -0.03325349 [MPa] 0.000000e+00 0.000000e+00
#> 9   2.08853049 [L/m^2] -0.03359361 [MPa] 0.000000e+00 0.000000e+00
#> 10  1.16044155 [L/m^2] -0.03434796 [MPa] 0.000000e+00 0.000000e+00
#> 11  0.87313959 [L/m^2] -0.03300000 [MPa] 2.905929e-03 3.681226e-09
#> 12  1.59099654 [L/m^2] -0.03300000 [MPa] 2.905929e-03 3.681226e-09
#> 13  1.28192745 [L/m^2] -0.03304085 [MPa] 2.910741e-03 3.697476e-09
#> 14  0.56671680 [L/m^2] -0.03389435 [MPa] 3.011734e-03 4.048890e-09
#> 15  1.63354269 [L/m^2] -0.03300000 [MPa] 3.011734e-03 4.048890e-09
#> 16  1.21374928 [L/m^2] -0.03331743 [MPa] 3.011734e-03 4.048890e-09
#> 17  0.65412428 [L/m^2] -0.03283351 [MPa] 3.011734e-03 4.048890e-09
#> 18  0.83326703 [L/m^2] -0.03306369 [MPa] 3.011734e-03 4.048890e-09
#> 19  1.84697725 [L/m^2] -0.03334556 [MPa] 3.011734e-03 4.048890e-09
#> 20  0.82124762 [L/m^2] -0.03380572 [MPa] 3.011734e-03 4.048890e-09
#> 21  0.05005525 [L/m^2] -0.03300000 [MPa] 3.473351e-05 4.394058e-08
#> 22  0.08989595 [L/m^2] -0.03300000 [MPa] 3.473351e-05 4.394058e-08
#> 23  0.07354416 [L/m^2] -0.03305546 [MPa] 3.486528e-05 4.416958e-08
#> 24  0.03292077 [L/m^2] -0.03418210 [MPa] 3.760235e-05 4.899806e-08
#> 25  0.09263463 [L/m^2] -0.03300000 [MPa] 3.760235e-05 4.899806e-08
#> 26  0.07119407 [L/m^2] -0.03343102 [MPa] 3.760235e-05 4.899806e-08
#> 27  0.03924413 [L/m^2] -0.03279136 [MPa] 3.760235e-05 4.899806e-08
#> 28  0.04870336 [L/m^2] -0.03308026 [MPa] 3.760235e-05 4.899806e-08
#> 29  0.09940022 [L/m^2] -0.03344343 [MPa] 3.760235e-05 4.899806e-08
#> 30  0.04745604 [L/m^2] -0.03402042 [MPa] 3.760235e-05 4.899806e-08
#>        PlantWaterBalance       LeafRWC       StemRWC          LFMC PlantStress
#> 1   0.000000e+00 [L/m^2] 0.9979679 [%] 0.9998350 [%] 125.86544 [%] 0.004613739
#> 2   0.000000e+00 [L/m^2] 0.9979679 [%] 0.9998350 [%] 125.86544 [%] 0.004613739
#> 3  -3.486537e-06 [L/m^2] 0.9979631 [%] 0.9998346 [%] 125.86505 [%] 0.004630005
#> 4  -6.918596e-05 [L/m^2] 0.9978685 [%] 0.9998269 [%] 125.85736 [%] 0.004956690
#> 5   7.267250e-05 [L/m^2] 0.9979679 [%] 0.9998350 [%] 125.86544 [%] 0.004613739
#> 6  -2.707994e-05 [L/m^2] 0.9979309 [%] 0.9998320 [%] 125.86243 [%] 0.004740578
#> 7   2.730681e-05 [L/m^2] 0.9979682 [%] 0.9998350 [%] 125.86546 [%] 0.004612681
#> 8  -1.163619e-05 [L/m^2] 0.9979523 [%] 0.9998337 [%] 125.86417 [%] 0.004667040
#> 9  -1.530789e-05 [L/m^2] 0.9979314 [%] 0.9998320 [%] 125.86247 [%] 0.004738871
#> 10 -3.395188e-05 [L/m^2] 0.9978849 [%] 0.9998282 [%] 125.85869 [%] 0.004899483
#> 11  0.000000e+00 [L/m^2] 0.9986797 [%] 0.9997500 [%]  93.05186 [%] 0.006282261
#> 12  0.000000e+00 [L/m^2] 0.9986797 [%] 0.9997500 [%]  92.91408 [%] 0.006282261
#> 13 -4.289616e-07 [L/m^2] 0.9986781 [%] 0.9997497 [%]  92.91395 [%] 0.006291170
#> 14 -8.961833e-06 [L/m^2] 0.9986439 [%] 0.9997432 [%]  92.91111 [%] 0.006477652
#> 15  9.390232e-06 [L/m^2] 0.9986797 [%] 0.9997500 [%]  92.90906 [%] 0.006282261
#> 16 -3.332823e-06 [L/m^2] 0.9986670 [%] 0.9997476 [%]  92.90809 [%] 0.006351525
#> 17  5.080943e-06 [L/m^2] 0.9986864 [%] 0.9997512 [%]  92.90957 [%] 0.006245969
#> 18 -2.416870e-06 [L/m^2] 0.9986772 [%] 0.9997495 [%]  92.90886 [%] 0.006296152
#> 19 -2.959525e-06 [L/m^2] 0.9986659 [%] 0.9997473 [%]  92.90800 [%] 0.006357670
#> 20 -4.831417e-06 [L/m^2] 0.9986475 [%] 0.9997439 [%]  92.90659 [%] 0.006458257
#> 21  3.469447e-18 [L/m^2] 0.9987207 [%] 0.9983684 [%]  96.38420 [%] 0.003088161
#> 22  0.000000e+00 [L/m^2] 0.9987207 [%] 0.9983684 [%]  96.38552 [%] 0.003088161
#> 23 -5.275953e-08 [L/m^2] 0.9987185 [%] 0.9983657 [%]  96.38527 [%] 0.003094899
#> 24 -1.071742e-06 [L/m^2] 0.9986748 [%] 0.9983100 [%]  96.38015 [%] 0.003232501
#> 25  1.124500e-06 [L/m^2] 0.9987207 [%] 0.9983684 [%]  96.38563 [%] 0.003088161
#> 26 -4.100197e-07 [L/m^2] 0.9987040 [%] 0.9983471 [%]  96.38367 [%] 0.003140616
#> 27  6.084966e-07 [L/m^2] 0.9987287 [%] 0.9983787 [%]  96.38658 [%] 0.003062842
#> 28 -2.748240e-07 [L/m^2] 0.9987175 [%] 0.9983644 [%]  96.38527 [%] 0.003097913
#> 29 -3.454849e-07 [L/m^2] 0.9987035 [%] 0.9983465 [%]  96.38362 [%] 0.003142130
#> 30 -5.488693e-07 [L/m^2] 0.9986811 [%] 0.9983180 [%]  96.38099 [%] 0.003212671

#Select the output tables/variables to be extracted
extract(S1, level ="cohort", output="Plants", vars = c("PlantStress", "StemPLC"))
#>          date cohort           species PlantStress      StemPLC
#> 1  2001-01-01 T1_148  Pinus halepensis 0.004613739 0.000000e+00
#> 2  2001-01-02 T1_148  Pinus halepensis 0.004613739 0.000000e+00
#> 3  2001-01-03 T1_148  Pinus halepensis 0.004630005 0.000000e+00
#> 4  2001-01-04 T1_148  Pinus halepensis 0.004956690 0.000000e+00
#> 5  2001-01-05 T1_148  Pinus halepensis 0.004613739 0.000000e+00
#> 6  2001-01-06 T1_148  Pinus halepensis 0.004740578 0.000000e+00
#> 7  2001-01-07 T1_148  Pinus halepensis 0.004612681 0.000000e+00
#> 8  2001-01-08 T1_148  Pinus halepensis 0.004667040 0.000000e+00
#> 9  2001-01-09 T1_148  Pinus halepensis 0.004738871 0.000000e+00
#> 10 2001-01-10 T1_148  Pinus halepensis 0.004899483 0.000000e+00
#> 11 2001-01-01 T2_168      Quercus ilex 0.006282261 3.681226e-09
#> 12 2001-01-02 T2_168      Quercus ilex 0.006282261 3.681226e-09
#> 13 2001-01-03 T2_168      Quercus ilex 0.006291170 3.697476e-09
#> 14 2001-01-04 T2_168      Quercus ilex 0.006477652 4.048890e-09
#> 15 2001-01-05 T2_168      Quercus ilex 0.006282261 4.048890e-09
#> 16 2001-01-06 T2_168      Quercus ilex 0.006351525 4.048890e-09
#> 17 2001-01-07 T2_168      Quercus ilex 0.006245969 4.048890e-09
#> 18 2001-01-08 T2_168      Quercus ilex 0.006296152 4.048890e-09
#> 19 2001-01-09 T2_168      Quercus ilex 0.006357670 4.048890e-09
#> 20 2001-01-10 T2_168      Quercus ilex 0.006458257 4.048890e-09
#> 21 2001-01-01 S1_165 Quercus coccifera 0.003088161 4.394058e-08
#> 22 2001-01-02 S1_165 Quercus coccifera 0.003088161 4.394058e-08
#> 23 2001-01-03 S1_165 Quercus coccifera 0.003094899 4.416958e-08
#> 24 2001-01-04 S1_165 Quercus coccifera 0.003232501 4.899806e-08
#> 25 2001-01-05 S1_165 Quercus coccifera 0.003088161 4.899806e-08
#> 26 2001-01-06 S1_165 Quercus coccifera 0.003140616 4.899806e-08
#> 27 2001-01-07 S1_165 Quercus coccifera 0.003062842 4.899806e-08
#> 28 2001-01-08 S1_165 Quercus coccifera 0.003097913 4.899806e-08
#> 29 2001-01-09 S1_165 Quercus coccifera 0.003142130 4.899806e-08
#> 30 2001-01-10 S1_165 Quercus coccifera 0.003212671 4.899806e-08