Extracts model outputs

Function extract() extracts daily or subdaily output and returns it as a tidy data frame.

Usage

extract(
  x,
  level = "forest",
  output = NULL,
  vars = NULL,
  dates = NULL,
  subdaily = FALSE,
  addunits = FALSE
)

Arguments

x

An object returned by simulation functions spwb, aspwb, pwb, growth or fordyn.

level

Level of simulation output, either "forest" (stand-level results), "soillayer" (soil layer-level results), "cohort" (cohort-level results), "sunlitleaf" or "shadeleaf" (leaf-level results)

output

Section of the model output to be explored. See details.

vars

Variables to be extracted (by default, all of them).

dates

A date vector indicating the subset of simulated days for which output is desired.

subdaily

A flag to indicate that subdaily values are desired (see details).

addunits

A flag to indicate that variable units should be added whenever possible.

Value

Function extract() returns a data frame:

If level = "forest", columns are "date" and variable names.

If level = "soillayer", columns are "date", "soillayer" and variable names.

If level = "cohort", level = "sunlitleaf" or level = "shadeleaf", columns are "date", "cohorts", "species" and variable names.

If subdaily = TRUE, columns are "datetime", "cohorts", "species" and variable names.

Details

When subdaily = FALSE, parameter output is used to restrict the section in x where variables are located. For example output = "Plants" will correspond to variables "LAI", "LAIlive", "Transpiration", "StemPLC",... as returned by a call names(x$Plants).

Option subdaily = TRUE only works when simulations have been carried using control option 'subdailyResults = TRUE' (see defaultControl). When using subdaily = TRUE, parameter output is not taken into account, and options for parameter vars are the following:

Variables for level = "forest" or level = "soillayer": Not allowed. An error is raised.

Variables for level = "cohort": "E","Ag","An","dEdP","RootPsi","StemPsi","LeafPsi","StemPLC","StemRWC","LeafRWC","StemSympRWC","LeafSympRWC","PWB".

Variables for level = "shadeleaf" and level="sunlitleaf": "Abs_SWR","Abs_PAR","Net_LWR","E","Ag","An","Ci","Gsw","VPD","Temp","Psi","iWUE".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

#Define soil with default soil params (4 layers)
examplesoil <- defaultSoilParams(4)

#Initialize control parameters
control <- defaultControl("Granier")

#Initialize input
x <- spwbInput(exampleforest,examplesoil, SpParamsMED, control)

#Call simulation function (ten days)
S1<-spwb(x, examplemeteo[1:10, ], latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 4.73001
#> Initial soil water content (mm): 290.875
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:
#> 
#> Final plant water content (mm): 4.72995
#> Final soil water content (mm): 294.611
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): -6.10983e-05
#> Plant water balance result (mm): -6.10844e-05
#> Change in soil water content (mm): 3.7356
#> Soil water balance result (mm): 3.7356
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 39 Rain (mm) 34 Snow (mm) 5
#>   Interception (mm) 9 Net rainfall (mm) 25
#>   Infiltration (mm) 30 Infiltration excess (mm) 0 Saturation excess (mm) 0 Capillarity rise (mm) 0
#>   Soil evaporation (mm) 3  Herbaceous transpiration (mm) 0 Woody plant transpiration (mm) 3
#>   Plant extraction from soil (mm) 3  Plant water balance (mm) -0 Hydraulic redistribution (mm) 0
#>   Runoff (mm) 0 Deep drainage (mm) 21

#Extracts daily forest-level output as a data frame
extract(S1, level = "forest", addunits = TRUE)
#>          date               PET     Precipitation              Rain
#> 1  2001-01-01 0.8828475 [L/m^2]  4.869109 [L/m^2]  4.869109 [L/m^2]
#> 2  2001-01-02 1.6375337 [L/m^2]  2.498292 [L/m^2]  2.498292 [L/m^2]
#> 3  2001-01-03 1.3017026 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 4  2001-01-04 0.5690790 [L/m^2]  5.796973 [L/m^2]  5.796973 [L/m^2]
#> 5  2001-01-05 1.6760567 [L/m^2]  1.884401 [L/m^2]  1.884401 [L/m^2]
#> 6  2001-01-06 1.2077028 [L/m^2] 13.359801 [L/m^2] 13.359801 [L/m^2]
#> 7  2001-01-07 0.6368075 [L/m^2]  5.377400 [L/m^2]  0.000000 [L/m^2]
#> 8  2001-01-08 0.8318207 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 9  2001-01-09 1.9823797 [L/m^2]  0.000000 [L/m^2]  0.000000 [L/m^2]
#> 10 2001-01-10 0.8286873 [L/m^2]  5.117768 [L/m^2]  5.117768 [L/m^2]
#>              Snow            NetRain       Snowmelt       Infiltration
#> 1  0.0000 [L/m^2]  3.4241795 [L/m^2] 0.0000 [L/m^2]  3.4241795 [L/m^2]
#> 2  0.0000 [L/m^2]  1.0728002 [L/m^2] 0.0000 [L/m^2]  1.0728002 [L/m^2]
#> 3  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 4  0.0000 [L/m^2]  4.3636213 [L/m^2] 0.0000 [L/m^2]  4.3636213 [L/m^2]
#> 5  0.0000 [L/m^2]  0.7547843 [L/m^2] 0.0000 [L/m^2]  0.7547843 [L/m^2]
#> 6  0.0000 [L/m^2] 11.7252817 [L/m^2] 0.0000 [L/m^2] 11.7252817 [L/m^2]
#> 7  5.3774 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 8  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 9  0.0000 [L/m^2]  0.0000000 [L/m^2] 0.0000 [L/m^2]  0.0000000 [L/m^2]
#> 10 0.0000 [L/m^2]  3.6738416 [L/m^2] 5.3774 [L/m^2]  9.0512413 [L/m^2]
#>    InfiltrationExcess SaturationExcess    Runoff      DeepDrainage
#> 1           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 2.7617207 [L/m^2]
#> 2           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.1899867 [L/m^2]
#> 3           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.0000000 [L/m^2]
#> 4           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 3.2495151 [L/m^2]
#> 5           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.1001772 [L/m^2]
#> 6           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#> 7           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#> 8           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 2.5926340 [L/m^2]
#> 9           0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 0.0000000 [L/m^2]
#> 10          0 [L/m^2]        0 [L/m^2] 0 [L/m^2] 4.1214138 [L/m^2]
#>    CapillarityRise Evapotranspiration     Interception   SoilEvaporation
#> 1        0 [L/m^2]  2.1073881 [L/m^2] 1.444929 [L/m^2] 0.4478948 [L/m^2]
#> 2        0 [L/m^2]  2.3231089 [L/m^2] 1.425491 [L/m^2] 0.5000000 [L/m^2]
#> 3        0 [L/m^2]  0.8014863 [L/m^2] 0.000000 [L/m^2] 0.4854151 [L/m^2]
#> 4        0 [L/m^2]  1.7311676 [L/m^2] 1.433352 [L/m^2] 0.1596697 [L/m^2]
#> 5        0 [L/m^2]  2.0365726 [L/m^2] 1.129617 [L/m^2] 0.5000000 [L/m^2]
#> 6        0 [L/m^2]  2.2354533 [L/m^2] 1.634519 [L/m^2] 0.3077228 [L/m^2]
#> 7        0 [L/m^2]  0.1546240 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 8        0 [L/m^2]  0.2019654 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 9        0 [L/m^2]  0.4812894 [L/m^2] 0.000000 [L/m^2] 0.0000000 [L/m^2]
#> 10       0 [L/m^2]  1.8368128 [L/m^2] 1.443926 [L/m^2] 0.1917182 [L/m^2]
#>      HerbTranspiration   PlantExtraction     Transpiration
#> 1  0.011023432 [L/m^2] 0.2035406 [L/m^2] 0.2035406 [L/m^2]
#> 2  0.020463709 [L/m^2] 0.3771538 [L/m^2] 0.3771538 [L/m^2]
#> 3  0.016266929 [L/m^2] 0.2998042 [L/m^2] 0.2998042 [L/m^2]
#> 4  0.007111348 [L/m^2] 0.1310349 [L/m^2] 0.1310349 [L/m^2]
#> 5  0.020945746 [L/m^2] 0.3860097 [L/m^2] 0.3860097 [L/m^2]
#> 6  0.015092529 [L/m^2] 0.2781190 [L/m^2] 0.2781190 [L/m^2]
#> 7  0.007958208 [L/m^2] 0.1466657 [L/m^2] 0.1466657 [L/m^2]
#> 8  0.010395243 [L/m^2] 0.1915701 [L/m^2] 0.1915701 [L/m^2]
#> 9  0.024773584 [L/m^2] 0.4565159 [L/m^2] 0.4565159 [L/m^2]
#> 10 0.010355858 [L/m^2] 0.1908126 [L/m^2] 0.1908126 [L/m^2]
#>    HydraulicRedistribution                LAI             LAIherb
#> 1      0.000000000 [L/m^2] 1.758585 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 2      0.000000000 [L/m^2] 1.756533 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 3      0.000000000 [L/m^2] 1.756533 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 4      0.000000000 [L/m^2] 1.756532 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 5      0.000000000 [L/m^2] 1.756459 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 6      0.000000000 [L/m^2] 1.756459 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 7      0.001658730 [L/m^2] 1.756459 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 8      0.001891912 [L/m^2] 1.756459 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 9      0.000000000 [L/m^2] 1.756459 [m^2/m^2] 0.1736369 [m^2/m^2]
#> 10     0.000000000 [L/m^2] 1.756459 [m^2/m^2] 0.1736369 [m^2/m^2]
#>               LAIlive        LAIexpanded     LAIdead               Cm
#> 1  1.584948 [m^2/m^2] 1.584948 [m^2/m^2] 0 [m^2/m^2] 1.390485 [L/m^2]
#> 2  1.584948 [m^2/m^2] 1.582896 [m^2/m^2] 0 [m^2/m^2] 1.389459 [L/m^2]
#> 3  1.584948 [m^2/m^2] 1.582896 [m^2/m^2] 0 [m^2/m^2] 1.389459 [L/m^2]
#> 4  1.584948 [m^2/m^2] 1.582895 [m^2/m^2] 0 [m^2/m^2] 1.389458 [L/m^2]
#> 5  1.584948 [m^2/m^2] 1.582823 [m^2/m^2] 0 [m^2/m^2] 1.389422 [L/m^2]
#> 6  1.584948 [m^2/m^2] 1.582823 [m^2/m^2] 0 [m^2/m^2] 1.389422 [L/m^2]
#> 7  1.584948 [m^2/m^2] 1.582823 [m^2/m^2] 0 [m^2/m^2] 1.389422 [L/m^2]
#> 8  1.584948 [m^2/m^2] 1.582823 [m^2/m^2] 0 [m^2/m^2] 1.389422 [L/m^2]
#> 9  1.584948 [m^2/m^2] 1.582823 [m^2/m^2] 0 [m^2/m^2] 1.389422 [L/m^2]
#> 10 1.584948 [m^2/m^2] 1.582823 [m^2/m^2] 0 [m^2/m^2] 1.389422 [L/m^2]
#>      LgroundPAR   LgroundSWR            SWE
#> 1  40.00754 [%] 50.73297 [%] 0.0000 [L/m^2]
#> 2  40.05271 [%] 50.77539 [%] 0.0000 [L/m^2]
#> 3  40.05271 [%] 50.77539 [%] 0.0000 [L/m^2]
#> 4  40.05274 [%] 50.77541 [%] 0.0000 [L/m^2]
#> 5  40.05433 [%] 50.77691 [%] 0.0000 [L/m^2]
#> 6  40.05433 [%] 50.77691 [%] 0.0000 [L/m^2]
#> 7  40.05433 [%] 50.77691 [%] 5.3774 [L/m^2]
#> 8  40.05433 [%] 50.77691 [%] 5.3774 [L/m^2]
#> 9  40.05433 [%] 50.77691 [%] 5.3774 [L/m^2]
#> 10 40.05433 [%] 50.77691 [%] 0.0000 [L/m^2]

#Extracts daily soil layer-level output as a data frame
extract(S1, level = "soillayer", addunits = TRUE)
#>          date soillayer       SWC       RWC       REW                ML
#> 1  2001-01-01         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 2  2001-01-02         1 0.3029289 0.9997828 0.9996743  68.15901 [L/m^2]
#> 3  2001-01-03         1 0.2998701 0.9896877 0.9845344  67.47078 [L/m^2]
#> 4  2001-01-04         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 5  2001-01-05         1 0.3018732 0.9962985 0.9944487  67.92147 [L/m^2]
#> 6  2001-01-06         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 7  2001-01-07         1 0.3025167 0.9984223 0.9976339  68.06625 [L/m^2]
#> 8  2001-01-08         1 0.3018899 0.9963536 0.9945314  67.92522 [L/m^2]
#> 9  2001-01-09         1 0.3005230 0.9918423 0.9877658  67.61767 [L/m^2]
#> 10 2001-01-10         1 0.3029947 1.0000000 1.0000000  68.17381 [L/m^2]
#> 11 2001-01-01         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 12 2001-01-02         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 13 2001-01-03         2 0.3027773 0.9992825 0.9989240 116.56927 [L/m^2]
#> 14 2001-01-04         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 15 2001-01-05         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 16 2001-01-06         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 17 2001-01-07         2 0.3029154 0.9997381 0.9996072 116.62242 [L/m^2]
#> 18 2001-01-08         2 0.3028091 0.9993874 0.9990813 116.58151 [L/m^2]
#> 19 2001-01-09         2 0.3024933 0.9983452 0.9975182 116.45993 [L/m^2]
#> 20 2001-01-10         2 0.3029947 1.0000000 1.0000000 116.65297 [L/m^2]
#> 21 2001-01-01         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 22 2001-01-02         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 23 2001-01-03         3 0.3029242 0.9997671 0.9996508  75.73104 [L/m^2]
#> 24 2001-01-04         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 25 2001-01-05         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 26 2001-01-06         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 27 2001-01-07         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 28 2001-01-08         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 29 2001-01-09         3 0.3028728 0.9995975 0.9993963  75.71819 [L/m^2]
#> 30 2001-01-10         3 0.3029947 1.0000000 1.0000000  75.74868 [L/m^2]
#> 31 2001-01-01         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 32 2001-01-02         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 33 2001-01-03         4 0.3028755 0.9996065 0.9994098  30.28755 [L/m^2]
#> 34 2001-01-04         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 35 2001-01-05         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 36 2001-01-06         4 0.3705006 1.2227955 1.3341314  37.05006 [L/m^2]
#> 37 2001-01-07         4 0.3291213 1.0862278 1.1293177  32.91213 [L/m^2]
#> 38 2001-01-08         4 0.3029947 1.0000000 1.0000000  30.29947 [L/m^2]
#> 39 2001-01-09         4 0.3027780 0.9992847 0.9989273  30.27780 [L/m^2]
#> 40 2001-01-10         4 0.3403507 1.1232893 1.1848997  34.03507 [L/m^2]
#> 41 2001-01-01   Overall 0.3029947 1.0000000 1.0000000 290.87494 [L/m^2]
#> 42 2001-01-02   Overall 0.3029793 0.9999491 0.9999237 290.86014 [L/m^2]
#> 43 2001-01-03   Overall 0.3021444 0.9971937 0.9957913 290.05865 [L/m^2]
#> 44 2001-01-04   Overall 0.3029947 1.0000000 1.0000000 290.87494 [L/m^2]
#> 45 2001-01-05   Overall 0.3027319 0.9991325 0.9986989 290.62259 [L/m^2]
#> 46 2001-01-06   Overall 0.3100266 1.0232079 1.0348053 297.62552 [L/m^2]
#> 47 2001-01-07   Overall 0.3055724 1.0085073 1.0127585 293.34949 [L/m^2]
#> 48 2001-01-08   Overall 0.3026613 0.9988997 0.9983498 290.55489 [L/m^2]
#> 49 2001-01-09   Overall 0.3021600 0.9972451 0.9958684 290.07360 [L/m^2]
#> 50 2001-01-10   Overall 0.3068860 1.0128426 1.0192604 294.61054 [L/m^2]
#>                  Psi             PlantExt      HydraulicInput
#> 1  -0.03300000 [MPa]  0.147990816 [L/m^2] 0.000000000 [L/m^2]
#> 2  -0.03303397 [MPa]  0.274284815 [L/m^2] 0.000000000 [L/m^2]
#> 3  -0.03465208 [MPa]  0.187989165 [L/m^2] 0.000000000 [L/m^2]
#> 4  -0.03300000 [MPa]  0.080956152 [L/m^2] 0.000000000 [L/m^2]
#> 5  -0.03358379 [MPa]  0.280725302 [L/m^2] 0.000000000 [L/m^2]
#> 6  -0.03300000 [MPa]  0.173452219 [L/m^2] 0.000000000 [L/m^2]
#> 7  -0.03324759 [MPa]  0.100307500 [L/m^2] 0.000000000 [L/m^2]
#> 8  -0.03357502 [MPa]  0.131558125 [L/m^2] 0.000000000 [L/m^2]
#> 9  -0.03430021 [MPa]  0.284975093 [L/m^2] 0.000000000 [L/m^2]
#> 10 -0.03300000 [MPa]  0.118452195 [L/m^2] 0.000000000 [L/m^2]
#> 11 -0.03300000 [MPa]  0.049247162 [L/m^2] 0.000000000 [L/m^2]
#> 12 -0.03300000 [MPa]  0.091195256 [L/m^2] 0.000000000 [L/m^2]
#> 13 -0.03311237 [MPa]  0.082541033 [L/m^2] 0.000000000 [L/m^2]
#> 14 -0.03300000 [MPa]  0.036627132 [L/m^2] 0.000000000 [L/m^2]
#> 15 -0.03300000 [MPa]  0.093336613 [L/m^2] 0.000000000 [L/m^2]
#> 16 -0.03300000 [MPa]  0.077398069 [L/m^2] 0.000000000 [L/m^2]
#> 17 -0.03304097 [MPa]  0.029984218 [L/m^2] 0.000000000 [L/m^2]
#> 18 -0.03309592 [MPa]  0.040173749 [L/m^2] 0.000000000 [L/m^2]
#> 19 -0.03325974 [MPa]  0.119816786 [L/m^2] 0.000000000 [L/m^2]
#> 20 -0.03300000 [MPa]  0.050998413 [L/m^2] 0.000000000 [L/m^2]
#> 21 -0.03300000 [MPa]  0.005050851 [L/m^2] 0.000000000 [L/m^2]
#> 22 -0.03300000 [MPa]  0.009354689 [L/m^2] 0.000000000 [L/m^2]
#> 23 -0.03303643 [MPa]  0.017435555 [L/m^2] 0.000000000 [L/m^2]
#> 24 -0.03300000 [MPa]  0.008144219 [L/m^2] 0.000000000 [L/m^2]
#> 25 -0.03300000 [MPa]  0.009574346 [L/m^2] 0.000000000 [L/m^2]
#> 26 -0.03300000 [MPa]  0.016263052 [L/m^2] 0.000000000 [L/m^2]
#> 27 -0.03300000 [MPa] -0.001310085 [L/m^2] 0.001658730 [L/m^2]
#> 28 -0.03300000 [MPa] -0.001225646 [L/m^2] 0.001891912 [L/m^2]
#> 29 -0.03306299 [MPa]  0.030181021 [L/m^2] 0.000000000 [L/m^2]
#> 30 -0.03300000 [MPa]  0.013062366 [L/m^2] 0.000000000 [L/m^2]
#> 31 -0.03300000 [MPa]  0.001251763 [L/m^2] 0.000000000 [L/m^2]
#> 32 -0.03300000 [MPa]  0.002319017 [L/m^2] 0.000000000 [L/m^2]
#> 33 -0.03306158 [MPa]  0.011838459 [L/m^2] 0.000000000 [L/m^2]
#> 34 -0.03300000 [MPa]  0.005307351 [L/m^2] 0.000000000 [L/m^2]
#> 35 -0.03300000 [MPa]  0.002373470 [L/m^2] 0.000000000 [L/m^2]
#> 36 -0.01035437 [MPa]  0.011005677 [L/m^2] 0.000000000 [L/m^2]
#> 37 -0.02181811 [MPa]  0.017684115 [L/m^2] 0.000000000 [L/m^2]
#> 38 -0.03300000 [MPa]  0.021063890 [L/m^2] 0.000000000 [L/m^2]
#> 39 -0.03311202 [MPa]  0.021542956 [L/m^2] 0.000000000 [L/m^2]
#> 40 -0.01812380 [MPa]  0.008299633 [L/m^2] 0.000000000 [L/m^2]
#> 41 -0.03300000 [MPa]  0.203540592 [L/m^2] 0.000000000 [L/m^2]
#> 42 -0.03300796 [MPa]  0.377153777 [L/m^2] 0.000000000 [L/m^2]
#> 43 -0.03344817 [MPa]  0.299804212 [L/m^2] 0.000000000 [L/m^2]
#> 44 -0.03300000 [MPa]  0.131034853 [L/m^2] 0.000000000 [L/m^2]
#> 45 -0.03313682 [MPa]  0.386009732 [L/m^2] 0.000000000 [L/m^2]
#> 46 -0.03064108 [MPa]  0.278119017 [L/m^2] 0.000000000 [L/m^2]
#> 47 -0.03190968 [MPa]  0.146665748 [L/m^2] 0.000000000 [L/m^2]
#> 48 -0.03317324 [MPa]  0.191570119 [L/m^2] 0.000000000 [L/m^2]
#> 49 -0.03343697 [MPa]  0.456515856 [L/m^2] 0.000000000 [L/m^2]
#> 50 -0.03145040 [MPa]  0.190812607 [L/m^2] 0.000000000 [L/m^2]

#Extracts daily cohort-level output as a data frame
extract(S1, level = "cohort", addunits = TRUE)
#>          date cohort           species                  LAI
#> 1  2001-01-01 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 2  2001-01-02 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 3  2001-01-03 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 4  2001-01-04 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 5  2001-01-05 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 6  2001-01-06 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 7  2001-01-07 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 8  2001-01-08 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 9  2001-01-09 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 10 2001-01-10 T1_148  Pinus halepensis 0.84874773 [m^2/m^2]
#> 11 2001-01-01 T2_168      Quercus ilex 0.70557382 [m^2/m^2]
#> 12 2001-01-02 T2_168      Quercus ilex 0.70352348 [m^2/m^2]
#> 13 2001-01-03 T2_168      Quercus ilex 0.70352348 [m^2/m^2]
#> 14 2001-01-04 T2_168      Quercus ilex 0.70352206 [m^2/m^2]
#> 15 2001-01-05 T2_168      Quercus ilex 0.70344991 [m^2/m^2]
#> 16 2001-01-06 T2_168      Quercus ilex 0.70344991 [m^2/m^2]
#> 17 2001-01-07 T2_168      Quercus ilex 0.70344991 [m^2/m^2]
#> 18 2001-01-08 T2_168      Quercus ilex 0.70344991 [m^2/m^2]
#> 19 2001-01-09 T2_168      Quercus ilex 0.70344991 [m^2/m^2]
#> 20 2001-01-10 T2_168      Quercus ilex 0.70344991 [m^2/m^2]
#> 21 2001-01-01 S1_165 Quercus coccifera 0.03062604 [m^2/m^2]
#> 22 2001-01-02 S1_165 Quercus coccifera 0.03062498 [m^2/m^2]
#> 23 2001-01-03 S1_165 Quercus coccifera 0.03062498 [m^2/m^2]
#> 24 2001-01-04 S1_165 Quercus coccifera 0.03062498 [m^2/m^2]
#> 25 2001-01-05 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 26 2001-01-06 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 27 2001-01-07 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 28 2001-01-08 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 29 2001-01-09 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#> 30 2001-01-10 S1_165 Quercus coccifera 0.03062489 [m^2/m^2]
#>                 LAIlive         FPAR AbsorbedSWRFraction       Transpiration
#> 1  0.84874773 [m^2/m^2] 92.18285 [%]           35.076344 0.125013855 [L/m^2]
#> 2  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.231814863 [L/m^2]
#> 3  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.184272331 [L/m^2]
#> 4  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.080536010 [L/m^2]
#> 5  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.237258112 [L/m^2]
#> 6  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.170941003 [L/m^2]
#> 7  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.090146466 [L/m^2]
#> 8  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.117746145 [L/m^2]
#> 9  0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.280590291 [L/m^2]
#> 10 0.84874773 [m^2/m^2] 92.18285 [%]           35.082685 0.117278013 [L/m^2]
#> 11 0.70557382 [m^2/m^2] 72.36365 [%]           30.444383 0.072474080 [L/m^2]
#> 12 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.134110679 [L/m^2]
#> 13 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.106606397 [L/m^2]
#> 14 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.046597341 [L/m^2]
#> 15 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.137259734 [L/m^2]
#> 16 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.098897800 [L/m^2]
#> 17 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.052152899 [L/m^2]
#> 18 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.068120642 [L/m^2]
#> 19 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.162334069 [L/m^2]
#> 20 0.70557382 [m^2/m^2] 72.38566 [%]           30.386677 0.067853365 [L/m^2]
#> 21 0.03062604 [m^2/m^2] 44.32407 [%]            2.366131 0.006052657 [L/m^2]
#> 22 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.011228235 [L/m^2]
#> 23 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.008925484 [L/m^2]
#> 24 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.003901502 [L/m^2]
#> 25 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.011491885 [L/m^2]
#> 26 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.008280215 [L/m^2]
#> 27 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.004366382 [L/m^2]
#> 28 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.005703332 [L/m^2]
#> 29 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.013591496 [L/m^2]
#> 30 0.03062604 [m^2/m^2] 44.37408 [%]            2.367553 0.005681228 [L/m^2]
#>    GrossPhotosynthesis          PlantPsi      LeafPLC      StemPLC
#> 1   1.21341578 [L/m^2] -0.03300000 [MPa] 0.000000e+00 0.000000e+00
#> 2   2.10348703 [L/m^2] -0.03300000 [MPa] 0.000000e+00 0.000000e+00
#> 3   1.77396699 [L/m^2] -0.03302923 [MPa] 0.000000e+00 0.000000e+00
#> 4   0.81308802 [L/m^2] -0.03443797 [MPa] 0.000000e+00 0.000000e+00
#> 5   2.18356243 [L/m^2] -0.03300000 [MPa] 0.000000e+00 0.000000e+00
#> 6   1.79060063 [L/m^2] -0.03350265 [MPa] 0.000000e+00 0.000000e+00
#> 7   1.02742267 [L/m^2] -0.03291028 [MPa] 0.000000e+00 0.000000e+00
#> 8   1.21703794 [L/m^2] -0.03316782 [MPa] 0.000000e+00 0.000000e+00
#> 9   2.08857051 [L/m^2] -0.03350647 [MPa] 0.000000e+00 0.000000e+00
#> 10  1.16049105 [L/m^2] -0.03415235 [MPa] 0.000000e+00 0.000000e+00
#> 11  0.87313959 [L/m^2] -0.03300000 [MPa] 2.905929e-03 3.681226e-09
#> 12  1.59099654 [L/m^2] -0.03300000 [MPa] 2.905929e-03 3.681226e-09
#> 13  1.28193415 [L/m^2] -0.03301702 [MPa] 2.907933e-03 3.687988e-09
#> 14  0.56671908 [L/m^2] -0.03388137 [MPa] 3.010191e-03 4.043372e-09
#> 15  1.63354373 [L/m^2] -0.03300000 [MPa] 3.010191e-03 4.043372e-09
#> 16  1.21375667 [L/m^2] -0.03329260 [MPa] 3.010191e-03 4.043372e-09
#> 17  0.65412471 [L/m^2] -0.03283336 [MPa] 3.010191e-03 4.043372e-09
#> 18  0.83326878 [L/m^2] -0.03305700 [MPa] 3.010191e-03 4.043372e-09
#> 19  1.84698419 [L/m^2] -0.03333137 [MPa] 3.010191e-03 4.043372e-09
#> 20  0.82125418 [L/m^2] -0.03377234 [MPa] 3.010191e-03 4.043372e-09
#> 21  0.05005525 [L/m^2] -0.03300000 [MPa] 3.473351e-05 4.394058e-08
#> 22  0.08989595 [L/m^2] -0.03300000 [MPa] 3.473351e-05 4.394058e-08
#> 23  0.07354445 [L/m^2] -0.03302310 [MPa] 3.478835e-05 4.403586e-08
#> 24  0.03292091 [L/m^2] -0.03415838 [MPa] 3.754355e-05 4.889290e-08
#> 25  0.09263469 [L/m^2] -0.03300000 [MPa] 3.754355e-05 4.889290e-08
#> 26  0.07119441 [L/m^2] -0.03339730 [MPa] 3.754355e-05 4.889290e-08
#> 27  0.03924416 [L/m^2] -0.03279120 [MPa] 3.754355e-05 4.889290e-08
#> 28  0.04870346 [L/m^2] -0.03306831 [MPa] 3.754355e-05 4.889290e-08
#> 29  0.09940060 [L/m^2] -0.03341750 [MPa] 3.754355e-05 4.889290e-08
#> 30  0.04745642 [L/m^2] -0.03395976 [MPa] 3.754355e-05 4.889290e-08
#>        PlantWaterBalance       LeafRWC       StemRWC          LFMC PlantStress
#> 1   0.000000e+00 [L/m^2] 0.9979679 [%] 0.9998350 [%] 125.86544 [%] 0.004613739
#> 2   0.000000e+00 [L/m^2] 0.9979679 [%] 0.9998350 [%] 125.86544 [%] 0.004613739
#> 3  -1.319457e-06 [L/m^2] 0.9979661 [%] 0.9998348 [%] 125.86529 [%] 0.004619875
#> 4  -6.358830e-05 [L/m^2] 0.9978794 [%] 0.9998278 [%] 125.85824 [%] 0.004918765
#> 5   6.490775e-05 [L/m^2] 0.9979679 [%] 0.9998350 [%] 125.86544 [%] 0.004613739
#> 6  -2.268898e-05 [L/m^2] 0.9979370 [%] 0.9998325 [%] 125.86292 [%] 0.004719625
#> 7   2.673892e-05 [L/m^2] 0.9979734 [%] 0.9998354 [%] 125.86589 [%] 0.004594923
#> 8  -1.162512e-05 [L/m^2] 0.9979576 [%] 0.9998341 [%] 125.86460 [%] 0.004649003
#> 9  -1.528648e-05 [L/m^2] 0.9979367 [%] 0.9998324 [%] 125.86290 [%] 0.004720434
#> 10 -2.915406e-05 [L/m^2] 0.9978970 [%] 0.9998292 [%] 125.85967 [%] 0.004857666
#> 11  0.000000e+00 [L/m^2] 0.9986797 [%] 0.9997500 [%]  93.05186 [%] 0.006282261
#> 12  0.000000e+00 [L/m^2] 0.9986797 [%] 0.9997500 [%]  92.91408 [%] 0.006282261
#> 13 -1.793122e-07 [L/m^2] 0.9986790 [%] 0.9997498 [%]  92.91402 [%] 0.006285972
#> 14 -9.108700e-06 [L/m^2] 0.9986445 [%] 0.9997433 [%]  92.91128 [%] 0.006474811
#> 15  9.287453e-06 [L/m^2] 0.9986797 [%] 0.9997500 [%]  92.90913 [%] 0.006282261
#> 16 -3.083267e-06 [L/m^2] 0.9986680 [%] 0.9997477 [%]  92.90824 [%] 0.006346104
#> 17  4.839227e-06 [L/m^2] 0.9986864 [%] 0.9997512 [%]  92.90964 [%] 0.006245937
#> 18 -2.356565e-06 [L/m^2] 0.9986774 [%] 0.9997495 [%]  92.90896 [%] 0.006294691
#> 19 -2.891243e-06 [L/m^2] 0.9986665 [%] 0.9997475 [%]  92.90812 [%] 0.006354571
#> 20 -4.646707e-06 [L/m^2] 0.9986488 [%] 0.9997441 [%]  92.90677 [%] 0.006450953
#> 21  0.000000e+00 [L/m^2] 0.9987207 [%] 0.9983684 [%]  96.38420 [%] 0.003088161
#> 22  0.000000e+00 [L/m^2] 0.9987207 [%] 0.9983684 [%]  96.38552 [%] 0.003088161
#> 23 -2.237386e-08 [L/m^2] 0.9987198 [%] 0.9983673 [%]  96.38542 [%] 0.003090967
#> 24 -1.099604e-06 [L/m^2] 0.9986758 [%] 0.9983112 [%]  96.38026 [%] 0.003229591
#> 25  1.121976e-06 [L/m^2] 0.9987207 [%] 0.9983684 [%]  96.38563 [%] 0.003088161
#> 26 -3.848146e-07 [L/m^2] 0.9987053 [%] 0.9983488 [%]  96.38382 [%] 0.003136505
#> 27  5.870562e-07 [L/m^2] 0.9987288 [%] 0.9983787 [%]  96.38658 [%] 0.003062823
#> 28 -2.684088e-07 [L/m^2] 0.9987180 [%] 0.9983650 [%]  96.38532 [%] 0.003096461
#> 29 -3.382177e-07 [L/m^2] 0.9987045 [%] 0.9983478 [%]  96.38373 [%] 0.003138968
#> 30 -5.252128e-07 [L/m^2] 0.9986835 [%] 0.9983210 [%]  96.38126 [%] 0.003205238

#Select the output tables/variables to be extracted
extract(S1, level ="cohort", output="Plants", vars = c("PlantStress", "StemPLC"))
#>          date cohort           species PlantStress      StemPLC
#> 1  2001-01-01 T1_148  Pinus halepensis 0.004613739 0.000000e+00
#> 2  2001-01-02 T1_148  Pinus halepensis 0.004613739 0.000000e+00
#> 3  2001-01-03 T1_148  Pinus halepensis 0.004619875 0.000000e+00
#> 4  2001-01-04 T1_148  Pinus halepensis 0.004918765 0.000000e+00
#> 5  2001-01-05 T1_148  Pinus halepensis 0.004613739 0.000000e+00
#> 6  2001-01-06 T1_148  Pinus halepensis 0.004719625 0.000000e+00
#> 7  2001-01-07 T1_148  Pinus halepensis 0.004594923 0.000000e+00
#> 8  2001-01-08 T1_148  Pinus halepensis 0.004649003 0.000000e+00
#> 9  2001-01-09 T1_148  Pinus halepensis 0.004720434 0.000000e+00
#> 10 2001-01-10 T1_148  Pinus halepensis 0.004857666 0.000000e+00
#> 11 2001-01-01 T2_168      Quercus ilex 0.006282261 3.681226e-09
#> 12 2001-01-02 T2_168      Quercus ilex 0.006282261 3.681226e-09
#> 13 2001-01-03 T2_168      Quercus ilex 0.006285972 3.687988e-09
#> 14 2001-01-04 T2_168      Quercus ilex 0.006474811 4.043372e-09
#> 15 2001-01-05 T2_168      Quercus ilex 0.006282261 4.043372e-09
#> 16 2001-01-06 T2_168      Quercus ilex 0.006346104 4.043372e-09
#> 17 2001-01-07 T2_168      Quercus ilex 0.006245937 4.043372e-09
#> 18 2001-01-08 T2_168      Quercus ilex 0.006294691 4.043372e-09
#> 19 2001-01-09 T2_168      Quercus ilex 0.006354571 4.043372e-09
#> 20 2001-01-10 T2_168      Quercus ilex 0.006450953 4.043372e-09
#> 21 2001-01-01 S1_165 Quercus coccifera 0.003088161 4.394058e-08
#> 22 2001-01-02 S1_165 Quercus coccifera 0.003088161 4.394058e-08
#> 23 2001-01-03 S1_165 Quercus coccifera 0.003090967 4.403586e-08
#> 24 2001-01-04 S1_165 Quercus coccifera 0.003229591 4.889290e-08
#> 25 2001-01-05 S1_165 Quercus coccifera 0.003088161 4.889290e-08
#> 26 2001-01-06 S1_165 Quercus coccifera 0.003136505 4.889290e-08
#> 27 2001-01-07 S1_165 Quercus coccifera 0.003062823 4.889290e-08
#> 28 2001-01-08 S1_165 Quercus coccifera 0.003096461 4.889290e-08
#> 29 2001-01-09 S1_165 Quercus coccifera 0.003138968 4.889290e-08
#> 30 2001-01-10 S1_165 Quercus coccifera 0.003205238 4.889290e-08