Extracts model outputs

Function extract() extracts daily or subdaily output and returns it as a tidy data frame.

extract(
  x,
  level = "forest",
  output = NULL,
  vars = NULL,
  dates = NULL,
  subdaily = FALSE
)

Arguments

x

An object returned by simulation functions spwb, pwb or growth.

level

Level of simulation output, either "forest" (stand-level results), "cohort" (cohort-level results), "sunlitleaf" or "shadeleaf" (leaf-level results)

output

Section of the model output to be explored. See details.

vars

Variables to be extracted (by default, all of them).

dates

A date vector indicating the subset of simulated days for which output is desired.

subdaily

A flag to indicate that subdaily values are desired (see details).

Value

Function extract() returns a data frame:

If level = "forest", columns are "date" and variable names.

If level = "cohort", level = "sunlitleaf" or level = "shadeleaf", columns are "date", "cohorts", "species" and variable names.

If subdaily = TRUE, columns are "datetime", "cohorts", "species" and variable names.

Details

When subdaily = FALSE, parameter output is used to restrict the section in x where variables are located. For example output = "Plants" will correspond to variables "LAI", "LAIlive", "Transpiration", "StemPLC",... as returned by a call names(x$Plants).

Option subdaily = TRUE only works when simulations have been carried using control option 'subdailyResults = TRUE' (see defaultControl). When using subdaily = TRUE, parameter output is not taken into account, and options for parameter vars are the following:

Variables for level = "forest": Not allowed. An error is raised.

Variables for level = "cohort": "E","Ag","An","dEdP","RootPsi","StemPsi","LeafPsi","StemPLC","StemRWC","LeafRWC","StemSympRWC","LeafSympRWC","PWB".

Variables for level = "shadeleaf" and level="sunlitleaf": "Abs_SWR","Abs_PAR","Net_LWR","E","Ag","An","Ci","Gsw","VPD","Temp","Psi","iWUE".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforest,examplesoil, SpParamsMED, control)

#Call simulation function (ten days)
S1<-spwb(x, examplemeteo[1:10, ], latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 5.25158
#> Initial soil water content (mm): 184.827
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:
#> 
#> Final plant water content (mm): 5.2526
#> Final soil water content (mm): 211.868
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): 0.00102106
#> Plant water balance result (mm): 0.00102106
#> Change in soil water content (mm): 27.0411
#> Soil water balance result (mm): 27.0411
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 39
#>   Rain (mm) 34 Snow (mm) 5
#>   Interception (mm) 9 Net rainfall (mm) 25
#>   Infiltration (mm) 25  Saturation excess (mm) 0 Runoff (mm) 0 Deep drainage (mm) 1
#>   Soil evaporation (mm) 1  Herbaceous transpiration (mm) 0 Woody plant transpiration (mm) 2
#>   Plant extraction from soil (mm) 2  Plant water balance (mm) 0 Hydraulic redistribution (mm) 0

#Extracts daily forest-level output as a data frame
extract(S1, level = "forest")
#>          date       PET Precipitation      Rain   Snow    NetRain Snowmelt
#> 1  2001-01-01 0.8828475      4.869109  4.869109 0.0000  3.4241795   0.0000
#> 2  2001-01-02 1.6375337      2.498292  2.498292 0.0000  1.0717469   0.0000
#> 3  2001-01-03 1.3017026      0.000000  0.000000 0.0000  0.0000000   0.0000
#> 4  2001-01-04 0.5690790      5.796973  5.796973 0.0000  4.3625616   0.0000
#> 5  2001-01-05 1.6760567      1.884401  1.884401 0.0000  0.7539027   0.0000
#> 6  2001-01-06 1.2077028     13.359801 13.359801 0.0000 11.7240275   0.0000
#> 7  2001-01-07 0.6368075      5.377400  0.000000 5.3774  0.0000000   0.0000
#> 8  2001-01-08 0.8318207      0.000000  0.000000 0.0000  0.0000000   0.0000
#> 9  2001-01-09 1.9823797      0.000000  0.000000 0.0000  0.0000000   0.0000
#> 10 2001-01-10 0.8286873      5.117768  5.117768 0.0000  3.6727363   5.3774
#>    Infiltration SaturationExcess       Runoff  DeepDrainage Evapotranspiration
#> 1     3.4241795     0.000000e+00 0.000000e+00  6.248286e-02          2.0551277
#> 2     1.0717469     0.000000e+00 0.000000e+00  6.279543e-02          1.7946355
#> 3     0.0000000     2.125854e-05 2.125854e-05  6.024147e-02          0.2954225
#> 4     4.3625616     0.000000e+00 0.000000e+00  6.576128e-02          1.6048387
#> 5     0.7539027     4.794808e-06 4.794808e-06  7.875170e-02          1.4653077
#> 6    11.7240275     0.000000e+00 0.000000e+00  9.072904e-02          1.8836962
#> 7     0.0000000     0.000000e+00 0.000000e+00 -1.080821e-09          0.1175354
#> 8     0.0000000     0.000000e+00 0.000000e+00 -2.418830e-09          0.1535263
#> 9     0.0000000     0.000000e+00 0.000000e+00 -3.304118e-08          0.3658728
#> 10    3.6727363     0.000000e+00 0.000000e+00  9.783856e-02          1.6080882
#>    Interception SoilEvaporation HerbTranspiration PlantExtraction Transpiration
#> 1      1.444929      0.44789475       0.011023432      0.15128020    0.15128020
#> 2      1.426545      0.06679272       0.020448162      0.28084991    0.28084991
#> 3      0.000000      0.05587694       0.016254795      0.22329079    0.22329079
#> 4      1.434411      0.06571597       0.007106194      0.09760519    0.09760519
#> 5      1.130499      0.02608214       0.020931049      0.28779575    0.28779575
#> 6      1.635773      0.02545435       0.015082163      0.20738652    0.20738652
#> 7      0.000000      0.00000000       0.007953597      0.10958185    0.10958185
#> 8      0.000000      0.00000000       0.010389263      0.14313708    0.14313708
#> 9      0.000000      0.00000000       0.024759469      0.34111332    0.34111332
#> 10     1.445031      0.01012068       0.010350084      0.14258599    0.14258599
#>    HydraulicRedistribution      LAI   LAIherb  LAIlive LAIexpanded LAIdead
#> 1              0.000000000 1.758585 0.1736369 1.584948    1.584948       0
#> 2              0.000000000 1.758585 0.1736369 1.584948    1.584948       0
#> 3              0.000000000 1.758585 0.1736369 1.584948    1.584948       0
#> 4              0.000000000 1.758585 0.1736369 1.584948    1.584948       0
#> 5              0.002462785 1.758585 0.1736369 1.584948    1.584948       0
#> 6              0.001748480 1.758585 0.1736369 1.584948    1.584948       0
#> 7              0.003282991 1.758585 0.1736369 1.584948    1.584948       0
#> 8              0.004248779 1.758585 0.1736369 1.584948    1.584948       0
#> 9              0.010001021 1.758585 0.1736369 1.584948    1.584948       0
#> 10             0.004053762 1.758585 0.1736369 1.584948    1.584948       0
#>          Cm LgroundPAR LgroundSWR      W.1      W.2     ML.1     ML.2    MLTot
#> 1  1.390485   40.00754   50.73297 1.000000 1.000000 68.17381 116.6530 184.8268
#> 2  1.390485   40.00754   50.73297 1.026961 1.007830 70.01188 117.5664 187.5783
#> 3  1.390485   40.00754   50.73297 1.032404 1.010143 70.38293 117.8362 188.2191
#> 4  1.390485   40.00754   50.73297 1.027419 1.010008 70.04308 117.8204 187.8635
#> 5  1.390485   40.00754   50.73297 1.070108 1.020433 72.95332 119.0365 191.9898
#> 6  1.390485   40.00754   50.73297 1.071846 1.022334 73.07184 119.2583 192.3302
#> 7  1.390485   40.00754   50.73297 1.189844 1.050975 81.11622 122.5993 203.7155
#> 8  1.390485   40.00754   50.73297 1.188227 1.050912 81.00597 122.5920 203.5980
#> 9  1.390485   40.00754   50.73297 1.186117 1.050829 80.86211 122.5824 203.4445
#> 10 1.390485   40.00754   50.73297 1.181094 1.050628 80.51970 122.5589 203.0786
#>    SaturatedDepth    SWE PlantExt.1  PlantExt.2 HydraulicInput.1
#> 1        965.9416 0.0000 0.11328728 0.037992928                0
#> 2        957.7798 0.0000 0.24532184 0.035528070                0
#> 3        961.7718 0.0000 0.19586573 0.027425058                0
#> 4        911.3125 0.0000 0.08504188 0.012563310                0
#> 5        906.6626 0.0000 0.26096146 0.026834289                0
#> 6        767.4948 0.0000 0.18797466 0.019411861                0
#> 7        768.8222 5.3774 0.10666326 0.002918585                0
#> 8        770.5567 5.3774 0.13918999 0.003947098                0
#> 9        774.6919 5.3774 0.33128384 0.009829478                0
#> 10       671.6071 0.0000 0.13804877 0.004537213                0
#>    HydraulicInput.2        psi.1       psi.2
#> 1       0.000000000 -0.029036340 -0.03179780
#> 2       0.000000000 -0.028293369 -0.03145097
#> 3       0.000000000 -0.028973165 -0.03147122
#> 4       0.000000000 -0.023605026 -0.02995184
#> 5       0.002462785 -0.023406529 -0.02968240
#> 6       0.001748480 -0.012645731 -0.02588878
#> 7       0.003282991 -0.012764438 -0.02589655
#> 8       0.004248779 -0.012920294 -0.02590685
#> 9       0.010001021 -0.013295644 -0.02593187
#> 10      0.004053762 -0.006391431 -0.02437821

#Extracts daily cohort-level output as a data frame
extract(S1, level = "cohort")
#>          date cohort           species        LAI    LAIlive     FPAR
#> 1  2001-01-01 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 2  2001-01-02 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 3  2001-01-03 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 4  2001-01-04 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 5  2001-01-05 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 6  2001-01-06 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 7  2001-01-07 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 8  2001-01-08 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 9  2001-01-09 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 10 2001-01-10 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 11 2001-01-01 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 12 2001-01-02 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 13 2001-01-03 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 14 2001-01-04 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 15 2001-01-05 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 16 2001-01-06 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 17 2001-01-07 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 18 2001-01-08 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 19 2001-01-09 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 20 2001-01-10 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 21 2001-01-01 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 22 2001-01-02 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 23 2001-01-03 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 24 2001-01-04 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 25 2001-01-05 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 26 2001-01-06 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 27 2001-01-07 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 28 2001-01-08 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 29 2001-01-09 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 30 2001-01-10 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#>    AbsorbedSWRFraction Transpiration GrossPhotosynthesis    PlantPsi
#> 1            35.076344   0.092472560          0.89344066 -0.03300000
#> 2            35.076344   0.171678099          1.56122963 -0.02937932
#> 3            35.076344   0.136492862          1.30902232 -0.02868678
#> 4            35.076344   0.059663328          0.59732714 -0.02928290
#> 5            35.076344   0.175922428          1.62025722 -0.02441845
#> 6            35.076344   0.126768943          1.31230799 -0.02421076
#> 7            35.076344   0.066974213          0.74878280 -0.01453682
#> 8            35.076344   0.087482584          0.89470957 -0.01463577
#> 9            35.076344   0.208482060          1.59049819 -0.01476587
#> 10           35.076344   0.087146140          0.85676583 -0.01508000
#> 11           30.444383   0.052754988          0.62572744 -0.03300000
#> 12           30.444383   0.097940885          1.14626048 -0.03034579
#> 13           30.444383   0.077869617          0.92207345 -0.02979100
#> 14           30.444383   0.034038806          0.40703728 -0.03015758
#> 15           30.444383   0.100372391          1.17794967 -0.02662126
#> 16           30.444383   0.072330237          0.87175289 -0.02638904
#> 17           30.444383   0.038233872          0.46958944 -0.01898594
#> 18           30.444383   0.049941378          0.60018700 -0.01905067
#> 19           30.444383   0.119016001          1.34349553 -0.01913575
#> 20           30.444383   0.049748452          0.59240741 -0.01934091
#> 21            2.366131   0.006052657          0.05005525 -0.03300000
#> 22            2.366131   0.011230924          0.08988054 -0.02983673
#> 23            2.366131   0.008928307          0.07353749 -0.02921004
#> 24            2.366131   0.003903058          0.03292051 -0.02969658
#> 25            2.366131   0.011500927          0.09266950 -0.02547398
#> 26            2.366131   0.008287336          0.07122637 -0.02525448
#> 27            2.366131   0.004373764          0.03929435 -0.01675208
#> 28            2.366131   0.005713121          0.04876701 -0.01683249
#> 29            2.366131   0.013615261          0.09953350 -0.01693824
#> 30            2.366131   0.005691395          0.04752182 -0.01719371
#>         LeafPLC StemPLC PlantWaterBalance   LeafRWC   StemRWC      LFMC
#> 1  0.0030695571       0      0.000000e+00 0.9979679 0.9998350 125.89003
#> 2  0.0024906271       0      1.712759e-04 0.9981908 0.9998531 125.90545
#> 3  0.0023833945       0      3.276221e-05 0.9982334 0.9998565 125.90840
#> 4  0.0024756292       0     -2.820108e-05 0.9981967 0.9998536 125.90586
#> 5  0.0018275916       0      2.301365e-04 0.9984962 0.9998779 125.92658
#> 6  0.0017190506       0      9.826154e-06 0.9985089 0.9998789 125.92747
#> 7  0.0010666732       0      4.577547e-04 0.9991046 0.9999273 125.96868
#> 8  0.0004857033       0     -4.682593e-06 0.9990985 0.9999268 125.96826
#> 9  0.0005005142       0     -6.156794e-06 0.9990905 0.9999261 125.96770
#> 10 0.0005365186       0     -1.486546e-05 0.9990712 0.9999246 125.96637
#> 11 0.0000000000       0      0.000000e+00 0.9986794 0.9997489  93.07448
#> 12 0.0000000000       0      2.984111e-05 0.9987857 0.9997692  93.08081
#> 13 0.0000000000       0      6.234841e-06 0.9988079 0.9997735  93.08213
#> 14 0.0000000000       0     -4.119536e-06 0.9987932 0.9997707  93.08126
#> 15 0.0000000000       0      3.972457e-05 0.9989347 0.9997977  93.08968
#> 16 0.0000000000       0      2.607405e-06 0.9989440 0.9997995  93.09024
#> 17 0.0000000000       0      8.305276e-05 0.9992403 0.9998559  93.10786
#> 18 0.0000000000       0     -7.257402e-07 0.9992377 0.9998554  93.10771
#> 19 0.0000000000       0     -9.537736e-07 0.9992343 0.9998548  93.10751
#> 20 0.0000000000       0     -2.299916e-06 0.9992261 0.9998532  93.10702
#> 21 0.0000000000       0      3.469447e-18 0.9987207 0.9983684  96.39176
#> 22 0.0000000000       0      3.970602e-06 0.9988433 0.9985247  96.40543
#> 23 0.0000000000       0      7.866760e-07 0.9988676 0.9985557  96.40814
#> 24 0.0000000000       0     -6.107490e-07 0.9988487 0.9985316  96.40604
#> 25 0.0000000000       0      5.300903e-06 0.9990124 0.9987403  96.42428
#> 26 0.0000000000       0      2.755711e-07 0.9990209 0.9987511  96.42523
#> 27 0.0000000000       0      1.067578e-05 0.9993505 0.9991714  96.46198
#> 28 0.0000000000       0     -1.009722e-07 0.9993474 0.9991675  96.46163
#> 29 0.0000000000       0     -1.328052e-07 0.9993433 0.9991622  96.46118
#> 30 0.0000000000       0     -3.208056e-07 0.9993334 0.9991496  96.46007
#>    PlantStress
#> 1  0.005808769
#> 2  0.004898439
#> 3  0.004729977
#> 4  0.004874875
#> 5  0.003734107
#> 6  0.003687581
#> 7  0.001743281
#> 8  0.001760744
#> 9  0.001783788
#> 10 0.001839821
#> 11 0.010612890
#> 12 0.009710418
#> 13 0.009522312
#> 14 0.009646581
#> 15 0.008451425
#> 16 0.008373240
#> 17 0.005903427
#> 18 0.005924810
#> 19 0.005952917
#> 20 0.006020724
#> 21 0.003088161
#> 22 0.002709510
#> 23 0.002635858
#> 24 0.002692999
#> 25 0.002206732
#> 26 0.002182072
#> 27 0.001280308
#> 28 0.001288297
#> 29 0.001298823
#> 30 0.001324329

#Select the output tables/variables to be extracted
extract(S1, level ="cohort", output="Plants", vars = c("PlantStress", "StemPLC"))
#>          date cohort           species PlantStress StemPLC
#> 1  2001-01-01 T1_148  Pinus halepensis 0.005808769       0
#> 2  2001-01-02 T1_148  Pinus halepensis 0.004898439       0
#> 3  2001-01-03 T1_148  Pinus halepensis 0.004729977       0
#> 4  2001-01-04 T1_148  Pinus halepensis 0.004874875       0
#> 5  2001-01-05 T1_148  Pinus halepensis 0.003734107       0
#> 6  2001-01-06 T1_148  Pinus halepensis 0.003687581       0
#> 7  2001-01-07 T1_148  Pinus halepensis 0.001743281       0
#> 8  2001-01-08 T1_148  Pinus halepensis 0.001760744       0
#> 9  2001-01-09 T1_148  Pinus halepensis 0.001783788       0
#> 10 2001-01-10 T1_148  Pinus halepensis 0.001839821       0
#> 11 2001-01-01 T2_168      Quercus ilex 0.010612890       0
#> 12 2001-01-02 T2_168      Quercus ilex 0.009710418       0
#> 13 2001-01-03 T2_168      Quercus ilex 0.009522312       0
#> 14 2001-01-04 T2_168      Quercus ilex 0.009646581       0
#> 15 2001-01-05 T2_168      Quercus ilex 0.008451425       0
#> 16 2001-01-06 T2_168      Quercus ilex 0.008373240       0
#> 17 2001-01-07 T2_168      Quercus ilex 0.005903427       0
#> 18 2001-01-08 T2_168      Quercus ilex 0.005924810       0
#> 19 2001-01-09 T2_168      Quercus ilex 0.005952917       0
#> 20 2001-01-10 T2_168      Quercus ilex 0.006020724       0
#> 21 2001-01-01 S1_165 Quercus coccifera 0.003088161       0
#> 22 2001-01-02 S1_165 Quercus coccifera 0.002709510       0
#> 23 2001-01-03 S1_165 Quercus coccifera 0.002635858       0
#> 24 2001-01-04 S1_165 Quercus coccifera 0.002692999       0
#> 25 2001-01-05 S1_165 Quercus coccifera 0.002206732       0
#> 26 2001-01-06 S1_165 Quercus coccifera 0.002182072       0
#> 27 2001-01-07 S1_165 Quercus coccifera 0.001280308       0
#> 28 2001-01-08 S1_165 Quercus coccifera 0.001288297       0
#> 29 2001-01-09 S1_165 Quercus coccifera 0.001298823       0
#> 30 2001-01-10 S1_165 Quercus coccifera 0.001324329       0