Extracts model outputs

Function extract() extracts daily or subdaily output and returns it as a tidy data frame.

extract(
  x,
  level = "forest",
  output = NULL,
  vars = NULL,
  dates = NULL,
  subdaily = FALSE
)

Arguments

x

An object returned by simulation functions spwb, pwb or growth.

level

Level of simulation output, either "forest" (stand-level results), "cohort" (cohort-level results), "sunlitleaf" or "shadeleaf" (leaf-level results)

output

Section of the model output to be explored. See details.

vars

Variables to be extracted (by default, all of them).

dates

A date vector indicating the subset of simulated days for which output is desired.

subdaily

A flag to indicate that subdaily values are desired (see details).

Value

Function extract() returns a data frame:

If level = "forest", columns are "date" and variable names.

If level = "cohort", level = "sunlitleaf" or level = "shadeleaf", columns are "date", "cohorts", "species" and variable names.

If subdaily = TRUE, columns are "datetime", "cohorts", "species" and variable names.

Details

When subdaily = FALSE, parameter output is used to restrict the section in x where variables are located. For example output = "Plants" will correspond to variables "LAI", "LAIlive", "Transpiration", "StemPLC",... as returned by a call names(x$Plants).

Option subdaily = TRUE only works when simulations have been carried using control option 'subdailyResults = TRUE' (see defaultControl). When using subdaily = TRUE, parameter output is not taken into account, and options for parameter vars are the following:

Variables for level = "forest": Not allowed. An error is raised.

Variables for level = "cohort": "E","Ag","An","dEdP","RootPsi","StemPsi","LeafPsi","StemPLC","StemRWC","LeafRWC","StemSympRWC","LeafSympRWC","PWB".

Variables for level = "shadeleaf" and level="sunlitleaf": "Abs_SWR","Abs_PAR","Net_LWR","E","Ag","An","Ci","Gsw","VPD","Temp","Psi","iWUE".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Call simulation function (ten days)
S1<-spwb(x, examplemeteo[1:10, ], latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 5.25158
#> Initial soil water content (mm): 184.827
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:
#> 
#> Final plant water content (mm): 5.25251
#> Final soil water content (mm): 210.986
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): 0.000935515
#> Plant water balance result (mm): 0.000935515
#> Change in soil water content (mm): 26.159
#> Soil water balance result (mm): 26.1486
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 39
#>   Rain (mm) 34 Snow (mm) 5
#>   Interception (mm) 9 Net rainfall (mm) 25
#>   Infiltration (mm) 25 Runoff (mm) 0 Deep drainage (mm) 1
#>   Soil evaporation (mm) 1  Herbaceous transpiration (mm) 0 Woody plant transpiration (mm) 2
#>   Plant extraction from soil (mm) 2  Plant water balance (mm) 0 Hydraulic redistribution (mm) 0

#Extracts daily forest-level output as a data frame
extract(S1, level = "forest")
#>          date       PET Precipitation      Rain   Snow    NetRain Snowmelt
#> 1  2001-01-01 0.8828475      4.869109  4.869109 0.0000  3.4241795   0.0000
#> 2  2001-01-02 1.6375337      2.498292  2.498292 0.0000  1.0717469   0.0000
#> 3  2001-01-03 1.3017026      0.000000  0.000000 0.0000  0.0000000   0.0000
#> 4  2001-01-04 0.5690790      5.796973  5.796973 0.0000  4.3625616   0.0000
#> 5  2001-01-05 1.6760567      1.884401  1.884401 0.0000  0.7539027   0.0000
#> 6  2001-01-06 1.2077028     13.359801 13.359801 0.0000 11.7240275   0.0000
#> 7  2001-01-07 0.6368075      5.377400  0.000000 5.3774  0.0000000   0.0000
#> 8  2001-01-08 0.8318207      0.000000  0.000000 0.0000  0.0000000   0.0000
#> 9  2001-01-09 1.9823797      0.000000  0.000000 0.0000  0.0000000   0.0000
#> 10 2001-01-10 0.8286873      5.117768  5.117768 0.0000  3.6727363   5.3774
#>    Infiltration Runoff DeepDrainage Evapotranspiration Interception
#> 1     3.4241795      0   0.09717336          2.0551277     1.444929
#> 2     1.0717469      0   0.10303069          1.7962051     1.426545
#> 3     0.0000000      0   0.10444170          0.2979600     0.000000
#> 4     4.3625616      0   0.11089501          1.6106766     1.434411
#> 5     0.7539027      0   0.11908009          1.4665876     1.130499
#> 6    11.7240275      0   0.14215116          1.8852671     1.635773
#> 7     0.0000000      0   0.16989592          0.1175294     0.000000
#> 8     0.0000000      0   0.17520834          0.1535093     0.000000
#> 9     0.0000000      0   0.17972934          0.3658095     0.000000
#> 10    3.6727363      0   0.19502315          1.6098532     1.445031
#>    SoilEvaporation HerbTranspiration PlantExtraction Transpiration
#> 1       0.44789475       0.011023432      0.15128020    0.15128020
#> 2       0.06836698       0.020448130      0.28084523    0.28084523
#> 3       0.05842254       0.016254740      0.22328271    0.22328271
#> 4       0.07155976       0.007106153      0.09759933    0.09759933
#> 5       0.02738124       0.020930931      0.28777669    0.28777669
#> 6       0.02704258       0.015082056      0.20736932    0.20736932
#> 7       0.00000000       0.007953570      0.10957584    0.10957584
#> 8       0.00000000       0.010389176      0.14312011    0.14312011
#> 9       0.00000000       0.024759134      0.34105037    0.34105037
#> 10      0.01192238       0.010349885      0.14254946    0.14254946
#>    HydraulicRedistribution      LAI   LAIherb  LAIlive LAIexpanded LAIdead
#> 1              0.000000000 1.758585 0.1736369 1.584948    1.584948       0
#> 2              0.000000000 1.758585 0.1736369 1.584948    1.584948       0
#> 3              0.000000000 1.758585 0.1736369 1.584948    1.584948       0
#> 4              0.000000000 1.758585 0.1736369 1.584948    1.584948       0
#> 5              0.002058693 1.758585 0.1736369 1.584948    1.584948       0
#> 6              0.001374571 1.758585 0.1736369 1.584948    1.584948       0
#> 7              0.003137253 1.758585 0.1736369 1.584948    1.584948       0
#> 8              0.003725416 1.758585 0.1736369 1.584948    1.584948       0
#> 9              0.007962851 1.758585 0.1736369 1.584948    1.584948       0
#> 10             0.002859254 1.758585 0.1736369 1.584948    1.584948       0
#>          Cm LgroundPAR LgroundSWR      W.1      W.2     ML.1     ML.2    MLTot
#> 1  1.390485   40.00754   50.73297 1.000000 1.000000 68.17381 116.6530 184.8268
#> 2  1.390485   40.00754   50.73297 1.026318 1.007910 69.96800 117.5756 187.5436
#> 3  1.390485   40.00754   50.73297 1.030956 1.010335 70.28419 117.8585 188.1427
#> 4  1.390485   40.00754   50.73297 1.025098 1.010308 69.88483 117.8554 187.7403
#> 5  1.390485   40.00754   50.73297 1.066763 1.020897 72.72529 119.0907 191.8160
#> 6  1.390485   40.00754   50.73297 1.067604 1.022967 72.78263 119.3321 192.1147
#> 7  1.390485   40.00754   50.73297 1.184027 1.052105 80.71965 122.7311 203.4508
#> 8  1.390485   40.00754   50.73297 1.174583 1.055150 80.07577 123.0864 203.1621
#> 9  1.390485   40.00754   50.73297 1.165260 1.057772 79.44020 123.3922 202.8324
#> 10 1.390485   40.00754   50.73297 1.153703 1.059841 78.65233 123.6336 202.2859
#>         WTD    SWE PlantExt.1  PlantExt.2 HydraulicInput.1 HydraulicInput.2
#> 1  966.2871 0.0000  0.1132873 0.037992928                0      0.000000000
#> 2  958.5343 0.0000  0.2450673 0.035777895                0      0.000000000
#> 3  962.9916 0.0000  0.1954148 0.027867943                0      0.000000000
#> 4  913.0145 0.0000  0.0847182 0.012881123                0      0.000000000
#> 5  908.7460 0.0000  0.2597996 0.027977118                0      0.002058693
#> 6  769.8894 0.0000  0.1869025 0.020466829                0      0.001374571
#> 7  771.6510 5.3774  0.1061631 0.003412745                0      0.003137253
#> 8  774.0644 5.3774  0.1374214 0.005698725                0      0.003725416
#> 9  779.1309 5.3774  0.3244930 0.016557415                0      0.007962851
#> 10 676.7841 0.0000  0.1341420 0.008407459                0      0.002859254
#>           psi.1       psi.2
#> 1  -0.029125406 -0.03178586
#> 2  -0.028489331 -0.03142245
#> 3  -0.029294933 -0.03142638
#> 4  -0.023991035 -0.02988578
#> 5  -0.023893459 -0.02959333
#> 6  -0.013075699 -0.02574864
#> 7  -0.013791744 -0.02537441
#> 8  -0.014521418 -0.02505609
#> 9  -0.015459423 -0.02480737
#> 10 -0.008109836 -0.02325312

#Extracts daily cohort-level output as a data frame
extract(S1, level = "cohort")
#>          date cohort           species        LAI    LAIlive     FPAR
#> 1  2001-01-01 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 2  2001-01-02 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 3  2001-01-03 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 4  2001-01-04 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 5  2001-01-05 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 6  2001-01-06 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 7  2001-01-07 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 8  2001-01-08 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 9  2001-01-09 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 10 2001-01-10 T1_148  Pinus halepensis 0.84874773 0.84874773 92.18285
#> 11 2001-01-01 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 12 2001-01-02 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 13 2001-01-03 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 14 2001-01-04 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 15 2001-01-05 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 16 2001-01-06 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 17 2001-01-07 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 18 2001-01-08 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 19 2001-01-09 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 20 2001-01-10 T2_168      Quercus ilex 0.70557382 0.70557382 72.36365
#> 21 2001-01-01 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 22 2001-01-02 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 23 2001-01-03 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 24 2001-01-04 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 25 2001-01-05 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 26 2001-01-06 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 27 2001-01-07 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 28 2001-01-08 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 29 2001-01-09 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#> 30 2001-01-10 S1_165 Quercus coccifera 0.03062604 0.03062604 44.32407
#>    AbsorbedSWRFraction Transpiration GrossPhotosynthesis    PlantPsi
#> 1            35.076344   0.092472560          0.89344066 -0.03300000
#> 2            35.076344   0.171674878          1.56120034 -0.02945560
#> 3            35.076344   0.136487301          1.30896899 -0.02885417
#> 4            35.076344   0.059659284          0.59728665 -0.02955853
#> 5            35.076344   0.175909504          1.62013819 -0.02474351
#> 6            35.076344   0.126757251          1.31218695 -0.02462016
#> 7            35.076344   0.066970262          0.74873863 -0.01486922
#> 8            35.076344   0.087470514          0.89458613 -0.01540554
#> 9            35.076344   0.208436667          1.59015189 -0.01596779
#> 10           35.076344   0.087119803          0.85650689 -0.01672186
#> 11           30.444383   0.052754988          0.62572744 -0.03300000
#> 12           30.444383   0.097939506          1.14624433 -0.03038690
#> 13           30.444383   0.077867237          0.92204527 -0.02988033
#> 14           30.444383   0.034037083          0.40701669 -0.03030534
#> 15           30.444383   0.100366581          1.17788149 -0.02679166
#> 16           30.444383   0.072325025          0.87169008 -0.02660128
#> 17           30.444383   0.038231921          0.46956548 -0.01913945
#> 18           30.444383   0.049936770          0.60013163 -0.01932821
#> 19           30.444383   0.118999520          1.34330949 -0.01955208
#> 20           30.444383   0.049738887          0.59229350 -0.01991846
#> 21            2.366131   0.006052657          0.05005525 -0.03300000
#> 22            2.366131   0.011230845          0.08987991 -0.02989623
#> 23            2.366131   0.008928170          0.07353636 -0.02934012
#> 24            2.366131   0.003902959          0.03291967 -0.02991136
#> 25            2.366131   0.011500603          0.09266689 -0.02572342
#> 26            2.366131   0.008287044          0.07122386 -0.02556753
#> 27            2.366131   0.004373660          0.03929343 -0.01698910
#> 28            2.366131   0.005712830          0.04876452 -0.01734261
#> 29            2.366131   0.013614180          0.09952559 -0.01772881
#> 30            2.366131   0.005690768          0.04751658 -0.01828328
#>         LeafPLC      StemPLC PlantWaterBalance   LeafRWC   StemRWC      LFMC
#> 1  3.713826e-03 1.320577e-11      0.000000e+00 0.9979679 0.9998350 125.89003
#> 2  3.148335e-03 1.320577e-11      1.676671e-04 0.9981861 0.9998527 125.90513
#> 3  3.055298e-03 1.320577e-11      2.845186e-05 0.9982231 0.9998557 125.90769
#> 4  3.164344e-03 1.320577e-11     -3.332107e-05 0.9981797 0.9998522 125.90469
#> 5  2.443324e-03 1.320577e-11      2.277960e-04 0.9984761 0.9998762 125.92520
#> 6  2.425627e-03 1.320577e-11      5.835946e-06 0.9984837 0.9998769 125.92572
#> 7  1.164553e-03 1.320577e-11      4.613944e-04 0.9990841 0.9999256 125.96726
#> 8  1.226188e-03 1.320577e-11     -2.538042e-05 0.9990511 0.9999229 125.96498
#> 9  1.291858e-03 1.320577e-11     -2.660723e-05 0.9990165 0.9999201 125.96258
#> 10 1.381591e-03 1.320577e-11     -3.568385e-05 0.9989700 0.9999164 125.95937
#> 11 3.473351e-05 1.691266e-06      0.000000e+00 0.9986794 0.9997489  93.07448
#> 12 2.883760e-05 1.691266e-06      2.937910e-05 0.9987840 0.9997689  93.08071
#> 13 2.776487e-05 1.691266e-06      5.692987e-06 0.9988043 0.9997728  93.08192
#> 14 2.866336e-05 1.691266e-06     -4.776342e-06 0.9987873 0.9997695  93.08091
#> 15 2.170913e-05 1.691266e-06      3.947190e-05 0.9989279 0.9997964  93.08928
#> 16 2.136282e-05 1.691266e-06      2.137718e-06 0.9989355 0.9997979  93.08973
#> 17 1.016853e-05 1.691266e-06      8.371480e-05 0.9992342 0.9998547  93.10750
#> 18 1.039607e-05 1.691266e-06     -2.116106e-06 0.9992266 0.9998533  93.10705
#> 19 1.066958e-05 1.691266e-06     -2.509865e-06 0.9992176 0.9998516  93.10652
#> 20 1.112573e-05 1.691266e-06     -4.107721e-06 0.9992030 0.9998488  93.10564
#> 21 3.473351e-05 4.394058e-08      3.469447e-18 0.9987207 0.9983684  96.39176
#> 22 2.779821e-05 4.394058e-08      3.895903e-06 0.9988410 0.9985218  96.40517
#> 23 2.664578e-05 4.394058e-08      6.980821e-07 0.9988625 0.9985492  96.40758
#> 24 2.782993e-05 4.394058e-08     -7.170689e-07 0.9988404 0.9985210  96.40511
#> 25 1.980592e-05 4.394058e-08      5.257363e-06 0.9990027 0.9987280  96.42320
#> 26 1.953630e-05 4.394058e-08      1.957016e-07 0.9990088 0.9987357  96.42388
#> 27 7.772166e-06 4.394058e-08      1.077116e-05 0.9993413 0.9991597  96.46096
#> 28 8.141624e-06 4.394058e-08     -4.439322e-07 0.9993276 0.9991422  96.45943
#> 29 8.556182e-06 4.394058e-08     -4.849673e-07 0.9993126 0.9991231  96.45776
#> 30 9.171479e-06 4.394058e-08     -6.962704e-07 0.9992911 0.9990957  96.45536
#>    PlantStress
#> 1  0.005808769
#> 2  0.004917109
#> 3  0.004770526
#> 4  0.004942336
#> 5  0.003807299
#> 6  0.003779472
#> 7  0.001802163
#> 8  0.001898471
#> 9  0.002001133
#> 10 0.002141490
#> 11 0.010612890
#> 12 0.009724363
#> 13 0.009552588
#> 14 0.009696696
#> 15 0.008508820
#> 16 0.008444696
#> 17 0.005954140
#> 18 0.006016526
#> 19 0.006090566
#> 20 0.006211841
#> 21 0.003088161
#> 22 0.002716528
#> 23 0.002651108
#> 24 0.002718312
#> 25 0.002234831
#> 26 0.002217261
#> 27 0.001303891
#> 28 0.001339248
#> 29 0.001378121
#> 30 0.001434372

#Select the output tables/variables to be extracted
extract(S1, level ="cohort", output="Plants", vars = c("PlantStress", "StemPLC"))
#>          date cohort           species PlantStress      StemPLC
#> 1  2001-01-01 T1_148  Pinus halepensis 0.005808769 1.320577e-11
#> 2  2001-01-02 T1_148  Pinus halepensis 0.004917109 1.320577e-11
#> 3  2001-01-03 T1_148  Pinus halepensis 0.004770526 1.320577e-11
#> 4  2001-01-04 T1_148  Pinus halepensis 0.004942336 1.320577e-11
#> 5  2001-01-05 T1_148  Pinus halepensis 0.003807299 1.320577e-11
#> 6  2001-01-06 T1_148  Pinus halepensis 0.003779472 1.320577e-11
#> 7  2001-01-07 T1_148  Pinus halepensis 0.001802163 1.320577e-11
#> 8  2001-01-08 T1_148  Pinus halepensis 0.001898471 1.320577e-11
#> 9  2001-01-09 T1_148  Pinus halepensis 0.002001133 1.320577e-11
#> 10 2001-01-10 T1_148  Pinus halepensis 0.002141490 1.320577e-11
#> 11 2001-01-01 T2_168      Quercus ilex 0.010612890 1.691266e-06
#> 12 2001-01-02 T2_168      Quercus ilex 0.009724363 1.691266e-06
#> 13 2001-01-03 T2_168      Quercus ilex 0.009552588 1.691266e-06
#> 14 2001-01-04 T2_168      Quercus ilex 0.009696696 1.691266e-06
#> 15 2001-01-05 T2_168      Quercus ilex 0.008508820 1.691266e-06
#> 16 2001-01-06 T2_168      Quercus ilex 0.008444696 1.691266e-06
#> 17 2001-01-07 T2_168      Quercus ilex 0.005954140 1.691266e-06
#> 18 2001-01-08 T2_168      Quercus ilex 0.006016526 1.691266e-06
#> 19 2001-01-09 T2_168      Quercus ilex 0.006090566 1.691266e-06
#> 20 2001-01-10 T2_168      Quercus ilex 0.006211841 1.691266e-06
#> 21 2001-01-01 S1_165 Quercus coccifera 0.003088161 4.394058e-08
#> 22 2001-01-02 S1_165 Quercus coccifera 0.002716528 4.394058e-08
#> 23 2001-01-03 S1_165 Quercus coccifera 0.002651108 4.394058e-08
#> 24 2001-01-04 S1_165 Quercus coccifera 0.002718312 4.394058e-08
#> 25 2001-01-05 S1_165 Quercus coccifera 0.002234831 4.394058e-08
#> 26 2001-01-06 S1_165 Quercus coccifera 0.002217261 4.394058e-08
#> 27 2001-01-07 S1_165 Quercus coccifera 0.001303891 4.394058e-08
#> 28 2001-01-08 S1_165 Quercus coccifera 0.001339248 4.394058e-08
#> 29 2001-01-09 S1_165 Quercus coccifera 0.001378121 4.394058e-08
#> 30 2001-01-10 S1_165 Quercus coccifera 0.001434372 4.394058e-08