Fire behaviour functions

Function fire_FCCS() implements a modification of the fire behavior models described for the Fuel Characteristics Classification System (FCCS) in Prichard et al. (2013). Function fire_Rothermel() implements Rothermel's (1972) fire behaviour model (modified from package 'Rothermel' (Giorgio Vacchiano, Davide Ascoli)).

fire_FCCS(FCCSpropsSI, MliveSI = as.numeric(c(90, 90, 60)), 
          MdeadSI = as.numeric(c(6, 6, 6, 6, 6)), 
          slope = 0, windSpeedSI = 11)
fire_Rothermel(modeltype, wSI, sSI, delta, mx_dead,
               hSI, mSI, u, windDir, slope, aspect)

Arguments

FCCSpropsSI

A data frame describing the properties of five fuel strata (canopy, shrub, herbs, dead woody and litter) returned by fuel_FCCS.

MliveSI

Moisture of live fuels (in percent of dry weight) for canopy, shrub, and herb strata. Live moisture values are drawn from column ActFCM in FCCSpropsSI if available (see fuel_FCCS). Otherwise, moisture values supplied for MliveSI are used.

MdeadSI

Moisture of dead fuels (in percent of dry weight) for canopy, shrub, herb, woody and litter strata.

slope

Slope (in degrees).

windSpeedSI

Wind speed (in m/s) at 20 ft (6 m) over vegetation (default 11 m/s = 40 km/h)

modeltype

'S'(tatic) or 'D'(ynamic)

wSI

A vector of fuel load (t/ha) for five fuel classes.

sSI

A vector of surface-to-volume ratio (m2/m3) for five fuel classes.

delta

A value of fuel bed depth (cm).

mx_dead

A value of dead fuel moisture of extinction (percent).

hSI

A vector of heat content (kJ/kg) for five fuel classes.

mSI

A vector of percent moisture on a dry weight basis (percent) for five fuel classes.

u

A value of windspeed (m/s) at midflame height.

windDir

Wind direction (in degrees from north). North means blowing from north to south.

aspect

Aspect (in degrees from north).

Details

Default moisture, slope and windspeed values are benchmark conditions used to calculate fire potentials (Sandberg et al. 2007) and map vulnerability to fire.

Value

Both functions return list with fire behavior variables. In the case of fire_FCCS, the function returns the variables in three blocks (lists SurfaceFire, CrownFire and FirePotentials), and the values are:

SurfaceFire$`midflame_WindSpeed [m/s]`: Midflame wind speed in the surface fire.

SurfaceFire$phi_wind: Spread rate modifier due to wind.

SurfaceFire$phi_slope: Spread rate modifier due to slope.

SurfaceFire$`I_R_surf [kJ/m2/min]`: Intensity of the surface fire reaction.

SurfaceFire$`I_R_litter [kJ/m2/min]`: Intensity of the litter fire reaction.

SurfaceFire$`q_surf [kJ/m2]`: Heat sink of the surface fire.

SurfaceFire$`q_litter [kJ/m2]`: Heat sink of the litter fire.

SurfaceFire$xi_surf: Propagating flux ratio of the surface fire.

SurfaceFire$xi_litter: Propagating flux ratio of the litter fire.

SurfaceFire$`ROS_surf [m/min]`: Spread rate of the surface fire(without accounting for faster spread in the litter layer).

SurfaceFire$`ROS_litter [m/min]`: Spread rate of the litter fire.

SurfaceFire$`ROS_windslopecap [m/min]`: Maximum surface fire spread rate according to wind speed.

SurfaceFire$`ROS [m/min]`: Final spread rate of the surface fire.

SurfaceFire$`I_b [kW/m]`: Fireline intensity of the surface fire.

SurfaceFire$`FL [m]`: Flame length of the surface fire.

CrownFire$`I_R_canopy [kJ/m2/min]`: Intensity of the canopy fire reaction.

CrownFire$`I_R_crown [kJ/m2/min]`: Intensity of the crown fire reaction (adding surface and canopy reactions).

CrownFire$`q_canopy [kJ/m2]`: Heat sink of the canopy fire.

CrownFire$`q_crown [kJ/m2]`: Heat sink of the crown fire (adding surface and canopy heat sinks).

CrownFire$xi_surf: Propagating flux ratio of the crown fire.

CrownFire$`canopy_WindSpeed [m/s]`: Wind speed in the canopy fire (canopy top wind speed).

CrownFire$WAF: Wind speed adjustment factor for crown fires.

CrownFire$`ROS [m/min]`: Spread rate of the crown fire.

CrownFire$Ic_ratio: Crown initiation ratio.

CrownFire$`I_b [kW/m]`: Fireline intensity of the crown fire.

CrownFire$`FL [m]`: Flame length of the crown fire.

FirePotentials$RP: Surface fire reaction potential ([0-9]).

FirePotentials$SP: Surface fire spread rate potential ([0-9]).

FirePotentials$FP: Surface fire flame length potential ([0-9]).

FirePotentials$SFP: Surface fire potential ([0-9]).

FirePotentials$IC: Crown initiation potential ([0-9]).

FirePotentials$TC: Crown-to-crown transmission potential ([0-9]).

FirePotentials$RC: Crown fire spread rate potential ([0-9]).

FirePotentials$CFC: Crown fire potential ([0-9]).

References

Albini, F. A. (1976). Computer-based models of wildland fire behavior: A users' manual. Ogden, UT: US Department of Agriculture, Forest Service, Intermountain Forest and Range Experiment Station.

Rothermel, R. C. 1972. A mathematical model for predicting fire spread in wildland fuels. USDA Forest Service Research Paper INT USA.

Prichard, S. J., D. V Sandberg, R. D. Ottmar, E. Eberhardt, A. Andreu, P. Eagle, and K. Swedin. 2013. Classification System Version 3.0: Technical Documentation.

Author

Miquel De Cáceres Ainsa, CREAF

Note

Default moisture, slope and windspeed values are benchmark conditions used to calculate fire potentials (Sandberg et al. 2007) and map vulnerability to fire.

See also

Examples

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Calculate fuel properties according to FCCS
fccs = fuel_FCCS(exampleforestMED, SpParamsMED)

#Calculate fire behavior according to FCCS
fire_FCCS(fccs)
#> $SurfaceFire
#> $SurfaceFire$`midflame_WindSpeed [m/s]`
#> [1] 2.188184
#> 
#> $SurfaceFire$phi_wind
#> [1] 17.1111
#> 
#> $SurfaceFire$phi_slope
#> [1] 0
#> 
#> $SurfaceFire$`I_R_surf [kJ/m2/min]`
#> [1] 21044.02
#> 
#> $SurfaceFire$`I_R_litter [kJ/m2/min]`
#> [1] 2854.286
#> 
#> $SurfaceFire$`q_surf [kJ/m2]`
#> [1] 16760.95
#> 
#> $SurfaceFire$`q_litter [kJ/m2]`
#> [1] 1617.702
#> 
#> $SurfaceFire$xi_surf
#> [1] 0.09755156
#> 
#> $SurfaceFire$xi_litter
#> [1] 0.1612053
#> 
#> $SurfaceFire$`ROS_surf [m/min]`
#> [1] 2.219832
#> 
#> $SurfaceFire$`ROS_litter [m/min]`
#> [1] 5.155064
#> 
#> $SurfaceFire$`ROS_windslopecap [m/min]`
#> [1] 131.291
#> 
#> $SurfaceFire$`ROS [m/min]`
#> [1] 5.155064
#> 
#> $SurfaceFire$`I_b [kW/m]`
#> [1] 574.4353
#> 
#> $SurfaceFire$`FL [m]`
#> [1] 1.440186
#> 
#> 
#> $CrownFire
#> $CrownFire$`I_R_canopy [kJ/m2/min]`
#> [1] 26080.5
#> 
#> $CrownFire$`I_R_crown [kJ/m2/min]`
#> [1] 47124.52
#> 
#> $CrownFire$`q_canopy [kJ/m2]`
#> [1] 402.1434
#> 
#> $CrownFire$`q_crown [kJ/m2]`
#> [1] 17163.1
#> 
#> $CrownFire$xi_crown
#> [1] 0.08844873
#> 
#> $CrownFire$`canopy_WindSpeed [m/s]`
#> [1] 5.072246
#> 
#> $CrownFire$WAF
#> [1] 2.039257
#> 
#> $CrownFire$`ROS_crown [m/min]`
#> [1] 21.15147
#> 
#> $CrownFire$`I_b_crown [kW/m]`
#> [1] 3987.826
#> 
#> $CrownFire$Ic_ratio
#> [1] 1.193898
#> 
#> $CrownFire$`FL_crown [m]`
#> [1] 7.21161
#> 
#> 
#> $FirePotentials
#> $FirePotentials$RP
#> [1] 3.444983
#> 
#> $FirePotentials$SP
#> [1] 9
#> 
#> $FirePotentials$FP
#> [1] 5.434277
#> 
#> $FirePotentials$SFP
#> [1] 9
#> 
#> $FirePotentials$IC
#> [1] 4.144322
#> 
#> $FirePotentials$TC
#> [1] 9
#> 
#> $FirePotentials$RC
#> [1] 8.330341
#> 
#> $FirePotentials$CFP
#> [1] 5.87568
#> 
#> 

#Load fuel model parameter data
data(SFM_metric)

#Fuel stratification (returns heights in cm)
fs = fuel_stratification(exampleforestMED, SpParamsMED)

#Correct windspeed (transform heights to m)
u = 11 #m/s
umf = u*fuel_windAdjustmentFactor(fs$surfaceLayerTopHeight/100, fs$canopyBaseHeight/100, 
                                  fs$canopyTopHeight/100, 60)

#Call Rothermel function using fuel model 'A6'
fire_Rothermel(modeltype="D", wSI = as.numeric(SFM_metric["A6",2:6]), 
                    sSI = as.numeric(SFM_metric["A6",7:11]), 
                    delta = as.numeric(SFM_metric["A6",12]),
                    mx_dead = as.numeric(SFM_metric["A6",13]),
                    hSI = as.numeric(SFM_metric["A6",14:18]),
                    mSI = c(10,10,10,30,60),
                    u=umf, windDir=0, slope=0, aspect=0)
#> $`Characteristic dead fuel moisture [%]`
#> [1] 10
#> 
#> $`Characteristic live fuel moisture [%]`
#> [1] 0
#> 
#> $`Live fuel moisture of extinction [%]`
#> [1] 25
#> 
#> $`Characteristic SAV [m2/m3]`
#> [1] 5129.366
#> 
#> $`Bulk density [kg/m3]`
#> [1] 1.95
#> 
#> $`Packing ratio [dimensionless]`
#> [1] 0.003804203
#> 
#> $`Relative packing ratio [dimensionless]`
#> [1] 0.4689126
#> 
#> $`Dead fuel Reaction intensity [kW/m2]`
#> [1] 365.233
#> 
#> $`Live fuel Reaction intensity [kW/m2]`
#> [1] 0
#> 
#> $`Reaction intensity [kW/m2]`
#> [1] 365.233
#> 
#> $`Fireline intensity [kW/m]`
#> [1] 1070.755
#> 
#> $`Wind factor [dimensionless]`
#> [1] 24.44011
#> 
#> $`Slope factor [dimensionless]`
#> [1] 0
#> 
#> $`Slope-wind vector`
#> [1] 24.440107  3.141593
#> 
#> $`Virtual wind speed [m/s]`
#> [1] 2.824933
#> 
#> $`Heat source [kW/m2]`
#> [1] 276.182
#> 
#> $`Heat sink [kJ/m3]`
#> [1] 1388.106
#> 
#> $`ROS [m/min]`
#> [1] 11.93624
#>