Forest dynamics

Function fordyn implements a forest dynamics model that simulates growth, mortality and recruitment for plant cohorts in a given forest stand during a period specified in the input climatic data.

fordyn(forest, soil, SpParams, meteo, control,
       latitude , elevation = NA, slope = NA, aspect = NA)

Arguments

forest
An object of class forest.

soil
An object of class soil.

SpParams
A data frame with species parameters (see SpParamsMED and SpParamsMED).

meteo
A data frame with daily meteorological data series. Row names of the data frame should correspond to date strings with format "yyyy-mm-dd" (see Date).

control
A list with default control parameters (see defaultControl).

latitude
Latitude (in degrees). Required when x$TranspirationMode = "Sperry".

elevation, slope, aspect
Elevation above sea level (in m), slope (in degrees) and aspect (in degrees from North). Required when x$TranspirationMode = "Sperry". Elevation is also required for 'Granier' if snowpack dynamics are simulated.

Details

Function fordyn simulates forest dynamics for annual time steps, building on other simulation functions. For each simulated year, the function performs the following steps:

Calls function growth to simulate daily water/carbon balance, growth and mortality processes.

Simulate recruitment (for species present in the stand).

Prepares the input of function growth for the next annual time step.

Store forest status and summaries.

Value

A list of class 'fordyn' with the following elements:

"StandSummary": A data frame with stand-level summaries (leaf area index, tree basal area, tree density, shrub cover, etc.) at the beginning of the simulation and after each simulated year.

"SpeciesSummary": A data frame with species-level summaries (leaf area index, tree basal area, tree density, shrub cover, etc.) at the beginning of the simulation and after each simulated year.

"CohortSummary": A data frame with cohort-level summaries (leaf area index, tree basal area, tree density, shrub cover, etc.) at the beginning of the simulation and after each simulated year.

"TreeTable": A data frame with tree-cohort data (species, density, diameter, height, etc.) at the beginning of the simulation (if any) and after each simulated year.

"DeadTreeTable": A data frame with dead tree-cohort data (species, density, diameter, height, etc.) at the beginning of the simulation and after each simulated year.

"ShrubTable": A data frame with shrub-cohort data (species, density, cover, height, etc.) at the beginning of the simulation and after each simulated year.

"DeadShrubTable": A data frame with dead shrub-cohort data (species, density, cover, height, etc.) at the beginning of the simulation (if any) and after each simulated year.

"ForestStructures": A list with the forest object of the stand at the beginning of the simulation and after each simulated year.

"GrowthResults": A list with the results of calling function growth for each simulated year.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

#Load example daily meteorological data
data(examplemeteo)
#Prepare a two-year meteorological data with half precipitation during 
#the second year
meteo2001 = examplemeteo
meteo2002 = examplemeteo
meteo2002$Precipitation = meteo2002$Precipitation/2
row.names(meteo2002) = seq(as.Date("2002-01-01"), 
                           as.Date("2002-12-31"), by="day")
meteo_01_02 = rbind(meteo2001, meteo2002)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize control parameters
control = defaultControl()

#Initialize soil with default soil params (4 layers)
examplesoil = soil(defaultSoilParams(4))

#Call simulation function
fd<-fordyn(exampleforestMED, examplesoil, 
           SpParamsMED, meteo_01_02, control,
           latitude = 41.82592, elevation = 100)

#> Simulating forest dynamics for year 2001 (1/2)
#>    (a) Growth/mortality
#>    (b) Recruitment
#>        Coldest month mean temp. (Celsius): 0.05   Moisture index: 0.36   FPAR (%): 39.8
#>        Tree species with seed rain: 54,68 recruited: <none>
#>    (c) Summaries
#> Simulating forest dynamics for year 2002 (2/2)
#>    (a) Growth/mortality
#>    (b) Recruitment
#>        Coldest month mean temp. (Celsius): 0.05   Moisture index: 0.18   FPAR (%): 38.9
#>        Tree species with seed rain: 54,68 recruited: <none>
#>    (c) Summaries

#Stand-level summaries
fd$StandSummary

#>   Step LeafAreaIndex TreeDensityLive TreeBasalAreaLive ShrubCoverLive MaxHeight
#> 1    0      1.662671        552.0000          25.03330           3.75  800.0000
#> 2    1      1.920717        546.3412          26.85657           3.75  824.6774
#> 3    2      2.085433        540.7403          27.79923           3.75  838.4610

#Tree table by annual steps
fd$TreeTable

#>   Step Year Cohort Species             Name        N      DBH   Height Z50  Z95
#> 1    0   NA  T1_54      54 Pinus halepensis 168.0000 37.55000 800.0000 750 3000
#> 2    0   NA  T2_68      68     Quercus ilex 384.0000 14.60000 660.0000 750 3000
#> 3    1 2001  T1_54      54 Pinus halepensis 166.2777 38.76854 824.6774 750 3000
#> 4    1 2001  T2_68      68     Quercus ilex 380.0634 15.56123 685.5711 750 3000
#> 5    2 2002  T1_54      54 Pinus halepensis 164.5731 39.45864 838.4610 750 3000
#> 6    2 2002  T2_68      68     Quercus ilex 376.1672 16.11705 700.1935 750 3000

#Dead tree table by annual steps
fd$DeadTreeTable

#>   Step Year Cohort Species             Name        N      DBH   Height Z50  Z95
#> 1    1 2001  T1_54      54 Pinus halepensis 1.722253 38.76854 824.6774 750 3000
#> 2    1 2001  T2_68      68     Quercus ilex 3.936579 15.56123 685.5711 750 3000
#> 3    2 2002  T1_54      54 Pinus halepensis 1.704598 39.45864 838.4610 750 3000
#> 4    2 2002  T2_68      68     Quercus ilex 3.896223 16.11705 700.1935 750 3000