Forest dynamics

Function fordyn implements a forest dynamics model that simulates growth, mortality, recruitment and (optionally) management actions in a given forest stand during a period specified in the input climatic data.

fordyn(forest, soil, SpParams, meteo, control,
       latitude , elevation = NA, slope = NA, aspect = NA,
       management_function = NULL, management_args = NULL)

Arguments

forest

An object of class forest. Alternatively, the output of a previous run, if continuing a previous simulation.

soil

An object of class soil.

SpParams

A data frame with species parameters (see SpParamsMED and SpParamsDefinition).

meteo

A data frame with daily meteorological data series. Row names of the data frame should correspond to date strings with format "yyyy-mm-dd" (see Date).

control

A list with default control parameters (see defaultControl).

latitude

Latitude (in degrees). Required when x$TranspirationMode = "Sperry".

elevation, slope, aspect

Elevation above sea level (in m), slope (in degrees) and aspect (in degrees from North). Required when x$TranspirationMode = "Sperry". Elevation is also required for 'Granier' if snowpack dynamics are simulated.

management_function

A function that implements forest management actions (see details).

management_args

A list of additional arguments to be passed to the management_function.

Details

Function fordyn simulates forest dynamics for annual time steps, building on other simulation functions. For each simulated year, the function performs the following steps:

Calls function growth to simulate daily water/carbon balance, growth and mortality processes and update the forest object.

If required, calls function management_function, using as parameters the forest object and management_args, which may result in a density reduction for existing plant cohorts and/or a set of new planted cohorts.

Simulate natural recruitment (for species present in the stand or given in a seed rain input).

Prepares the input of function growth for the next annual time step.

Store forest status, management arguments, and summaries.

To enable forest management, the user needs to provide a function that implements it, which is passed to fordyn via its argument management_function. Such function should have the following arguments:

"x": the forest object representing the stand to be managed.

"args": a list of parameters regulating the behavior of the management function.

"verbose": a logical flag to enable console output during the execution of the management function.

and return a list with the following elements:

"action": A string identifying the action performed (e.g. "thinning").

"N_tree_cut": A vector with the density of trees removed.

"Cover_shrub_cut": A vector with the cover of shrubs removed.

"planted_forest": An object of class forest with the new plant cohorts resulting from tree/shrub planting.

"management_args": A list of management arguments to be used in the next call to the management function.

An example of management function is provided in defaultManagementFunction.

Value

A list of class 'fordyn' with the following elements:

"StandSummary": A data frame with stand-level summaries (tree basal area, tree density, shrub cover, etc.) at the beginning of the simulation and after each simulated year.

"SpeciesSummary": A data frame with species-level summaries (tree basal area, tree density, shrub cover, etc.) at the beginning of the simulation and after each simulated year.

"CohortSummary": A data frame with cohort-level summaries (tree basal area, tree density, shrub cover, etc.) at the beginning of the simulation and after each simulated year.

"TreeTable": A data frame with tree-cohort data (species, density, diameter, height, etc.) at the beginning of the simulation (if any) and after each simulated year.

"DeadTreeTable": A data frame with dead tree-cohort data (species, density, diameter, height, etc.) at the beginning of the simulation and after each simulated year.

"CutTreeTable": A data frame with cut tree data (species, density, diameter, height, etc.) after each simulated year.

"ShrubTable": A data frame with shrub-cohort data (species, density, cover, height, etc.) at the beginning of the simulation and after each simulated year.

"DeadShrubTable": A data frame with dead shrub-cohort data (species, density, cover, height, etc.) at the beginning of the simulation (if any) and after each simulated year.

"CutShrubTable": A data frame with cut shrub data (species, density, cover, height, etc.) after each simulated year.

"ForestStructures": A list with the forest object of the stand at the beginning of the simulation and after each simulated year.

"GrowthResults": A list with the results of calling function growth for each simulated year.

"ManagementArgs": A list of management arguments to be used in another call to fordyn.

"NextInputObject": An object of class growthInput to be used in a subsequent simulation.

"NextForestObject": An object of class forest to be used in a subsequent simulation.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

#Load example daily meteorological data
data(examplemeteo)
#Prepare a two-year meteorological data with half precipitation during 
#the second year
meteo2001 = examplemeteo
meteo2002 = examplemeteo
meteo2002$Precipitation = meteo2002$Precipitation/2
row.names(meteo2002) = seq(as.Date("2002-01-01"), 
                           as.Date("2002-12-31"), by="day")
meteo_01_02 = rbind(meteo2001, meteo2002)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize control parameters
control = defaultControl("Granier")

#Initialize soil with default soil params (4 layers)
examplesoil = soil(defaultSoilParams(4))

#Call simulation function
fd<-fordyn(exampleforestMED, examplesoil, 
           SpParamsMED, meteo_01_02, control,
           latitude = 41.82592, elevation = 100)
#> Simulating year 2001 (1/2):  (a) Growth/mortality, (b) Recruitment
#> Simulating year 2002 (2/2):  (a) Growth/mortality, (b) Recruitment

#Stand-level summaries
fd$StandSummary
#>   Step NumTreeSpecies NumTreeCohorts NumShrubSpecies NumShrubCohorts
#> 1    0              2              2               1               1
#> 2    1              2              2               1               1
#> 3    2              2              2               1               1
#>   TreeDensityLive TreeBasalAreaLive DominantTreeHeight DominantTreeDiameter
#> 1        552.0000          25.03330           800.0000             37.55000
#> 2        551.1388          25.17084           809.1730             37.71939
#> 3        550.2821          25.23594           815.0852             37.82921
#>   QuadraticMeanTreeDiameter HartBeckingIndex ShrubCoverLive BasalAreaDead
#> 1                  24.02949         53.20353       3.750000    0.00000000
#> 2                  24.11423         52.64148       2.737354    0.09623075
#> 3                  24.16418         52.30032       2.824083    0.09629575
#>   ShrubCoverDead BasalAreaCut ShrubCoverCut
#> 1    0.000000000            0             0
#> 2    0.004097987            0             0
#> 3    0.004271120            0             0

#Tree table by annual steps
fd$TreeTable
#>   Step Year Cohort Species             Name        N      DBH   Height Z50  Z95
#> 1    0   NA T1_148     148 Pinus halepensis 168.0000 37.55000 800.0000 100  600
#> 2    0   NA T2_168     168     Quercus ilex 384.0000 14.60000 660.0000 300 1000
#> 3    1 2001 T1_148     148 Pinus halepensis 167.1388 37.71939 809.1730 100  600
#> 4    1 2001 T2_168     168     Quercus ilex 384.0000 14.67423 662.4111 300 1000
#> 5    2 2002 T1_148     148 Pinus halepensis 166.2821 37.82921 815.0852 100  600
#> 6    2 2002 T2_168     168     Quercus ilex 384.0000 14.73341 664.3390 300 1000

#Dead tree table by annual steps
fd$DeadTreeTable
#>   Step Year Cohort Species             Name         N      DBH   Height Z50 Z95
#> 1    1 2001 T1_148     148 Pinus halepensis 0.8611814 37.71939 809.1730 100 600
#> 2    2 2002 T1_148     148 Pinus halepensis 0.8567669 37.82921 815.0852 100 600