Forest description

Description of a forest stand.

# S3 method for forest
summary(object, SpParams, detailed=FALSE, ...)
# S3 method for summary.forest
print(x, digits = getOption("digits"), ...)
emptyforest(ID="", patchsize=10000, ntree = 0, nshrub = 0)

Arguments

object
An object of class forest has the following structure:

ID: An identifier of the forest stand (a string).

patchsize: The area that forest stand description represents, in square meters.

treeData: A data frame of tree cohorts (in rows) and the following columns:

Species: Non-negative integer for tree species identity (i.e., 0,1,2,...).

Height: Total height (in cm).

DBH: Diameter at breast height (in cm).

N: Density (number of individuals/cell).

Z50: Depth (in mm) corresponding to 50% of fine roots.

Z95: Depth (in mm) corresponding to 95% of fine roots.

shrubData: A data frame of shrub cohorts (in rows) and the following columns:

Species: Non-negative integer for shrub species identity (i.e., 0,1,2,...).

Height: Total height (in cm).

Cover: Percent cover.

Z50: Depth (in mm) corresponding to 50% of fine roots.

Z95: Depth (in mm) corresponding to 95% of fine roots.

seedBank: A data frame containing the abundance of seeds for each species (in rows) and the following columns:

Species: Non-negative integer for shrub species identity (i.e., 0,1,2,...).

Abundance: Abundance class (0 - none; 1 - low; 2 - medium; 3 - high; 4 - very high).

herbCover: Percent cover of the herb layer.

herbHeight: Mean height (in cm) of the herb layer.

SpParams
A data frame with species parameters (see SpParamsMED).

detailed
A boolean flag to indicate that a detailed summary is desired.

x
The object returned by summary.forest.

digits
Minimal number of significant digits.

...
Additional parameters for functions summary and print.

ID
An identifier of the forest stand (a string).

patchsize
The area of the forest stand, in square meters.

ntree, nshrub
Number of tree and shrub cohorts, respectively.

Details

Function summary.forest can be used to summarize a forest object in the console. Function emptyforest creates an empty forest object.

Value

Function summary.forest returns a data frame with the basal area and LAI of the forest, either expressed as totals or divided among life stages and species. Function emptyforest returns an empty forest object.

See also

Examples

data(exampleforest)
data(SpParamsMED)

summary(exampleforest, SpParamsMED)

#> Forest summary:
#> 
#>   Tree density (ind/ha): 552 
#>   BA (m2/ha): 25.0333016 
#>   Shrub crown phytovolume (m3/m2): 0.0109563 
#>   LAI (m2/m2) total: 1.7036297  trees: 1.6144964  shrubs: 0.0891333 
#>   Live fine fuel (kg/m2) total: 0.7033713  trees: 0.6642999  shrubs: 0.0390714