Forest description

Description of a forest stand

emptyforest(ntree = 0, nshrub = 0)

# S3 method for forest
summary(object, SpParams, ...)

# S3 method for summary.forest
print(x, digits = getOption("digits"), ...)

Arguments

ntree, nshrub

Number of tree and shrub cohorts, respectively.

object

An object of class forest has the following structure:

treeData: A data frame of tree cohorts (in rows) and the following columns:

Species: String with species (taxon) name or a non-negative integer for tree species identity (i.e., 0,1,2,...) matching SpParams.

Height: Total tree height (in cm).

DBH: Tree diameter at breast height (in cm).

N: Density (number of individuals/hectare) that the measured tree represents.

Z50: Depth (in mm) corresponding to 50% of fine roots.

Z95: Depth (in mm) corresponding to 95% of fine roots.

shrubData: A data frame of shrub cohorts (in rows) and the following columns:

Species: String with species (taxon) name or a non-negative integer for shrub species identity (i.e., 0,1,2,...) matching SpParams.

Height: Average total height of plants (in cm).

Cover: Percent cover.

Z50: Depth (in mm) corresponding to 50% of fine roots.

Z95: Depth (in mm) corresponding to 95% of fine roots.

herbCover: Percent cover of the herb layer.

herbHeight: Mean height (in cm) of the herb layer.

SpParams

A data frame with species parameters (see SpParamsMED).

...

Additional parameters for functions summary and print.

x

The object returned by summary.forest.

digits

Minimal number of significant digits.

Value

Function summary.forest returns a list with several structural attributes, such as the basal area and LAI of the forest. Function emptyforest returns an empty forest object.

Details

Function summary.forest can be used to summarize a forest object in the console. Function emptyforest creates an empty forest object.

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

data(exampleforestMED)
data(SpParamsMED)

# Example forest
exampleforestMED
#> $treeData
#>            Species   N   DBH Height Z50  Z95
#> 1 Pinus halepensis 168 37.55    800 100  600
#> 2     Quercus ilex 384 14.60    660 300 1000
#> 
#> $shrubData
#>             Species Cover Height Z50  Z95
#> 1 Quercus coccifera  3.75     80 200 1000
#> 
#> $herbCover
#> [1] 10
#> 
#> $herbHeight
#> [1] 20
#> 
#> attr(,"class")
#> [1] "forest" "list"  

# Summary of example forest
summary(exampleforestMED, SpParamsMED)
#> Tree BA (m2/ha): 25.0333016  adult trees: 25.0333016  saplings: 0 
#> Density (ind/ha) adult trees: 552  saplings: 0  shrubs (estimated): 749.4923076 
#> Cover (%) adult trees: 100  saplings: 0  shrubs: 3.75  herbs: 10 
#> LAI (m2/m2) total: 1.8335096  adult trees: 1.5543216  saplings: 0  shrubs: 0.027188  herbs: 0.252 
#> Fuel (kg/m2) total: 0.5699197  adult trees: 0.5255004  saplings: 0  shrubs: 0.0164193  herbs: 0.028 
#> PAR ground (%): 38.5431077  SWR ground (%): 49.3507632