Fuel stratification and fuel characteristics

Function fuel_stratification provides a stratification of the stand into understory and canopy strata. Function fuel_FCCS calculates fuel characteristics from a forest object following an adaptation of the protocols described for the Fuel Characteristics Classification System (Prichard et al. 2013). Function fuel_windAdjustmentFactor determines the adjustment factor of wind for surface fires, according to Andrews (2012). Function fuel_cohortFineFMC calculates the fuel moisture content of leaves and twigs of each cohort, from the results of soil water balance.

fuel_stratification(object, SpParams, gdd = NA,
                      heightProfileStep = 10.0, maxHeightProfile = 5000.0,
                      bulkDensityThreshold = 0.05)
fuel_FCCS(object, ShrubCover, CanopyCover, SpParams, cohortFMC = as.numeric(c()),
          gdd = NA, heightProfileStep = 10, maxHeightProfile = 5000,
          bulkDensityThreshold = 0.05)
fuel_cohortFineFMC(spwb, x)
fuel_windAdjustmentFactor(topShrubHeight, bottomCanopyHeight, topCanopyHeight,
                          canopyCover)

Arguments

object
An object of class forest

ShrubCover
Total shrub cover (in percent) of the stand.

CanopyCover
Total canopy cover (in percent) of the stand.

SpParams
A data frame with species parameters (see SpParamsMED).

cohortFMC
A numeric vector of (actual) fuel moisture content by cohort (e.g. taken from the result of fuel_cohortFineFMC).

gdd
Growth degree-days.

heightProfileStep
Precision for the fuel bulk density profile.

maxHeightProfile
Maximum height for the fuel bulk density profile.

bulkDensityThreshold
Minimum fuel bulk density to delimit fuel strata.

spwb
Object returned by function spwb.

topShrubHeight
Shrub stratum top height (in m).

bottomCanopyHeight
Canopy base height (in m).

topCanopyHeight
Canopy top height (in m).

canopyCover
Canopy percent cover.

x
An object of class spwbInput.

Details

Details are described in a vignette.

Value

Function fuel_FCCS returns a data frame with five rows corresponding to fuel layers: canopy, shrub, herb, woody and litter. Columns correspond fuel properties:

w: Fine fuel loading (in kg/m2).

cover: Percent cover.

hbc: Height to base of crowns (in m).

htc: Height to top of crowns (in m).

delta: Fuel depth (in m).

rhob: Fuel bulk density (in kg/m3).

rhop: Fuel particle density (in kg/m3).

PV: Particle volume (in m3/m2).

beta: Packing ratio (unitless).

betarel: Relative packing ratio (unitless).

etabetarel: Reaction efficiency (unitless).

sigma: Surface area-to-volume ratio (m2/m3).

pDead: Proportion of dead fuels.

FAI: Fuel area index (unitless).

h: High heat content (in kJ/kg).

etaF: Flammability modifier (between 1 and 2).

RV: Reactive volume (in m3/m2).

MinFMC: Minimum fuel moisture content (as percent over dry weight).

MaxFMC: Maximum fuel moisture content (as percent over dry weight).

Function fuel_stratification returns a list with the following items:

surfaceLayerBaseHeight: Base height of crowns of shrubs in the surface layer (in cm).

surfaceLayerTopHeight: Top height of crowns of shrubs in the surface layer (in cm).

understoryLAI: Cumulated LAI of the understory layer (i.e. leaf area comprised between surface layer base and top heights).

canopyBaseHeight: Base height of tree crowns in the canopy (in cm).

canopyTopHeight: Top height of tree crowns in the canopy (in cm).

canopyLAI: Cumulated LAI of the canopy (i.e. leaf area comprised between canopy base and top heights).

Function fuel_cohortFineFMC returns a list with three matrices (for leaves, twigs and fine fuels). Each of them contains live moisture content values for each day (in rows) and plant cohort (in columns).

Function fuel_windAdjustmentFactor returns a value between 0 and 1.

References

Andrews, P. L. 2012. Modeling wind adjustment factor and midflame wind speed for Rothermel’s surface fire spread model. USDA Forest Service - General Technical Report RMRS-GTR:1–39.

Prichard, S. J., D. V Sandberg, R. D. Ottmar, E. Eberhardt, A. Andreu, P. Eagle, and K. Swedin. 2013. Classification System Version 3.0: Technical Documentation.

Reinhardt, E., D. Lutes, and J. Scott. 2006. FuelCalc: A method for estimating fuel characteristics. Pages 273–282.

See also

Examples

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

#Show stratification of fuels
fuel_stratification(exampleforest, SpParamsMED)

#> $surfaceLayerBaseHeight
#> [1] 0
#> 
#> $surfaceLayerTopHeight
#> [1] 30
#> 
#> $understoryLAI
#> [1] 0.08913325
#> 
#> $canopyBaseHeight
#> [1] 260
#> 
#> $canopyTopHeight
#> [1] 740
#> 
#> $canopyLAI
#> [1] 1.543083
#> 

#Calculate fuel properties according to FCCS
fccs = fuel_FCCS(exampleforest, 50,100, SpParamsMED)
fccs

#>                 w cover hbc htc       delta       rhob rhop           PV
#> canopy 0.66429990   100 2.6 7.4 5.051677308  0.1315009  400 1.660750e-03
#> shrub  0.03907141    50 0.0 0.3 0.292166666  0.1337299  400 9.767853e-05
#> herb   0.02800000    10 0.0  NA 0.200000000  0.1400000  400 7.000000e-05
#> woody  0.19685510    NA 0.0  NA 0.007448169 26.4300000  400 4.921377e-04
#> litter 0.17473831    NA 0.0  NA 0.008653036 20.1938720  400 4.368458e-04
#>                beta   betarel etabetarel     sigma      pDead       FAI
#> canopy 0.0003287521 0.2117231  0.4657068  5274.119 0.08060296 8.7589911
#> shrub  0.0003343247 0.4334359  0.7638016  4141.000 0.10856686 0.4044868
#> herb   0.0003500000 0.8228426  0.9823251 11483.000 0.00000000 0.8038100
#> woody  0.0660750000 0.8228426  0.9823251  1601.050 1.00000000 0.7879371
#> litter 0.0504846799 9.1128920  0.1467533  7340.922 1.00000000 3.2068508
#>               h etaF           RV   MinFMC    MaxFMC ActFMC
#> canopy 21044.37    1 1.660750e-03 78.34051 101.81002     NA
#> shrub  21100.00    1 9.767853e-05 43.00000  98.35104     NA
#> herb   18608.00    1 7.000000e-05       NA        NA     NA
#> woody  18608.00    1 4.921377e-04       NA        NA     NA
#> litter 18608.00    1 4.368458e-04       NA        NA     NA

fuel_windAdjustmentFactor(fccs$htc[2], fccs$hbc[1], fccs$htc[1], fccs$cover[1])

#> [1] 0.1968428