Fuel stratification and fuel characteristics

Function fuel_stratification provides a stratification of the stand into understory and canopy strata. Function fuel_FCCS calculates fuel characteristics from a forest object following an adaptation of the protocols described for the Fuel Characteristics Classification System (Prichard et al. 2013). Function fuel_windAdjustmentFactor determines the adjustment factor of wind for surface fires, according to Andrews (2012).

fuel_stratification(object, SpParams, gdd = NA, mode = "MED",
                      heightProfileStep = 10.0, maxHeightProfile = 5000.0,
                      bulkDensityThreshold = 0.05)
fuel_FCCS(object, ShrubCover, CanopyCover, SpParams, cohortFMC = as.numeric(c()),
          gdd = NA, mode = "MED", heightProfileStep = 10, maxHeightProfile = 5000, 
          bulkDensityThreshold = 0.05, depthMode = "crownaverage")
fuel_windAdjustmentFactor(topShrubHeight, bottomCanopyHeight, topCanopyHeight,
                          canopyCover)

Arguments

object
An object of class forest

ShrubCover
Total shrub cover (in percent) of the stand.

CanopyCover
Total canopy cover (in percent) of the stand.

SpParams
A data frame with species parameters (see SpParamsMED).

cohortFMC
A numeric vector of (actual) fuel moisture content by cohort (e.g. can be taken as a row of the matrix returned by moisture_cohortFMC).

gdd
Growth degree-days.

mode
Calculation mode, either "MED" or "US".

heightProfileStep
Precision for the fuel bulk density profile.

maxHeightProfile
Maximum height for the fuel bulk density profile.

bulkDensityThreshold
Minimum fuel bulk density to delimit fuel strata.

depthMode
Specifies how fuel depth (and therefore canopy and understory bulk density) should be estimated:

"crownaverage": As weighed average of crown lengths using loadings as weights.

"profile": As the difference of base and top heights in bulk density profiles.

"absoluteprofile": As the difference of absolute base and absolute top heights in bulk density profiles.

topShrubHeight
Shrub stratum top height (in m).

bottomCanopyHeight
Canopy base height (in m).

topCanopyHeight
Canopy top height (in m).

canopyCover
Canopy percent cover.

Details

Details are described in a vignette.

Value

Function fuel_FCCS returns a data frame with five rows corresponding to fuel layers: canopy, shrub, herb, woody and litter. Columns correspond fuel properties:

w: Fine fuel loading (in kg/m2).

cover: Percent cover.

hbc: Height to base of crowns (in m).

htc: Height to top of crowns (in m).

delta: Fuel depth (in m).

rhob: Fuel bulk density (in kg/m3).

rhop: Fuel particle density (in kg/m3).

PV: Particle volume (in m3/m2).

beta: Packing ratio (unitless).

betarel: Relative packing ratio (unitless).

etabetarel: Reaction efficiency (unitless).

sigma: Surface area-to-volume ratio (m2/m3).

pDead: Proportion of dead fuels.

FAI: Fuel area index (unitless).

h: High heat content (in kJ/kg).

RV: Reactive volume (in m3/m2).

MinFMC: Minimum fuel moisture content (as percent over dry weight).

MaxFMC: Maximum fuel moisture content (as percent over dry weight).

Function fuel_stratification returns a list with the following items:

surfaceLayerBaseHeight: Base height of crowns of shrubs in the surface layer (in cm).

surfaceLayerTopHeight: Top height of crowns of shrubs in the surface layer (in cm).

understoryLAI: Cumulated LAI of the understory layer (i.e. leaf area comprised between surface layer base and top heights).

canopyBaseHeight: Base height of tree crowns in the canopy (in cm).

canopyTopHeight: Top height of tree crowns in the canopy (in cm).

canopyLAI: Cumulated LAI of the canopy (i.e. leaf area comprised between canopy base and top heights).

Function fuel_cohortFineFMC returns a list with three matrices (for leaves, twigs and fine fuels). Each of them contains live moisture content values for each day (in rows) and plant cohort (in columns).

Function fuel_windAdjustmentFactor returns a value between 0 and 1.

References

Andrews, P. L. 2012. Modeling wind adjustment factor and midflame wind speed for Rothermel’s surface fire spread model. USDA Forest Service - General Technical Report RMRS-GTR:1–39.

Prichard, S. J., D. V Sandberg, R. D. Ottmar, E. Eberhardt, A. Andreu, P. Eagle, and K. Swedin. 2013. Classification System Version 3.0: Technical Documentation.

Reinhardt, E., D. Lutes, and J. Scott. 2006. FuelCalc: A method for estimating fuel characteristics. Pages 273–282.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Show stratification of fuels
fuel_stratification(exampleforestMED, SpParamsMED)

#> $surfaceLayerBaseHeight
#> [1] 40
#> 
#> $surfaceLayerTopHeight
#> [1] 50
#> 
#> $surfaceLayerAbsoluteBaseHeight
#> [1] 0
#> 
#> $surfaceLayerAbsoluteTopHeight
#> [1] 80
#> 
#> $understoryLAI
#> [1] 0.0083744
#> 
#> $canopyBaseHeight
#> [1] 270
#> 
#> $canopyTopHeight
#> [1] 740
#> 
#> $canopyAbsoluteBaseHeight
#> [1] 260
#> 
#> $canopyAbsoluteTopHeight
#> [1] 800
#> 
#> $canopyLAI
#> [1] 1.856981
#> 

#Calculate fuel properties according to FCCS
fccs = fuel_FCCS(exampleforestMED, 50,100, SpParamsMED)
fccs

#>                 w cover hbc htc habc hatc       delta        rhob     rhop
#> canopy 0.61531386   100 2.7 7.4  2.6  8.0 4.770490365  0.12898336 593.5150
#> shrub  0.02372309    50 0.4 0.5  0.0  0.8 0.642625347  0.03691589 308.7209
#> herb   0.02800000    10 0.0  NA  0.0   NA 0.200000000  0.14000000 400.0000
#> woody  0.20319943    NA 0.0  NA  0.0   NA 0.007688211 26.43000000 171.3280
#> litter 0.27544624    NA 0.0  NA  0.0   NA 0.014186176 19.41652394 387.1262
#>                  PV         beta   betarel etabetarel     sigma        pDead
#> canopy 1.069526e-03 0.0002241962 0.1483048  0.3475696  5252.030 0.0004042436
#> shrub  7.684315e-05 0.0001195769 1.1271871  0.9925659  4141.000 0.1448400000
#> herb   7.000000e-05 0.0003500000 3.1390396  0.3696781 11483.000 0.0000000000
#> woody  1.186026e-03 0.1542654897 3.1390396  0.3696781  1601.050 1.0000000000
#> litter 7.115153e-04 0.0501555387 9.2478770  0.1426090  7438.017 1.0000000000
#>              FAI        h           RV   MinFMC    MaxFMC ActFMC
#> canopy 5.7941052 21012.89 1.036728e-03 74.74129 112.91297     NA
#> shrub  0.3182075 20117.67 7.684315e-05 63.64891  96.53441     NA
#> herb   0.8038100 18608.00 7.000000e-05       NA        NA     NA
#> woody  1.8988864 18608.00 1.186026e-03       NA        NA     NA
#> litter 5.2922629 18608.00 7.115153e-04       NA        NA     NA

fuel_windAdjustmentFactor(fccs$htc[2], fccs$hbc[1], fccs$htc[1], fccs$cover[1])

#> [1] 0.1989258