Fuel stratification and fuel characteristics

Function fuel_stratification provides a stratification of the stand into understory and canopy strata. Function fuel_FCCS calculates fuel characteristics from a forest object following an adaptation of the protocols described for the Fuel Characteristics Classification System (Prichard et al. 2013). Function fuel_windAdjustmentFactor determines the adjustment factor of wind for surface fires, according to Andrews (2012).

fuel_stratification(object, SpParams, gdd = NA, mode = "MED",
                      heightProfileStep = 10.0, maxHeightProfile = 5000.0,
                      bulkDensityThreshold = 0.05)
fuel_FCCS(object, SpParams, cohortFMC = as.numeric(c()),
          gdd = NA, mode = "MED", heightProfileStep = 10, maxHeightProfile = 5000, 
          bulkDensityThreshold = 0.05, depthMode = "crownaverage")
fuel_windAdjustmentFactor(topShrubHeight, bottomCanopyHeight, topCanopyHeight,
                          canopyCover)

Arguments

object

An object of class forest

SpParams

A data frame with species parameters (see SpParamsMED).

cohortFMC

A numeric vector of (actual) fuel moisture content by cohort.

gdd

Growth degree-days.

mode

Calculation mode, either "MED" or "US".

heightProfileStep

Precision for the fuel bulk density profile.

maxHeightProfile

Maximum height for the fuel bulk density profile.

bulkDensityThreshold

Minimum fuel bulk density to delimit fuel strata.

depthMode

Specifies how fuel depth (and therefore canopy and understory bulk density) should be estimated:

"crownaverage": As weighed average of crown lengths using loadings as weights.

"profile": As the difference of base and top heights in bulk density profiles.

"absoluteprofile": As the difference of absolute base and absolute top heights in bulk density profiles.

topShrubHeight

Shrub stratum top height (in m).

bottomCanopyHeight

Canopy base height (in m).

topCanopyHeight

Canopy top height (in m).

canopyCover

Canopy percent cover.

Details

Details are described in a vignette.

Value

Function fuel_FCCS returns a data frame with five rows corresponding to fuel layers: canopy, shrub, herb, woody and litter. Columns correspond fuel properties:

w: Fine fuel loading (in kg/m2).

cover: Percent cover.

hbc: Height to base of crowns (in m).

htc: Height to top of crowns (in m).

delta: Fuel depth (in m).

rhob: Fuel bulk density (in kg/m3).

rhop: Fuel particle density (in kg/m3).

PV: Particle volume (in m3/m2).

beta: Packing ratio (unitless).

betarel: Relative packing ratio (unitless).

etabetarel: Reaction efficiency (unitless).

sigma: Surface area-to-volume ratio (m2/m3).

pDead: Proportion of dead fuels.

FAI: Fuel area index (unitless).

h: High heat content (in kJ/kg).

RV: Reactive volume (in m3/m2).

MinFMC: Minimum fuel moisture content (as percent over dry weight).

MaxFMC: Maximum fuel moisture content (as percent over dry weight).

ActFMC: Actual fuel moisture content (as percent over dry weight). These are set to NA if parameter cohortFMC is empty.

Function fuel_stratification returns a list with the following items:

surfaceLayerBaseHeight: Base height of crowns of shrubs in the surface layer (in cm).

surfaceLayerTopHeight: Top height of crowns of shrubs in the surface layer (in cm).

understoryLAI: Cumulated LAI of the understory layer (i.e. leaf area comprised between surface layer base and top heights).

canopyBaseHeight: Base height of tree crowns in the canopy (in cm).

canopyTopHeight: Top height of tree crowns in the canopy (in cm).

canopyLAI: Cumulated LAI of the canopy (i.e. leaf area comprised between canopy base and top heights).

Function fuel_cohortFineFMC returns a list with three matrices (for leaves, twigs and fine fuels). Each of them contains live moisture content values for each day (in rows) and plant cohort (in columns).

Function fuel_windAdjustmentFactor returns a value between 0 and 1.

References

Andrews, P. L. 2012. Modeling wind adjustment factor and midflame wind speed for Rothermel’s surface fire spread model. USDA Forest Service - General Technical Report RMRS-GTR:1–39.

Prichard, S. J., D. V Sandberg, R. D. Ottmar, E. Eberhardt, A. Andreu, P. Eagle, and K. Swedin. 2013. Classification System Version 3.0: Technical Documentation.

Reinhardt, E., D. Lutes, and J. Scott. 2006. FuelCalc: A method for estimating fuel characteristics. Pages 273–282.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Show stratification of fuels
fuel_stratification(exampleforestMED, SpParamsMED)
#> $surfaceLayerBaseHeight
#> [1] 40
#> 
#> $surfaceLayerTopHeight
#> [1] 50
#> 
#> $surfaceLayerAbsoluteBaseHeight
#> [1] 0
#> 
#> $surfaceLayerAbsoluteTopHeight
#> [1] 80
#> 
#> $understoryLAI
#> [1] 0.00829473
#> 
#> $canopyBaseHeight
#> [1] 270
#> 
#> $canopyTopHeight
#> [1] 740
#> 
#> $canopyAbsoluteBaseHeight
#> [1] 260
#> 
#> $canopyAbsoluteTopHeight
#> [1] 800
#> 
#> $canopyLAI
#> [1] 1.756232
#> 

#Calculate fuel properties according to FCCS
fccs = fuel_FCCS(exampleforestMED, SpParamsMED)
fccs
#>                 w  cover hbc htc habc hatc       delta        rhob     rhop
#> canopy 0.61531386 100.00 2.7 7.4  2.6  8.0 4.770490365  0.12898336 595.8159
#> shrub  0.02372309   3.75 0.4 0.5  0.0  0.8 0.642625347  0.03691589 412.0091
#> herb   0.02800000  10.00 0.0  NA  0.0   NA 0.200000000  0.14000000 400.0000
#> woody  0.20185195     NA 0.0  NA  0.0   NA 0.007637228 26.43000000 438.9106
#> litter 0.27309512     NA 0.0  NA  0.0   NA 0.014027046 19.46918201 370.9679
#>                  PV         beta   betarel etabetarel    sigma        pDead
#> canopy 1.068691e-03 2.240211e-04 0.1483959  0.3477514  5252.03 0.0004042436
#> shrub  5.757903e-05 8.959969e-05 0.2912218  0.5916193  4141.00 0.1448400000
#> herb   7.000000e-05 3.500000e-04 0.7375491  0.9588969 11483.00 0.0000000000
#> woody  4.598931e-04 6.021728e-02 0.7375491  0.9588969  1601.05 1.0000000000
#> litter 7.361692e-04 5.248213e-02 9.2387331  0.1428898  7431.44 1.0000000000
#>              FAI        h           RV   MinFMC    MaxFMC ActFMC
#> canopy 5.7976757 21012.89 1.032725e-03 74.74129 112.91297     NA
#> shrub  0.2384348 20117.67 5.757903e-05 63.64891  96.53441     NA
#> herb   0.8038100 18608.00 7.000000e-05       NA        NA     NA
#> woody  0.7363118 18608.00 4.598931e-04       NA        NA     NA
#> litter 5.4707970 18608.00 7.361692e-04       NA        NA     NA

fuel_windAdjustmentFactor(fccs$htc[2], fccs$hbc[1], fccs$htc[1], fccs$cover[1])
#> [1] 0.1989258