Forest growth

Function growth is a process-based model that performs energy, water and carbon balances; and determines changes in water/carbon pools, functional variables (leaf area, sapwood area, root area) and structural ones (tree diameter, tree height, shrub cover) for woody plant cohorts in a given forest stand during a period specified in the input climatic data.

growth(
  x,
  meteo,
  latitude,
  elevation = NA_real_,
  slope = NA_real_,
  aspect = NA_real_,
  CO2ByYear = numeric(0)
)

Arguments

x

An object of class growthInput.

meteo

A data frame with daily meteorological data series (see spwb).

latitude

Latitude (in degrees).

elevation, slope, aspect

Elevation above sea level (in m), slope (in degrees) and aspect (in degrees from North).

CO2ByYear

A named numeric vector with years as names and atmospheric CO2 concentration (in ppm) as values. Used to specify annual changes in CO2 concentration along the simulation (as an alternative to specifying daily values in meteo).

Value

A list of class 'growth' with the following elements:

"latitude": Latitude (in degrees) given as input.

"topography": Vector with elevation, slope and aspect given as input.

"weather": A copy of the input weather data frame.

"growthInput": A copy of the object x of class growthInput given as input.

"growthOutput": An copy of the final state of the object x of class growthInput.

"WaterBalance": A data frame where different water balance variables (see spwb).

"EnergyBalance": A data frame with the daily values of energy balance components for the soil and the canopy (only for transpirationMode = "Sperry" or transpirationMode = "Cochard"; see spwb).

"CarbonBalance": A data frame where different stand-level carbon balance components (gross primary production, maintenance respiration, synthesis respiration and net primary production), all in g C · m-2.

"BiomassBalance": A data frame with the daily values of stand biomass balance components (in g dry · m-2.

"Temperature": A data frame with the daily values of minimum/mean/maximum temperatures for the atmosphere (input), canopy and soil (only for transpirationMode = "Sperry" or transpirationMode = "Cochard"; see spwb).

"Soil": A data frame where different soil variables (see spwb).

"Stand": A data frame where different stand-level variables (see spwb).

"Plants": A list of daily results for plant cohorts (see spwb).

"SunlitLeaves" and "ShadeLeaves": A list with daily results for sunlit and shade leaves (only for transpirationMode = "Sperry" or transpirationMode = "Cochard"; see spwb).

"LabileCarbonBalance": A list of daily labile carbon balance results for plant cohorts, with elements:

"GrossPhotosynthesis": Daily gross photosynthesis per dry weight of living biomass (g gluc · g dry-1).

"MaintentanceRespiration": Daily maintenance respiration per dry weight of living biomass (g gluc · g dry-1).

"GrowthCosts": Daily growth costs per dry weight of living biomass (g gluc · g dry-1).

"RootExudation": Root exudation per dry weight of living biomass (g gluc · g dry-1).

"LabileCarbonBalance": Daily labile carbon balance (photosynthesis - maintenance respiration - growth costs - root exudation) per dry weight of living biomass (g gluc · g dry-1).

"SugarLeaf": Sugar concentration (mol·l-1) in leaves.

"StarchLeaf": Starch concentration (mol·l-1) in leaves.

"SugarSapwood": Sugar concentration (mol·l-1) in sapwood.

"StarchSapwood": Starch concentration (mol·l-1) in sapwood.

"SugarTransport": Average instantaneous rate of carbon transferred between leaves and stem compartments via floem (mol gluc·s-1).

"PlantBiomassBalance": A list of daily plant biomass balance results for plant cohorts, with elements:

"StructuralBiomassBalance": Daily structural biomass balance (g dry · m-2).

"LabileBiomassBalance": Daily labile biomass balance (g dry · m-2).

"PlantBiomassBalance": Daily plant biomass balance, i.e. labile change + structural change (g dry · m-2).

"MortalityBiomassLoss": Biomass loss due to mortality (g dry · m-2).

"CohortBiomassBalance": Daily cohort biomass balance (including mortality) (g dry · m-2).

"PlantStructure": A list of daily area and biomass values for compartments of plant cohorts, with elements:

"LeafBiomass": Daily amount of leaf structural biomass (in g dry) for an average individual of each plant cohort.

"SapwoodBiomass": Daily amount of sapwood structural biomass (in g dry) for an average individual of each plant cohort.

"FineRootBiomass": Daily amount of fine root biomass (in g dry) for an average individual of each plant cohort.

"LeafArea": Daily amount of leaf area (in m2) for an average individual of each plant cohort.

"SapwoodArea": Daily amount of sapwood area (in cm2) for an average individual of each plant cohort.

"FineRootArea": Daily amount of fine root area (in m2) for an average individual of each plant cohort.

"HuberValue": The ratio of sapwood area to (target) leaf area (in cm2/m2).

"RootAreaLeafArea": The ratio of fine root area to (target) leaf area (in m2/m2).

"DBH": Diameter at breast height (in cm) for an average individual of each plant cohort.

"Height": Height (in cm) for an average individual of each plant cohort.

"GrowthMortality": A list of daily growth and mortality rates for plant cohorts, with elements:

"LAgrowth": Leaf area growth (in m2·day-1) for an average individual of each plant cohort.

"SAgrowth": Sapwood area growth rate (in cm2·day-1) for an average individual of each plant cohort.

"FRAgrowth": Fine root area growth (in m2·day-1) for an average individual of each plant cohort.

"StarvationRate": Daily mortality rate from starvation (ind/d-1).

"DessicationRate": Daily mortality rate from dessication (ind/d-1).

"MortalityRate": Daily mortality rate (any cause) (ind/d-1).

"subdaily": A list of objects of class growth_day, one per day simulated (only if required in control parameters, see defaultControl).

Details

Detailed model description is available in the medfate book. Simulations using the 'Sperry' or 'Cochard' transpiration modes are computationally much more expensive than those using the 'Granier' transpiration mode.

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)
  
#Initialize control parameters
control <- defaultControl("Granier")
  
#Initialize soil with default soil params (4 layers)
examplesoil <- soil(defaultSoilParams(4))

#Initialize vegetation input
x1 <- forest2growthInput(exampleforestMED, examplesoil, SpParamsMED, control)

#Call simulation function
G1 <- growth(x1, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial plant cohort biomass (g/m2): 6108.49
#> Initial plant water content (mm): 7.05319
#> Initial soil water content (mm): 286.862
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  Year 2001:....................................
#> 
#> Final plant biomass (g/m2): 6263.13
#> Change in plant biomass (g/m2): 154.639
#> Plant biomass balance result (g/m2): 154.639
#> Plant biomass balance components:
#>   Structural balance (g/m2) 83 Labile balance (g/m2) 96
#>   Plant individual balance (g/m2) 179 Mortality loss (g/m2) 25
#> Final plant water content (mm): 7.06479
#> Final soil water content (mm): 269.417
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): 0.0115975
#> Plant water balance result (mm): -0.00174902
#> Change in soil water content (mm): -17.4447
#> Soil water balance result (mm): -17.4447
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 97 Net rainfall (mm) 365
#>   Infiltration (mm) 405 Runoff (mm) 11 Deep drainage (mm) 138
#>   Soil evaporation (mm) 27  Herbaceous transpiration (mm) 26 Woody plant transpiration (mm) 232
#>   Plant extraction from soil (mm) 232  Plant water balance (mm) -0 Hydraulic redistribution (mm) 2
 
# \donttest{
#Switch to 'Sperry' transpiration mode
control <- defaultControl("Sperry")

#Initialize vegetation input
x2 <- forest2growthInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Call simulation function
G2 <-growth(x2, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial plant cohort biomass (g/m2): 7414.04
#> Initial plant water content (mm): 8.34165
#> Initial soil water content (mm): 286.862
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  Year 2001:....................................
#> 
#> Final plant biomass (g/m2): 7595.29
#> Change in plant biomass (g/m2): 181.255
#> Plant biomass balance result (g/m2): 181.255
#> Plant biomass balance components:
#>   Structural balance (g/m2) 108 Labile balance (g/m2) 103
#>   Plant individual balance (g/m2) 211 Mortality loss (g/m2) 30
#> Final plant water content (mm): 8.35934
#> Final soil water content (mm): 271.202
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): 0.0176939
#> Plant water balance result (mm): 9.40856e-17
#> Change in soil water content (mm): -15.6596
#> Soil water balance result (mm): -15.6596
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 7.10543e-15
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 97 Net rainfall (mm) 365
#>   Infiltration (mm) 405 Runoff (mm) 11 Deep drainage (mm) 159
#>   Soil evaporation (mm) 24  Herbaceous transpiration (mm) 26 Woody plant transpiration (mm) 212
#>   Plant extraction from soil (mm) 212  Plant water balance (mm) 0 Hydraulic redistribution (mm) 2

#Switch to 'Cochard' transpiration mode
control <- defaultControl("Cochard")

#Initialize vegetation input
x3 <- forest2growthInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Call simulation function
G3 <-growth(x3, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial plant cohort biomass (g/m2): 7414.04
#> Initial plant water content (mm): 8.34165
#> Initial soil water content (mm): 286.862
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  Year 2001:....................................
#> 
#> Final plant biomass (g/m2): 7552.67
#> Change in plant biomass (g/m2): 138.631
#> Plant biomass balance result (g/m2): 138.631
#> Plant biomass balance components:
#>   Structural balance (g/m2) 68 Labile balance (g/m2) 100
#>   Plant individual balance (g/m2) 169 Mortality loss (g/m2) 30
#> Final plant water content (mm): 8.26153
#> Final soil water content (mm): 276.351
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): -0.0801187
#> Plant water balance result (mm): 5.02546
#> Change in soil water content (mm): -10.5113
#> Soil water balance result (mm): -10.5113
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 95 Net rainfall (mm) 367
#>   Infiltration (mm) 407 Runoff (mm) 11 Deep drainage (mm) 216
#>   Soil evaporation (mm) 31  Herbaceous transpiration (mm) 26 Woody plant transpiration (mm) 139
#>   Plant extraction from soil (mm) 144  Plant water balance (mm) 5 Hydraulic redistribution (mm) 0
# }