Soil vertical inputs

High-level functions for hydrological processes. Function hydrology_soilWaterInputs performs canopy water interception and snow accumulation/melt. Function hydrology_soilInfiltrationPercolation performs soil infiltration and percolation from the input given by the previous function.

hydrology_soilWaterInputs(
  soil,
  soilFunctions,
  prec,
  er,
  tday,
  rad,
  elevation,
  Cm,
  LgroundPAR,
  LgroundSWR,
  runon = 0,
  snowpack = TRUE,
  modifySoil = TRUE
)

hydrology_soilInfiltrationPercolation(
  soil,
  soilFunctions,
  waterInput,
  rockyLayerDrainage = TRUE,
  modifySoil = TRUE
)

Arguments

soil

A list containing the description of the soil (see soil).

soilFunctions

Soil water retention curve and conductivity functions, either 'SX' (for Saxton) or 'VG' (for Van Genuchten).

prec

Precipitation for a given day (mm)

er

The ratio of evaporation rate to rainfall rate.

tday

Average day temperature (ºC).

rad

Solar radiation (in MJ/m2/day).

elevation

Altitude above sea level (m).

Cm

Canopy water storage capacity.

LgroundPAR

Percentage of photosynthetically-acvive radiation (PAR) reaching the ground.

LgroundSWR

Percentage of short-wave radiation (SWR) reaching the ground.

runon

Surface water amount running on the target area from upslope (in mm).

snowpack

Boolean flag to indicate the simulation of snow accumulation and melting.

modifySoil

Boolean flag to indicate that the input soil object should be modified during the simulation.

waterInput

Soil water input for a given day (mm).

rockyLayerDrainage

Boolean flag to indicate the simulation of drainage from rocky layers (> 95% of rocks).

Value

Function hydrology_soilWaterInputs returns a named vector with the following elements, all in mm:

Rain

Precipitation as rainfall.

Snow

Precipitation as snow.

Interception

Rainfall water intercepted by the canopy and evaporated.

NetRain

Rainfall reaching the ground.

Snowmelt

Snow melted during the day, and added to the water infiltrated.

Runon

Surface water amount running on the target area from upslope.

Input

Total soil input, including runon, snowmelt and net rain.

Function hydrology_soilInfiltrationPercolation returns a named vector with the following elements, all in mm:

Infiltration

Water infiltrated into the soil (i.e. throughfall + runon + snowmelt - runoff).

Runoff

Surface water leaving the target area.

DeepDrainage

Water leaving the target soil towards the water table.

Details

The function simulates different vertical hydrological processes, which are described separately in other functions. If modifySoil = TRUE the function will modify the soil object (including both soil moisture and the snowpack on its surface) as a result of simulating hydrological processes.

See also

Author

Miquel De Cáceres Ainsa, CREAF