Input for simulation models

Functions forest2spwbInput and forest2growthInput take an object of class forest and calculate input data for functions spwb, pwb and growth, respectively. Functions spwbInput and growthInput do the same but starting from different input data. Function forest2aboveground calculates aboveground variables that may be used in spwbInput and growthInput functions. Function forest2belowground calculates belowground fine root distribution.

forest2aboveground(x, SpParams, gdd = NA, mode = "MED")
forest2belowground(x, soil)
forest2growthInput(x, soil, SpParams, control)
forest2spwbInput(x, soil, SpParams, control, mode = "MED")
growthInput(above,  Z50, Z95, soil, SpParams, control)
spwbInput(above,  Z50, Z95, soil, SpParams, control)

Arguments

x

An object of class forest.

SpParams

A data frame with species parameters (see SpParamsMED and SpParamsMED).

gdd

Growth degree days to account for leaf phenology effects (in Celsius). This should be left NA in most applications.

mode

Calculation mode, either "MED" or "US".

soil

An object of class soil.

control

A list with default control parameters (see defaultControl).

above

A data frame with aboveground plant information (see the return value of forest2aboveground below). In the case of spwbInput the variables should include SP, N, LAI_live, LAI_dead, H and CR. In the case of growthInput variables should include DBH and Cover.

Z50, Z95

Numeric vectors with cohort depths (in mm) corresponding to 50% and 95% of fine roots.

Details

Functions forest2spwbInput and forest2abovegroundInput extracts height and species identity from plant cohorts of x, and calculate leaf area index and crown ratio.forest2spwbInput also calculates the distribution of fine roots across soil. Both forest2spwbInput and spwbInput find parameter values for each plant cohort according to the parameters of its species as specified in SpParams. If control$transpirationMode = "Sperry" the functions also estimate the maximum conductance of rhizosphere, root xylem and stem xylem elements.

Value

Function forest2aboveground() returns a data frame with the following columns (rows are identified as specified by function plant_ID):

SP: Species identity (an integer) (first species is 0).

N: Cohort density (ind/ha) (see function plant_density).

DBH: Tree diameter at breast height (cm).

H: Plant total height (cm).

CR: Crown ratio (crown length to total height) (between 0 and 1).

LAI_live: Live leaf area index (m2/m2) (one-side leaf area relative to plot area), includes leaves in winter dormant buds.

LAI_expanded: Leaf area index of expanded leaves (m2/m2) (one-side leaf area relative to plot area).

LAI_dead: Dead leaf area index (m2/m2) (one-side leaf area relative to plot area).

Functions forest2spwbInput() and spwbInput() return a list of class spwbInput with the following elements (rows of data frames are identified as specified by function plant_ID):

control: List with control parameters (see defaultControl).

canopy: A list of stand-level state variables.

cohorts: A data frame with cohort information, with columns SP and Name.

above: A data frame with columns H, CR and LAI (see function forest2aboveground).

below: A data frame with columns Z50, Z95. If control$transpirationMode = "Sperry" additional columns are fineRootBiomass and coarseRootSoilVolume.

belowLayers: A list. If control$transpirationMode = "Granier" it contains elements:

V: A matrix with the proportion of fine roots of each cohort (in rows) in each soil layer (in columns).

L: A matrix with the length of coarse roots of each cohort (in rows) in each soil layer (in columns).

Wpool: A matrix with the soil moisture relative to field capacity around the rhizosphere of each cohort (in rows) in each soil layer (in columns).

If control$transpirationMode = "Sperry" there are the following additional elements:

VGrhizo_kmax: A matrix with maximum rhizosphere conductance values of each cohort (in rows) in each soil layer (in columns).

VGroot_kmax: A matrix with maximum root xylem conductance values of each cohort (in rows) in each soil layer (in columns).

RhizoPsi: A matrix with the water potential around the rhizosphere of each cohort (in rows) in each soil layer (in columns).

paramsPhenology: A data frame with leaf phenology parameters:

PhenologyType: Leaf phenology type.

LeafDuration: Leaf duration (in years).

Sgdd: Degree days needed for leaf budburst (for winter decideous species).

Tbgdd: Base temperature for the calculation of degree days to leaf budburst.

Ssen: Degree days corresponding to leaf senescence.

Phsen: Photoperiod corresponding to start counting senescence degree-days.

Tbsen: Base temperature for the calculation of degree days to leaf senescence.

paramsAnatomy: A data frame with plant anatomy parameters for each cohort (only if control$transpirationMode = "Sperry"):

Hmax: Maximum plant height (cm).

Hmed: Median plant height (cm).

Al2As: Leaf area to sapwood area ratio (in m2·m-2).

SLA: Specific leaf area (mm2/mg = m2/kg).

LeafWidth: Leaf width (in cm).

LeafDensity: Density of leaf tissue (dry weight over volume).

WoodDensity: Density of wood tissue (dry weight over volume).

FineRootDensity: Density of fine root tissue (dry weight over volume).

SRL: Specific Root length (cm·g-1).

RLD: Root length density (cm·cm-3).

r635: Ratio between the weight of leaves plus branches and the weight of leaves alone for branches of 6.35 mm.

paramsInterception: A data frame with rain interception and light extinction parameters for each cohort:

kPAR: PAR extinction coefficient.

g: Canopy water retention capacity per LAI unit (mm/LAI).

If control$transpirationMode = "Sperry" additional columns are:

gammaSWR: Reflectance (albedo) coefficient for SWR .

alphaSWR: Absorbance coefficient for SWR .

paramsTranspiration: A data frame with parameters for transpiration and photosynthesis. If control$transpirationMode = "Granier", columns are:

Tmax_LAI: Coefficient relating LAI with the ratio of maximum transpiration over potential evapotranspiration.

Tmax_LAIsq: Coefficient relating squared LAI with the ratio of maximum transpiration over potential evapotranspiration.

Psi_Extract: Water potential corresponding to 50% relative conductance (in MPa).

Psi_Critic: Water potential corresponding to 50% of stem cavitation (in MPa).

WUE: Water use efficiency for carbon assimilation (g C/mm water).

pRootDisc: Proportion of whole-plant conductance leading to disconnection from soil.

If control$transpirationMode = "Sperry" columns are:

Gswmax: Maximum stomatal conductance to water vapor (in mol H2O·m-2·s-1).

Vmax298: Maximum Rubisco carboxilation rate at 25ºC (in micromol CO2·s-1·m-2).

Jmax298: Maximum rate of electron transport at 25ºC (in micromol photons·s-1·m-2).

Kmax_stemxylem: Sapwood-specific hydraulic conductivity of stem xylem (in kg H2O·s-1·m-2).

Kmax_rootxylem: Sapwood-specific hydraulic conductivity of root xylem (in kg H2O·s-1·m-2).

VCleaf_kmax: Maximum leaf hydraulic conductance.

VCleaf_c, VCleaf_d: Parameters of the leaf vulnerability curve.

VCstem_kmax: Maximum stem xylem conductance.

VCstem_c, VCstem_d: Parameters of the stem xylem vulnerability curve.

VCroot_c, VCroot_d: Parameters of the root xylem vulnerability curve.

Plant_kmax: Maximum whole-plant conductance.

paramsWaterStorage: A data frame with plant water storage parameters for each cohort (only if control$transpirationMode = "Sperry"):

LeafPI0: Osmotic potential at full turgor of leaves (MPa).

LeafEPS: Modulus of elasticity (capacity of the cell wall to resist changes in volume in response to changes in turgor) of leaves (MPa).

LeafAF: Apoplastic fraction (proportion of water outside the living cells) in leaves.

Vleaf: Storage water capacity in leaves, per leaf area (L/m2).

StemPI0: Osmotic potential at full turgor of symplastic xylem tissue (MPa).

StemEPS: Modulus of elasticity (capacity of the cell wall to resist changes in volume in response to changes in turgor) of symplastic xylem tissue (Mpa).

StemAF: Apoplastic fraction (proportion of water outside the living cells) in stem xylem.

Vstem: Storage water capacity in sapwood, per leaf area (L/m2).

internalPhenology and internalWater: data frames to store internal state variables.

Functions forest2growthInput and growthInput return a list of class growthInput with the same elements as spwbInput, but with additional information.

Element above includes the following additional columns:

LA_live: Live leaf area per individual (m2/ind).

LA_dead: Dead leaf area per individual (m2/ind).

SA: Live sapwood area per individual (cm2/ind).

paramsGrowth: A data frame with growth parameters for each cohort:

WoodC: Wood carbon content per dry weight (g C /g dry).

RGRsapwoodmax: Maximum relative growth rate (in basal area or sapwood area) (in cm2·cm-2).

fHDmin: Minimum value of the height-to-diameter ratio (dimensionless).

fHDmax: Maximum value of the height-to-diameter ratio (dimensionless).

paramsAllometry: A data frame with allometric parameters for each cohort:

Aash: Regression coefficient relating the square of shrub height with shrub area.

Absh, Bbsh: Allometric coefficients relating phytovolume with dry weight of shrub individuals.

Acr, B1cr, B2cr, B3cr, C1cr, C2cr: Regression coefficients used to calculate crown ratio of trees.

Acw, Bcw: Regression coefficients used to calculated crown width of trees.

internalAllocation: A data frame with internal allocation variables for each cohort:

allocationTarget: Value of the allocation target variable.

leafAreaTarget: Target leaf area (m2) per individual.

fineRootBiomassTarget: Target fine root biomass (g dry) per individual (only if transpirationMode = "Sperry").

internalCarbon and internalRings: data structures to store other internal state variables.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

# Aboveground parameters
above = forest2aboveground(exampleforestMED, SpParamsMED)
above
#>         SP        N   DBH Cover   H        CR   LAI_live LAI_expanded LAI_dead
#> T1_148 148 168.0000 37.55    NA 800 0.6605196 0.96734365   0.96734365        0
#> T2_168 168 384.0000 14.60    NA 660 0.6055642 0.93111690   0.93111690        0
#> S1_165 165 749.4923    NA  3.75  80 0.8032817 0.03928201   0.03928201        0

# Initialize soil with default soil params
examplesoil = soil(defaultSoilParams())

# Rooting depths
Z50 = c(exampleforestMED$treeData$Z50, exampleforestMED$shrubData$Z50)
Z95 = c(exampleforestMED$treeData$Z95, exampleforestMED$shrubData$Z95)

# Initialize control parameters
control = defaultControl("Granier")

# Prepare spwb input
spwbInput(above, Z50, Z95, examplesoil,SpParamsMED, control)
#> $control
#> $control$modifyInput
#> [1] TRUE
#> 
#> $control$fillMissingSpParams
#> [1] TRUE
#> 
#> $control$verbose
#> [1] TRUE
#> 
#> $control$subdailyResults
#> [1] FALSE
#> 
#> $control$transpirationMode
#> [1] "Granier"
#> 
#> $control$soilFunctions
#> [1] "SX"
#> 
#> $control$defaultWindSpeed
#> [1] 2.5
#> 
#> $control$snowpack
#> [1] TRUE
#> 
#> $control$leafPhenology
#> [1] TRUE
#> 
#> $control$rockyLayerDrainage
#> [1] TRUE
#> 
#> $control$unlimitedSoilWater
#> [1] FALSE
#> 
#> $control$plantWaterPools
#> [1] FALSE
#> 
#> $control$unfoldingDD
#> [1] 300
#> 
#> $control$verticalLayerSize
#> [1] 100
#> 
#> $control$windMeasurementHeight
#> [1] 200
#> 
#> $control$cavitationRefill
#> [1] "total"
#> 
#> $control$ndailysteps
#> [1] 24
#> 
#> $control$nsubsteps
#> [1] 3600
#> 
#> $control$cochard
#> [1] FALSE
#> 
#> $control$capacitance
#> [1] FALSE
#> 
#> $control$taper
#> [1] TRUE
#> 
#> $control$multiLayerBalance
#> [1] FALSE
#> 
#> $control$gainModifier
#> [1] 1
#> 
#> $control$costModifier
#> [1] 1
#> 
#> $control$costWater
#> [1] "dEdP"
#> 
#> $control$klatstem
#> [1] 0.01
#> 
#> $control$klatleaf
#> [1] 0.01
#> 
#> $control$numericParams
#> $control$numericParams$maxNsteps
#> [1] 400
#> 
#> $control$numericParams$ntrial
#> [1] 200
#> 
#> $control$numericParams$psiTol
#> [1] 1e-04
#> 
#> $control$numericParams$ETol
#> [1] 1e-07
#> 
#> 
#> $control$fracLeafResistance
#> [1] NA
#> 
#> $control$fracRootResistance
#> [1] 0.4
#> 
#> $control$averageFracRhizosphereResistance
#> [1] 0.15
#> 
#> $control$Catm
#> [1] 386
#> 
#> $control$thermalCapacityLAI
#> [1] 1e+06
#> 
#> $control$boundaryLayerSize
#> [1] 2000
#> 
#> $control$refillMaximumRate
#> [1] 0.05
#> 
#> $control$allowDessication
#> [1] TRUE
#> 
#> $control$allowStarvation
#> [1] TRUE
#> 
#> $control$allowDefoliation
#> [1] TRUE
#> 
#> $control$sinkLimitation
#> [1] TRUE
#> 
#> $control$shrubDynamics
#> [1] FALSE
#> 
#> $control$allocationStrategy
#> [1] "Plant_kmax"
#> 
#> $control$nonStomatalPhotosynthesisLimitation
#> [1] TRUE
#> 
#> $control$phloemConductanceFactor
#> [1] 0.2
#> 
#> $control$nonSugarConcentration
#> [1] 0.25
#> 
#> $control$equilibriumOsmoticConcentration
#> $control$equilibriumOsmoticConcentration$leaf
#> [1] 0.8
#> 
#> $control$equilibriumOsmoticConcentration$sapwood
#> [1] 0.6
#> 
#> 
#> $control$minimumRelativeSugarForGrowth
#> [1] 0.5
#> 
#> $control$respirationRates
#> $control$respirationRates$leaf
#> [1] 0.00260274
#> 
#> $control$respirationRates$sapwood
#> [1] 6.849315e-05
#> 
#> $control$respirationRates$fineroot
#> [1] 0.002054795
#> 
#> 
#> $control$senescenceRates
#> $control$senescenceRates$sapwood
#> [1] 0.0001261398
#> 
#> $control$senescenceRates$fineroot
#> [1] 0.001897231
#> 
#> 
#> $control$constructionCosts
#> $control$constructionCosts$leaf
#> [1] 1.5
#> 
#> $control$constructionCosts$sapwood
#> [1] 1.47
#> 
#> $control$constructionCosts$fineroot
#> [1] 1.3
#> 
#> 
#> $control$maximumRelativeGrowthRates
#> $control$maximumRelativeGrowthRates$leaf
#> [1] 0.01
#> 
#> $control$maximumRelativeGrowthRates$sapwood
#> [1] 0.005
#> 
#> $control$maximumRelativeGrowthRates$fineroot
#> [1] 0.1
#> 
#> 
#> $control$mortalityMode
#> [1] "density/deterministic"
#> 
#> $control$mortalityBaselineRate
#> [1] 0.01
#> 
#> $control$mortalityRelativeSugarThreshold
#> [1] 0.3
#> 
#> $control$mortalityRWCThreshold
#> [1] 0.3
#> 
#> $control$recruitmentMode
#> [1] "deterministic"
#> 
#> $control$removeEmptyCohorts
#> [1] TRUE
#> 
#> $control$minimumCohortDensity
#> [1] 1
#> 
#> $control$seedRain
#> NULL
#> 
#> $control$seedProductionTreeHeight
#> [1] 300
#> 
#> $control$seedProductionShrubHeight
#> [1] 30
#> 
#> $control$minTempRecr
#> [1] 0
#> 
#> $control$minMoistureRecr
#> [1] 0.3
#> 
#> $control$minFPARRecr
#> [1] 10
#> 
#> $control$recrTreeDBH
#> [1] 1
#> 
#> $control$recrTreeDensity
#> [1] 100
#> 
#> $control$recrTreeHeight
#> [1] 100
#> 
#> $control$recrShrubCover
#> [1] 1
#> 
#> $control$recrShrubHeight
#> [1] 10
#> 
#> $control$recrTreeZ50
#> [1] 100
#> 
#> $control$recrShrubZ50
#> [1] 50
#> 
#> $control$recrTreeZ95
#> [1] 1000
#> 
#> $control$recrShrubZ95
#> [1] 500
#> 
#> 
#> $soil
#> Soil depth (mm): 4000 
#> 
#> Layer  1  [ 0  to  300 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 25 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 32 Vol. FC (mm): 68 Vol. SAT (mm): 111 Vol. current (mm): 68 
#>     Temperature (Celsius): NA 
#> 
#> Layer  2  [ 300  to  1000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 45 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 55 Vol. FC (mm): 117 Vol. SAT (mm): 190 Vol. current (mm): 117 
#>     Temperature (Celsius): NA 
#> 
#> Layer  3  [ 1000  to  2000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 75 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 36 Vol. FC (mm): 76 Vol. SAT (mm): 123 Vol. current (mm): 76 
#>     Temperature (Celsius): NA 
#> 
#> Layer  4  [ 2000  to  4000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 95 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 14 Vol. FC (mm): 30 Vol. SAT (mm): 49 Vol. current (mm): 30 
#>     Temperature (Celsius): NA 
#> 
#> Total soil saturated capacity (mm): 473 
#> Total soil water holding capacity (mm): 291 
#> Total soil extractable water (mm): 183 
#> Total soil current Volume (mm): 291 
#> 
#> Snow pack water equivalent (mm): 0 
#> Soil water table depth (mm): 4000 
#> 
#> 
#> $canopy
#> list()
#> 
#> $cohorts
#>         SP              Name
#> T1_148 148  Pinus halepensis
#> T2_168 168      Quercus ilex
#> S1_165 165 Quercus coccifera
#> 
#> $above
#>          H        CR        N   LAI_live LAI_expanded LAI_dead
#> T1_148 800 0.6605196 168.0000 0.96734365   0.96734365        0
#> T2_168 660 0.6055642 384.0000 0.93111690   0.93111690        0
#> S1_165  80 0.8032817 749.4923 0.03928201   0.03928201        0
#> 
#> $below
#>        Z50  Z95 fineRootBiomass coarseRootSoilVolume
#> T1_148 750 3000              NA            16.359912
#> T2_168 750 3000              NA            16.359912
#> S1_165 300 1500              NA             7.473966
#> 
#> $belowLayers
#> $belowLayers$V
#>                1         2          3          4
#> T1_148 0.1288916 0.5376838 0.24791124 0.08551338
#> T2_168 0.1288916 0.5376838 0.24791124 0.08551338
#> S1_165 0.5044057 0.4037153 0.07011387 0.02176513
#> 
#> $belowLayers$L
#>               1        2        3        4
#> T1_148 1645.778 2650.000 2636.223 3471.865
#> T2_168 1645.778 2650.000 2636.223 3471.865
#> S1_165 2150.000 1821.357 1908.416 3160.904
#> 
#> $belowLayers$Wpool
#>        1 2 3 4
#> T1_148 1 1 1 1
#> T2_168 1 1 1 1
#> S1_165 1 1 1 1
#> 
#> 
#> $paramsPhenology
#>             PhenologyType LeafDuration t0gdd  Sgdd Tbgdd  Ssen Phsen Tbsen xsen
#> T1_148 oneflush-evergreen     2.536875  50.0 200.0  0.00  8268  12.5  28.5    2
#> T2_168 oneflush-evergreen     2.145402  54.5 240.7  4.34 10178  12.5  28.5    2
#> S1_165 oneflush-evergreen     1.227917  54.5 240.7  4.34 10178  12.5  28.5    2
#>        ysen
#> T1_148    2
#> T2_168    2
#> S1_165    2
#> 
#> $paramsInterception
#>        kPAR   g
#> T1_148 0.50 1.0
#> T2_168 0.55 0.5
#> S1_165 0.55 0.5
#> 
#> $paramsTranspiration
#>        Tmax_LAI Tmax_LAIsq Psi_Extract Psi_Critic      WUE pRootDisc
#> T1_148    0.134     -0.006   -2.600000  -3.816852 3.613308         0
#> T2_168    0.134     -0.006   -3.001667  -3.001667 1.475715         0
#> S1_165    0.134     -0.006   -3.455000  -6.950920 2.433101         0
#> 
#> $internalPhenology
#>        gdd sen budFormation leafUnfolding leafSenescence leafDormancy phi
#> T1_148   0   0        FALSE         FALSE          FALSE        FALSE   0
#> T2_168   0   0        FALSE         FALSE          FALSE        FALSE   0
#> S1_165   0   0        FALSE         FALSE          FALSE        FALSE   0
#> 
#> $internalWater
#>        PlantPsi StemPLC
#> T1_148        0       0
#> T2_168        0       0
#> S1_165        0       0
#> 
#> attr(,"class")
#> [1] "spwbInput" "list"     

# When starting from an object of class 'forest' the whole process
# can be simplified:
forest2spwbInput(exampleforestMED, examplesoil, SpParamsMED, control)
#> $control
#> $control$modifyInput
#> [1] TRUE
#> 
#> $control$fillMissingSpParams
#> [1] TRUE
#> 
#> $control$verbose
#> [1] TRUE
#> 
#> $control$subdailyResults
#> [1] FALSE
#> 
#> $control$transpirationMode
#> [1] "Granier"
#> 
#> $control$soilFunctions
#> [1] "SX"
#> 
#> $control$defaultWindSpeed
#> [1] 2.5
#> 
#> $control$snowpack
#> [1] TRUE
#> 
#> $control$leafPhenology
#> [1] TRUE
#> 
#> $control$rockyLayerDrainage
#> [1] TRUE
#> 
#> $control$unlimitedSoilWater
#> [1] FALSE
#> 
#> $control$plantWaterPools
#> [1] FALSE
#> 
#> $control$unfoldingDD
#> [1] 300
#> 
#> $control$verticalLayerSize
#> [1] 100
#> 
#> $control$windMeasurementHeight
#> [1] 200
#> 
#> $control$cavitationRefill
#> [1] "total"
#> 
#> $control$ndailysteps
#> [1] 24
#> 
#> $control$nsubsteps
#> [1] 3600
#> 
#> $control$cochard
#> [1] FALSE
#> 
#> $control$capacitance
#> [1] FALSE
#> 
#> $control$taper
#> [1] TRUE
#> 
#> $control$multiLayerBalance
#> [1] FALSE
#> 
#> $control$gainModifier
#> [1] 1
#> 
#> $control$costModifier
#> [1] 1
#> 
#> $control$costWater
#> [1] "dEdP"
#> 
#> $control$klatstem
#> [1] 0.01
#> 
#> $control$klatleaf
#> [1] 0.01
#> 
#> $control$numericParams
#> $control$numericParams$maxNsteps
#> [1] 400
#> 
#> $control$numericParams$ntrial
#> [1] 200
#> 
#> $control$numericParams$psiTol
#> [1] 1e-04
#> 
#> $control$numericParams$ETol
#> [1] 1e-07
#> 
#> 
#> $control$fracLeafResistance
#> [1] NA
#> 
#> $control$fracRootResistance
#> [1] 0.4
#> 
#> $control$averageFracRhizosphereResistance
#> [1] 0.15
#> 
#> $control$Catm
#> [1] 386
#> 
#> $control$thermalCapacityLAI
#> [1] 1e+06
#> 
#> $control$boundaryLayerSize
#> [1] 2000
#> 
#> $control$refillMaximumRate
#> [1] 0.05
#> 
#> $control$allowDessication
#> [1] TRUE
#> 
#> $control$allowStarvation
#> [1] TRUE
#> 
#> $control$allowDefoliation
#> [1] TRUE
#> 
#> $control$sinkLimitation
#> [1] TRUE
#> 
#> $control$shrubDynamics
#> [1] FALSE
#> 
#> $control$allocationStrategy
#> [1] "Plant_kmax"
#> 
#> $control$nonStomatalPhotosynthesisLimitation
#> [1] TRUE
#> 
#> $control$phloemConductanceFactor
#> [1] 0.2
#> 
#> $control$nonSugarConcentration
#> [1] 0.25
#> 
#> $control$equilibriumOsmoticConcentration
#> $control$equilibriumOsmoticConcentration$leaf
#> [1] 0.8
#> 
#> $control$equilibriumOsmoticConcentration$sapwood
#> [1] 0.6
#> 
#> 
#> $control$minimumRelativeSugarForGrowth
#> [1] 0.5
#> 
#> $control$respirationRates
#> $control$respirationRates$leaf
#> [1] 0.00260274
#> 
#> $control$respirationRates$sapwood
#> [1] 6.849315e-05
#> 
#> $control$respirationRates$fineroot
#> [1] 0.002054795
#> 
#> 
#> $control$senescenceRates
#> $control$senescenceRates$sapwood
#> [1] 0.0001261398
#> 
#> $control$senescenceRates$fineroot
#> [1] 0.001897231
#> 
#> 
#> $control$constructionCosts
#> $control$constructionCosts$leaf
#> [1] 1.5
#> 
#> $control$constructionCosts$sapwood
#> [1] 1.47
#> 
#> $control$constructionCosts$fineroot
#> [1] 1.3
#> 
#> 
#> $control$maximumRelativeGrowthRates
#> $control$maximumRelativeGrowthRates$leaf
#> [1] 0.01
#> 
#> $control$maximumRelativeGrowthRates$sapwood
#> [1] 0.005
#> 
#> $control$maximumRelativeGrowthRates$fineroot
#> [1] 0.1
#> 
#> 
#> $control$mortalityMode
#> [1] "density/deterministic"
#> 
#> $control$mortalityBaselineRate
#> [1] 0.01
#> 
#> $control$mortalityRelativeSugarThreshold
#> [1] 0.3
#> 
#> $control$mortalityRWCThreshold
#> [1] 0.3
#> 
#> $control$recruitmentMode
#> [1] "deterministic"
#> 
#> $control$removeEmptyCohorts
#> [1] TRUE
#> 
#> $control$minimumCohortDensity
#> [1] 1
#> 
#> $control$seedRain
#> NULL
#> 
#> $control$seedProductionTreeHeight
#> [1] 300
#> 
#> $control$seedProductionShrubHeight
#> [1] 30
#> 
#> $control$minTempRecr
#> [1] 0
#> 
#> $control$minMoistureRecr
#> [1] 0.3
#> 
#> $control$minFPARRecr
#> [1] 10
#> 
#> $control$recrTreeDBH
#> [1] 1
#> 
#> $control$recrTreeDensity
#> [1] 100
#> 
#> $control$recrTreeHeight
#> [1] 100
#> 
#> $control$recrShrubCover
#> [1] 1
#> 
#> $control$recrShrubHeight
#> [1] 10
#> 
#> $control$recrTreeZ50
#> [1] 100
#> 
#> $control$recrShrubZ50
#> [1] 50
#> 
#> $control$recrTreeZ95
#> [1] 1000
#> 
#> $control$recrShrubZ95
#> [1] 500
#> 
#> 
#> $soil
#> Soil depth (mm): 4000 
#> 
#> Layer  1  [ 0  to  300 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 25 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 32 Vol. FC (mm): 68 Vol. SAT (mm): 111 Vol. current (mm): 68 
#>     Temperature (Celsius): NA 
#> 
#> Layer  2  [ 300  to  1000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 45 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 55 Vol. FC (mm): 117 Vol. SAT (mm): 190 Vol. current (mm): 117 
#>     Temperature (Celsius): NA 
#> 
#> Layer  3  [ 1000  to  2000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 75 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 36 Vol. FC (mm): 76 Vol. SAT (mm): 123 Vol. current (mm): 76 
#>     Temperature (Celsius): NA 
#> 
#> Layer  4  [ 2000  to  4000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 95 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 14 Vol. FC (mm): 30 Vol. SAT (mm): 49 Vol. current (mm): 30 
#>     Temperature (Celsius): NA 
#> 
#> Total soil saturated capacity (mm): 473 
#> Total soil water holding capacity (mm): 291 
#> Total soil extractable water (mm): 183 
#> Total soil current Volume (mm): 291 
#> 
#> Snow pack water equivalent (mm): 0 
#> Soil water table depth (mm): 4000 
#> 
#> 
#> $canopy
#> list()
#> 
#> $cohorts
#>         SP              Name
#> T1_148 148  Pinus halepensis
#> T2_168 168      Quercus ilex
#> S1_165 165 Quercus coccifera
#> 
#> $above
#>          H        CR        N   LAI_live LAI_expanded LAI_dead
#> T1_148 800 0.6605196 168.0000 0.96734365   0.96734365        0
#> T2_168 660 0.6055642 384.0000 0.93111690   0.93111690        0
#> S1_165  80 0.8032817 749.4923 0.03928201   0.03928201        0
#> 
#> $below
#>        Z50  Z95 fineRootBiomass coarseRootSoilVolume
#> T1_148 750 3000              NA            16.359912
#> T2_168 750 3000              NA            16.359912
#> S1_165 300 1500              NA             7.473966
#> 
#> $belowLayers
#> $belowLayers$V
#>                1         2          3          4
#> T1_148 0.1288916 0.5376838 0.24791124 0.08551338
#> T2_168 0.1288916 0.5376838 0.24791124 0.08551338
#> S1_165 0.5044057 0.4037153 0.07011387 0.02176513
#> 
#> $belowLayers$L
#>               1        2        3        4
#> T1_148 1645.778 2650.000 2636.223 3471.865
#> T2_168 1645.778 2650.000 2636.223 3471.865
#> S1_165 2150.000 1821.357 1908.416 3160.904
#> 
#> $belowLayers$Wpool
#>        1 2 3 4
#> T1_148 1 1 1 1
#> T2_168 1 1 1 1
#> S1_165 1 1 1 1
#> 
#> 
#> $paramsPhenology
#>             PhenologyType LeafDuration t0gdd  Sgdd Tbgdd  Ssen Phsen Tbsen xsen
#> T1_148 oneflush-evergreen     2.536875  50.0 200.0  0.00  8268  12.5  28.5    2
#> T2_168 oneflush-evergreen     2.145402  54.5 240.7  4.34 10178  12.5  28.5    2
#> S1_165 oneflush-evergreen     1.227917  54.5 240.7  4.34 10178  12.5  28.5    2
#>        ysen
#> T1_148    2
#> T2_168    2
#> S1_165    2
#> 
#> $paramsInterception
#>        kPAR   g
#> T1_148 0.50 1.0
#> T2_168 0.55 0.5
#> S1_165 0.55 0.5
#> 
#> $paramsTranspiration
#>        Tmax_LAI Tmax_LAIsq Psi_Extract Psi_Critic      WUE pRootDisc
#> T1_148    0.134     -0.006   -2.600000  -3.816852 3.613308         0
#> T2_168    0.134     -0.006   -3.001667  -3.001667 1.475715         0
#> S1_165    0.134     -0.006   -3.455000  -6.950920 2.433101         0
#> 
#> $internalPhenology
#>        gdd sen budFormation leafUnfolding leafSenescence leafDormancy phi
#> T1_148   0   0        FALSE         FALSE          FALSE        FALSE   0
#> T2_168   0   0        FALSE         FALSE          FALSE        FALSE   0
#> S1_165   0   0        FALSE         FALSE          FALSE        FALSE   0
#> 
#> $internalWater
#>        PlantPsi StemPLC
#> T1_148        0       0
#> T2_168        0       0
#> S1_165        0       0
#> 
#> attr(,"class")
#> [1] "spwbInput" "list"     

# Prepare input for Sperry transpiration mode
control = defaultControl("Sperry")
forest2spwbInput(exampleforestMED,examplesoil,SpParamsMED, control)
#> $control
#> $control$modifyInput
#> [1] TRUE
#> 
#> $control$fillMissingSpParams
#> [1] TRUE
#> 
#> $control$verbose
#> [1] TRUE
#> 
#> $control$subdailyResults
#> [1] FALSE
#> 
#> $control$transpirationMode
#> [1] "Sperry"
#> 
#> $control$soilFunctions
#> [1] "VG"
#> 
#> $control$defaultWindSpeed
#> [1] 2.5
#> 
#> $control$snowpack
#> [1] TRUE
#> 
#> $control$leafPhenology
#> [1] TRUE
#> 
#> $control$rockyLayerDrainage
#> [1] TRUE
#> 
#> $control$unlimitedSoilWater
#> [1] FALSE
#> 
#> $control$plantWaterPools
#> [1] FALSE
#> 
#> $control$unfoldingDD
#> [1] 300
#> 
#> $control$verticalLayerSize
#> [1] 100
#> 
#> $control$windMeasurementHeight
#> [1] 200
#> 
#> $control$cavitationRefill
#> [1] "total"
#> 
#> $control$ndailysteps
#> [1] 24
#> 
#> $control$nsubsteps
#> [1] 3600
#> 
#> $control$cochard
#> [1] FALSE
#> 
#> $control$capacitance
#> [1] FALSE
#> 
#> $control$taper
#> [1] TRUE
#> 
#> $control$multiLayerBalance
#> [1] FALSE
#> 
#> $control$gainModifier
#> [1] 1
#> 
#> $control$costModifier
#> [1] 1
#> 
#> $control$costWater
#> [1] "dEdP"
#> 
#> $control$klatstem
#> [1] 0.01
#> 
#> $control$klatleaf
#> [1] 0.01
#> 
#> $control$numericParams
#> $control$numericParams$maxNsteps
#> [1] 400
#> 
#> $control$numericParams$ntrial
#> [1] 200
#> 
#> $control$numericParams$psiTol
#> [1] 1e-04
#> 
#> $control$numericParams$ETol
#> [1] 1e-07
#> 
#> 
#> $control$fracLeafResistance
#> [1] NA
#> 
#> $control$fracRootResistance
#> [1] 0.4
#> 
#> $control$averageFracRhizosphereResistance
#> [1] 0.15
#> 
#> $control$Catm
#> [1] 386
#> 
#> $control$thermalCapacityLAI
#> [1] 1e+06
#> 
#> $control$boundaryLayerSize
#> [1] 2000
#> 
#> $control$refillMaximumRate
#> [1] 0.05
#> 
#> $control$allowDessication
#> [1] TRUE
#> 
#> $control$allowStarvation
#> [1] TRUE
#> 
#> $control$allowDefoliation
#> [1] TRUE
#> 
#> $control$sinkLimitation
#> [1] TRUE
#> 
#> $control$shrubDynamics
#> [1] FALSE
#> 
#> $control$allocationStrategy
#> [1] "Plant_kmax"
#> 
#> $control$nonStomatalPhotosynthesisLimitation
#> [1] TRUE
#> 
#> $control$phloemConductanceFactor
#> [1] 0.2
#> 
#> $control$nonSugarConcentration
#> [1] 0.25
#> 
#> $control$equilibriumOsmoticConcentration
#> $control$equilibriumOsmoticConcentration$leaf
#> [1] 0.8
#> 
#> $control$equilibriumOsmoticConcentration$sapwood
#> [1] 0.6
#> 
#> 
#> $control$minimumRelativeSugarForGrowth
#> [1] 0.5
#> 
#> $control$respirationRates
#> $control$respirationRates$leaf
#> [1] 0.00260274
#> 
#> $control$respirationRates$sapwood
#> [1] 6.849315e-05
#> 
#> $control$respirationRates$fineroot
#> [1] 0.002054795
#> 
#> 
#> $control$senescenceRates
#> $control$senescenceRates$sapwood
#> [1] 0.0001261398
#> 
#> $control$senescenceRates$fineroot
#> [1] 0.001897231
#> 
#> 
#> $control$constructionCosts
#> $control$constructionCosts$leaf
#> [1] 1.5
#> 
#> $control$constructionCosts$sapwood
#> [1] 1.47
#> 
#> $control$constructionCosts$fineroot
#> [1] 1.3
#> 
#> 
#> $control$maximumRelativeGrowthRates
#> $control$maximumRelativeGrowthRates$leaf
#> [1] 0.01
#> 
#> $control$maximumRelativeGrowthRates$sapwood
#> [1] 0.005
#> 
#> $control$maximumRelativeGrowthRates$fineroot
#> [1] 0.1
#> 
#> 
#> $control$mortalityMode
#> [1] "density/deterministic"
#> 
#> $control$mortalityBaselineRate
#> [1] 0.01
#> 
#> $control$mortalityRelativeSugarThreshold
#> [1] 0.3
#> 
#> $control$mortalityRWCThreshold
#> [1] 0.3
#> 
#> $control$recruitmentMode
#> [1] "deterministic"
#> 
#> $control$removeEmptyCohorts
#> [1] TRUE
#> 
#> $control$minimumCohortDensity
#> [1] 1
#> 
#> $control$seedRain
#> NULL
#> 
#> $control$seedProductionTreeHeight
#> [1] 300
#> 
#> $control$seedProductionShrubHeight
#> [1] 30
#> 
#> $control$minTempRecr
#> [1] 0
#> 
#> $control$minMoistureRecr
#> [1] 0.3
#> 
#> $control$minFPARRecr
#> [1] 10
#> 
#> $control$recrTreeDBH
#> [1] 1
#> 
#> $control$recrTreeDensity
#> [1] 100
#> 
#> $control$recrTreeHeight
#> [1] 100
#> 
#> $control$recrShrubCover
#> [1] 1
#> 
#> $control$recrShrubHeight
#> [1] 10
#> 
#> $control$recrTreeZ50
#> [1] 100
#> 
#> $control$recrShrubZ50
#> [1] 50
#> 
#> $control$recrTreeZ95
#> [1] 1000
#> 
#> $control$recrShrubZ95
#> [1] 500
#> 
#> 
#> $soil
#> Soil depth (mm): 4000 
#> 
#> Layer  1  [ 0  to  300 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 25 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 32 Vol. FC (mm): 68 Vol. SAT (mm): 111 Vol. current (mm): 68 
#>     Temperature (Celsius): NA 
#> 
#> Layer  2  [ 300  to  1000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 45 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 55 Vol. FC (mm): 117 Vol. SAT (mm): 190 Vol. current (mm): 117 
#>     Temperature (Celsius): NA 
#> 
#> Layer  3  [ 1000  to  2000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 75 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 36 Vol. FC (mm): 76 Vol. SAT (mm): 123 Vol. current (mm): 76 
#>     Temperature (Celsius): NA 
#> 
#> Layer  4  [ 2000  to  4000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 95 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 14 Vol. FC (mm): 30 Vol. SAT (mm): 49 Vol. current (mm): 30 
#>     Temperature (Celsius): NA 
#> 
#> Total soil saturated capacity (mm): 473 
#> Total soil water holding capacity (mm): 291 
#> Total soil extractable water (mm): 183 
#> Total soil current Volume (mm): 291 
#> 
#> Snow pack water equivalent (mm): 0 
#> Soil water table depth (mm): 4000 
#> 
#> 
#> $canopy
#>    zlow zmid  zup Tair Cair VPair
#> 1     0   50  100   NA   NA    NA
#> 2   100  150  200   NA   NA    NA
#> 3   200  250  300   NA   NA    NA
#> 4   300  350  400   NA   NA    NA
#> 5   400  450  500   NA   NA    NA
#> 6   500  550  600   NA   NA    NA
#> 7   600  650  700   NA   NA    NA
#> 8   700  750  800   NA   NA    NA
#> 9   800  850  900   NA   NA    NA
#> 10  900  950 1000   NA   NA    NA
#> 11 1000 1050 1100   NA   NA    NA
#> 12 1100 1150 1200   NA   NA    NA
#> 13 1200 1250 1300   NA   NA    NA
#> 14 1300 1350 1400   NA   NA    NA
#> 15 1400 1450 1500   NA   NA    NA
#> 16 1500 1550 1600   NA   NA    NA
#> 17 1600 1650 1700   NA   NA    NA
#> 18 1700 1750 1800   NA   NA    NA
#> 19 1800 1850 1900   NA   NA    NA
#> 20 1900 1950 2000   NA   NA    NA
#> 21 2000 2050 2100   NA   NA    NA
#> 22 2100 2150 2200   NA   NA    NA
#> 23 2200 2250 2300   NA   NA    NA
#> 24 2300 2350 2400   NA   NA    NA
#> 25 2400 2450 2500   NA   NA    NA
#> 26 2500 2550 2600   NA   NA    NA
#> 27 2600 2650 2700   NA   NA    NA
#> 28 2700 2750 2800   NA   NA    NA
#> 
#> $cohorts
#>         SP              Name
#> T1_148 148  Pinus halepensis
#> T2_168 168      Quercus ilex
#> S1_165 165 Quercus coccifera
#> 
#> $above
#>          H        CR        N   LAI_live LAI_expanded LAI_dead
#> T1_148 800 0.6605196 168.0000 0.96734365   0.96734365        0
#> T2_168 660 0.6055642 384.0000 0.93111690   0.93111690        0
#> S1_165  80 0.8032817 749.4923 0.03928201   0.03928201        0
#> 
#> $below
#>        Z50  Z95 fineRootBiomass coarseRootSoilVolume
#> T1_148 750 3000        7328.012               19.000
#> T2_168 750 3000        2955.316               21.000
#> S1_165 300 1500         202.888                0.125
#> 
#> $belowLayers
#> $belowLayers$V
#>                1         2          3          4
#> T1_148 0.1288916 0.5376838 0.24791124 0.08551338
#> T2_168 0.1288916 0.5376838 0.24791124 0.08551338
#> S1_165 0.5044057 0.4037153 0.07011387 0.02176513
#> 
#> $belowLayers$L
#>                1        2        3        4
#> T1_148 2161.3279 3906.263 4448.950 6274.149
#> T2_168 2256.8419 4055.398 4574.618 6390.847
#> S1_165  598.6612 1204.262 2105.636 4093.059
#> 
#> $belowLayers$VGrhizo_kmax
#>                1        2        3       4
#> T1_148  33998988 50571077 23316936 8042838
#> T2_168  35078232 52998034 24435939 8428822
#> S1_165 299837332 86916167 15094867 4685831
#> 
#> $belowLayers$VCroot_kmax
#>                1         2         3          4
#> T1_148 0.3786338 0.8739386 0.3537970 0.08653558
#> T2_168 0.9764309 2.2667914 0.9265300 0.22876720
#> S1_165 3.2284705 1.2845535 0.1275906 0.02037563
#> 
#> $belowLayers$Wpool
#>        1 2 3 4
#> T1_148 1 1 1 1
#> T2_168 1 1 1 1
#> S1_165 1 1 1 1
#> 
#> $belowLayers$RhizoPsi
#>        1 2 3 4
#> T1_148 0 0 0 0
#> T2_168 0 0 0 0
#> S1_165 0 0 0 0
#> 
#> 
#> $paramsPhenology
#>             PhenologyType LeafDuration t0gdd  Sgdd Tbgdd  Ssen Phsen Tbsen xsen
#> T1_148 oneflush-evergreen     2.536875  50.0 200.0  0.00  8268  12.5  28.5    2
#> T2_168 oneflush-evergreen     2.145402  54.5 240.7  4.34 10178  12.5  28.5    2
#> S1_165 oneflush-evergreen     1.227917  54.5 240.7  4.34 10178  12.5  28.5    2
#>        ysen
#> T1_148    2
#> T2_168    2
#> S1_165    2
#> 
#> $paramsAnatomy
#>        Hmax Hmed    Al2As      SLA LeafWidth LeafDensity WoodDensity
#> T1_148 1900  850 1317.523 5.140523  0.700000   0.2982842   0.6077950
#> T2_168 1300  500 2512.563 6.850951  1.767436   0.5200298   0.9013662
#> S1_165  250   80 4189.325 4.980084  1.376108   0.3871262   0.4389106
#>        FineRootDensity conduit2sapwood      SRL RLD     r635
#> T1_148       0.2982842       0.9236406 3172.572  10 1.964226
#> T2_168       0.5200298       0.6238125 4066.803  10 1.805872
#> S1_165       0.3871262       0.6238125 4066.803  10 2.289452
#> 
#> $paramsInterception
#>        alphaSWR gammaSWR kPAR   g
#> T1_148      0.7     0.14 0.50 1.0
#> T2_168      0.7     0.18 0.55 0.5
#> S1_165      0.7     0.18 0.55 0.5
#> 
#> $paramsTranspiration
#>             Gswmin    Gswmax  Vmax298  Jmax298 Kmax_stemxylem Kmax_rootxylem
#> T1_148 0.002031250 0.1940833 62.81754 128.9875      0.1500000       0.600000
#> T2_168 0.006003126 0.2007222 41.01518 103.0690      0.7735834       3.094334
#> S1_165 0.010455247 0.2830167 63.91803 125.5370      0.2900000       1.160000
#>        VCleaf_kmax VCleaf_c  VCleaf_d VCstem_kmax VCstem_c  VCstem_d
#> T1_148    7.166471 1.497782 -2.142745    1.339563 7.458883 -4.481392
#> T2_168    7.354306 1.331076 -1.777291    4.876881 2.558358 -5.103492
#> S1_165    9.579077 1.844224 -3.030130    4.599269 3.537784 -4.126512
#>        VCroot_kmax VCroot_c  VCroot_d VGrhizo_kmax Plant_kmax
#> T1_148    1.692905 1.916254 -3.240714    115929839   0.677162
#> T2_168    4.398520 1.957824 -3.366992    120941026   1.759408
#> S1_165    4.660990 1.760307 -2.797092    406534197   1.864396
#> 
#> $paramsWaterStorage
#>          LeafPI0   LeafEPS LeafAF     Vleaf   StemPI0   StemEPS    StemAF
#> T1_148 -1.591429  8.918571 0.3525 0.5258525 -2.008427 13.261704 0.9236406
#> T2_168 -1.483333 19.260000 0.1700 0.1859035 -3.229683 46.527365 0.6238125
#> S1_165 -2.370000 17.230000 0.2400 0.3883039 -1.305868  6.297155 0.6238125
#>         Vsapwood
#> T1_148 3.6755513
#> T2_168 1.0893268
#> S1_165 0.1365362
#> 
#> $internalPhenology
#>        gdd sen budFormation leafUnfolding leafSenescence leafDormancy phi
#> T1_148   0   0        FALSE         FALSE          FALSE        FALSE   0
#> T2_168   0   0        FALSE         FALSE          FALSE        FALSE   0
#> S1_165   0   0        FALSE         FALSE          FALSE        FALSE   0
#> 
#> $internalWater
#>        Einst RootCrownPsi Stem1Psi Stem2Psi LeafPsi StemSympPsi LeafSympPsi
#> T1_148     0            0        0        0       0           0           0
#> T2_168     0            0        0        0       0           0           0
#> S1_165     0            0        0        0       0           0           0
#>        StemPLC NSPL
#> T1_148       0    1
#> T2_168       0    1
#> S1_165       0    1
#> 
#> attr(,"class")
#> [1] "spwbInput" "list"