Input for simulation models

Functions forest2spwbInput and forest2growthInput take an object of class forest and calculate input data for functions spwb, pwb and growth, respectively. Functions spwbInput and growthInput do the same but starting from different input data. Function forest2aboveground calculates aboveground variables that may be used in spwbInput and growthInput functions. Function forest2belowground calculates belowground fine root distribution.

forest2aboveground(x, SpParams, gdd = NA, mode = "MED")
forest2belowground(x, soil)
forest2growthInput(x, soil, SpParams, control)
forest2spwbInput(x, soil, SpParams, control, mode = "MED")
growthInput(above,  Z50, Z95, soil, SpParams, control)
spwbInput(above,  Z50, Z95, soil, SpParams, control)

Arguments

x
An object of class forest.

SpParams
A data frame with species parameters (see SpParamsMED and SpParamsMED).

gdd
Growth degree days to account for leaf phenology effects (in Celsius). This should be left NA in most applications.

mode
Calculation mode, either "MED" or "US".

soil
An object of class soil.

control
A list with default control parameters (see defaultControl).

above
A data frame with aboveground plant information (see the return value of forest2aboveground below). In the case of spwbInput the variables should include SP, N, LAI_live, LAI_dead, H and CR. In the case of growthInput variables should include DBH and Cover.

Z50, Z95
Numeric vectors with cohort depths (in mm) corresponding to 50% and 95% of fine roots.

Details

Functions forest2spwbInput and forest2abovegroundInput extracts height and species identity from plant cohorts of x, and calculate leaf area index and crown ratio.forest2spwbInput also calculates the distribution of fine roots across soil. Both forest2spwbInput and spwbInput find parameter values for each plant cohort according to the parameters of its species as specified in SpParams. If control$transpirationMode = "Sperry" the functions also estimate the maximum conductance of rhizosphere, root xylem and stem xylem elements.

Value

Function forest2aboveground() returns a data frame with the following columns (rows are identified as specified by function plant_ID):

SP: Species identity (an integer) (first species is 0).

N: Cohort density (ind/ha) (see function plant_density).

DBH: Tree diameter at breast height (cm).

H: Plant total height (cm).

CR: Crown ratio (crown length to total height) (between 0 and 1).

LAI_live: Live leaf area index (m2/m2) (one-side leaf area relative to plot area), includes leaves in winter dormant buds.

LAI_expanded: Leaf area index of expanded leaves (m2/m2) (one-side leaf area relative to plot area).

LAI_dead: Dead leaf area index (m2/m2) (one-side leaf area relative to plot area).

Functions forest2spwbInput() and spwbInput() return a list of class spwbInput with the following elements (rows of data frames are identified as specified by function plant_ID):

control: List with control parameters (see defaultControl).

canopy: A list of stand-level state variables.

cohorts: A data frame with cohort information, with columns SP and Name.

above: A data frame with columns H, CR and LAI (see function forest2aboveground).

below: A data frame with columns Z50, Z95. If control$transpirationMode = "Sperry" additional columns are fineRootBiomass and coarseRootSoilVolume.

belowLayers: A list. If control$transpirationMode = "Granier" it contains elements:

V: A matrix with the proportion of fine roots of each cohort (in rows) in each soil layer (in columns).

L: A matrix with the length of coarse roots of each cohort (in rows) in each soil layer (in columns).

Wpool: A matrix with the soil moisture relative to field capacity around the rhizosphere of each cohort (in rows) in each soil layer (in columns).

If control$transpirationMode = "Sperry" there are the following additional elements:

VGrhizo_kmax: A matrix with maximum rhizosphere conductance values of each cohort (in rows) in each soil layer (in columns).

VGroot_kmax: A matrix with maximum root xylem conductance values of each cohort (in rows) in each soil layer (in columns).

RhizoPsi: A matrix with the water potential around the rhizosphere of each cohort (in rows) in each soil layer (in columns).

paramsPhenology: A data frame with leaf phenology parameters:

PhenologyType: Leaf phenology type.

LeafDuration: Leaf duration (in years).

Sgdd: Degree days needed for leaf budburst (for winter decideous species).

Tbgdd: Base temperature for the calculation of degree days to leaf budburst.

Ssen: Degree days corresponding to leaf senescence.

Phsen: Photoperiod corresponding to start counting senescence degree-days.

Tbsen: Base temperature for the calculation of degree days to leaf senescence.

paramsAnatomy: A data frame with plant anatomy parameters for each cohort (only if control$transpirationMode = "Sperry"):

Hmax: Maximum plant height (cm).

Hmed: Median plant height (cm).

Al2As: Leaf area to sapwood area ratio (in m2·m-2).

SLA: Specific leaf area (mm2/mg = m2/kg).

LeafWidth: Leaf width (in cm).

LeafDensity: Density of leaf tissue (dry weight over volume).

WoodDensity: Density of wood tissue (dry weight over volume).

FineRootDensity: Density of fine root tissue (dry weight over volume).

SRL: Specific Root length (cm·g-1).

RLD: Root length density (cm·cm-3).

r635: Ratio between the weight of leaves plus branches and the weight of leaves alone for branches of 6.35 mm.

paramsInterception: A data frame with rain interception and light extinction parameters for each cohort:

kPAR: PAR extinction coefficient.

g: Canopy water retention capacity per LAI unit (mm/LAI).

If control$transpirationMode = "Sperry" additional columns are:

gammaSWR: Reflectance (albedo) coefficient for SWR .

alphaSWR: Absorbance coefficient for SWR .

paramsTranspiration: A data frame with parameters for transpiration and photosynthesis. If control$transpirationMode = "Granier", columns are:

Tmax_LAI: Coefficient relating LAI with the ratio of maximum transpiration over potential evapotranspiration.

Tmax_LAIsq: Coefficient relating squared LAI with the ratio of maximum transpiration over potential evapotranspiration.

Psi_Extract: Water potential corresponding to 50% relative conductance (in MPa).

Psi_Critic: Water potential corresponding to 50% of stem cavitation (in MPa).

WUE: Water use efficiency for carbon assimilation (g C/mm water).

pRootDisc: Proportion of whole-plant conductance leading to disconnection from soil.

If control$transpirationMode = "Sperry" columns are:

Gswmax: Maximum stomatal conductance to water vapor (in mol H2O·m-2·s-1).

Vmax298: Maximum Rubisco carboxilation rate at 25ºC (in micromol CO2·s-1·m-2).

Jmax298: Maximum rate of electron transport at 25ºC (in micromol photons·s-1·m-2).

Kmax_stemxylem: Sapwood-specific hydraulic conductivity of stem xylem (in kg H2O·s-1·m-2).

Kmax_rootxylem: Sapwood-specific hydraulic conductivity of root xylem (in kg H2O·s-1·m-2).

VCleaf_kmax: Maximum leaf hydraulic conductance.

VCleaf_c, VCleaf_d: Parameters of the leaf vulnerability curve.

VCstem_kmax: Maximum stem xylem conductance.

VCstem_c, VCstem_d: Parameters of the stem xylem vulnerability curve.

VCroot_c, VCroot_d: Parameters of the root xylem vulnerability curve.

Plant_kmax: Maximum whole-plant conductance.

paramsWaterStorage: A data frame with plant water storage parameters for each cohort (only if control$transpirationMode = "Sperry"):

LeafPI0: Osmotic potential at full turgor of leaves (MPa).

LeafEPS: Modulus of elasticity (capacity of the cell wall to resist changes in volume in response to changes in turgor) of leaves (MPa).

LeafAF: Apoplastic fraction (proportion of water outside the living cells) in leaves.

Vleaf: Storage water capacity in leaves, per leaf area (L/m2).

StemPI0: Osmotic potential at full turgor of symplastic xylem tissue (MPa).

StemEPS: Modulus of elasticity (capacity of the cell wall to resist changes in volume in response to changes in turgor) of symplastic xylem tissue (Mpa).

StemAF: Apoplastic fraction (proportion of water outside the living cells) in stem xylem.

Vstem: Storage water capacity in sapwood, per leaf area (L/m2).

internalPhenology and internalWater: data frames to store internal state variables.

Functions forest2growthInput and growthInput return a list of class growthInput with the same elements as spwbInput, but with additional information.

Element above includes the following additional columns:

LA_live: Live leaf area per individual (m2/ind).

LA_dead: Dead leaf area per individual (m2/ind).

SA: Live sapwood area per individual (cm2/ind).

paramsGrowth: A data frame with growth parameters for each cohort:

WoodC: Wood carbon content per dry weight (g C /g dry).

RGRsapwoodmax: Maximum relative growth rate (in basal area or sapwood area) (in cm2·cm-2).

fHDmin: Minimum value of the height-to-diameter ratio (dimensionless).

fHDmax: Maximum value of the height-to-diameter ratio (dimensionless).

paramsAllometry: A data frame with allometric parameters for each cohort:

Aash: Regression coefficient relating the square of shrub height with shrub area.

Absh, Bbsh: Allometric coefficients relating phytovolume with dry weight of shrub individuals.

Acr, B1cr, B2cr, B3cr, C1cr, C2cr: Regression coefficients used to calculate crown ratio of trees.

Acw, Bcw: Regression coefficients used to calculated crown width of trees.

internalAllocation: A data frame with internal allocation variables for each cohort:

allocationTarget: Value of the allocation target variable.

leafAreaTarget: Target leaf area (m2) per individual.

fineRootBiomassTarget: Target fine root biomass (g dry) per individual (only if transpirationMode = "Sperry").

internalCarbon and internalRings: data structures to store other internal state variables.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

# Aboveground parameters
above = forest2aboveground(exampleforestMED, SpParamsMED)
above

#>       SP        N   DBH Cover   H        CR   LAI_live LAI_expanded LAI_dead
#> T1_54 54 168.0000 37.55    NA 800 0.7150421 0.81670117   0.81670117        0
#> T2_68 68 384.0000 14.60    NA 660 0.6055642 0.79779523   0.79779523        0
#> S1_65 65 749.4923    NA  3.75  80 0.8032817 0.04817502   0.04817502        0

# Initialize soil with default soil params
examplesoil = soil(defaultSoilParams())

# Rooting depths
Z50 = c(exampleforestMED$treeData$Z50, exampleforestMED$shrubData$Z50)
Z95 = c(exampleforestMED$treeData$Z95, exampleforestMED$shrubData$Z95)

# Initialize control parameters
control = defaultControl("Granier")

# Prepare spwb input
spwbInput(above, Z50, Z95, examplesoil,SpParamsMED, control)

#> $control
#> $control$modifyInput
#> [1] TRUE
#> 
#> $control$verbose
#> [1] TRUE
#> 
#> $control$subdailyResults
#> [1] FALSE
#> 
#> $control$transpirationMode
#> [1] "Granier"
#> 
#> $control$soilFunctions
#> [1] "SX"
#> 
#> $control$defaultWindSpeed
#> [1] 2.5
#> 
#> $control$snowpack
#> [1] TRUE
#> 
#> $control$leafPhenology
#> [1] TRUE
#> 
#> $control$rockyLayerDrainage
#> [1] TRUE
#> 
#> $control$unlimitedSoilWater
#> [1] FALSE
#> 
#> $control$plantWaterPools
#> [1] FALSE
#> 
#> $control$unfoldingDD
#> [1] 300
#> 
#> $control$verticalLayerSize
#> [1] 100
#> 
#> $control$windMeasurementHeight
#> [1] 200
#> 
#> $control$cavitationRefill
#> [1] "total"
#> 
#> $control$ndailysteps
#> [1] 24
#> 
#> $control$nsubsteps
#> [1] 3600
#> 
#> $control$cochard
#> [1] FALSE
#> 
#> $control$capacitance
#> [1] FALSE
#> 
#> $control$taper
#> [1] TRUE
#> 
#> $control$multiLayerBalance
#> [1] FALSE
#> 
#> $control$gainModifier
#> [1] 1
#> 
#> $control$costModifier
#> [1] 1
#> 
#> $control$costWater
#> [1] "dEdP"
#> 
#> $control$klatstem
#> [1] 0.01
#> 
#> $control$klatleaf
#> [1] 0.01
#> 
#> $control$numericParams
#> $control$numericParams$maxNsteps
#> [1] 400
#> 
#> $control$numericParams$ntrial
#> [1] 200
#> 
#> $control$numericParams$psiTol
#> [1] 1e-04
#> 
#> $control$numericParams$ETol
#> [1] 1e-07
#> 
#> 
#> $control$fracLeafResistance
#> [1] NA
#> 
#> $control$fracRootResistance
#> [1] 0.4
#> 
#> $control$averageFracRhizosphereResistance
#> [1] 0.15
#> 
#> $control$Catm
#> [1] 386
#> 
#> $control$thermalCapacityLAI
#> [1] 1e+06
#> 
#> $control$boundaryLayerSize
#> [1] 2000
#> 
#> $control$refillMaximumRate
#> [1] 0.05
#> 
#> $control$allowDessication
#> [1] TRUE
#> 
#> $control$allowStarvation
#> [1] TRUE
#> 
#> $control$allowDefoliation
#> [1] TRUE
#> 
#> $control$sinkLimitation
#> [1] TRUE
#> 
#> $control$shrubDynamics
#> [1] FALSE
#> 
#> $control$allocationStrategy
#> [1] "Plant_kmax"
#> 
#> $control$nonStomatalPhotosynthesisLimitation
#> [1] TRUE
#> 
#> $control$phloemConductanceFactor
#> [1] 0.2
#> 
#> $control$nonSugarConcentration
#> [1] 0.25
#> 
#> $control$equilibriumOsmoticConcentration
#> $control$equilibriumOsmoticConcentration$leaf
#> [1] 0.8
#> 
#> $control$equilibriumOsmoticConcentration$sapwood
#> [1] 0.6
#> 
#> 
#> $control$minimumRelativeSugarForGrowth
#> [1] 0.5
#> 
#> $control$respirationRates
#> $control$respirationRates$leaf
#> [1] 0.00260274
#> 
#> $control$respirationRates$sapwood
#> [1] 6.849315e-05
#> 
#> $control$respirationRates$fineroot
#> [1] 0.002054795
#> 
#> 
#> $control$turnoverRates
#> $control$turnoverRates$sapwood
#> [1] 0.0001261398
#> 
#> $control$turnoverRates$fineroot
#> [1] 0.001897231
#> 
#> 
#> $control$constructionCosts
#> $control$constructionCosts$leaf
#> [1] 1.5
#> 
#> $control$constructionCosts$sapwood
#> [1] 1.47
#> 
#> $control$constructionCosts$fineroot
#> [1] 1.3
#> 
#> 
#> $control$maximumRelativeGrowthRates
#> $control$maximumRelativeGrowthRates$leaf
#> [1] 0.01
#> 
#> $control$maximumRelativeGrowthRates$sapwood
#> [1] 0.002
#> 
#> $control$maximumRelativeGrowthRates$fineroot
#> [1] 0.1
#> 
#> 
#> $control$mortalityMode
#> [1] "density/deterministic"
#> 
#> $control$mortalityBaselineRate
#> [1] 0.01
#> 
#> $control$mortalityRelativeSugarThreshold
#> [1] 0.3
#> 
#> $control$mortalityRWCThreshold
#> [1] 0.3
#> 
#> $control$recruitmentMode
#> [1] "deterministic"
#> 
#> $control$removeDeadCohorts
#> [1] TRUE
#> 
#> $control$minimumCohortDensity
#> [1] 1
#> 
#> $control$seedRain
#> NULL
#> 
#> $control$seedProductionTreeHeight
#> [1] 300
#> 
#> $control$seedProductionShrubHeight
#> [1] 30
#> 
#> $control$minTempRecr
#> [1] 0
#> 
#> $control$minMoistureRecr
#> [1] 0.3
#> 
#> $control$minFPARRecr
#> [1] 10
#> 
#> $control$recrTreeDBH
#> [1] 1
#> 
#> $control$recrTreeDensity
#> [1] 100
#> 
#> $control$recrTreeHeight
#> [1] 100
#> 
#> $control$recrShrubCover
#> [1] 1
#> 
#> $control$recrShrubHeight
#> [1] 10
#> 
#> $control$recrTreeZ50
#> [1] 100
#> 
#> $control$recrShrubZ50
#> [1] 50
#> 
#> $control$recrTreeZ95
#> [1] 1000
#> 
#> $control$recrShrubZ95
#> [1] 500
#> 
#> 
#> $soil
#> Soil depth (mm): 4000 
#> 
#> Layer  1  [ 0  to  300 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 25 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 32 Vol. FC (mm): 68 Vol. SAT (mm): 111 Vol. current (mm): 68 
#>     Temperature (Celsius): NA 
#> 
#> Layer  2  [ 300  to  1000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 45 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 55 Vol. FC (mm): 117 Vol. SAT (mm): 190 Vol. current (mm): 117 
#>     Temperature (Celsius): NA 
#> 
#> Layer  3  [ 1000  to  2000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 75 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 36 Vol. FC (mm): 76 Vol. SAT (mm): 123 Vol. current (mm): 76 
#>     Temperature (Celsius): NA 
#> 
#> Layer  4  [ 2000  to  4000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 95 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 14 Vol. FC (mm): 30 Vol. SAT (mm): 49 Vol. current (mm): 30 
#>     Temperature (Celsius): NA 
#> 
#> Total soil saturated capacity (mm): 473 
#> Total soil water holding capacity (mm): 291 
#> Total soil extractable water (mm): 183 
#> Total soil current Volume (mm): 291 
#> 
#> Snow pack water equivalent (mm): 0 
#> Soil water table depth (mm): 4000 
#> 
#> 
#> $canopy
#> list()
#> 
#> $cohorts
#>       SP              Name
#> T1_54 54  Pinus halepensis
#> T2_68 68      Quercus ilex
#> S1_65 65 Quercus coccifera
#> 
#> $above
#>         H        CR        N   LAI_live LAI_expanded LAI_dead
#> T1_54 800 0.7150421 168.0000 0.81670117   0.81670117        0
#> T2_68 660 0.6055642 384.0000 0.79779523   0.79779523        0
#> S1_65  80 0.8032817 749.4923 0.04817502   0.04817502        0
#> 
#> $below
#>       Z50  Z95 coarseRootSoilVolume
#> T1_54 750 3000            16.359912
#> T2_68 750 3000            16.359912
#> S1_65 300 1500             7.473966
#> 
#> $belowLayers
#> $belowLayers$V
#>               1         2          3          4
#> T1_54 0.1288916 0.5376838 0.24791124 0.08551338
#> T2_68 0.1288916 0.5376838 0.24791124 0.08551338
#> S1_65 0.5044057 0.4037153 0.07011387 0.02176513
#> 
#> $belowLayers$L
#>              1        2        3        4
#> T1_54 1645.778 2650.000 2636.223 3471.865
#> T2_68 1645.778 2650.000 2636.223 3471.865
#> S1_65 2150.000 1821.357 1908.416 3160.904
#> 
#> $belowLayers$Wpool
#>       1 2 3 4
#> T1_54 1 1 1 1
#> T2_68 1 1 1 1
#> S1_65 1 1 1 1
#> 
#> 
#> $paramsPhenology
#>            PhenologyType LeafDuration Sgdd Tbgdd Ssen Phsen Tbsen
#> T1_54 oneflush-evergreen         3.00  200     5 8268  12.5  28.5
#> T2_68 oneflush-evergreen         3.00  200     5 8268  12.5  28.5
#> S1_65 oneflush-evergreen         1.33  200     5 8268  12.5  28.5
#> 
#> $paramsInterception
#>       kPAR    g
#> T1_54 0.50 1.00
#> T2_68 0.55 0.50
#> S1_65 0.55 0.25
#> 
#> $paramsTranspiration
#>       Tmax_LAI Tmax_LAIsq Psi_Extract Psi_Critic WUE pRootDisc
#> T1_54    0.134     -0.006       -3.00  -5.140000   4         0
#> T2_68    0.134     -0.006       -3.92  -7.130000   4         0
#> S1_65    0.134     -0.006       -3.96  -7.936105   4         0
#> 
#> $internalPhenology
#>       gdd sen budFormation leafUnfolding leafSenescence leafDormancy phi
#> T1_54   0   0        FALSE         FALSE          FALSE        FALSE   0
#> T2_68   0   0        FALSE         FALSE          FALSE        FALSE   0
#> S1_65   0   0        FALSE         FALSE          FALSE        FALSE   0
#> 
#> $internalWater
#>       PlantPsi StemPLC
#> T1_54        0       0
#> T2_68        0       0
#> S1_65        0       0
#> 
#> attr(,"class")
#> [1] "spwbInput" "list"     

# When starting from an object of class 'forest' the whole process
# can be simplified:
forest2spwbInput(exampleforestMED, examplesoil, SpParamsMED, control)

#> $control
#> $control$modifyInput
#> [1] TRUE
#> 
#> $control$verbose
#> [1] TRUE
#> 
#> $control$subdailyResults
#> [1] FALSE
#> 
#> $control$transpirationMode
#> [1] "Granier"
#> 
#> $control$soilFunctions
#> [1] "SX"
#> 
#> $control$defaultWindSpeed
#> [1] 2.5
#> 
#> $control$snowpack
#> [1] TRUE
#> 
#> $control$leafPhenology
#> [1] TRUE
#> 
#> $control$rockyLayerDrainage
#> [1] TRUE
#> 
#> $control$unlimitedSoilWater
#> [1] FALSE
#> 
#> $control$plantWaterPools
#> [1] FALSE
#> 
#> $control$unfoldingDD
#> [1] 300
#> 
#> $control$verticalLayerSize
#> [1] 100
#> 
#> $control$windMeasurementHeight
#> [1] 200
#> 
#> $control$cavitationRefill
#> [1] "total"
#> 
#> $control$ndailysteps
#> [1] 24
#> 
#> $control$nsubsteps
#> [1] 3600
#> 
#> $control$cochard
#> [1] FALSE
#> 
#> $control$capacitance
#> [1] FALSE
#> 
#> $control$taper
#> [1] TRUE
#> 
#> $control$multiLayerBalance
#> [1] FALSE
#> 
#> $control$gainModifier
#> [1] 1
#> 
#> $control$costModifier
#> [1] 1
#> 
#> $control$costWater
#> [1] "dEdP"
#> 
#> $control$klatstem
#> [1] 0.01
#> 
#> $control$klatleaf
#> [1] 0.01
#> 
#> $control$numericParams
#> $control$numericParams$maxNsteps
#> [1] 400
#> 
#> $control$numericParams$ntrial
#> [1] 200
#> 
#> $control$numericParams$psiTol
#> [1] 1e-04
#> 
#> $control$numericParams$ETol
#> [1] 1e-07
#> 
#> 
#> $control$fracLeafResistance
#> [1] NA
#> 
#> $control$fracRootResistance
#> [1] 0.4
#> 
#> $control$averageFracRhizosphereResistance
#> [1] 0.15
#> 
#> $control$Catm
#> [1] 386
#> 
#> $control$thermalCapacityLAI
#> [1] 1e+06
#> 
#> $control$boundaryLayerSize
#> [1] 2000
#> 
#> $control$refillMaximumRate
#> [1] 0.05
#> 
#> $control$allowDessication
#> [1] TRUE
#> 
#> $control$allowStarvation
#> [1] TRUE
#> 
#> $control$allowDefoliation
#> [1] TRUE
#> 
#> $control$sinkLimitation
#> [1] TRUE
#> 
#> $control$shrubDynamics
#> [1] FALSE
#> 
#> $control$allocationStrategy
#> [1] "Plant_kmax"
#> 
#> $control$nonStomatalPhotosynthesisLimitation
#> [1] TRUE
#> 
#> $control$phloemConductanceFactor
#> [1] 0.2
#> 
#> $control$nonSugarConcentration
#> [1] 0.25
#> 
#> $control$equilibriumOsmoticConcentration
#> $control$equilibriumOsmoticConcentration$leaf
#> [1] 0.8
#> 
#> $control$equilibriumOsmoticConcentration$sapwood
#> [1] 0.6
#> 
#> 
#> $control$minimumRelativeSugarForGrowth
#> [1] 0.5
#> 
#> $control$respirationRates
#> $control$respirationRates$leaf
#> [1] 0.00260274
#> 
#> $control$respirationRates$sapwood
#> [1] 6.849315e-05
#> 
#> $control$respirationRates$fineroot
#> [1] 0.002054795
#> 
#> 
#> $control$turnoverRates
#> $control$turnoverRates$sapwood
#> [1] 0.0001261398
#> 
#> $control$turnoverRates$fineroot
#> [1] 0.001897231
#> 
#> 
#> $control$constructionCosts
#> $control$constructionCosts$leaf
#> [1] 1.5
#> 
#> $control$constructionCosts$sapwood
#> [1] 1.47
#> 
#> $control$constructionCosts$fineroot
#> [1] 1.3
#> 
#> 
#> $control$maximumRelativeGrowthRates
#> $control$maximumRelativeGrowthRates$leaf
#> [1] 0.01
#> 
#> $control$maximumRelativeGrowthRates$sapwood
#> [1] 0.002
#> 
#> $control$maximumRelativeGrowthRates$fineroot
#> [1] 0.1
#> 
#> 
#> $control$mortalityMode
#> [1] "density/deterministic"
#> 
#> $control$mortalityBaselineRate
#> [1] 0.01
#> 
#> $control$mortalityRelativeSugarThreshold
#> [1] 0.3
#> 
#> $control$mortalityRWCThreshold
#> [1] 0.3
#> 
#> $control$recruitmentMode
#> [1] "deterministic"
#> 
#> $control$removeDeadCohorts
#> [1] TRUE
#> 
#> $control$minimumCohortDensity
#> [1] 1
#> 
#> $control$seedRain
#> NULL
#> 
#> $control$seedProductionTreeHeight
#> [1] 300
#> 
#> $control$seedProductionShrubHeight
#> [1] 30
#> 
#> $control$minTempRecr
#> [1] 0
#> 
#> $control$minMoistureRecr
#> [1] 0.3
#> 
#> $control$minFPARRecr
#> [1] 10
#> 
#> $control$recrTreeDBH
#> [1] 1
#> 
#> $control$recrTreeDensity
#> [1] 100
#> 
#> $control$recrTreeHeight
#> [1] 100
#> 
#> $control$recrShrubCover
#> [1] 1
#> 
#> $control$recrShrubHeight
#> [1] 10
#> 
#> $control$recrTreeZ50
#> [1] 100
#> 
#> $control$recrShrubZ50
#> [1] 50
#> 
#> $control$recrTreeZ95
#> [1] 1000
#> 
#> $control$recrShrubZ95
#> [1] 500
#> 
#> 
#> $soil
#> Soil depth (mm): 4000 
#> 
#> Layer  1  [ 0  to  300 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 25 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 32 Vol. FC (mm): 68 Vol. SAT (mm): 111 Vol. current (mm): 68 
#>     Temperature (Celsius): NA 
#> 
#> Layer  2  [ 300  to  1000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 45 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 55 Vol. FC (mm): 117 Vol. SAT (mm): 190 Vol. current (mm): 117 
#>     Temperature (Celsius): NA 
#> 
#> Layer  3  [ 1000  to  2000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 75 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 36 Vol. FC (mm): 76 Vol. SAT (mm): 123 Vol. current (mm): 76 
#>     Temperature (Celsius): NA 
#> 
#> Layer  4  [ 2000  to  4000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 95 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 14 Vol. FC (mm): 30 Vol. SAT (mm): 49 Vol. current (mm): 30 
#>     Temperature (Celsius): NA 
#> 
#> Total soil saturated capacity (mm): 473 
#> Total soil water holding capacity (mm): 291 
#> Total soil extractable water (mm): 183 
#> Total soil current Volume (mm): 291 
#> 
#> Snow pack water equivalent (mm): 0 
#> Soil water table depth (mm): 4000 
#> 
#> 
#> $canopy
#> list()
#> 
#> $cohorts
#>       SP              Name
#> T1_54 54  Pinus halepensis
#> T2_68 68      Quercus ilex
#> S1_65 65 Quercus coccifera
#> 
#> $above
#>         H        CR        N   LAI_live LAI_expanded LAI_dead
#> T1_54 800 0.7150421 168.0000 0.81670117   0.81670117        0
#> T2_68 660 0.6055642 384.0000 0.79779523   0.79779523        0
#> S1_65  80 0.8032817 749.4923 0.04817502   0.04817502        0
#> 
#> $below
#>       Z50  Z95 coarseRootSoilVolume
#> T1_54 750 3000            16.359912
#> T2_68 750 3000            16.359912
#> S1_65 300 1500             7.473966
#> 
#> $belowLayers
#> $belowLayers$V
#>               1         2          3          4
#> T1_54 0.1288916 0.5376838 0.24791124 0.08551338
#> T2_68 0.1288916 0.5376838 0.24791124 0.08551338
#> S1_65 0.5044057 0.4037153 0.07011387 0.02176513
#> 
#> $belowLayers$L
#>              1        2        3        4
#> T1_54 1645.778 2650.000 2636.223 3471.865
#> T2_68 1645.778 2650.000 2636.223 3471.865
#> S1_65 2150.000 1821.357 1908.416 3160.904
#> 
#> $belowLayers$Wpool
#>       1 2 3 4
#> T1_54 1 1 1 1
#> T2_68 1 1 1 1
#> S1_65 1 1 1 1
#> 
#> 
#> $paramsPhenology
#>            PhenologyType LeafDuration Sgdd Tbgdd Ssen Phsen Tbsen
#> T1_54 oneflush-evergreen         3.00  200     5 8268  12.5  28.5
#> T2_68 oneflush-evergreen         3.00  200     5 8268  12.5  28.5
#> S1_65 oneflush-evergreen         1.33  200     5 8268  12.5  28.5
#> 
#> $paramsInterception
#>       kPAR    g
#> T1_54 0.50 1.00
#> T2_68 0.55 0.50
#> S1_65 0.55 0.25
#> 
#> $paramsTranspiration
#>       Tmax_LAI Tmax_LAIsq Psi_Extract Psi_Critic WUE pRootDisc
#> T1_54    0.134     -0.006       -3.00  -5.140000   4         0
#> T2_68    0.134     -0.006       -3.92  -7.130000   4         0
#> S1_65    0.134     -0.006       -3.96  -7.936105   4         0
#> 
#> $internalPhenology
#>       gdd sen budFormation leafUnfolding leafSenescence leafDormancy phi
#> T1_54   0   0        FALSE         FALSE          FALSE        FALSE   0
#> T2_68   0   0        FALSE         FALSE          FALSE        FALSE   0
#> S1_65   0   0        FALSE         FALSE          FALSE        FALSE   0
#> 
#> $internalWater
#>       PlantPsi StemPLC
#> T1_54        0       0
#> T2_68        0       0
#> S1_65        0       0
#> 
#> attr(,"class")
#> [1] "spwbInput" "list"     

# Prepare input for Sperry transpiration mode
control = defaultControl("Sperry")
forest2spwbInput(exampleforestMED,examplesoil,SpParamsMED, control)

#> $control
#> $control$modifyInput
#> [1] TRUE
#> 
#> $control$verbose
#> [1] TRUE
#> 
#> $control$subdailyResults
#> [1] FALSE
#> 
#> $control$transpirationMode
#> [1] "Sperry"
#> 
#> $control$soilFunctions
#> [1] "VG"
#> 
#> $control$defaultWindSpeed
#> [1] 2.5
#> 
#> $control$snowpack
#> [1] TRUE
#> 
#> $control$leafPhenology
#> [1] TRUE
#> 
#> $control$rockyLayerDrainage
#> [1] TRUE
#> 
#> $control$unlimitedSoilWater
#> [1] FALSE
#> 
#> $control$plantWaterPools
#> [1] FALSE
#> 
#> $control$unfoldingDD
#> [1] 300
#> 
#> $control$verticalLayerSize
#> [1] 100
#> 
#> $control$windMeasurementHeight
#> [1] 200
#> 
#> $control$cavitationRefill
#> [1] "total"
#> 
#> $control$ndailysteps
#> [1] 24
#> 
#> $control$nsubsteps
#> [1] 3600
#> 
#> $control$cochard
#> [1] FALSE
#> 
#> $control$capacitance
#> [1] FALSE
#> 
#> $control$taper
#> [1] TRUE
#> 
#> $control$multiLayerBalance
#> [1] FALSE
#> 
#> $control$gainModifier
#> [1] 1
#> 
#> $control$costModifier
#> [1] 1
#> 
#> $control$costWater
#> [1] "dEdP"
#> 
#> $control$klatstem
#> [1] 0.01
#> 
#> $control$klatleaf
#> [1] 0.01
#> 
#> $control$numericParams
#> $control$numericParams$maxNsteps
#> [1] 400
#> 
#> $control$numericParams$ntrial
#> [1] 200
#> 
#> $control$numericParams$psiTol
#> [1] 1e-04
#> 
#> $control$numericParams$ETol
#> [1] 1e-07
#> 
#> 
#> $control$fracLeafResistance
#> [1] NA
#> 
#> $control$fracRootResistance
#> [1] 0.4
#> 
#> $control$averageFracRhizosphereResistance
#> [1] 0.15
#> 
#> $control$Catm
#> [1] 386
#> 
#> $control$thermalCapacityLAI
#> [1] 1e+06
#> 
#> $control$boundaryLayerSize
#> [1] 2000
#> 
#> $control$refillMaximumRate
#> [1] 0.05
#> 
#> $control$allowDessication
#> [1] TRUE
#> 
#> $control$allowStarvation
#> [1] TRUE
#> 
#> $control$allowDefoliation
#> [1] TRUE
#> 
#> $control$sinkLimitation
#> [1] TRUE
#> 
#> $control$shrubDynamics
#> [1] FALSE
#> 
#> $control$allocationStrategy
#> [1] "Plant_kmax"
#> 
#> $control$nonStomatalPhotosynthesisLimitation
#> [1] TRUE
#> 
#> $control$phloemConductanceFactor
#> [1] 0.2
#> 
#> $control$nonSugarConcentration
#> [1] 0.25
#> 
#> $control$equilibriumOsmoticConcentration
#> $control$equilibriumOsmoticConcentration$leaf
#> [1] 0.8
#> 
#> $control$equilibriumOsmoticConcentration$sapwood
#> [1] 0.6
#> 
#> 
#> $control$minimumRelativeSugarForGrowth
#> [1] 0.5
#> 
#> $control$respirationRates
#> $control$respirationRates$leaf
#> [1] 0.00260274
#> 
#> $control$respirationRates$sapwood
#> [1] 6.849315e-05
#> 
#> $control$respirationRates$fineroot
#> [1] 0.002054795
#> 
#> 
#> $control$turnoverRates
#> $control$turnoverRates$sapwood
#> [1] 0.0001261398
#> 
#> $control$turnoverRates$fineroot
#> [1] 0.001897231
#> 
#> 
#> $control$constructionCosts
#> $control$constructionCosts$leaf
#> [1] 1.5
#> 
#> $control$constructionCosts$sapwood
#> [1] 1.47
#> 
#> $control$constructionCosts$fineroot
#> [1] 1.3
#> 
#> 
#> $control$maximumRelativeGrowthRates
#> $control$maximumRelativeGrowthRates$leaf
#> [1] 0.01
#> 
#> $control$maximumRelativeGrowthRates$sapwood
#> [1] 0.002
#> 
#> $control$maximumRelativeGrowthRates$fineroot
#> [1] 0.1
#> 
#> 
#> $control$mortalityMode
#> [1] "density/deterministic"
#> 
#> $control$mortalityBaselineRate
#> [1] 0.01
#> 
#> $control$mortalityRelativeSugarThreshold
#> [1] 0.3
#> 
#> $control$mortalityRWCThreshold
#> [1] 0.3
#> 
#> $control$recruitmentMode
#> [1] "deterministic"
#> 
#> $control$removeDeadCohorts
#> [1] TRUE
#> 
#> $control$minimumCohortDensity
#> [1] 1
#> 
#> $control$seedRain
#> NULL
#> 
#> $control$seedProductionTreeHeight
#> [1] 300
#> 
#> $control$seedProductionShrubHeight
#> [1] 30
#> 
#> $control$minTempRecr
#> [1] 0
#> 
#> $control$minMoistureRecr
#> [1] 0.3
#> 
#> $control$minFPARRecr
#> [1] 10
#> 
#> $control$recrTreeDBH
#> [1] 1
#> 
#> $control$recrTreeDensity
#> [1] 100
#> 
#> $control$recrTreeHeight
#> [1] 100
#> 
#> $control$recrShrubCover
#> [1] 1
#> 
#> $control$recrShrubHeight
#> [1] 10
#> 
#> $control$recrTreeZ50
#> [1] 100
#> 
#> $control$recrShrubZ50
#> [1] 50
#> 
#> $control$recrTreeZ95
#> [1] 1000
#> 
#> $control$recrShrubZ95
#> [1] 500
#> 
#> 
#> $soil
#> Soil depth (mm): 4000 
#> 
#> Layer  1  [ 0  to  300 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 25 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 32 Vol. FC (mm): 68 Vol. SAT (mm): 111 Vol. current (mm): 68 
#>     Temperature (Celsius): NA 
#> 
#> Layer  2  [ 300  to  1000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 45 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 55 Vol. FC (mm): 117 Vol. SAT (mm): 190 Vol. current (mm): 117 
#>     Temperature (Celsius): NA 
#> 
#> Layer  3  [ 1000  to  2000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 75 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 36 Vol. FC (mm): 76 Vol. SAT (mm): 123 Vol. current (mm): 76 
#>     Temperature (Celsius): NA 
#> 
#> Layer  4  [ 2000  to  4000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 95 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 14 Vol. FC (mm): 30 Vol. SAT (mm): 49 Vol. current (mm): 30 
#>     Temperature (Celsius): NA 
#> 
#> Total soil saturated capacity (mm): 473 
#> Total soil water holding capacity (mm): 291 
#> Total soil extractable water (mm): 183 
#> Total soil current Volume (mm): 291 
#> 
#> Snow pack water equivalent (mm): 0 
#> Soil water table depth (mm): 4000 
#> 
#> 
#> $canopy
#>    zlow zmid  zup Tair Cair VPair
#> 1     0   50  100   NA  386    NA
#> 2   100  150  200   NA  386    NA
#> 3   200  250  300   NA  386    NA
#> 4   300  350  400   NA  386    NA
#> 5   400  450  500   NA  386    NA
#> 6   500  550  600   NA  386    NA
#> 7   600  650  700   NA  386    NA
#> 8   700  750  800   NA  386    NA
#> 9   800  850  900   NA  386    NA
#> 10  900  950 1000   NA  386    NA
#> 11 1000 1050 1100   NA  386    NA
#> 12 1100 1150 1200   NA  386    NA
#> 13 1200 1250 1300   NA  386    NA
#> 14 1300 1350 1400   NA  386    NA
#> 15 1400 1450 1500   NA  386    NA
#> 16 1500 1550 1600   NA  386    NA
#> 17 1600 1650 1700   NA  386    NA
#> 18 1700 1750 1800   NA  386    NA
#> 19 1800 1850 1900   NA  386    NA
#> 20 1900 1950 2000   NA  386    NA
#> 21 2000 2050 2100   NA  386    NA
#> 22 2100 2150 2200   NA  386    NA
#> 23 2200 2250 2300   NA  386    NA
#> 24 2300 2350 2400   NA  386    NA
#> 25 2400 2450 2500   NA  386    NA
#> 26 2500 2550 2600   NA  386    NA
#> 27 2600 2650 2700   NA  386    NA
#> 28 2700 2750 2800   NA  386    NA
#> 
#> $cohorts
#>       SP              Name
#> T1_54 54  Pinus halepensis
#> T2_68 68      Quercus ilex
#> S1_65 65 Quercus coccifera
#> 
#> $above
#>         H        CR        N   LAI_live LAI_expanded LAI_dead
#> T1_54 800 0.7150421 168.0000 0.81670117   0.81670117        0
#> T2_68 660 0.6055642 384.0000 0.79779523   0.79779523        0
#> S1_65  80 0.8032817 749.4923 0.04817502   0.04817502        0
#> 
#> $below
#>       Z50  Z95 fineRootBiomass coarseRootSoilVolume
#> T1_54 750 3000        1178.680            23.500000
#> T2_68 750 3000        2534.841            24.000000
#> S1_65 300 1500        1146.417             0.390625
#> 
#> $belowLayers
#> $belowLayers$V
#>               1         2          3          4
#> T1_54 0.1288916 0.5376838 0.24791124 0.08551338
#> T2_68 0.1288916 0.5376838 0.24791124 0.08551338
#> S1_65 0.5044057 0.4037153 0.07011387 0.02176513
#> 
#> $belowLayers$L
#>               1        2        3        4
#> T1_54 2370.0461 4232.154 4723.561 6529.158
#> T2_68 2391.9519 4266.358 4752.383 6555.923
#> S1_65  827.9634 1361.087 2186.740 4164.508
#> 
#> $belowLayers$VGrhizo_kmax
#>                1         2        3        4
#> T1_54    6488820   9502024  4381122  1511204
#> T2_68   31939094  48255259 22249175  7674530
#> S1_65 1167494294 360257501 62566484 19422230
#> 
#> $belowLayers$VCroot_kmax
#>               1        2         3          4
#> T1_54 0.3432826 0.801954 0.3312918 0.08267246
#> T2_68 0.9221872 2.156834 0.8927540 0.22322726
#> S1_65 3.9179693 1.907574 0.2062049 0.03361165
#> 
#> $belowLayers$Wpool
#>       1 2 3 4
#> T1_54 1 1 1 1
#> T2_68 1 1 1 1
#> S1_65 1 1 1 1
#> 
#> $belowLayers$RhizoPsi
#>       1 2 3 4
#> T1_54 0 0 0 0
#> T2_68 0 0 0 0
#> S1_65 0 0 0 0
#> 
#> 
#> $paramsPhenology
#>            PhenologyType LeafDuration Sgdd Tbgdd Ssen Phsen Tbsen
#> T1_54 oneflush-evergreen         3.00  200     5 8268  12.5  28.5
#> T2_68 oneflush-evergreen         3.00  200     5 8268  12.5  28.5
#> S1_65 oneflush-evergreen         1.33  200     5 8268  12.5  28.5
#> 
#> $paramsAnatomy
#>         Hmax Hmed    Al2As   SLA LeafWidth LeafDensity WoodDensity
#> T1_54 2000.0  970 1317.523 4.340       0.1         0.7     0.55253
#> T2_68 1421.8  650 2512.563 5.870       3.0         0.7     0.90000
#> S1_65  180.0   70 2512.563 5.859       1.0         0.7     0.92000
#>       FineRootDensity conduit2sapwood  SRL RLD     r635
#> T1_54             0.7           0.925 3870  10 1.964226
#> T2_68             0.7           0.700 3870  10 1.805872
#> S1_65             0.7           0.700 3870  10 2.289452
#> 
#> $paramsInterception
#>       alphaSWR gammaSWR kPAR    g
#> T1_54      0.7     0.14 0.50 1.00
#> T2_68      0.7     0.18 0.55 0.50
#> S1_65      0.7     0.18 0.55 0.25
#> 
#> $paramsTranspiration
#>        Gswmin    Gswmax Vmax298  Jmax298 Kmax_stemxylem Kmax_rootxylem
#> T1_54 0.00203 0.1900000    62.5 129.5000      0.1500000       0.893000
#> T2_68 0.00450 0.2100000    66.2 130.0000      0.7735834       3.094334
#> S1_65 0.01045 0.4518345   100.0 163.6253      0.2900000       1.160000
#>       VCleaf_kmax  VCleaf_c  VCleaf_d VCstem_kmax  VCstem_c  VCstem_d
#> T1_54           6 12.712961 -3.526895    1.257285 12.712961 -5.290342
#> T2_68           8  2.188265 -4.138168    4.299811  4.581685 -7.723800
#> S1_65           8  2.511383 -6.122057    8.176212  2.511383 -9.183085
#>       VCroot_kmax VCroot_c  VCroot_d VGrhizo_kmax Plant_kmax
#> T1_54    1.559201 1.056898 -1.244767     21883171  0.6236803
#> T2_68    4.195002 1.494098 -2.134283    110118057  1.6780008
#> S1_65    6.065360 2.655734 -6.198418   1609740510  2.4261439
#> 
#> $paramsWaterStorage
#>       LeafPI0 LeafEPS LeafAF     Vleaf   StemPI0  StemEPS StemAF Vsapwood
#> T1_54   -2.11   12.18  0.162 0.1795440 -1.778525 10.43383    0.8 3.893453
#> T2_68   -2.39   19.26  0.170 0.1327463 -3.224000 46.25763    0.8 1.091657
#> S1_65   -2.37   17.23  0.240 0.1329955 -3.307200 50.36679    0.8 0.128187
#> 
#> $internalPhenology
#>       gdd sen budFormation leafUnfolding leafSenescence leafDormancy phi
#> T1_54   0   0        FALSE         FALSE          FALSE        FALSE   0
#> T2_68   0   0        FALSE         FALSE          FALSE        FALSE   0
#> S1_65   0   0        FALSE         FALSE          FALSE        FALSE   0
#> 
#> $internalWater
#>       Einst RootCrownPsi Stem1Psi Stem2Psi LeafPsi StemSympPsi LeafSympPsi
#> T1_54     0            0        0        0       0           0           0
#> T2_68     0            0        0        0       0           0           0
#> S1_65     0            0        0        0       0           0           0
#>       StemPLC NSPL
#> T1_54       0    1
#> T2_68       0    1
#> S1_65       0    1
#> 
#> attr(,"class")
#> [1] "spwbInput" "list"