Soil initialization

Initializes soil parameters and state variables for its use in simulations.

soil(SoilParams, VG_PTF = "Toth", W = as.numeric(c(1)), SWE = 0)
# S3 method for soil
print(x, model="SX", ...)
redefineSoilLayers(SoilParams, widths = c(300, 700, 1000, 2000))

Arguments

SoilParams

A data frame of soil parameters (see an example in defaultSoilParams).

VG_PTF

Pedotransfer functions to obtain parameters for the van Genuchten-Mualem equations. Either "Carsel" (Carsel & Parrish 1988) or "Toth" (Toth et al. 2015).

W

A numerical vector with the initial relative water content of each soil layer.

SWE

Initial snow water equivalent of the snow pack on the soil surface (mm).

x

An object of class soil.

model

Either 'SX' or 'VG' for Saxton or Van Genuchten pedotransfer models.

widths

A numeric vector indicating the desired layer widths, in mm.

...

Additional parameters to print.

Details

Function print prompts a description of soil characteristics and state variables (water content and temperature) according to a water retention curve (either Saxton's or Van Genuchten's). Volume at field capacity is calculated assuming a soil water potential equal to -0.033 MPa. Parameter Temp is initialized as missing for all soil layers. Function redefineSoilLayers allows redefining soil layer widths of an input data frame of soil parameters.

Value

Function soil returns a list of class soil with the following elements:

SoilDepth: Soil depth (in mm).

W: State variable with relative water content of each layer (in as proportion relative to FC).

Temp: State variable with temperature (in ºC) of each layer.

Ksoil: Kappa parameter for infiltration.

Gsoil: Gamma parameter for infiltration.

dVec: Width of soil layers (in mm).

sand: Sand percentage for each layer (in percent volume).

clay: Clay percentage for each layer (in percent volume).

om: Organic matter percentage for each layer (in percent volume).

VG_alpha, VG_n, VG_theta_res, VG_theta_sat: Parameters for van Genuchten's pedotransfer functions, for each layer, corresponding to the USDA texture type.

Ksat: Saturated soil conductivity for each layer (estimated using function soil_saturatedConductivitySX.

macro: Macroporosity for each layer (estimated using Stolf et al. 2011).

rfc: Percentage of rock fragment content for each layer.

Kdrain: Saturated vertical hydraulic conductivity (mm/day) (i.e. how easy is deep drainage towards groundwater). Function soil estimates it as a function of soil saturated hydraulic conductivity, but should be parametrized as a function of bedrock material.

Author

Miquel De Cáceres Ainsa, CREAF

References

Carsel, R.F., & Parrish, R.S. 1988. Developing joint probability distributions of soil water retention characteristics. Water Resources Research 24: 755–769.

Tóth, B., Weynants, M., Nemes, A., Makó, A., Bilas, G., & Tóth, G. 2015. New generation of hydraulic pedotransfer functions for Europe. European Journal of Soil Science 66: 226–238.

Stolf, R., Thurler, A., Oliveira, O., Bacchi, S., Reichardt, K., 2011. Method to estimate soil macroporosity and microporosity based on sand content and bulk density. Rev. Bras. Ciencias do Solo 35, 447–459.

See also

Examples

# Initializes soil
s = soil(defaultSoilParams())

# Prints soil characteristics according to Saxton's water retention curve
print(s, model="SX")
#> Soil depth (mm): 4000 
#> 
#> Layer  1  [ 0  to  300 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 25 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 32 Vol. FC (mm): 68 Vol. SAT (mm): 111 Vol. current (mm): 68 
#>     Temperature (Celsius): NA 
#> 
#> Layer  2  [ 300  to  1000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 45 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 55 Vol. FC (mm): 117 Vol. SAT (mm): 190 Vol. current (mm): 117 
#>     Temperature (Celsius): NA 
#> 
#> Layer  3  [ 1000  to  2000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 75 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 36 Vol. FC (mm): 76 Vol. SAT (mm): 123 Vol. current (mm): 76 
#>     Temperature (Celsius): NA 
#> 
#> Layer  4  [ 2000  to  4000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 95 Macroporosity (%): 5 
#>     Theta WP (%): 14 Theta FC (%): 30 Theta SAT (%): 49 Theta current (%) 30 
#>     Vol. WP (mm): 14 Vol. FC (mm): 30 Vol. SAT (mm): 49 Vol. current (mm): 30 
#>     Temperature (Celsius): NA 
#> 
#> Total soil saturated capacity (mm): 473 
#> Total soil water holding capacity (mm): 291 
#> Total soil extractable water (mm): 183 
#> Total soil current Volume (mm): 291 
#> 
#> Snow pack water equivalent (mm): 0 
#> Soil water table depth (mm): 4000 
#> 

# Prints soil characteristics according to Van Genuchten's water retention curve
print(s, model="VG")
#> Soil depth (mm): 4000 
#> 
#> Layer  1  [ 0  to  300 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 25 Macroporosity (%): 5 
#>     Theta WP (%): 13 Theta FC (%): 29 Theta SAT (%): 42 Theta current (%) 29 
#>     Vol. WP (mm): 30 Vol. FC (mm): 64 Vol. SAT (mm): 95 Vol. current (mm): 64 
#>     Temperature (Celsius): NA 
#> 
#> Layer  2  [ 300  to  1000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 45 Macroporosity (%): 5 
#>     Theta WP (%): 13 Theta FC (%): 30 Theta SAT (%): 42 Theta current (%) 30 
#>     Vol. WP (mm): 49 Vol. FC (mm): 117 Vol. SAT (mm): 163 Vol. current (mm): 117 
#>     Temperature (Celsius): NA 
#> 
#> Layer  3  [ 1000  to  2000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 75 Macroporosity (%): 5 
#>     Theta WP (%): 13 Theta FC (%): 30 Theta SAT (%): 42 Theta current (%) 30 
#>     Vol. WP (mm): 32 Vol. FC (mm): 76 Vol. SAT (mm): 106 Vol. current (mm): 76 
#>     Temperature (Celsius): NA 
#> 
#> Layer  4  [ 2000  to  4000 mm ] 
#>     clay (%): 25 silt (%): 50 sand (%): 25 organic matter (%): NA [ Silt loam ]
#>     Rock fragment content (%): 95 Macroporosity (%): 5 
#>     Theta WP (%): 13 Theta FC (%): 30 Theta SAT (%): 42 Theta current (%) 30 
#>     Vol. WP (mm): 13 Vol. FC (mm): 30 Vol. SAT (mm): 42 Vol. current (mm): 30 
#>     Temperature (Celsius): NA 
#> 
#> Total soil saturated capacity (mm): 407 
#> Total soil water holding capacity (mm): 287 
#> Total soil extractable water (mm): 189 
#> Total soil current Volume (mm): 287 
#> 
#> Snow pack water equivalent (mm): 0 
#> Soil water table depth (mm): 4000 
#>