Summarize simulation results

Function summary summarizes the model's output in different temporal steps (i.e. weekly, annual, ...).

# S3 method for spwb
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for pwb
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for growth
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for fordyn
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

Arguments

object

An object of class spwb, pwb, growth or fordyn.

freq

Frequency of summary statistics (see cut.Date).

output

The data table to be summarized. Accepted values are the path to data tables in object, such as 'WaterBalance', 'Soil', 'Stand' or 'Plants$LAI'. It is also possible to use strings like 'Transpiration' and the function will interpret it as 'Plants$Transpiration'.

FUN

The function to summarize results (e.g., sum, mean, ...)

bySpecies

Allows aggregating output by species before calculating summaries (only has an effect with some values of output). Aggregation can involve a sum (as for plant lai or transpiration) or a LAI-weighted mean (as for plant stress or plant water potential).

months

A vector of month numbers (1 to 12) to subset the season where summaries apply.

...

Additional parameters for function summary.

Value

A matrix with dates as row names and the desired summaries in columns

Note

When applied to fordyn objects, the summary function can be used to gather the results of different yearly steps into a single table while keeping a daily resolution (i.e. using freq = "days".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforest,examplesoil, SpParamsMED, control)

#Call simulation function
S1<-spwb(x, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 5.25158
#> Initial soil water content (mm): 184.827
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:....................................
#> 
#> Final plant water content (mm): 5.2515
#> Final soil water content (mm): 205.078
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): -7.99418e-05
#> Plant water balance result (mm): -7.99418e-05
#> Change in soil water content (mm): 20.2516
#> Soil water balance result (mm): 20.2516
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): -7.10543e-15
#> Water balance components:
#>   Precipitation (mm) 513 Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 92 Net rainfall (mm) 370
#>   Infiltration (mm) 375 InfiltrationExcess (mm) 46 SaturationExcess (mm) 0 Runoff (mm) 46 Deep drainage (mm) 141
#>   Soil evaporation (mm) 14  Herbaceous transpiration (mm) 14 Woody plant transpiration (mm) 187
#>   Plant extraction from soil (mm) 187  Plant water balance (mm) -0 Hydraulic redistribution (mm) 1

#Monthly summary (averages) of soil status
summary(S1, freq="months",FUN=mean, output="Soil")
#>                  W.1       W.2     ML.1     ML.2    MLTot SaturatedDepth
#> 2001-01-01 1.2429153 1.1144982 84.73428 130.0095 214.7438       596.7733
#> 2001-02-01 1.1553995 1.2157917 78.76799 141.8257 220.5937       512.1004
#> 2001-03-01 1.1741455 1.2681372 80.04598 147.9320 227.9780       395.6482
#> 2001-04-01 0.9726177 1.1611430 66.30706 135.4508 201.7578       715.1048
#> 2001-05-01 1.0470167 1.1206984 71.37912 130.7328 202.1119       744.4997
#> 2001-06-01 0.8322089 0.9831077 56.73486 114.6824 171.4173       950.2320
#> 2001-07-01 1.0676038 0.9944090 72.78262 116.0008 188.7834       829.8818
#> 2001-08-01 1.0773362 1.0808821 73.44612 126.0881 199.5342       793.7349
#> 2001-09-01 1.0696298 1.1200027 72.92075 130.6516 203.5724       735.6449
#> 2001-10-01 1.0979368 1.1265708 74.85054 131.4178 206.2684       702.7754
#> 2001-11-01 1.1710015 1.1962015 79.83164 139.5405 219.3721       516.8676
#> 2001-12-01 1.0405052 1.2326340 70.93521 143.7904 214.7256       566.8797
#>                   SWE PlantExt.1 PlantExt.2 HydraulicInput.1 HydraulicInput.2
#> 2001-01-01 1.65619239  0.1631759 0.01094778                0     3.364730e-03
#> 2001-02-01 0.27608905  0.2356073 0.15666322                0     2.000489e-05
#> 2001-03-01 0.01762496  0.2261468 0.18161066                0     1.465066e-04
#> 2001-04-01 0.58072652  0.1964653 0.33248082                0     0.000000e+00
#> 2001-05-01 0.00000000  0.3395982 0.28886928                0     0.000000e+00
#> 2001-06-01 0.00000000  0.3049327 0.58697788                0     0.000000e+00
#> 2001-07-01 0.00000000  0.6586968 0.10237022                0     1.356623e-02
#> 2001-08-01 0.00000000  0.5564444 0.21030351                0     1.577554e-03
#> 2001-09-01 0.00000000  0.3620543 0.23148864                0     7.182487e-06
#> 2001-10-01 0.00000000  0.2648783 0.14868283                0     2.994093e-04
#> 2001-11-01 2.60802857  0.1916250 0.10382167                0     9.542350e-04
#> 2001-12-01 0.00000000  0.1124425 0.16387120                0     0.000000e+00
#>                   psi.1        psi.2
#> 2001-01-01 -0.009666023 -0.019300067
#> 2001-02-01 -0.016622727 -0.010958357
#> 2001-03-01 -0.015238742 -0.007524482
#> 2001-04-01 -0.038810848 -0.015345516
#> 2001-05-01 -0.028825335 -0.018665143
#> 2001-06-01 -0.077896134 -0.037591011
#> 2001-07-01 -0.027744647 -0.034449597
#> 2001-08-01 -0.023744328 -0.022404644
#> 2001-09-01 -0.024297569 -0.018468786
#> 2001-10-01 -0.021330446 -0.017828389
#> 2001-11-01 -0.015605692 -0.012407260
#> 2001-12-01 -0.028729536 -0.009941158

#Queries the tables in 'Plants'
names(S1$Plants)
#>  [1] "LAI"                 "LAIlive"             "FPAR"               
#>  [4] "AbsorbedSWRFraction" "Transpiration"       "GrossPhotosynthesis"
#>  [7] "PlantPsi"            "LeafPLC"             "StemPLC"            
#> [10] "PlantWaterBalance"   "LeafRWC"             "StemRWC"            
#> [13] "LFMC"                "PlantStress"        

#Monthly summary (averages) of plant stress
summary(S1, freq="months",FUN=mean, output="Plants$PlantStress", 
        bySpecies = TRUE)
#>            Pinus halepensis Quercus coccifera Quercus ilex
#> 2001-01-01      0.001521178      0.0010354372  0.004659389
#> 2001-02-01      0.001978141      0.0010907108  0.004144834
#> 2001-03-01      0.001876975      0.0009820811  0.003532698
#> 2001-04-01      0.006792051      0.0030588885  0.008809910
#> 2001-05-01      0.004548663      0.0022631170  0.007468184
#> 2001-06-01      0.018513368      0.0076251717  0.019215872
#> 2001-07-01      0.006119515      0.0031440608  0.010553579
#> 2001-08-01      0.003688143      0.0020168478  0.007334312
#> 2001-09-01      0.003622953      0.0019135724  0.006758571
#> 2001-10-01      0.003043821      0.0016671546  0.006145929
#> 2001-11-01      0.002135652      0.0011746633  0.004427947
#> 2001-12-01      0.004182694      0.0019734231  0.006066690