Summarize simulation results

Function summary summarizes the model's output in different temporal steps (i.e. weekly, annual, ...).

# S3 method for spwb
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for pwb
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for growth
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for fordyn
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

Arguments

object

An object of class spwb, pwb, growth or fordyn.

freq

Frequency of summary statistics (see cut.Date).

output

The data table to be summarized. Accepted values are the path to data tables in object, such as 'WaterBalance', 'Soil', 'Stand' or 'Plants$LAI'. It is also possible to use strings like 'Transpiration' and the function will interpret it as 'Plants$Transpiration'.

FUN

The function to summarize results (e.g., sum, mean, ...)

bySpecies

Allows aggregating output by species before calculating summaries (only has an effect with some values of output). Aggregation can involve a sum (as for plant lai or transpiration) or a LAI-weighted mean (as for plant stress or plant water potential).

months

A vector of month numbers (1 to 12) to subset the season where summaries apply.

...

Additional parameters for function summary.

Value

A matrix with dates as row names and the desired summaries in columns

Note

When applied to fordyn objects, the summary function can be used to gather the results of different yearly steps into a single table while keeping a daily resolution (i.e. using freq = "days".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforest,examplesoil, SpParamsMED, control)

#Call simulation function
S1<-spwb(x, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 5.25158
#> Initial soil water content (mm): 184.827
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:....................................
#> 
#> Final plant water content (mm): 5.25147
#> Final soil water content (mm): 204.803
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): -0.000101742
#> Plant water balance result (mm): -0.000101742
#> Change in soil water content (mm): 19.9762
#> Soil water balance result (mm): 19.9762
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): -7.10543e-15
#> Water balance components:
#>   Precipitation (mm) 513 Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 92 Net rainfall (mm) 370
#>   Infiltration (mm) 375 Infiltration excess (mm) 46 Saturation excess (mm) 0 Capillarity rise (mm) 0
#>   Soil evaporation (mm) 14  Herbaceous transpiration (mm) 14 Woody plant transpiration (mm) 186
#>   Plant extraction from soil (mm) 186  Plant water balance (mm) -0 Hydraulic redistribution (mm) 1
#>   Runoff (mm) 46 Deep drainage (mm) 141

#Monthly summary (averages) of soil status
summary(S1, freq="months",FUN=mean, output="Soil")
#>                  W.1       W.2     ML.1     ML.2    MLTot SaturatedDepth
#> 2001-01-01 1.2350326 1.1186192 84.19688 130.4903 214.6871       595.5052
#> 2001-02-01 1.1423878 1.2189422 77.88093 142.1932 220.0742       516.5952
#> 2001-03-01 1.1713416 1.2669861 79.85483 147.7977 227.6525       399.4984
#> 2001-04-01 0.9646740 1.1661587 65.76551 136.0359 201.8014       711.3311
#> 2001-05-01 1.0205102 1.1252703 69.57207 131.2661 200.8382       750.9525
#> 2001-06-01 0.8224338 0.9788055 56.06845 114.1806 170.2490       951.5383
#> 2001-07-01 1.0640025 0.9885435 72.53711 115.3165 187.8537       840.0196
#> 2001-08-01 1.0770271 1.0762992 73.42504 125.5535 198.9785       801.7555
#> 2001-09-01 1.0691477 1.1180523 72.88788 130.4241 203.3120       739.3219
#> 2001-10-01 1.0980425 1.1253790 74.85774 131.2788 206.1365       704.7372
#> 2001-11-01 1.1659611 1.1976159 79.48802 139.7054 219.1935       518.3050
#> 2001-12-01 1.0371793 1.2315822 70.70847 143.6677 214.3762       570.4767
#>                   SWE PlantExt.1 PlantExt.2 HydraulicInput.1 HydraulicInput.2
#> 2001-01-01 1.65619239  0.1621314 0.01198852                0     3.082736e-03
#> 2001-02-01 0.27608905  0.2251667 0.16702751                0     0.000000e+00
#> 2001-03-01 0.01762496  0.2260070 0.18174297                0     1.145253e-04
#> 2001-04-01 0.58072652  0.1860935 0.34276190                0     0.000000e+00
#> 2001-05-01 0.00000000  0.3011620 0.32685715                0     0.000000e+00
#> 2001-06-01 0.00000000  0.2899867 0.60091887                0     0.000000e+00
#> 2001-07-01 0.00000000  0.6600035 0.10085771                0     1.412384e-02
#> 2001-08-01 0.00000000  0.5638249 0.20288612                0     1.861686e-03
#> 2001-09-01 0.00000000  0.3630927 0.23043641                0     1.582672e-05
#> 2001-10-01 0.00000000  0.2655063 0.14805218                0     3.130401e-04
#> 2001-11-01 2.60802857  0.1880451 0.10739507                0     6.873115e-04
#> 2001-12-01 0.00000000  0.1112731 0.16501545                0     0.000000e+00
#>                  psi.1        psi.2
#> 2001-01-01 -0.01010808 -0.018917784
#> 2001-02-01 -0.01785644 -0.010746833
#> 2001-03-01 -0.01536686 -0.007592415
#> 2001-04-01 -0.04009639 -0.014839643
#> 2001-05-01 -0.03227451 -0.018343139
#> 2001-06-01 -0.08165658 -0.038433226
#> 2001-07-01 -0.02837905 -0.035429008
#> 2001-08-01 -0.02379062 -0.022909776
#> 2001-09-01 -0.02434841 -0.018649330
#> 2001-10-01 -0.02131900 -0.017936824
#> 2001-11-01 -0.01590422 -0.012322239
#> 2001-12-01 -0.02918379 -0.010009925

#Queries the tables in 'Plants'
names(S1$Plants)
#>  [1] "LAI"                 "LAIlive"             "FPAR"               
#>  [4] "AbsorbedSWRFraction" "Transpiration"       "GrossPhotosynthesis"
#>  [7] "PlantPsi"            "LeafPLC"             "StemPLC"            
#> [10] "PlantWaterBalance"   "LeafRWC"             "StemRWC"            
#> [13] "LFMC"                "PlantStress"        

#Monthly summary (averages) of plant stress
summary(S1, freq="months",FUN=mean, output="Plants$PlantStress", 
        bySpecies = TRUE)
#>            Pinus halepensis Quercus coccifera Quercus ilex
#> 2001-01-01      0.001551774      0.0010453252  0.004669239
#> 2001-02-01      0.002175805      0.0011705241  0.004317186
#> 2001-03-01      0.001904558      0.0009954855  0.003574484
#> 2001-04-01      0.006995769      0.0031307345  0.008924484
#> 2001-05-01      0.005311195      0.0025508699  0.008042406
#> 2001-06-01      0.019802074      0.0080744860  0.020051063
#> 2001-07-01      0.006364981      0.0032541728  0.010851148
#> 2001-08-01      0.003712393      0.0020371681  0.007424809
#> 2001-09-01      0.003637879      0.0019231883  0.006796687
#> 2001-10-01      0.003044694      0.0016696564  0.006161481
#> 2001-11-01      0.002155357      0.0011847235  0.004459490
#> 2001-12-01      0.004280728      0.0020127830  0.006157395