Summarize simulation results

Function summary summarizes the model's output in different temporal steps (i.e. weekly, annual, ...).

# S3 method for spwb
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for pwb
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for growth
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for fordyn
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

Arguments

object

An object of class spwb, pwb, growth or fordyn.

freq

Frequency of summary statistics (see cut.Date).

output

The data table to be summarized. Accepted values are the path to data tables in object, such as 'WaterBalance', 'Soil', 'Stand' or 'Plants$LAI'. It is also possible to use strings like 'Transpiration' and the function will interpret it as 'Plants$Transpiration'.

FUN

The function to summarize results (e.g., sum, mean, ...)

bySpecies

Allows aggregating output by species before calculating summaries (only has an effect with some values of output). Aggregation can involve a sum (as for plant lai or transpiration) or a LAI-weighted mean (as for plant stress or plant water potential).

months

A vector of month numbers (1 to 12) to subset the season where summaries apply.

...

Additional parameters for function summary.

Value

A matrix with dates as row names and the desired summaries in columns

Note

When applied to fordyn objects, the summary function can be used to gather the results of different yearly steps into a single table while keeping a daily resolution (i.e. using freq = "days".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforest,examplesoil, SpParamsMED, control)

#Call simulation function
S1<-spwb(x, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 5.25158
#> Initial soil water content (mm): 184.827
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:....................................
#> 
#> Final plant water content (mm): 5.25151
#> Final soil water content (mm): 205.22
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): -6.12944e-05
#> Plant water balance result (mm): -6.12944e-05
#> Change in soil water content (mm): 20.3935
#> Soil water balance result (mm): 20.3935
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): -7.10543e-15
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 92 Net rainfall (mm) 370
#>   Infiltration (mm) 324 Runoff (mm) 46 Deep drainage (mm) 141
#>   Soil evaporation (mm) 13  Herbaceous transpiration (mm) 14 Woody plant transpiration (mm) 187
#>   Plant extraction from soil (mm) 187  Plant water balance (mm) -0 Hydraulic redistribution (mm) 1

#Monthly summary (averages) of soil status
summary(S1, freq="months",FUN=mean, output="Soil")
#>                  W.1      W.2     ML.1     ML.2    MLTot SaturatedDepth
#> 2001-01-01 1.2429153 1.114498 84.73428 130.0095 214.7438       596.7733
#> 2001-02-01 1.1553995 1.215792 78.76799 141.8257 220.5937       512.1004
#> 2001-03-01 1.1741528 1.268129 80.04648 147.9310 227.9774       395.6576
#> 2001-04-01 0.9726456 1.161145 66.30896 135.4510 201.7599       715.0951
#> 2001-05-01 1.0470344 1.120707 71.38033 130.7338 202.1142       744.4750
#> 2001-06-01 0.8322184 0.983118 56.73550 114.6836 171.4191       950.2140
#> 2001-07-01 1.0676084 0.994419 72.78293 116.0019 188.7849       829.8656
#> 2001-08-01 1.0773365 1.080889 73.44614 126.0890 199.5351       793.7221
#> 2001-09-01 1.0696286 1.120006 72.92066 130.6521 203.5727       735.6393
#> 2001-10-01 1.0979353 1.126572 74.85044 131.4180 206.2685       702.7737
#> 2001-11-01 1.1713966 1.195843 79.85857 139.4987 219.3573       517.1947
#> 2001-12-01 1.0421653 1.232522 71.04838 143.7774 214.8257       566.2472
#>                   SWE PlantExt.1 PlantExt.2 HydraulicInput.1 HydraulicInput.2
#> 2001-01-01 1.65619239  0.1631759 0.01094778                0     3.364730e-03
#> 2001-02-01 0.27608905  0.2356073 0.15666322                0     2.000489e-05
#> 2001-03-01 0.01762496  0.2261563 0.18160121                0     1.464577e-04
#> 2001-04-01 0.58072652  0.1964924 0.33245441                0     0.000000e+00
#> 2001-05-01 0.00000000  0.3396070 0.28886083                0     0.000000e+00
#> 2001-06-01 0.00000000  0.3049374 0.58697425                0     0.000000e+00
#> 2001-07-01 0.00000000  0.6586928 0.10237450                0     1.356512e-02
#> 2001-08-01 0.00000000  0.5564380 0.21031000                0     1.577149e-03
#> 2001-09-01 0.00000000  0.3620506 0.23149231                0     7.149711e-06
#> 2001-10-01 0.00000000  0.2648767 0.14868438                0     2.993803e-04
#> 2001-11-01 2.60802857  0.1918149 0.10363231                0     9.760892e-04
#> 2001-12-01 0.00000000  0.1132366 0.16308949                0     0.000000e+00
#>                   psi.1        psi.2
#> 2001-01-01 -0.009666023 -0.019300067
#> 2001-02-01 -0.016622727 -0.010958357
#> 2001-03-01 -0.015237790 -0.007524970
#> 2001-04-01 -0.038805610 -0.015345339
#> 2001-05-01 -0.028823020 -0.018664279
#> 2001-06-01 -0.077892483 -0.037589159
#> 2001-07-01 -0.027743863 -0.034447949
#> 2001-08-01 -0.023744286 -0.022403836
#> 2001-09-01 -0.024297721 -0.018468448
#> 2001-10-01 -0.021330602 -0.017828243
#> 2001-11-01 -0.015575267 -0.012428051
#> 2001-12-01 -0.028485712 -0.009946958

#Queries the tables in 'Plants'
names(S1$Plants)
#>  [1] "LAI"                 "LAIlive"             "FPAR"               
#>  [4] "AbsorbedSWRFraction" "Transpiration"       "GrossPhotosynthesis"
#>  [7] "PlantPsi"            "LeafPLC"             "StemPLC"            
#> [10] "PlantWaterBalance"   "LeafRWC"             "StemRWC"            
#> [13] "LFMC"                "PlantStress"        

#Monthly summary (averages) of plant stress
summary(S1, freq="months",FUN=mean, output="Plants$PlantStress", 
        bySpecies = TRUE)
#>            Pinus halepensis Quercus coccifera Quercus ilex
#> 2001-01-01      0.001521178      0.0010354372  0.004659389
#> 2001-02-01      0.001978141      0.0010907108  0.004144834
#> 2001-03-01      0.001876802      0.0009820193  0.003532614
#> 2001-04-01      0.006790745      0.0030584035  0.008808939
#> 2001-05-01      0.004548142      0.0022628980  0.007467638
#> 2001-06-01      0.018512099      0.0076247038  0.019214902
#> 2001-07-01      0.006119201      0.0031439097  0.010553139
#> 2001-08-01      0.003688111      0.0020168177  0.007334173
#> 2001-09-01      0.003622975      0.0019135744  0.006758543
#> 2001-10-01      0.003043847      0.0016671624  0.006145933
#> 2001-11-01      0.002133317      0.0011737718  0.004426598
#> 2001-12-01      0.004131452      0.0019537123  0.006025717