Summarize simulation results

Function summary summarizes the model's output in different temporal steps (i.e. weekly, annual, ...).

# S3 method for spwb
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  ...
)

# S3 method for pwb
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  ...
)

# S3 method for growth
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  ...
)

# S3 method for fordyn
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  ...
)

Arguments

object

An object of class spwb, pwb, growth or fordyn.

freq

Frequency of summary statistics (see cut.Date).

output

The data table to be summarized. Accepted values are the path to data tables in object, such as 'WaterBalance', 'Soil', 'Stand' or 'Plants$LAI'. It is also possible to use strings like 'Transpiration' and the function will interpret it as 'Plants$Transpiration'.

FUN

The function to summarize results (e.g., sum, mean, ...)

bySpecies

Allows aggregating output by species before calculating summaries (only has an effect with some values of output). Aggregation can involve a sum (as for plant lai or transpiration) or a LAI-weighted mean (as for plant stress or plant water potential).

...

Additional parameters for function summary.

Value

A matrix with dates as row names and the desired summaries in columns

Note

When applied to fordyn objects, the summary function can be used to gather the results of different yearly steps into a single table while keeping a daily resolution (i.e. using freq = "days".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Call simulation function
S1<-spwb(x, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 5.24843
#> Initial soil water content (mm): 180.814
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:....................................
#> 
#> Final plant water content (mm): 5.24653
#> Final soil water content (mm): 165.066
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): -0.00189302
#> Plant water balance result (mm): -0.00189302
#> Change in soil water content (mm): -15.7481
#> Soil water balance result (mm): -15.7481
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): -7.10543e-15
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 84 Net rainfall (mm) 378
#>   Infiltration (mm) 418 Runoff (mm) 12 Deep drainage (mm) 167
#>   Soil evaporation (mm) 28 Transpiration (mm) 238
#>   Plant extraction from soil (mm) 238  Plant water balance (mm) -0 Hydraulic redistribution (mm) 0

#Monthly summary (averages) of soil status
summary(S1, freq="months",FUN=mean, output="Soil")
#>                  W.1       W.2     ML.1      ML.2    MLTot       WTD        SWE
#> 2001-01-01 0.9878614 1.0020937 63.38197 116.89721 180.2792  995.0012 1.63052532
#> 2001-02-01 0.9417692 0.9830714 60.42466 114.67820 175.1029 1000.0000 0.27022147
#> 2001-03-01 0.9521763 0.9913322 61.09239 115.64185 176.7342  996.2334 0.01762496
#> 2001-04-01 0.8825729 0.9430289 56.62658 110.00712 166.6337 1000.0000 0.57265738
#> 2001-05-01 0.9038850 0.9646507 57.99398 112.52937 170.5234  999.9492 0.00000000
#> 2001-06-01 0.7147850 0.8205028 45.86117  95.71409 141.5753 1000.0000 0.00000000
#> 2001-07-01 0.9017732 0.9003682 57.85848 105.03063 162.8891  999.6553 0.00000000
#> 2001-08-01 0.9204971 0.9711890 59.05983 113.29208 172.3519  998.7708 0.00000000
#> 2001-09-01 0.9239069 0.9708433 59.27860 113.25176 172.5304  999.7093 0.00000000
#> 2001-10-01 0.9536245 0.9865249 61.18531 115.08106 176.2664  997.9094 0.00000000
#> 2001-11-01 0.9471795 0.9865112 60.77179 115.07946 175.8513  994.8612 2.54346786
#> 2001-12-01 0.8948944 0.9674168 57.41714 112.85204 170.2692 1000.0000 0.00000000
#>            PlantExt.1 PlantExt.2 HydraulicInput.1 HydraulicInput.2       psi.1
#> 2001-01-01  0.1539543 0.09117518                0       0.00000000 -0.03525776
#> 2001-02-01  0.2965052 0.21356236                0       0.00000000 -0.04555643
#> 2001-03-01  0.3126607 0.21775961                0       0.00000000 -0.04233010
#> 2001-04-01  0.3783968 0.29506140                0       0.00000000 -0.06187302
#> 2001-05-01  0.4119697 0.38278590                0       0.00000000 -0.05842904
#> 2001-06-01  0.3858437 0.70215437                0       0.00000000 -0.17598380
#> 2001-07-01  0.6727943 0.27983021                0       0.00878272 -0.06409041
#> 2001-08-01  0.5635902 0.40582093                0       0.00000000 -0.05075655
#> 2001-09-01  0.4407633 0.32019455                0       0.00000000 -0.04965088
#> 2001-10-01  0.3280538 0.21878270                0       0.00000000 -0.04224043
#> 2001-11-01  0.2288457 0.16582880                0       0.00000000 -0.04386082
#> 2001-12-01  0.1921260 0.17589524                0       0.00000000 -0.05812452
#>                  psi.2
#> 2001-01-01 -0.03270981
#> 2001-02-01 -0.03604391
#> 2001-03-01 -0.03439069
#> 2001-04-01 -0.04365533
#> 2001-05-01 -0.03960643
#> 2001-06-01 -0.08507060
#> 2001-07-01 -0.06017357
#> 2001-08-01 -0.03802744
#> 2001-09-01 -0.03796308
#> 2001-10-01 -0.03529464
#> 2001-11-01 -0.03526094
#> 2001-12-01 -0.03885259

#Queries the tables in 'Plants'
names(S1$Plants)
#>  [1] "LAI"                 "LAIlive"             "FPAR"               
#>  [4] "AbsorbedSWRFraction" "Transpiration"       "GrossPhotosynthesis"
#>  [7] "PlantPsi"            "StemPLC"             "PlantWaterBalance"  
#> [10] "LeafRWC"             "StemRWC"             "LFMC"               
#> [13] "PlantStress"        

#Monthly summary (averages) of plant stress
summary(S1, freq="months",FUN=mean, output="Plants$PlantStress", 
        bySpecies = TRUE)
#>            Pinus halepensis Quercus coccifera Quercus ilex
#> 2001-01-01      0.006219651       0.003234789   0.01087673
#> 2001-02-01      0.008762839       0.004242376   0.01313671
#> 2001-03-01      0.007889923       0.003884532   0.01228873
#> 2001-04-01      0.013434662       0.006069779   0.01719522
#> 2001-05-01      0.012461956       0.005598605   0.01585864
#> 2001-06-01      0.058388612       0.021321107   0.04423149
#> 2001-07-01      0.016201882       0.007272283   0.02010738
#> 2001-08-01      0.009990978       0.004713050   0.01415017
#> 2001-09-01      0.009822741       0.004655513   0.01405405
#> 2001-10-01      0.007839930       0.003892959   0.01240521
#> 2001-11-01      0.008406322       0.004090989   0.01276317
#> 2001-12-01      0.012389011       0.005600295   0.01593545