Summarises simulation results

Function summary summarizes the model's output in different temporal steps (i.e. weekly, annual, ...).

# S3 method for pwb
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for spwb
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for growth
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for fordyn
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)

Arguments

object

An object of class spwb, pwb, growth or fordyn.

freq

Frequency of summary statistics (see cut.Date).

output

The data table to be summarized. Accepted values are the path to data tables in object, such as 'WaterBalance', 'Soil', 'Stand' or 'Plants$LAI'. It is also possible to use strings like 'Transpiration' and the function will interpret it as 'Plants$Transpiration'.

FUN

The function to summarize results (e.g., sum, mean, ...)

bySpecies

Allows aggregating output by species before calculating summaries (only has an effect with some values of output). Aggregation can involve a sum (as for plant lai or transpiration) or a LAI-weighted mean (as for plant stress or plant water potential).

...

Additional parameters for function summary.

Author

Miquel De Cáceres Ainsa, CREAF

Note

When applied to fordyn objects, the summary function can be used to gather the results of different yearly steps into a single table while keeping a daily resolution (i.e. using freq = "days".

See also

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Call simulation function
S1<-spwb(x, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial soil water content (mm): 185.069
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:....................................
#> 
#> Final soil water content (mm): 162.999
#> Final snowpack content (mm): 0
#> Change in soil water content (mm): -22.07
#> Soil water balance result (mm): -22.07
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 7.10543e-15
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 98 Net rainfall (mm) 364
#>   Infiltration (mm) 406 Runoff (mm) 9 Deep drainage (mm) 67
#>   Soil evaporation (mm) 38 Transpiration (mm) 323
        
#Monthly summary (averages) of soil status
summary(S1, freq="months",FUN=mean, output="Soil")
#>                  W.1       W.2     ML.1      ML.2    MLTot       WTD        SWE
#> 2001-01-01 0.9913915 0.9981842 67.67564 116.59397 184.2696  997.6859 1.68350931
#> 2001-02-01 0.9686275 0.9380917 66.12170 109.57481 175.6965 1000.0000 0.28233380
#> 2001-03-01 0.9722558 0.9643247 66.36938 112.63898 179.0084  998.7165 0.01762496
#> 2001-04-01 0.9468318 0.8289361 64.63385  96.82478 161.4586 1000.0000 0.58931434
#> 2001-05-01 0.9453553 0.7912195 64.53306  92.41924 156.9523 1000.0000 0.00000000
#> 2001-06-01 0.8201550 0.5159670 55.98648  60.26808 116.2546 1000.0000 0.00000000
#> 2001-07-01 0.9528458 0.5940383 65.04439  69.38728 134.4317 1000.0000 0.00000000
#> 2001-08-01 0.9601295 0.6273318 65.54160  73.27617 138.8178 1000.0000 0.00000000
#> 2001-09-01 0.9614921 0.6252447 65.63461  73.03238 138.6670 1000.0000 0.00000000
#> 2001-10-01 0.9741574 0.6163589 66.49919  71.99446 138.4936 1000.0000 0.00000000
#> 2001-11-01 0.9696225 0.7818339 66.18962  91.32295 157.5126 1000.0000 2.67673914
#> 2001-12-01 0.9359150 0.9143177 63.88864 106.79786 170.6865 1000.0000 0.00000000
#>            PlantExt.1 PlantExt.2       psi.1       psi.2
#> 2001-01-01 0.06754553  0.2726358 -0.03461061 -0.03344571
#> 2001-02-01 0.14089772  0.5687079 -0.03919761 -0.04785382
#> 2001-03-01 0.14640284  0.5909292 -0.03807724 -0.04030147
#> 2001-04-01 0.18682627  0.7540756 -0.04389376 -0.09099102
#> 2001-05-01 0.22051801  0.8899767 -0.04542827 -0.12588801
#> 2001-06-01 0.31437956  1.0443542 -0.09421692 -1.45868540
#> 2001-07-01 0.26536399  0.9191011 -0.04365587 -0.99605206
#> 2001-08-01 0.26809804  1.0798166 -0.04074617 -0.38628039
#> 2001-09-01 0.21041105  0.8477380 -0.04068425 -0.38646229
#> 2001-10-01 0.15092845  0.6064069 -0.03793903 -0.45211952
#> 2001-11-01 0.10901764  0.4397916 -0.03885597 -0.18866400
#> 2001-12-01 0.10196384  0.4115585 -0.04682423 -0.05388003

#Queries the tables in 'Plants'
names(S1$Plants)
#> [1] "LAI"                 "AbsorbedSWRFraction" "Transpiration"      
#> [4] "GrossPhotosynthesis" "PlantPsi"            "StemPLC"            
#> [7] "PlantStress"        

#Monthly summary (averages) of plant stress
summary(S1, freq="months",FUN=mean, output="Plants$PlantStress", 
        bySpecies = TRUE)
#>            Pinus halepensis Quercus coccifera Quercus ilex
#> 2001-01-01     1.465921e-06      6.528932e-07 9.526716e-07
#> 2001-02-01     4.111547e-06      1.413059e-06 2.672012e-06
#> 2001-03-01     2.604064e-06      1.022797e-06 1.692329e-06
#> 2001-04-01     2.531000e-05      6.577862e-06 1.644856e-05
#> 2001-05-01     9.337284e-05      2.267016e-05 6.068460e-05
#> 2001-06-01     1.478104e-01      4.296831e-02 1.089763e-01
#> 2001-07-01     1.234626e-01      3.779792e-02 9.377075e-02
#> 2001-08-01     2.171432e-03      5.113801e-04 1.412640e-03
#> 2001-09-01     2.008400e-03      4.727706e-04 1.306136e-03
#> 2001-10-01     4.425445e-03      1.043334e-03 2.881948e-03
#> 2001-11-01     5.104809e-04      1.203917e-04 3.318247e-04
#> 2001-12-01     5.866276e-06      2.150647e-06 3.812376e-06