Summarises simulation results

Function summary summarizes the model's output in different temporal steps (i.e. weekly, annual, ...).

# S3 method for pwb
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for spwb
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for growth
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for fordyn
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)

Arguments

object

An object of class spwb, pwb, growth or fordyn.

freq

Frequency of summary statistics (see cut.Date).

output

The data table to be summarized. Accepted values are the path to data tables in object, such as 'WaterBalance', 'Soil', 'Stand' or 'Plants$LAI'. It is also possible to use strings like 'Transpiration' and the function will interpret it as 'Plants$Transpiration'.

FUN

The function to summarize results (e.g., sum, mean, ...)

bySpecies

Allows aggregating output by species before calculating summaries (only has an effect with some values of output). Aggregation can involve a sum (as for plant lai or transpiration) or a LAI-weighted mean (as for plant stress or plant water potential).

...

Additional parameters for function summary.

Author

Miquel De Cáceres Ainsa, CREAF

Note

When applied to fordyn objects, the summary function can be used to gather the results of different yearly steps into a single table while keeping a daily resolution (i.e. using freq = "days".

See also

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Call simulation function
S1<-spwb(x, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial soil water content (mm): 185.069
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:....................................
#> 
#> Final soil water content (mm): 163.428
#> Final snowpack content (mm): 0
#> Change in soil water content (mm): -21.6413
#> Soil water balance result (mm): -21.6413
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 7.10543e-15
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 98 Net rainfall (mm) 364
#>   Infiltration (mm) 406 Runoff (mm) 9 Deep drainage (mm) 75
#>   Soil evaporation (mm) 19 Transpiration (mm) 334
        
#Monthly summary (averages) of soil status
summary(S1, freq="months",FUN=mean, output="Soil")
#>                  W.1       W.2     ML.1      ML.2    MLTot       WTD        SWE
#> 2001-01-01 0.9887216 0.9996492 67.49339 116.76510 184.2585  997.5148 1.68350931
#> 2001-02-01 0.9464275 0.9542337 64.60625 111.46030 176.0665 1000.0000 0.28233380
#> 2001-03-01 0.9582412 0.9737577 65.41270 113.74082 179.1535  998.6065 0.01762496
#> 2001-04-01 0.8808286 0.8730534 60.12826 101.97794 162.1062 1000.0000 0.58931434
#> 2001-05-01 0.8749120 0.8537155 59.72437  99.71916 159.4435 1000.0000 0.00000000
#> 2001-06-01 0.6346222 0.6412969 43.32140  74.90737 118.2288 1000.0000 0.00000000
#> 2001-07-01 0.8016641 0.6293358 54.72423  73.51025 128.2345 1000.0000 0.00000000
#> 2001-08-01 0.7713983 0.7068250 52.65819  82.56145 135.2196 1000.0000 0.00000000
#> 2001-09-01 0.7730894 0.7366794 52.77362  86.04863 138.8223 1000.0000 0.00000000
#> 2001-10-01 0.8197846 0.7390651 55.96119  86.32730 142.2885 1000.0000 0.00000000
#> 2001-11-01 0.8942471 0.8605764 61.04425 100.52055 161.5648  999.5447 2.67673914
#> 2001-12-01 0.9106210 0.9317573 62.16199 108.83492 170.9969 1000.0000 0.00000000
#>            PlantExt.1 PlantExt.2       psi.1       psi.2
#> 2001-01-01  0.1333472 0.20683408 -0.03514255 -0.03315732
#> 2001-02-01  0.2905050 0.41910106 -0.04495338 -0.04306089
#> 2001-03-01  0.2771247 0.46020742 -0.04122974 -0.03808747
#> 2001-04-01  0.4656291 0.47528251 -0.06482327 -0.06792897
#> 2001-05-01  0.5593032 0.55127692 -0.07431900 -0.07822713
#> 2001-06-01  0.6257754 0.95157473 -0.41741626 -0.40307406
#> 2001-07-01  1.2537605 0.07775459 -0.18729988 -0.39954580
#> 2001-08-01  1.1023292 0.24765175 -0.13915341 -0.19444554
#> 2001-09-01  0.7186365 0.34095722 -0.12931259 -0.15805921
#> 2001-10-01  0.6034710 0.15658305 -0.10604410 -0.15428852
#> 2001-11-01  0.3499309 0.19910266 -0.06469047 -0.08515563
#> 2001-12-01  0.1753682 0.33815432 -0.05443934 -0.04849098

#Queries the tables in 'Plants'
names(S1$Plants)
#> [1] "LAI"                 "AbsorbedSWRFraction" "Transpiration"      
#> [4] "GrossPhotosynthesis" "PlantPsi"            "StemPLC"            
#> [7] "PlantStress"        

#Monthly summary (averages) of plant stress
summary(S1, freq="months",FUN=mean, output="Plants$PlantStress", 
        bySpecies = TRUE)
#>            Pinus halepensis Quercus coccifera Quercus ilex
#> 2001-01-01     1.519451e-06      6.602610e-07 9.874601e-07
#> 2001-02-01     3.359350e-06      1.452657e-06 2.183173e-06
#> 2001-03-01     2.453217e-06      1.073528e-06 1.594296e-06
#> 2001-04-01     1.159852e-05      4.852969e-06 7.537656e-06
#> 2001-05-01     2.163467e-05      9.267828e-06 1.406003e-05
#> 2001-06-01     4.181031e-03      1.779701e-03 2.722797e-03
#> 2001-07-01     4.011877e-03      1.572184e-03 2.616789e-03
#> 2001-08-01     2.283659e-04      8.902669e-05 1.484199e-04
#> 2001-09-01     1.290705e-04      5.211339e-05 8.388297e-05
#> 2001-10-01     1.133980e-04      4.426857e-05 7.369715e-05
#> 2001-11-01     3.149432e-05      1.200132e-05 2.046776e-05
#> 2001-12-01     5.359564e-06      2.376255e-06 3.483073e-06