Summarize simulation results

Function summary summarizes the model's output in different temporal steps (i.e. weekly, annual, ...).

# S3 method for spwb
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for pwb
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for growth
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for fordyn
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

Arguments

object

An object of class spwb, pwb, growth or fordyn.

freq

Frequency of summary statistics (see cut.Date).

output

The data table to be summarized. Accepted values are the path to data tables in object, such as 'WaterBalance', 'Soil', 'Stand' or 'Plants$LAI'. It is also possible to use strings like 'Transpiration' and the function will interpret it as 'Plants$Transpiration'.

FUN

The function to summarize results (e.g., sum, mean, ...)

bySpecies

Allows aggregating output by species before calculating summaries (only has an effect with some values of output). Aggregation can involve a sum (as for plant lai or transpiration) or a LAI-weighted mean (as for plant stress or plant water potential).

months

A vector of month numbers (1 to 12) to subset the season where summaries apply.

...

Additional parameters for function summary.

Value

A matrix with dates as row names and the desired summaries in columns

Note

When applied to fordyn objects, the summary function can be used to gather the results of different yearly steps into a single table while keeping a daily resolution (i.e. using freq = "days".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Call simulation function
S1<-spwb(x, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 5.25158
#> Initial soil water content (mm): 180.814
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:....................................
#> 
#> Final plant water content (mm): 5.24941
#> Final soil water content (mm): 163.985
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): -0.00216877
#> Plant water balance result (mm): -0.00216877
#> Change in soil water content (mm): -16.8285
#> Soil water balance result (mm): -16.8285
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): -7.10543e-15
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 93 Net rainfall (mm) 369
#>   Infiltration (mm) 409 Runoff (mm) 11 Deep drainage (mm) 144
#>   Soil evaporation (mm) 26  Herbaceous transpiration (mm) 18 Woody plant transpiration (mm) 238
#>   Plant extraction from soil (mm) 238  Plant water balance (mm) -0 Hydraulic redistribution (mm) 0

#Monthly summary (averages) of soil status
summary(S1, freq="months",FUN=mean, output="Soil")
#>                  W.1       W.2     ML.1      ML.2    MLTot       WTD        SWE
#> 2001-01-01 0.9865971 1.0019234 63.30085 116.87734 180.1782  995.2311 1.65619239
#> 2001-02-01 0.9353627 0.9821106 60.01362 114.56612 174.5797 1000.0000 0.27608905
#> 2001-03-01 0.9471080 0.9906257 60.76720 115.55943 176.3266  996.5180 0.01762496
#> 2001-04-01 0.8628512 0.9370395 55.36122 109.30844 164.6697 1000.0000 0.58072652
#> 2001-05-01 0.8921467 0.9545907 57.24084 111.35584 168.5967 1000.0000 0.00000000
#> 2001-06-01 0.6876293 0.8021609 44.11884  93.57445 137.6933 1000.0000 0.00000000
#> 2001-07-01 0.8861677 0.8718458 56.85723 101.70340 158.5606 1000.0000 0.00000000
#> 2001-08-01 0.9101232 0.9646157 58.39423 112.52529 170.9195  999.1205 0.00000000
#> 2001-09-01 0.9134073 0.9680610 58.60494 112.92719 171.5321  999.8309 0.00000000
#> 2001-10-01 0.9454854 0.9755010 60.66309 113.79509 174.4582  998.3117 0.00000000
#> 2001-11-01 0.9408182 0.9856342 60.36364 114.97716 175.3408  995.0667 2.60802857
#> 2001-12-01 0.8863424 0.9667605 56.86843 112.77548 169.6439 1000.0000 0.00000000
#>            PlantExt.1 PlantExt.2 HydraulicInput.1 HydraulicInput.2       psi.1
#> 2001-01-01  0.1538525 0.09186406                0     0.0000000000 -0.03550470
#> 2001-02-01  0.2909985 0.22008959                0     0.0000000000 -0.04734661
#> 2001-03-01  0.3083333 0.22321226                0     0.0000000000 -0.04361680
#> 2001-04-01  0.3430431 0.33074789                0     0.0000000000 -0.06916191
#> 2001-05-01  0.4101370 0.38520615                0     0.0000000000 -0.06320819
#> 2001-06-01  0.3220179 0.75053475                0     0.0000000000 -0.21469986
#> 2001-07-01  0.7137496 0.23667686                0     0.0100769793 -0.07459164
#> 2001-08-01  0.5473776 0.42351877                0     0.0001094234 -0.05411182
#> 2001-09-01  0.4282888 0.33398906                0     0.0000000000 -0.05257718
#> 2001-10-01  0.3315497 0.21630201                0     0.0000000000 -0.04417116
#> 2001-11-01  0.2248740 0.17060458                0     0.0000000000 -0.04546799
#> 2001-12-01  0.1857873 0.18286375                0     0.0000000000 -0.06111954
#>                  psi.2
#> 2001-01-01 -0.03273661
#> 2001-02-01 -0.03622713
#> 2001-03-01 -0.03450840
#> 2001-04-01 -0.04503304
#> 2001-05-01 -0.04157872
#> 2001-06-01 -0.09474213
#> 2001-07-01 -0.07068776
#> 2001-08-01 -0.03928111
#> 2001-09-01 -0.03855599
#> 2001-10-01 -0.03716809
#> 2001-11-01 -0.03540887
#> 2001-12-01 -0.03899233

#Queries the tables in 'Plants'
names(S1$Plants)
#>  [1] "LAI"                 "LAIlive"             "FPAR"               
#>  [4] "AbsorbedSWRFraction" "Transpiration"       "GrossPhotosynthesis"
#>  [7] "PlantPsi"            "StemPLC"             "PlantWaterBalance"  
#> [10] "LeafRWC"             "StemRWC"             "LFMC"               
#> [13] "PlantStress"        

#Monthly summary (averages) of plant stress
summary(S1, freq="months",FUN=mean, output="Plants$PlantStress", 
        bySpecies = TRUE)
#>            Pinus halepensis Quercus coccifera Quercus ilex
#> 2001-01-01      0.006277695       0.003257282   0.01092484
#> 2001-02-01      0.009254163       0.004421571   0.01348780
#> 2001-03-01      0.008235284       0.004011016   0.01253786
#> 2001-04-01      0.015731098       0.006886059   0.01874593
#> 2001-05-01      0.014051024       0.006179163   0.01703490
#> 2001-06-01      0.077638287       0.027296463   0.05299757
#> 2001-07-01      0.020946054       0.009048435   0.02369659
#> 2001-08-01      0.010997675       0.005092808   0.01494887
#> 2001-09-01      0.010660528       0.004964087   0.01467425
#> 2001-10-01      0.008386902       0.004136477   0.01304662
#> 2001-11-01      0.008845087       0.004251086   0.01307684
#> 2001-12-01      0.013294946       0.005915547   0.01650689