Summarises simulation results

Function summary summarizes the model's output in different temporal steps (i.e. weekly, annual, ...).

# S3 method for pwb
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for spwb
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for growth
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for fordyn
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)

Arguments

object

An object of class spwb, pwb, growth or fordyn.

freq

Frequency of summary statistics (see cut.Date).

output

The data table to be summarized. Accepted values are the path to data tables in object, such as 'WaterBalance', 'Soil', 'Stand' or 'Plants$LAI'. It is also possible to use strings like 'Transpiration' and the function will interpret it as 'Plants$Transpiration'.

FUN

The function to summarize results (e.g., sum, mean, ...)

bySpecies

Allows aggregating output by species before calculating summaries (only has an effect with some values of output). Aggregation can involve a sum (as for plant lai or transpiration) or a LAI-weighted mean (as for plant stress or plant water potential).

...

Additional parameters for function summary.

Author

Miquel De Cáceres Ainsa, CREAF

Note

When applied to fordyn objects, the summary function can be used to gather the results of different yearly steps into a single table while keeping a daily resolution (i.e. using freq = "days".

See also

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Call simulation function
S1<-spwb(x, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial soil water content (mm): 185.069
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:....................................
#> 
#> Final soil water content (mm): 162.655
#> Final snowpack content (mm): 0
#> Change in soil water content (mm): -22.4148
#> Soil water balance result (mm): -22.4148
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 96 Net rainfall (mm) 366
#>   Infiltration (mm) 408 Runoff (mm) 9 Deep drainage (mm) 69
#>   Soil evaporation (mm) 13 Transpiration (mm) 349
#>   Plant extraction from soil (mm) 349  Plant water balance (mm) -0
        
#Monthly summary (averages) of soil status
summary(S1, freq="months",FUN=mean, output="Soil")
#>                  W.1       W.2     ML.1      ML.2    MLTot       WTD        SWE
#> 2001-01-01 0.9832822 1.0016501 67.12208 116.99882 184.1209  997.1300 1.67344860
#> 2001-02-01 0.9065841 0.9769088 61.88642 114.10888 175.9953 1000.0000 0.28003388
#> 2001-03-01 0.9299882 0.9870444 63.48406 115.29278 178.7768  998.3204 0.01762496
#> 2001-04-01 0.8009750 0.9083463 54.67719 106.10037 160.7776 1000.0000 0.58615148
#> 2001-05-01 0.8213315 0.8723889 56.06679 101.90032 157.9671 1000.0000 0.00000000
#> 2001-06-01 0.5631482 0.6566966 38.44235  76.70615 115.1485 1000.0000 0.00000000
#> 2001-07-01 0.7653479 0.6231808 52.24517  72.79130 125.0365 1000.0000 0.00000000
#> 2001-08-01 0.7240714 0.6966422 49.42750  81.37205 130.7995 1000.0000 0.00000000
#> 2001-09-01 0.7078134 0.7240672 48.31767  84.57545 132.8931 1000.0000 0.00000000
#> 2001-10-01 0.7495555 0.7228324 51.16712  84.43121 135.5983 1000.0000 0.00000000
#> 2001-11-01 0.8535358 0.8419937 58.26516  98.34998 156.6151  999.9129 2.65143331
#> 2001-12-01 0.8567344 0.9586739 58.48351 111.97894 170.4624 1000.0000 0.00000000
#>            PlantExt.1 PlantExt.2       psi.1       psi.2
#> 2001-01-01  0.2575902 0.09885972 -0.03628895 -0.03277044
#> 2001-02-01  0.4609344 0.28253501 -0.05807667 -0.03771587
#> 2001-03-01  0.5053515 0.26731852 -0.04944538 -0.03532034
#> 2001-04-01  0.5094845 0.47633996 -0.10671551 -0.05541405
#> 2001-05-01  0.6786692 0.48466667 -0.11163276 -0.06912534
#> 2001-06-01  0.5503846 1.06320789 -0.74863890 -0.37480453
#> 2001-07-01  1.2736701 0.11092708 -0.27290215 -0.41348456
#> 2001-08-01  1.1501403 0.26405596 -0.19617138 -0.20865550
#> 2001-09-01  0.7776314 0.33220513 -0.21107919 -0.17184865
#> 2001-10-01  0.6116378 0.18466191 -0.17405725 -0.17285644
#> 2001-11-01  0.4412727 0.13415875 -0.09063391 -0.09736424
#> 2001-12-01  0.2972963 0.24068533 -0.07642070 -0.04153670

#Queries the tables in 'Plants'
names(S1$Plants)
#>  [1] "LAI"                 "LAIlive"             "FPAR"               
#>  [4] "AbsorbedSWRFraction" "Transpiration"       "GrossPhotosynthesis"
#>  [7] "PlantPsi"            "StemPLC"             "PlantWaterBalance"  
#> [10] "LeafRWC"             "StemRWC"             "LFMC"               
#> [13] "PlantStress"        

#Monthly summary (averages) of plant stress
summary(S1, freq="months",FUN=mean, output="Plants$PlantStress", 
        bySpecies = TRUE)
#>            Pinus halepensis Quercus coccifera Quercus ilex
#> 2001-01-01     2.505682e-06      3.079671e-06 1.222923e-06
#> 2001-02-01     1.108586e-05      1.224487e-05 4.186065e-06
#> 2001-03-01     7.106522e-06      7.981222e-06 2.800287e-06
#> 2001-04-01     6.286930e-05      6.765256e-05 2.215326e-05
#> 2001-05-01     1.218801e-04      1.314666e-04 4.323296e-05
#> 2001-06-01     2.788541e-02      2.998053e-02 1.024379e-02
#> 2001-07-01     1.215136e-02      1.424668e-02 5.746169e-03
#> 2001-08-01     5.523768e-04      7.020057e-04 2.902723e-04
#> 2001-09-01     6.238683e-04      7.163349e-04 2.599470e-04
#> 2001-10-01     4.360304e-04      5.258158e-04 2.039415e-04
#> 2001-11-01     7.720218e-05      1.038285e-04 4.561465e-05
#> 2001-12-01     2.414221e-05      2.606227e-05 8.580527e-06