Summarize simulation results

Function summary summarizes the model's output in different temporal steps (i.e. weekly, annual, ...).

# S3 method for spwb
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  ...
)

# S3 method for pwb
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  ...
)

# S3 method for growth
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  ...
)

# S3 method for fordyn
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  ...
)

Arguments

object

An object of class spwb, pwb, growth or fordyn.

freq

Frequency of summary statistics (see cut.Date).

output

The data table to be summarized. Accepted values are the path to data tables in object, such as 'WaterBalance', 'Soil', 'Stand' or 'Plants$LAI'. It is also possible to use strings like 'Transpiration' and the function will interpret it as 'Plants$Transpiration'.

FUN

The function to summarize results (e.g., sum, mean, ...)

bySpecies

Allows aggregating output by species before calculating summaries (only has an effect with some values of output). Aggregation can involve a sum (as for plant lai or transpiration) or a LAI-weighted mean (as for plant stress or plant water potential).

...

Additional parameters for function summary.

Value

A matrix with dates as row names and the desired summaries in columns

Note

When applied to fordyn objects, the summary function can be used to gather the results of different yearly steps into a single table while keeping a daily resolution (i.e. using freq = "days".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Call simulation function
S1<-spwb(x, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 5.2515
#> Initial soil water content (mm): 180.814
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:....................................
#> 
#> Final plant water content (mm): 5.24921
#> Final soil water content (mm): 163.518
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): -0.00229536
#> Plant water balance result (mm): -0.00229536
#> Change in soil water content (mm): -17.2955
#> Soil water balance result (mm): -17.2955
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 98 Net rainfall (mm) 365
#>   Infiltration (mm) 405 Runoff (mm) 11 Deep drainage (mm) 135
#>   Soil evaporation (mm) 24  Herbaceous transpiration (mm) 25 Woody plant transpiration (mm) 238
#>   Plant extraction from soil (mm) 238  Plant water balance (mm) -0 Hydraulic redistribution (mm) 0

#Monthly summary (averages) of soil status
summary(S1, freq="months",FUN=mean, output="Soil")
#>                  W.1       W.2     ML.1      ML.2    MLTot       WTD        SWE
#> 2001-01-01 0.9860826 1.0018515 63.26784 116.86895 180.1368  995.3304 1.66699065
#> 2001-02-01 0.9324581 0.9818352 59.82725 114.53399 174.3612 1000.0000 0.27855757
#> 2001-03-01 0.9448698 0.9903422 60.62359 115.52636 176.1500  996.6397 0.01762496
#> 2001-04-01 0.8537784 0.9347231 54.77910 109.03823 163.8173 1000.0000 0.58412125
#> 2001-05-01 0.8850188 0.9443839 56.78351 110.16518 166.9487 1000.0000 0.00000000
#> 2001-06-01 0.6726615 0.7892363 43.15850  92.06676 135.2253 1000.0000 0.00000000
#> 2001-07-01 0.8760547 0.8484897 56.20837  98.97884 155.1872 1000.0000 0.00000000
#> 2001-08-01 0.8974544 0.9441479 57.58139 110.13766 167.7190  999.5187 0.00000000
#> 2001-09-01 0.9085810 0.9667812 58.29528 112.77790 171.0732  999.9007 0.00000000
#> 2001-10-01 0.9411489 0.9710700 60.38486 113.27820 173.6631  998.4700 0.00000000
#> 2001-11-01 0.9379950 0.9852753 60.18250 114.93529 175.1178  995.1574 2.63518955
#> 2001-12-01 0.8825829 0.9665027 56.62722 112.74541 169.3726 1000.0000 0.00000000
#>            PlantExt.1 PlantExt.2 HydraulicInput.1 HydraulicInput.2       psi.1
#> 2001-01-01  0.1536527 0.09204362                0     0.0000000000 -0.03560758
#> 2001-02-01  0.2879824 0.22296670                0     0.0000000000 -0.04818785
#> 2001-03-01  0.3060295 0.22540347                0     0.0000000000 -0.04420613
#> 2001-04-01  0.3314256 0.34165041                0     0.0000000000 -0.07283462
#> 2001-05-01  0.4120807 0.38224269                0     0.0000000000 -0.06635934
#> 2001-06-01  0.2930631 0.76674549                0     0.0000000000 -0.24040425
#> 2001-07-01  0.7550046 0.19207672                0     0.0118857851 -0.08181582
#> 2001-08-01  0.5707472 0.39877925                0     0.0002327855 -0.05837790
#> 2001-09-01  0.4226708 0.33929623                0     0.0000000000 -0.05398254
#> 2001-10-01  0.3361690 0.21147843                0     0.0000000000 -0.04525563
#> 2001-11-01  0.2228126 0.17256617                0     0.0000000000 -0.04621143
#> 2001-12-01  0.1827207 0.18577998                0     0.0000000000 -0.06249831
#>                  psi.2
#> 2001-01-01 -0.03274790
#> 2001-02-01 -0.03628156
#> 2001-03-01 -0.03455637
#> 2001-04-01 -0.04557568
#> 2001-05-01 -0.04370594
#> 2001-06-01 -0.10204772
#> 2001-07-01 -0.07893883
#> 2001-08-01 -0.04336758
#> 2001-09-01 -0.03882920
#> 2001-10-01 -0.03795723
#> 2001-11-01 -0.03546966
#> 2001-12-01 -0.03904793

#Queries the tables in 'Plants'
names(S1$Plants)
#>  [1] "LAI"                 "LAIlive"             "FPAR"               
#>  [4] "AbsorbedSWRFraction" "Transpiration"       "GrossPhotosynthesis"
#>  [7] "PlantPsi"            "StemPLC"             "PlantWaterBalance"  
#> [10] "LeafRWC"             "StemRWC"             "LFMC"               
#> [13] "PlantStress"        

#Monthly summary (averages) of plant stress
summary(S1, freq="months",FUN=mean, output="Plants$PlantStress", 
        bySpecies = TRUE)
#>            Pinus halepensis Quercus coccifera Quercus ilex
#> 2001-01-01      0.006302119       0.003266732   0.01094500
#> 2001-02-01      0.009488413       0.004505748   0.01364841
#> 2001-03-01      0.008396285       0.004069498   0.01265155
#> 2001-04-01      0.016918687       0.007299559   0.01950458
#> 2001-05-01      0.015154990       0.006598156   0.01794225
#> 2001-06-01      0.091118952       0.031461696   0.05894628
#> 2001-07-01      0.024346780       0.010335656   0.02632144
#> 2001-08-01      0.012359095       0.005662965   0.01634723
#> 2001-09-01      0.011071029       0.005113927   0.01497132
#> 2001-10-01      0.008689950       0.004265078   0.01336480
#> 2001-11-01      0.009051526       0.004325788   0.01322125
#> 2001-12-01      0.013720098       0.006062308   0.01676942