Summarises simulation results

Function summary summarizes the model's output in different temporal steps (i.e. weekly, annual, ...).

# S3 method for pwb
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for spwb
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for growth
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for fordyn
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)

Arguments

object

An object of class spwb, pwb, growth or fordyn.

freq

Frequency of summary statistics (see cut.Date).

output

The data table to be summarized. Accepted values are the path to data tables in object, such as 'WaterBalance', 'Soil', 'Stand' or 'Plants$LAI'. It is also possible to use strings like 'Transpiration' and the function will interpret it as 'Plants$Transpiration'.

FUN

The function to summarize results (e.g., sum, mean, ...)

bySpecies

Allows aggregating output by species before calculating summaries (only has an effect with some values of output). Aggregation can involve a sum (as for plant lai or transpiration) or a LAI-weighted mean (as for plant stress or plant water potential).

...

Additional parameters for function summary.

Author

Miquel De Cáceres Ainsa, CREAF

Note

When applied to fordyn objects, the summary function can be used to gather the results of different yearly steps into a single table while keeping a daily resolution (i.e. using freq = "days".

See also

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Call simulation function
S1<-spwb(x, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial soil water content (mm): 185.069
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:....................................
#> 
#> Final soil water content (mm): 162.687
#> Final snowpack content (mm): 0
#> Change in soil water content (mm): -22.3827
#> Soil water balance result (mm): -22.3827
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 96 Net rainfall (mm) 366
#>   Infiltration (mm) 408 Runoff (mm) 9 Deep drainage (mm) 70
#>   Soil evaporation (mm) 13 Transpiration (mm) 348
#>   Plant extraction from soil (mm) 348  Plant water balance (mm) -0
        
#Monthly summary (averages) of soil status
summary(S1, freq="months",FUN=mean, output="Soil")
#>                  W.1       W.2     ML.1      ML.2    MLTot       WTD        SWE
#> 2001-01-01 0.9833027 1.0016505 67.12348 116.99886 184.1223  997.1295 1.67344860
#> 2001-02-01 0.9067659 0.9769280 61.89882 114.11113 176.0099 1000.0000 0.28003388
#> 2001-03-01 0.9300918 0.9870560 63.49113 115.29414 178.7853  998.3194 0.01762496
#> 2001-04-01 0.8012960 0.9084721 54.69910 106.11506 160.8142 1000.0000 0.58615148
#> 2001-05-01 0.8218570 0.8726796 56.10266 101.93428 158.0369 1000.0000 0.00000000
#> 2001-06-01 0.5639605 0.6578172 38.49780  76.83705 115.3348 1000.0000 0.00000000
#> 2001-07-01 0.7685003 0.6266564 52.46036  73.19727 125.6576 1000.0000 0.00000000
#> 2001-08-01 0.7279500 0.7001550 49.69226  81.78236 131.4746 1000.0000 0.00000000
#> 2001-09-01 0.7119019 0.7278587 48.59676  85.01831 133.6151 1000.0000 0.00000000
#> 2001-10-01 0.7536363 0.7269056 51.44569  84.90699 136.3527 1000.0000 0.00000000
#> 2001-11-01 0.8560153 0.8452438 58.43442  98.72961 157.1640  999.8512 2.65143331
#> 2001-12-01 0.8569346 0.9587503 58.49717 111.98786 170.4850 1000.0000 0.00000000
#>            PlantExt.1 PlantExt.2       psi.1       psi.2
#> 2001-01-01  0.2571489 0.09887605 -0.03628448 -0.03277039
#> 2001-02-01  0.4603187 0.28226274 -0.05800649 -0.03771116
#> 2001-03-01  0.5045547 0.26719294 -0.04941034 -0.03531818
#> 2001-04-01  0.5090163 0.47561910 -0.10648854 -0.05536979
#> 2001-05-01  0.6782404 0.48366654 -0.11121666 -0.06899378
#> 2001-06-01  0.5471352 1.05376867 -0.74105813 -0.36931314
#> 2001-07-01  1.2690456 0.10912518 -0.26647796 -0.40112354
#> 2001-08-01  1.1488466 0.26345375 -0.19069801 -0.20336521
#> 2001-09-01  0.7779964 0.33038847 -0.20484915 -0.16728941
#> 2001-10-01  0.6107793 0.18448739 -0.16898706 -0.16797605
#> 2001-11-01  0.4402009 0.13452328 -0.08854206 -0.09494223
#> 2001-12-01  0.2968699 0.24046772 -0.07632018 -0.04151901

#Queries the tables in 'Plants'
names(S1$Plants)
#>  [1] "LAI"                 "LAIlive"             "FPAR"               
#>  [4] "AbsorbedSWRFraction" "Transpiration"       "GrossPhotosynthesis"
#>  [7] "PlantPsi"            "StemPLC"             "PlantWaterBalance"  
#> [10] "LeafRWC"             "StemRWC"             "LFMC"               
#> [13] "PlantStress"        

#Monthly summary (averages) of plant stress
summary(S1, freq="months",FUN=mean, output="Plants$PlantStress", 
        bySpecies = TRUE)
#>            Pinus halepensis Quercus coccifera Quercus ilex
#> 2001-01-01     2.505006e-06      3.078966e-06 3.697016e-06
#> 2001-02-01     1.104158e-05      1.219857e-05 1.261348e-05
#> 2001-03-01     7.087200e-06      7.961023e-06 8.448014e-06
#> 2001-04-01     6.247136e-05      6.723312e-05 6.657982e-05
#> 2001-05-01     1.204181e-04      1.299174e-04 1.292010e-04
#> 2001-06-01     2.696097e-02      2.896988e-02 2.879190e-02
#> 2001-07-01     1.139841e-02      1.332840e-02 1.515067e-02
#> 2001-08-01     5.068992e-04      6.452444e-04 8.076880e-04
#> 2001-09-01     5.692514e-04      6.542855e-04 7.184451e-04
#> 2001-10-01     3.981012e-04      4.803965e-04 5.636365e-04
#> 2001-11-01     7.088444e-05      9.539399e-05 1.267902e-04
#> 2001-12-01     2.404216e-05      2.595703e-05 2.584391e-05