Summarises simulation results

Function summary summarizes the model's output in different temporal steps (i.e. weekly, annual, ...).

# S3 method for pwb
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for spwb
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for growth
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for fordyn
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)

Arguments

object

An object of class spwb, pwb, growth or fordyn.

freq

Frequency of summary statistics (see cut.Date).

output

The data table to be summarized. Accepted values are the path to data tables in object, such as 'WaterBalance', 'Soil', 'Stand' or 'Plants$LAI'. It is also possible to use strings like 'Transpiration' and the function will interpret it as 'Plants$Transpiration'.

FUN

The function to summarize results (e.g., sum, mean, ...)

bySpecies

Allows aggregating output by species before calculating summaries (only has an effect with some values of output). Aggregation can involve a sum (as for plant lai or transpiration) or a LAI-weighted mean (as for plant stress or plant water potential).

...

Additional parameters for function summary.

Author

Miquel De Cáceres Ainsa, CREAF

Note

When applied to fordyn objects, the summary function can be used to gather the results of different yearly steps into a single table while keeping a daily resolution (i.e. using freq = "days".

See also

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Call simulation function
S1<-spwb(x, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial soil water content (mm): 185.069
#> Initial snowpack content (mm): 0
#> Performing daily simulations 
#>  Year 2001:....................................done.
#> Final soil water content (mm): 162.249
#> Final snowpack content (mm): 0
#> Change in soil water content (mm): -22.82
#> Soil water balance result (mm): -22.82
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 0
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 100 Net rainfall (mm) 362
#>   Infiltration (mm) 404 Runoff (mm) 9 Deep drainage (mm) 63
#>   Soil evaporation (mm) 38 Transpiration (mm) 326
        
#Monthly summary (averages) of soil status
summary(S1, freq="months",FUN=mean, output="Soil")
#>                  W.1       W.2     ML.1      ML.2    MLTot       WTD
#> 2001-01-01 0.9913015 0.9980352 67.66950 116.57658 184.2461  997.7262
#> 2001-02-01 0.9682675 0.9364257 66.09712 109.38021 175.4773 1000.0000
#> 2001-03-01 0.9719232 0.9633419 66.34667 112.52419 178.8709  998.7720
#> 2001-04-01 0.9454933 0.8250240 64.54248  96.36782 160.9103 1000.0000
#> 2001-05-01 0.9447760 0.7819989 64.49352  91.34223 155.8357 1000.0000
#> 2001-06-01 0.8176867 0.5048067 55.81799  58.96449 114.7825 1000.0000
#> 2001-07-01 0.9513086 0.5878203 64.93946  68.66098 133.6004 1000.0000
#> 2001-08-01 0.9595725 0.6134178 65.50357  71.65092 137.1545 1000.0000
#> 2001-09-01 0.9610666 0.6048004 65.60557  70.64436 136.2499 1000.0000
#> 2001-10-01 0.9737840 0.5913550 66.47370  69.07386 135.5476 1000.0000
#> 2001-11-01 0.9693913 0.7604586 66.17384  88.82619 155.0000 1000.0000
#> 2001-12-01 0.9354059 0.9093367 63.85388 106.21605 170.0699 1000.0000
#>                   SWE PlantExt.1 PlantExt.2       psi.1       psi.2
#> 2001-01-01 1.68796802 0.06862091  0.2771057 -0.03462781 -0.03347430
#> 2001-02-01 0.28335308 0.14314093  0.5780320 -0.03927984 -0.04836369
#> 2001-03-01 0.01762496 0.14873370  0.6006176 -0.03815044 -0.04053473
#> 2001-04-01 0.59071605 0.18980069  0.7664371 -0.04422593 -0.09353383
#> 2001-05-01 0.00000000 0.22402884  0.9045384 -0.04560445 -0.13481545
#> 2001-06-01 0.00000000 0.31938419  1.0115551 -0.09582387 -1.60769546
#> 2001-07-01 0.00000000 0.26958872  0.9251424 -0.04413189 -1.03089299
#> 2001-08-01 0.00000000 0.27236638  1.0965811 -0.04087960 -0.43312505
#> 2001-09-01 0.00000000 0.21376098  0.8604834 -0.04078323 -0.46204112
#> 2001-10-01 0.00000000 0.15333136  0.6135933 -0.03801668 -0.56367102
#> 2001-11-01 2.68795419 0.11075329  0.4467669 -0.03890918 -0.23349606
#> 2001-12-01 0.00000000 0.10358719  0.4183059 -0.04696503 -0.05545887

#Queries the tables in 'Plants'
names(S1$Plants)
#> [1] "LAI"                 "AbsorbedSWRFraction" "Transpiration"      
#> [4] "GrossPhotosynthesis" "PlantPsi"            "StemPLC"            
#> [7] "PlantStress"        

#Monthly summary (averages) of plant stress
summary(S1, freq="months",FUN=mean, output="Plants$PlantStress", 
        bySpecies = TRUE)
#>            Pinus halepensis Quercus coccifera Quercus ilex
#> 2001-01-01     1.468709e-06      6.539347e-07 9.544836e-07
#> 2001-02-01     4.246538e-06      1.447573e-06 2.759740e-06
#> 2001-03-01     2.653684e-06      1.036851e-06 1.724576e-06
#> 2001-04-01     2.762424e-05      7.136898e-06 1.795256e-05
#> 2001-05-01     1.171952e-04      2.827651e-05 7.616851e-05
#> 2001-06-01     1.774825e-01      5.338952e-02 1.334371e-01
#> 2001-07-01     1.297993e-01      4.059602e-02 9.975341e-02
#> 2001-08-01     3.031035e-03      7.141245e-04 1.972700e-03
#> 2001-09-01     3.486202e-03      8.210120e-04 2.268459e-03
#> 2001-10-01     8.690910e-03      2.055770e-03 5.671831e-03
#> 2001-11-01     1.003865e-03      2.364025e-04 6.526789e-04
#> 2001-12-01     6.345690e-06      2.270343e-06 4.123938e-06