Summarises simulation results

Function summary summarizes the model's output in different temporal steps (i.e. weekly, annual, ...).

# S3 method for pwb
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for spwb
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for growth
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for fordyn
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)

Arguments

object

An object of class spwb, pwb, growth or fordyn.

freq

Frequency of summary statistics (see cut.Date).

output

The data table to be summarized. Accepted values are the path to data tables in object, such as 'WaterBalance', 'Soil', 'Stand' or 'Plants$LAI'. It is also possible to use strings like 'Transpiration' and the function will interpret it as 'Plants$Transpiration'.

FUN

The function to summarize results (e.g., sum, mean, ...)

bySpecies

Allows aggregating output by species before calculating summaries (only has an effect with some values of output). Aggregation can involve a sum (as for plant lai or transpiration) or a LAI-weighted mean (as for plant stress or plant water potential).

...

Additional parameters for function summary.

Author

Miquel De Cáceres Ainsa, CREAF

Note

When applied to fordyn objects, the summary function can be used to gather the results of different yearly steps into a single table while keeping a daily resolution (i.e. using freq = "days".

See also

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Call simulation function
S1<-spwb(x, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial soil water content (mm): 185.069
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:....................................
#> 
#> Final soil water content (mm): 163.654
#> Final snowpack content (mm): 0
#> Change in soil water content (mm): -21.4157
#> Soil water balance result (mm): -21.4157
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 7.10543e-15
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 98 Net rainfall (mm) 364
#>   Infiltration (mm) 406 Runoff (mm) 9 Deep drainage (mm) 80
#>   Soil evaporation (mm) 15 Transpiration (mm) 332
#>   Plant extraction from soil (mm) 332  Plant water balance (mm) -0
        
#Monthly summary (averages) of soil status
summary(S1, freq="months",FUN=mean, output="Soil")
#>                  W.1       W.2     ML.1      ML.2    MLTot       WTD        SWE
#> 2001-01-01 0.9843549 1.0014796 67.19531 116.97891 184.1742  997.1725 1.68350931
#> 2001-02-01 0.9168624 0.9754395 62.58804 113.93726 176.5253 1000.0000 0.28233380
#> 2001-03-01 0.9359619 0.9861992 63.89184 115.19406 179.0859  998.3017 0.01762496
#> 2001-04-01 0.8191907 0.9085926 55.92066 106.12914 162.0498 1000.0000 0.58931434
#> 2001-05-01 0.8465098 0.8768070 57.78555 102.41638 160.2019 1000.0000 0.00000000
#> 2001-06-01 0.5903182 0.6771158 40.29706  79.09124 119.3883 1000.0000 0.00000000
#> 2001-07-01 0.7998250 0.6524716 54.59868  76.21265 130.8113 1000.0000 0.00000000
#> 2001-08-01 0.7739631 0.7273650 52.83327  84.96065 137.7939 1000.0000 0.00000000
#> 2001-09-01 0.7683868 0.7614502 52.45261  88.94201 141.3946 1000.0000 0.00000000
#> 2001-10-01 0.8112528 0.7672127 55.37878  89.61511 144.9939 1000.0000 0.00000000
#> 2001-11-01 0.8923855 0.8782026 60.91717 102.57939 163.4966  999.2077 2.67673914
#> 2001-12-01 0.8681590 0.9576666 59.26339 111.86127 171.1247 1000.0000 0.00000000
#>            PlantExt.1 PlantExt.2       psi.1       psi.2
#> 2001-01-01  0.2313616  0.1088030 -0.03605620 -0.03280277
#> 2001-02-01  0.4228420  0.2866626 -0.05425082 -0.03801692
#> 2001-03-01  0.4567175  0.2806436 -0.04748287 -0.03548901
#> 2001-04-01  0.4833384  0.4574170 -0.09473135 -0.05520963
#> 2001-05-01  0.6578863  0.4522537 -0.09371889 -0.06695979
#> 2001-06-01  0.5381046  0.9936374 -0.57562453 -0.30561613
#> 2001-07-01  1.2026971  0.1171707 -0.20198804 -0.32318850
#> 2001-08-01  1.1014384  0.2480994 -0.13821035 -0.16742391
#> 2001-09-01  0.7684753  0.2907785 -0.13721660 -0.13230756
#> 2001-10-01  0.5866863  0.1732676 -0.11301958 -0.12757496
#> 2001-11-01  0.4078896  0.1412014 -0.06487909 -0.07439586
#> 2001-12-01  0.2721851  0.2412140 -0.07093450 -0.04176043

#Queries the tables in 'Plants'
names(S1$Plants)
#>  [1] "LAI"                 "LAIlive"             "FPAR"               
#>  [4] "AbsorbedSWRFraction" "Transpiration"       "GrossPhotosynthesis"
#>  [7] "PlantPsi"            "StemPLC"             "PlantWaterBalance"  
#> [10] "LeafRWC"             "StemRWC"             "LFMC"               
#> [13] "PlantStress"        

#Monthly summary (averages) of plant stress
summary(S1, freq="months",FUN=mean, output="Plants$PlantStress", 
        bySpecies = TRUE)
#>            Pinus halepensis Quercus coccifera Quercus ilex
#> 2001-01-01     7.427884e-06      3.046555e-06 4.441027e-06
#> 2001-02-01     2.665065e-05      9.962973e-06 1.279126e-05
#> 2001-03-01     1.833880e-05      6.949808e-06 9.107506e-06
#> 2001-04-01     1.327756e-04      4.822651e-05 5.914400e-05
#> 2001-05-01     2.090379e-04      7.664862e-05 9.543461e-05
#> 2001-06-01     3.644455e-02      1.359902e-02 1.652566e-02
#> 2001-07-01     1.362732e-02      5.835031e-03 8.301298e-03
#> 2001-08-01     6.077630e-04      2.767362e-04 4.522918e-04
#> 2001-09-01     5.136678e-04      2.086854e-04 3.004795e-04
#> 2001-10-01     3.589901e-04      1.534576e-04 2.347566e-04
#> 2001-11-01     7.079558e-05      3.416724e-05 5.898273e-05
#> 2001-12-01     5.749779e-05      2.088332e-05 2.561079e-05