Summarises simulation results

Function summary summarizes the model's output in different temporal steps (i.e. weekly, annual, ...).

# S3 method for pwb
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for spwb
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for growth
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)
# S3 method for fordyn
summary(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...)

Arguments

object

An object of class spwb, pwb, growth or fordyn.

freq

Frequency of summary statistics (see cut.Date).

output

The data table to be summarized. Accepted values are the path to data tables in object, such as 'WaterBalance', 'Soil', 'Stand' or 'Plants$LAI'. It is also possible to use strings like 'Transpiration' and the function will interpret it as 'Plants$Transpiration'.

FUN

The function to summarize results (e.g., sum, mean, ...)

bySpecies

Allows aggregating output by species before calculating summaries (only has an effect with some values of output). Aggregation can involve a sum (as for plant lai or transpiration) or a LAI-weighted mean (as for plant stress or plant water potential).

...

Additional parameters for function summary.

Author

Miquel De Cáceres Ainsa, CREAF

Note

When applied to fordyn objects, the summary function can be used to gather the results of different yearly steps into a single table while keeping a daily resolution (i.e. using freq = "days".

See also

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Call simulation function
S1<-spwb(x, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial soil water content (mm): 185.069
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:....................................
#> 
#> Final soil water content (mm): 163.474
#> Final snowpack content (mm): 0
#> Change in soil water content (mm): -21.5958
#> Soil water balance result (mm): -21.5958
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): 7.10543e-15
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 98 Net rainfall (mm) 364
#>   Infiltration (mm) 406 Runoff (mm) 9 Deep drainage (mm) 74
#>   Soil evaporation (mm) 19 Transpiration (mm) 334
        
#Monthly summary (averages) of soil status
summary(S1, freq="months",FUN=mean, output="Soil")
#>                  W.1       W.2     ML.1      ML.2    MLTot       WTD        SWE
#> 2001-01-01 0.9883340 0.9998059 67.46693 116.78340 184.2503  997.4786 1.68350931
#> 2001-02-01 0.9447282 0.9552808 64.49025 111.58260 176.0729 1000.0000 0.28233380
#> 2001-03-01 0.9557346 0.9750939 65.24159 113.89690 179.1385  998.5662 0.01762496
#> 2001-04-01 0.8777084 0.8748499 59.91526 102.18778 162.1030 1000.0000 0.58931434
#> 2001-05-01 0.8806006 0.8489488 60.11270  99.16237 159.2751 1000.0000 0.00000000
#> 2001-06-01 0.6293676 0.6426737 42.96270  75.06819 118.0309 1000.0000 0.00000000
#> 2001-07-01 0.8065016 0.6246904 55.05445  72.96763 128.0221 1000.0000 0.00000000
#> 2001-08-01 0.7921914 0.6923136 54.07759  80.86643 134.9440 1000.0000 0.00000000
#> 2001-09-01 0.7961754 0.7200734 54.34955  84.10894 138.4585 1000.0000 0.00000000
#> 2001-10-01 0.8337229 0.7270956 56.91267  84.92918 141.8418 1000.0000 0.00000000
#> 2001-11-01 0.9044009 0.8512540 61.73738  99.43164 161.1690  999.5422 2.67673914
#> 2001-12-01 0.9047210 0.9354367 61.75923 109.26469 171.0239 1000.0000 0.00000000
#>            PlantExt.1 PlantExt.2       psi.1       psi.2
#> 2001-01-01  0.1423623  0.1978190 -0.03522255 -0.03312624
#> 2001-02-01  0.2993235  0.4102826 -0.04542129 -0.04278837
#> 2001-03-01  0.2963786  0.4409535 -0.04187582 -0.03778011
#> 2001-04-01  0.4337665  0.5071451 -0.06595998 -0.06741009
#> 2001-05-01  0.5604305  0.5501495 -0.07265440 -0.08016513
#> 2001-06-01  0.6152754  0.9618847 -0.43282967 -0.40090116
#> 2001-07-01  1.1979311  0.1333958 -0.18337868 -0.41417564
#> 2001-08-01  1.0596437  0.2902989 -0.12234865 -0.21545866
#> 2001-09-01  0.7544459  0.3051333 -0.11336651 -0.17609451
#> 2001-10-01  0.5765861  0.1834577 -0.09512933 -0.16828753
#> 2001-11-01  0.3359155  0.2131148 -0.05960445 -0.09294212
#> 2001-12-01  0.1903973  0.3231252 -0.05643025 -0.04744702

#Queries the tables in 'Plants'
names(S1$Plants)
#> [1] "LAI"                 "LAIlive"             "AbsorbedSWRFraction"
#> [4] "Transpiration"       "GrossPhotosynthesis" "PlantPsi"           
#> [7] "StemPLC"             "PlantStress"        

#Monthly summary (averages) of plant stress
summary(S1, freq="months",FUN=mean, output="Plants$PlantStress", 
        bySpecies = TRUE)
#>            Pinus halepensis Quercus coccifera Quercus ilex
#> 2001-01-01     1.521784e-06      6.617867e-07 9.889759e-07
#> 2001-02-01     3.376880e-06      1.469092e-06 2.194566e-06
#> 2001-03-01     2.485219e-06      1.097125e-06 1.615093e-06
#> 2001-04-01     1.176946e-05      4.976282e-06 7.648746e-06
#> 2001-05-01     2.204461e-05      9.341584e-06 1.432645e-05
#> 2001-06-01     4.356450e-03      1.873911e-03 2.837283e-03
#> 2001-07-01     4.239795e-03      1.646548e-03 2.765670e-03
#> 2001-08-01     2.665201e-04      9.851466e-05 1.732197e-04
#> 2001-09-01     1.474896e-04      5.588921e-05 9.585405e-05
#> 2001-10-01     1.309040e-04      4.842195e-05 8.507504e-05
#> 2001-11-01     3.865967e-05      1.407354e-05 2.512456e-05
#> 2001-12-01     5.529363e-06      2.500144e-06 3.593423e-06