Summarize simulation results

Function summary summarizes the model's output in different temporal steps (i.e. weekly, annual, ...).

# S3 method for spwb
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for pwb
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for growth
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

# S3 method for fordyn
summary(
  object,
  freq = "years",
  output = "WaterBalance",
  FUN = sum,
  bySpecies = FALSE,
  months = NULL,
  ...
)

Arguments

object

An object of class spwb, pwb, growth or fordyn.

freq

Frequency of summary statistics (see cut.Date).

output

The data table to be summarized. Accepted values are the path to data tables in object, such as 'WaterBalance', 'Soil', 'Stand' or 'Plants$LAI'. It is also possible to use strings like 'Transpiration' and the function will interpret it as 'Plants$Transpiration'.

FUN

The function to summarize results (e.g., sum, mean, ...)

bySpecies

Allows aggregating output by species before calculating summaries (only has an effect with some values of output). Aggregation can involve a sum (as for plant lai or transpiration) or a LAI-weighted mean (as for plant stress or plant water potential).

months

A vector of month numbers (1 to 12) to subset the season where summaries apply.

...

Additional parameters for function summary.

Value

A matrix with dates as row names and the desired summaries in columns

Note

When applied to fordyn objects, the summary function can be used to gather the results of different yearly steps into a single table while keeping a daily resolution (i.e. using freq = "days".

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params (2 layers)
examplesoil = soil(defaultSoilParams(2))

#Initialize control parameters
control = defaultControl("Granier")

#Initialize input
x = forest2spwbInput(exampleforest,examplesoil, SpParamsMED, control)

#Call simulation function
S1<-spwb(x, examplemeteo, latitude = 41.82592, elevation = 100)
#> Initial plant water content (mm): 5.25158
#> Initial soil water content (mm): 184.827
#> Initial snowpack content (mm): 0
#> Performing daily simulations
#> 
#>  [Year 2001]:....................................
#> 
#> Final plant water content (mm): 5.25151
#> Final soil water content (mm): 205.22
#> Final snowpack content (mm): 0
#> Change in plant water content (mm): -6.13099e-05
#> Plant water balance result (mm): -6.13099e-05
#> Change in soil water content (mm): 20.3934
#> Soil water balance result (mm): 20.5291
#> Change in snowpack water content (mm): 0
#> Snowpack water balance result (mm): -7.10543e-15
#> Water balance components:
#>   Precipitation (mm) 513
#>   Rain (mm) 462 Snow (mm) 51
#>   Interception (mm) 92 Net rainfall (mm) 370
#>   Infiltration (mm) 324 Runoff (mm) 46 Deep drainage (mm) 141
#>   Soil evaporation (mm) 13  Herbaceous transpiration (mm) 14 Woody plant transpiration (mm) 187
#>   Plant extraction from soil (mm) 187  Plant water balance (mm) -0 Hydraulic redistribution (mm) 1

#Monthly summary (averages) of soil status
summary(S1, freq="months",FUN=mean, output="Soil")
#>                  W.1       W.2     ML.1     ML.2    MLTot SaturatedDepth
#> 2001-01-01 1.2429153 1.1144982 84.73428 130.0095 214.7438       596.7733
#> 2001-02-01 1.1553995 1.2157917 78.76799 141.8257 220.5937       512.1004
#> 2001-03-01 1.1742407 1.2681501 80.05247 147.9335 227.9859       395.5524
#> 2001-04-01 0.9726055 1.1611599 66.30623 135.4528 201.7590       715.0790
#> 2001-05-01 1.0470946 1.1206863 71.38443 130.7314 202.1158       744.4814
#> 2001-06-01 0.8322493 0.9831248 56.73761 114.6844 171.4220       950.2091
#> 2001-07-01 1.0676283 0.9944153 72.78430 116.0015 188.7858       829.8663
#> 2001-08-01 1.0773389 1.0808843 73.44630 126.0884 199.5347       793.7296
#> 2001-09-01 1.0696295 1.1200040 72.92072 130.6518 203.5725       735.6428
#> 2001-10-01 1.0979363 1.1265714 74.85051 131.4179 206.2684       702.7747
#> 2001-11-01 1.1714113 1.1958367 79.85958 139.4979 219.3575       517.1953
#> 2001-12-01 1.0421632 1.2325224 71.04824 143.7774 214.8256       566.2476
#>                   SWE PlantExt.1 PlantExt.2 HydraulicInput.1 HydraulicInput.2
#> 2001-01-01 1.65619239  0.1631759 0.01094778                0     3.364730e-03
#> 2001-02-01 0.27608905  0.2356073 0.15666322                0     2.000489e-05
#> 2001-03-01 0.01762496  0.2261700 0.18158776                0     1.503552e-04
#> 2001-04-01 0.58072652  0.1964418 0.33250414                0     0.000000e+00
#> 2001-05-01 0.00000000  0.3396790 0.28878988                0     0.000000e+00
#> 2001-06-01 0.00000000  0.3049947 0.58691987                0     0.000000e+00
#> 2001-07-01 0.00000000  0.6587087 0.10235913                0     1.356900e-02
#> 2001-08-01 0.00000000  0.5564455 0.21030249                0     1.577426e-03
#> 2001-09-01 0.00000000  0.3620530 0.23148987                0     7.171548e-06
#> 2001-10-01 0.00000000  0.2648777 0.14868336                0     2.993992e-04
#> 2001-11-01 2.60802857  0.1918170 0.10363018                0     9.767139e-04
#> 2001-12-01 0.00000000  0.1132356 0.16309051                0     0.000000e+00
#>                   psi.1        psi.2
#> 2001-01-01 -0.009666023 -0.019300067
#> 2001-02-01 -0.016622727 -0.010958357
#> 2001-03-01 -0.015232553 -0.007523739
#> 2001-04-01 -0.038813196 -0.015344208
#> 2001-05-01 -0.028814429 -0.018666156
#> 2001-06-01 -0.077881056 -0.037587695
#> 2001-07-01 -0.027741113 -0.034448183
#> 2001-08-01 -0.023743991 -0.022404393
#> 2001-09-01 -0.024297613 -0.018468666
#> 2001-10-01 -0.021330496 -0.017828335
#> 2001-11-01 -0.015574405 -0.012428539
#> 2001-12-01 -0.028485997 -0.009946933

#Queries the tables in 'Plants'
names(S1$Plants)
#>  [1] "LAI"                 "LAIlive"             "FPAR"               
#>  [4] "AbsorbedSWRFraction" "Transpiration"       "GrossPhotosynthesis"
#>  [7] "PlantPsi"            "LeafPLC"             "StemPLC"            
#> [10] "PlantWaterBalance"   "LeafRWC"             "StemRWC"            
#> [13] "LFMC"                "PlantStress"        

#Monthly summary (averages) of plant stress
summary(S1, freq="months",FUN=mean, output="Plants$PlantStress", 
        bySpecies = TRUE)
#>            Pinus halepensis Quercus coccifera Quercus ilex
#> 2001-01-01      0.001521178      0.0010354372  0.004659389
#> 2001-02-01      0.001978141      0.0010907108  0.004144834
#> 2001-03-01      0.001876291      0.0009817384  0.003531599
#> 2001-04-01      0.006792599      0.0030590646  0.008810122
#> 2001-05-01      0.004546440      0.0022622761  0.007466495
#> 2001-06-01      0.018508259      0.0076233851  0.019212537
#> 2001-07-01      0.006118434      0.0031436307  0.010552608
#> 2001-08-01      0.003688058      0.0020168092  0.007334207
#> 2001-09-01      0.003622959      0.0019135723  0.006758559
#> 2001-10-01      0.003043829      0.0016671569  0.006145929
#> 2001-11-01      0.002133262      0.0011737485  0.004426540
#> 2001-12-01      0.004131512      0.0019537350  0.006025763