Models for canopy turbulence

Models for canopy turbulence by Katul et al (2004).

wind_canopyTurbulenceModel(zm, Cx, hm, d0, z0, model = "k-epsilon")
wind_canopyTurbulence(zmid, LAD, canopyHeight, 
                      u, windMeasurementHeight = 200, model = "k-epsilon")

Arguments

zm

A numeric vector with height values (m).

Cx

Effective drag = Cd x leaf area density.

hm

Canopy height (m).

d0

Zero displacement height (m).

z0

Momentum roughness height (m).

zmid

A numeric vector of mid-point heights (in cm) for canopy layers.

LAD

A numeric vector of leaf area density values (m3/m2).

canopyHeight

Canopy height (in cm).

u

Measured wind speed (m/s).

windMeasurementHeight

Height of wind speed measurement with respect to canopy height (cm).

model

Closure model.

Value

Function wind_canopyTurbulenceModel returns a data frame of vertical profiles for variables:

z1: Height values.

U1: U/u*, where U is mean velocity and u* is friction velocity.

dU1: dUdz/u*, where dUdz is mean velocity gradient and u* is friction velocity.

epsilon1: epsilon/(u^3/h) where epsilon is the turbulent kinetic dissipation rate, u* is friction velocity and h is canopy height.

k1: k/(u*^2), where k is the turbulent kinetic energy and u* is friction velocity.

uw1: <uw>/(u*^2), where <uw> is the Reynolds stress and u* is friction velocity.

Lmix1: Mixing length.

Function wind_canopyTurbulence returns a data frame of vertical profiles for transformed variables:

zmid: Input mid-point heights (in cm) for canopy layers.

u: Wind speed (m/s).

du: Mean velocity gradient (1/s).

epsilon: Turbulent kinetic dissipation rate.

k: Turbulent kinetic energy.

uw: Reynolds stress.

Details

Implementation in Rcpp of the K-epsilon canopy turbulence models by Katul et al (2004) originally in Matlab code (https://nicholas.duke.edu/people/faculty/katul/k_epsilon_model.htm).

Author

Miquel De Cáceres Ainsa, CREAF

References

Katul GG, Mahrt L, Poggi D, Sanz C (2004) One- and two-equation models for canopy turbulence. Boundary-Layer Meteorol 113:81–109. https://doi.org/10.1023/B:BOUN.0000037333.48760.e5

See also

Examples

#Default species parameterization
data(SpParamsMED)

#Load example plot plant data
data(exampleforestMED)

#Canopy height (in m)
h= max(exampleforestMED$treeData$Height/100) 
d0 = 0.67*h
z0 = 0.08*h

#Height values (cm)
z = seq(50,1000, by=50)
zm = z/100 # (in m)

# Leaf area density
lad = vprofile_leafAreaDensity(exampleforestMED, SpParamsMED, draw = FALSE,
                               z = c(0,z))

# Effective drag
Cd = 0.2
Cx = Cd*lad

# canopy turbulence model
wind_canopyTurbulenceModel(zm, Cx,h,d0,z0)
#>      z1        U1        dU1    epsilon1        k1          uw1 Lmix1
#> 1   0.5 0.7203423 0.01611031 0.002113150 0.1567679 -0.002657900 1.056
#> 2   1.0 0.7275919 0.01611031 0.002703978 0.1650029 -0.002657900 1.056
#> 3   1.5 0.7391198 0.02448665 0.003526418 0.1778449 -0.004201616 1.056
#> 4   2.0 0.7504252 0.02403503 0.004270217 0.1955576 -0.004330546 1.056
#> 5   2.5 0.7615069 0.02358868 0.004904923 0.2190436 -0.004502665 1.056
#> 6   3.0 0.7723618 0.02314115 0.005358111 0.2482561 -0.004705714 1.056
#> 7   3.5 0.8000319 0.05627097 0.006739406 0.2911868 -0.012391109 1.056
#> 8   4.0 0.8607032 0.12142567 0.010292600 0.3595154 -0.029672625 1.056
#> 9   4.5 0.9656218 0.20902001 0.017547391 0.4677282 -0.058128155 1.056
#> 10  5.0 1.1263159 0.31977710 0.031373446 0.6380491 -0.103579445 1.056
#> 11  5.5 1.3528026 0.45085425 0.056457945 0.8998272 -0.172963709 1.056
#> 12  6.0 1.6511575 0.59449150 0.098928241 1.2841525 -0.271872172 1.056
#> 13  6.5 2.0213757 0.73852329 0.164211468 1.8122121 -0.400631461 1.056
#> 14  7.0 2.4568217 0.86954968 0.253166275 2.4813909 -0.551496870 1.056
#> 15  7.5 2.9201067 0.92611716 0.346754412 3.1922196 -0.665975735 1.056
#> 16  8.0 3.4060482 0.97193263 0.436335542 3.9178578 -0.774184947 1.056
#> 17  8.5 3.9116973 1.01160765 0.509977237 4.6288537 -1.041702790 1.256
#> 18  9.0 4.3170700 0.81156771 0.532272684 5.0767640 -1.014671039 1.456
#> 19  9.5 4.6577858 0.68235283 0.537536295 5.3913091 -1.000000000 1.656
#> 20 10.0 4.9525037 0.59036952 0.538793103 5.6312500 -1.000000000 1.856