Models for canopy turbulence

Models for canopy turbulence by Katul et al (2004).

wind_canopyTurbulenceModel(zm, Cx, hm, d0, z0, model = "k-epsilon")

wind_canopyTurbulence(
  zmid,
  LAD,
  canopyHeight,
  u,
  windMeasurementHeight = 200,
  model = "k-epsilon"
)

Arguments

zm

A numeric vector with height values (m).

Cx

Effective drag = Cd x leaf area density.

hm

Canopy height (m).

d0

Zero displacement height (m).

z0

Momentum roughness height (m).

model

Closure model.

zmid

A numeric vector of mid-point heights (in cm) for canopy layers.

LAD

A numeric vector of leaf area density values (m3/m2).

canopyHeight

Canopy height (in cm).

u

Measured wind speed (m/s).

windMeasurementHeight

Height of wind speed measurement with respect to canopy height (cm).

Value

Function wind_canopyTurbulenceModel returns a data frame of vertical profiles for variables:

z1: Height values.

U1: U/u*, where U is mean velocity and u* is friction velocity.

dU1: dUdz/u*, where dUdz is mean velocity gradient and u* is friction velocity.

epsilon1: epsilon/(u^3/h) where epsilon is the turbulent kinetic dissipation rate, u* is friction velocity and h is canopy height.

k1: k/(u*^2), where k is the turbulent kinetic energy and u* is friction velocity.

uw1: <uw>/(u*^2), where <uw> is the Reynolds stress and u* is friction velocity.

Lmix1: Mixing length.

Function wind_canopyTurbulence returns a data frame of vertical profiles for transformed variables:

zmid: Input mid-point heights (in cm) for canopy layers.

u: Wind speed (m/s).

du: Mean velocity gradient (1/s).

epsilon: Turbulent kinetic dissipation rate.

k: Turbulent kinetic energy.

uw: Reynolds stress.

Details

Implementation in Rcpp of the K-epsilon canopy turbulence models by Katul et al (2004) originally in Matlab code (https://nicholas.duke.edu/people/faculty/katul/k_epsilon_model.htm).

References

Katul GG, Mahrt L, Poggi D, Sanz C (2004) One- and two-equation models for canopy turbulence. Boundary-Layer Meteorol 113:81–109. https://doi.org/10.1023/B:BOUN.0000037333.48760.e5

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

#Default species parameterization
data(SpParamsMED)

#Load example plot plant data
data(exampleforestMED)

#Canopy height (in m)
h= max(exampleforestMED$treeData$Height/100) 
d0 = 0.67*h
z0 = 0.08*h

#Height values (cm)
z = seq(50,1000, by=50)
zm = z/100 # (in m)

# Leaf area density
lad = vprofile_leafAreaDensity(exampleforestMED, SpParamsMED, draw = FALSE,
                               z = c(0,z))
  
# Effective drag
Cd = 0.2
Cx = Cd*lad
  
# canopy turbulence model
wind_canopyTurbulenceModel(zm, Cx,h,d0,z0)
#>      z1        U1        dU1    epsilon1        k1          uw1 Lmix1
#> 1   0.5 0.9006853 0.01790035 0.003522799 0.2305301 -0.003778656 1.056
#> 2   1.0 0.9087404 0.01790035 0.004398260 0.2396801 -0.003778656 1.056
#> 3   1.5 0.9225229 0.02914276 0.005457062 0.2554134 -0.006360372 1.056
#> 4   2.0 0.9360357 0.02859149 0.006547763 0.2770912 -0.006507485 1.056
#> 5   2.5 0.9492759 0.02804364 0.007660643 0.3058583 -0.006712470 1.056
#> 6   3.0 0.9622362 0.02748766 0.008747462 0.3420332 -0.006962531 1.056
#> 7   3.5 0.9921301 0.06088537 0.011016101 0.3949826 -0.016574865 1.056
#> 8   4.0 1.0554387 0.12690096 0.015976863 0.4776254 -0.037959627 1.056
#> 9   4.5 1.1630241 0.21458313 0.025393680 0.6046140 -0.072104689 1.056
#> 10  5.0 1.3252152 0.32303014 0.042185304 0.7967182 -0.124349280 1.056
#> 11  5.5 1.5499981 0.44772361 0.070564477 1.0788050 -0.200146793 1.056
#> 12  6.0 1.8410580 0.58016210 0.115308837 1.4738179 -0.302627415 1.056
#> 13  6.5 2.1964458 0.70905633 0.179700902 1.9929829 -0.429586021 1.056
#> 14  7.0 2.6086753 0.82320648 0.262640303 2.6264961 -0.572127068 1.056
#> 15  7.5 3.0460205 0.87416465 0.348071988 3.2921888 -0.679962062 1.056
#> 16  8.0 3.5025464 0.91296563 0.428507980 3.9657209 -0.779295713 1.056
#> 17  8.5 3.9744795 0.94401726 0.494408950 4.6227044 -1.034723122 1.256
#> 18  9.0 4.3549492 0.76162651 0.518560757 5.0564112 -1.012206808 1.456
#> 19  9.5 4.6753103 0.64156526 0.529633172 5.3756691 -1.000000000 1.656
#> 20 10.0 4.9525037 0.55529197 0.538793103 5.6312500 -1.000000000 1.856