Models for canopy turbulence

Models for canopy turbulence by Katul et al (2004).

wind_canopyTurbulenceModel(zm, Cx, hm, d0, z0, model = "k-epsilon")
wind_canopyTurbulence(zmid, LAD, canopyHeight, 
                      u, windMeasurementHeight = 200, model = "k-epsilon")

Arguments

zm

A numeric vector with height values (m).

Cx

Effective drag = Cd x leaf area density.

hm

Canopy height (m).

d0

Zero displacement height (m).

z0

Momentum roughness height (m).

zmid

A numeric vector of mid-point heights (in cm) for canopy layers.

LAD

A numeric vector of leaf area density values (m3/m2).

canopyHeight

Canopy height (in cm).

u

Measured wind speed (m/s).

windMeasurementHeight

Height of wind speed measurement with respect to canopy height (cm).

model

Closure model.

Value

Function wind_canopyTurbulenceModel returns a data frame of vertical profiles for variables:

z1: Height values.

U1: U/u*, where U is mean velocity and u* is friction velocity.

dU1: dUdz/u*, where dUdz is mean velocity gradient and u* is friction velocity.

epsilon1: epsilon/(u^3/h) where epsilon is the turbulent kinetic dissipation rate, u* is friction velocity and h is canopy height.

k1: k/(u*^2), where k is the turbulent kinetic energy and u* is friction velocity.

uw1: <uw>/(u*^2), where <uw> is the Reynolds stress and u* is friction velocity.

Lmix1: Mixing length.

Function wind_canopyTurbulence returns a data frame of vertical profiles for transformed variables:

zmid: Input mid-point heights (in cm) for canopy layers.

u: Wind speed (m/s).

du: Mean velocity gradient (1/s).

epsilon: Turbulent kinetic dissipation rate.

k: Turbulent kinetic energy.

uw: Reynolds stress.

Details

Implementation in Rcpp of the K-epsilon canopy turbulence models by Katul et al (2004) originally in Matlab code (https://nicholas.duke.edu/people/faculty/katul/k_epsilon_model.htm).

Author

Miquel De Cáceres Ainsa, CREAF

References

Katul GG, Mahrt L, Poggi D, Sanz C (2004) One- and two-equation models for canopy turbulence. Boundary-Layer Meteorol 113:81–109. https://doi.org/10.1023/B:BOUN.0000037333.48760.e5

See also

Examples

#Default species parameterization
data(SpParamsMED)

#Load example plot plant data
data(exampleforestMED)

#Canopy height (in m)
h= max(exampleforestMED$treeData$Height/100) 
d0 = 0.67*h
z0 = 0.08*h

#Height values (cm)
z = seq(50,1000, by=50)
zm = z/100 # (in m)

# Leaf area density
lad = vprofile_leafAreaDensity(exampleforestMED, SpParamsMED, draw = FALSE,
                               z = c(0,z))

# Effective drag
Cd = 0.2
Cx = Cd*lad

# canopy turbulence model
wind_canopyTurbulenceModel(zm, Cx,h,d0,z0)
#>      z1        U1        dU1    epsilon1        k1          uw1 Lmix1
#> 1   0.5 0.7091733 0.01611031 0.002036478 0.1525059 -0.002603566 1.056
#> 2   1.0 0.7164229 0.01611031 0.002611852 0.1607409 -0.002603566 1.056
#> 3   1.5 0.7279020 0.02438884 0.003425568 0.1734750 -0.004102065 1.056
#> 4   2.0 0.7391585 0.02393733 0.004152352 0.1910208 -0.004230690 1.056
#> 5   2.5 0.7501917 0.02349191 0.004760244 0.2142700 -0.004401880 1.056
#> 6   3.0 0.7609988 0.02304623 0.005175039 0.2431534 -0.004603305 1.056
#> 7   3.5 0.7884288 0.05579311 0.006491989 0.2854526 -0.012073217 1.056
#> 8   4.0 0.8485683 0.12036626 0.009930831 0.3526502 -0.028912717 1.056
#> 9   4.5 0.9525938 0.20723819 0.016983623 0.4590272 -0.056663020 1.056
#> 10  5.0 1.1120533 0.31730813 0.030474151 0.6266318 -0.101081833 1.056
#> 11  5.5 1.3371183 0.44800129 0.055060479 0.8848367 -0.169126531 1.056
#> 12  6.0 1.6341646 0.59185350 0.096917001 1.2652198 -0.266591644 1.056
#> 13  6.5 2.0035882 0.73690544 0.161665167 1.7900953 -0.394227193 1.056
#> 14  7.0 2.4391527 0.86975802 0.250499212 2.4583901 -0.544794883 1.056
#> 15  7.5 2.9041484 0.92951589 0.344848663 3.1727889 -0.661187145 1.056
#> 16  8.0 3.3933409 0.97842148 0.435884439 3.9057739 -0.772077085 1.056
#> 17  8.5 3.9035163 1.02065740 0.511110033 4.6260231 -1.042495063 1.256
#> 18  9.0 4.3121619 0.81812405 0.533544735 5.0772888 -1.014933928 1.456
#> 19  9.5 4.6555228 0.68765294 0.538336431 5.3923845 -1.000000000 1.656
#> 20 10.0 4.9525037 0.59490216 0.538793103 5.6312500 -1.000000000 1.856