Models for canopy turbulence

Models for canopy turbulence by Katul et al (2004).

wind_canopyTurbulenceModel(zm, Cx, hm, d0, z0, model = "k-epsilon")
wind_canopyTurbulence(zmid, LAD, canopyHeight, 
                      u, windMeasurementHeight = 200, model = "k-epsilon")

Arguments

zm

A numeric vector with height values (m).

Cx

Effective drag = Cd x leaf area density.

hm

Canopy height (m).

d0

Zero displacement height (m).

z0

Momentum roughness height (m).

zmid

A numeric vector of mid-point heights (in cm) for canopy layers.

LAD

A numeric vector of leaf area density values (m3/m2).

canopyHeight

Canopy height (in cm).

u

Measured wind speed (m/s).

windMeasurementHeight

Height of wind speed measurement with respect to canopy height (cm).

model

Closure model.

Value

Function wind_canopyTurbulenceModel returns a data frame of vertical profiles for variables:

z1: Height values.

U1: U/u*, where U is mean velocity and u* is friction velocity.

dU1: dUdz/u*, where dUdz is mean velocity gradient and u* is friction velocity.

epsilon1: epsilon/(u^3/h) where epsilon is the turbulent kinetic dissipation rate, u* is friction velocity and h is canopy height.

k1: k/(u*^2), where k is the turbulent kinetic energy and u* is friction velocity.

uw1: <uw>/(u*^2), where <uw> is the Reynolds stress and u* is friction velocity.

Lmix1: Mixing length.

Function wind_canopyTurbulence returns a data frame of vertical profiles for transformed variables:

zmid: Input mid-point heights (in cm) for canopy layers.

u: Wind speed (m/s).

du: Mean velocity gradient (1/s).

epsilon: Turbulent kinetic dissipation rate.

k: Turbulent kinetic energy.

uw: Reynolds stress.

Details

Implementation in Rcpp of the K-epsilon canopy turbulence models by Katul et al (2004) originally in Matlab code (https://nicholas.duke.edu/people/faculty/katul/k_epsilon_model.htm).

Author

Miquel De Cáceres Ainsa, CREAF

References

Katul GG, Mahrt L, Poggi D, Sanz C (2004) One- and two-equation models for canopy turbulence. Boundary-Layer Meteorol 113:81–109. https://doi.org/10.1023/B:BOUN.0000037333.48760.e5

See also

Examples

#Default species parameterization
data(SpParamsMED)

#Load example plot plant data
data(exampleforestMED)

#Canopy height (in m)
h= max(exampleforestMED$treeData$Height/100) 
d0 = 0.67*h
z0 = 0.08*h

#Height values (cm)
z = seq(50,1000, by=50)
zm = z/100 # (in m)

# Leaf area density
lad = vprofile_leafAreaDensity(exampleforestMED, SpParamsMED, draw = FALSE,
                               z = c(0,z))

# Effective drag
Cd = 0.2
Cx = Cd*lad

# canopy turbulence model
wind_canopyTurbulenceModel(zm, Cx,h,d0,z0)
#>      z1        U1        dU1    epsilon1        k1          uw1 Lmix1
#> 1   0.5 0.7431863 0.01611031 0.002243710 0.1650870 -0.002732151 1.056
#> 2   1.0 0.7504359 0.01611031 0.002875687 0.1733220 -0.002732151 1.056
#> 3   1.5 0.7621553 0.02486681 0.003717292 0.1863958 -0.004380949 1.056
#> 4   2.0 0.7736466 0.02440344 0.004496007 0.2044602 -0.004508896 1.056
#> 5   2.5 0.7849080 0.02394394 0.005184473 0.2284445 -0.004681022 1.056
#> 6   3.0 0.7959354 0.02348156 0.005714064 0.2583519 -0.004885219 1.056
#> 7   3.5 0.8236759 0.05641816 0.007190491 0.3022726 -0.012695211 1.056
#> 8   4.0 0.8843430 0.12143645 0.010885712 0.3720067 -0.030278365 1.056
#> 9   4.5 0.9890339 0.20859727 0.018343993 0.4819393 -0.059071969 1.056
#> 10  5.0 1.1490230 0.31841354 0.032411102 0.6539162 -0.104757132 1.056
#> 11  5.5 1.3740127 0.44791577 0.057667947 0.9164797 -0.174011502 1.056
#> 12  6.0 1.6698290 0.58946890 0.100030398 1.2995514 -0.272134074 1.056
#> 13  6.5 2.0364316 0.73133213 0.164677904 1.8232927 -0.399348081 1.056
#> 14  7.0 2.4674124 0.86063767 0.252371955 2.4849581 -0.548176117 1.056
#> 15  7.5 2.9274727 0.91964725 0.345218985 3.1911653 -0.663558822 1.056
#> 16  8.0 3.4112104 0.96749508 0.434563974 3.9146917 -0.773066562 1.056
#> 17  8.5 3.9151725 1.00820589 0.508294461 4.6248782 -1.041425389 1.256
#> 18  9.0 4.3192150 0.80888614 0.531027356 5.0734727 -1.014566261 1.456
#> 19  9.5 4.6587954 0.68006606 0.536871655 5.3894379 -1.000000000 1.656
#> 20 10.0 4.9525037 0.58833820 0.538793103 5.6312500 -1.000000000 1.856