Models for canopy turbulence

Models for canopy turbulence by Katul et al (2004).

wind_canopyTurbulenceModel(zm, Cx, hm, d0, z0, model = "k-epsilon")
wind_canopyTurbulence(zmid, LAD, canopyHeight, 
                      u, windMeasurementHeight = 200, model = "k-epsilon")

Arguments

zm
A numeric vector with height values (m).

Cx
Effective drag = Cd x leaf area density.

hm
Canopy height (m).

d0
Zero displacement height (m).

z0
Momentum roughness height (m).

zmid
A numeric vector of mid-point heights (in cm) for canopy layers.

LAD
A numeric vector of leaf area density values (m3/m2).

canopyHeight
Canopy height (in cm).

u
Measured wind speed (m/s).

windMeasurementHeight
Height of wind speed measurement with respect to canopy height (cm).

model
Closure model.

Value

Function wind_canopyTurbulenceModel returns a data frame of vertical profiles for variables:

z1: Height values.

U1: U/u*, where U is mean velocity and u* is friction velocity.

dU1: dUdz/u*, where dUdz is mean velocity gradient and u* is friction velocity.

epsilon1: epsilon/(u^3/h) where epsilon is the turbulent kinetic dissipation rate, u* is friction velocity and h is canopy height.

k1: k/(u*^2), where k is the turbulent kinetic energy and u* is friction velocity.

uw1: <uw>/(u*^2), where <uw> is the Reynolds stress and u* is friction velocity.

Lmix1: Mixing length.

Function wind_canopyTurbulence returns a data frame of vertical profiles for transformed variables:

zmid: Input mid-point heights (in cm) for canopy layers.

u: Wind speed (m/s).

du: Mean velocity gradient (1/s).

epsilon: Turbulent kinetic dissipation rate.

k: Turbulent kinetic energy.

uw: Reynolds stress.

Details

Implementation in Rcpp of the K-epsilon canopy turbulence models by Katul et al (2004) originally in Matlab code (https://nicholas.duke.edu/people/faculty/katul/k_epsilon_model.htm).

Author

Miquel De Cáceres Ainsa, CREAF

References

Katul GG, Mahrt L, Poggi D, Sanz C (2004) One- and two-equation models for canopy turbulence. Boundary-Layer Meteorol 113:81–109. https://doi.org/10.1023/B:BOUN.0000037333.48760.e5

See also

Examples

#Default species parameterization
data(SpParamsMED)

#Load example plot plant data
data(exampleforestMED)

#Canopy height (in m)
h= max(exampleforestMED$treeData$Height/100) 
d0 = 0.67*h
z0 = 0.08*h

#Height values (cm)
z = seq(50,1000, by=50)
zm = z/100 # (in m)

# Leaf area density
lad = vprofile_leafAreaDensity(exampleforestMED, SpParamsMED, draw = FALSE, z = z)

# Effective drag
Cd = 0.2
Cx = Cd*lad

# canopy turbulence model
wind_canopyTurbulence(zm, Cx,h,d0,z0)

#>    zmid        u       du  epsilon        k        uw
#> 1   0.5 5.977306 3.229486 1002.700 8.148157 -1.013463
#> 2   1.0 6.004586 3.229486 1009.079 8.188812 -1.013463
#> 3   1.5 6.031865 3.229486 1036.810 8.273680 -1.018782
#> 4   2.0 6.059144 3.229486 1057.987 8.346818 -1.023305
#> 5   2.5 6.086447 3.231886 1083.021 8.425798 -1.028956
#> 6   3.0 6.113863 3.242739 1106.786 8.503385 -1.037188
#> 7   3.5 6.141335 3.247904 1132.276 8.583132 -1.043739
#> 8   4.0 6.168870 3.253907 1157.596 8.663041 -1.050554
#> 9   4.5 6.196445 3.257609 1184.075 8.744305 -1.056698
#> 10  5.0 6.224034 3.258872 1210.798 8.826135 -1.062062
#> 11  5.5 6.251601 3.256810 1238.471 8.909012 -1.066377
#> 12  6.0 6.279125 3.252622 1266.616 8.992608 -1.069999
#> 13  6.5 6.306593 3.247340 1295.648 9.077164 -1.073275
#> 14  7.0 6.333960 3.237538 1325.294 9.162486 -1.075048
#> 15  7.5 6.361312 3.236426 1355.860 9.248869 -1.079723
#> 16  8.0 6.388640 3.234065 1387.153 9.336138 -1.083997
#> 17  8.5 6.415908 3.228229 1419.357 9.424361 -1.293012
#> 18  9.0 6.443190 3.229855 1452.370 9.513292 -1.506678
#> 19  9.5 6.470468 3.229351 1486.274 9.603027 -1.721377
#> 20 10.0 6.497748 3.229544 1521.059 9.693506 -1.721377